% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add.plague.data.frame.columns}
\alias{add.plague.data.frame.columns}
\title{add columns to data frame so it looks like original plague data frame}
\usage{
add.plague.data.frame.columns(data)
}
\arguments{
\item{data}{data frame}
}
\description{
add columns to data frame so it looks like original plague data frame
}
