% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_const_incidence}
\alias{add_const_incidence}
\title{add a constant incidence term (b*t) to RHS of formula}
\usage{
add_const_incidence(cform)
}
\arguments{
\item{cform}{base formula}
}
\description{
add a constant incidence term (b*t) to RHS of formula
}
\keyword{internal}
