% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggsum}
\alias{aggsum}
\title{aggregate count data into a different time unit
(don't try increasing the frequency this way!)}
\usage{
aggsum(
  year,
  month,
  day,
  deaths,
  period = "1 months",
  time = NULL,
  percorrect = TRUE
)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{day}{day}

\item{deaths}{number of counts (of whatever)}

\item{period}{time period for aggregation}

\item{time}{decimal date (alternative to year/month/day)}

\item{percorrect}{rescale to original units}
}
\description{
aggregate count data into a different time unit
(don't try increasing the frequency this way!)
}
\examples{
agg1 <- with(london_bills,aggsum_monthly(year,month,day,plague.deaths))
agg2 <- with(london_bills,aggsum(year,month,day,plague.deaths))
agg3 <- with(london_parish_all,aggsum(time=time,deaths=total))
plot(deaths~time,data=agg3,type="l",log="y")
agg4 <- with(london_parish_all,aggsum(time=time,deaths=total,
                                      period="10 days"))
plot(deaths~time,data=agg4,type="l",log="y")
}
