% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_par}
\alias{boot_par}
\title{Parametric bootstrap}
\usage{
boot_par(
  object,
  n = 6L,
  method_outer = c("nlminb", "nlm", "BFGS", "Nelder-Mead"),
  method_inner = c("newton", "BFGS", "Nelder-Mead"),
  control_outer = list(maxit = 1000L),
  control_inner = list(maxit = 1000L),
  trace = TRUE,
  parallel = c("serial", "multicore", "snow"),
  cores = getOption("egf.cores", 2L),
  outfile = NULL,
  cl = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{n}{A positive integer. The desired number of samples.}

\item{method_outer}{Character strings specifying optimizers to be used
for the outer and inner optimizations, respectively.
The outer optimization permits \code{\link[stats:nlminb]{stats::nlminb()}},
\code{\link[stats:nlm]{stats::nlm()}}, and certain optimizers available
through \code{\link[stats:optim]{stats::optim()}}. The inner optimization
permits \code{\link[TMB:newton]{TMB::newton()}} and the same \code{\link[stats:optim]{stats::optim()}}
methods. \code{method_inner} may have length greater than 1;
in this case, the listed optimizers will be tried
in turn until one succeeds.}

\item{method_inner}{Character strings specifying optimizers to be used
for the outer and inner optimizations, respectively.
The outer optimization permits \code{\link[stats:nlminb]{stats::nlminb()}},
\code{\link[stats:nlm]{stats::nlm()}}, and certain optimizers available
through \code{\link[stats:optim]{stats::optim()}}. The inner optimization
permits \code{\link[TMB:newton]{TMB::newton()}} and the same \code{\link[stats:optim]{stats::optim()}}
methods. \code{method_inner} may have length greater than 1;
in this case, the listed optimizers will be tried
in turn until one succeeds.}

\item{control_outer}{Named lists of control parameters to be passed to optimizers.
If \code{length(method_inner) > 1L} and \code{names(control_inner)}
contains \code{method_inner[i]}, then \code{control_inner[[method_inner[i]]]}
is passed to optimizer \code{i} instead of \code{control_inner}.}

\item{control_inner}{Named lists of control parameters to be passed to optimizers.
If \code{length(method_inner) > 1L} and \code{names(control_inner)}
contains \code{method_inner[i]}, then \code{control_inner[[method_inner[i]]]}
is passed to optimizer \code{i} instead of \code{control_inner}.}

\item{trace}{A logical scalar. If \code{TRUE}, then basic tracing messages
are printed.}

\item{parallel}{A character string indicating a library for parallel computation.
\code{"serial"} indicates no parallelization. \code{"multicore"} forks and
is intended for use from a terminal rather than, e.g., RStudio.
On Windows, it is equivalent to \code{"serial"}. \code{"snow"} creates
socket clusters and parallelizes on both Unix-alikes and Windows.}

\item{cores}{A positive integer. The number of worker processes spawned when
\code{parallel != "serial"}. See also \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{outfile}{A character string containing a file path or otherwise \code{NULL}.
If a file path, then console output is diverted there. If \code{NULL},
then there is no diversion. When \code{parallel = "snow"}, diversion
may be necessary to view output.}

\item{cl}{A optional socket cluster created by
\code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}}, to be used
when \code{parallel = "snow"}. If non-\code{NULL},
then \code{cores} and \code{outfile} are ignored.}

\item{...}{Unused optional arguments.}
}
\value{
A numeric matrix inheriting from class \code{"egf_boot"},
with \code{length(object$best)} rows and \code{n} columns.
}
\description{
Sample from the conditional distribution of the full
parameter vector given the fitted epidemic model.
}
