% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_covid}
\alias{canada_covid}
\title{Canada COVID-19 cases, 2020}
\format{
A tibble containing columns:
\describe{
\item{\code{Date}}{reporting date}
\item{\code{Province}}{province}
\item{\code{newConfirmations}}{number of new confirmed cases}
}
}
\usage{
canada_covid
}
\description{
Canada COVID-19 cases, 2020
}
\keyword{datasets}
