% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_final_size.R
\name{compute_final_size}
\alias{compute_final_size}
\title{\loadmathjax
Compute the expected epidemic final size}
\usage{
compute_final_size(R0)
}
\arguments{
\item{R0}{A numeric vector listing (non-negative) values
for the basic reproduction number.}
}
\value{
A numeric vector \code{x} of length \code{length(R0)}. \code{x[i]} is the
expected epidemic final size implied by basic reproduction number
\code{R0[i]}. See Details.
}
\description{
Computes, as a function of the basic reproduction number,
the proportion of the population expected to be infected
over the course of an epidemic in which the entire population
is initially susceptible.
}
\details{
The basic reproduction number \mjseqn{\mathcal{R}_0} defines
the expected epidemic final size \mjseqn{Z \in \lbrack 0,1 \rbrack}

\mjsdeqn{Z = 1 - e^{-\mathcal{R}_0 Z}\,.}

\insertCite{MaEarn06;textual}{epigrowthfit} review and extend
earlier work showing that this relation is valid for a large
class of epidemic models, including SEIR models with arbitrarily
distributed infectious periods.

The explicit solution for \mjseqn{Z} in terms of \mjseqn{\mathcal{R}_0}
involves the non-elementary
\href{https://en.wikipedia.org/wiki/Lambert_W_function}{Lambert W function}:

\mjsdeqn{Z(\mathcal{R}_0) = 1 + \frac{1}{\mathcal{R}_0} W(-\mathcal{R}_0 e^{-\mathcal{R}_0})\,.}
}
\examples{
R0 <- seq(0, 40, by = 0.2)
final_size <- compute_final_size(R0)
plot(R0, final_size, type = "l", las = 1, ylab = "final size")

}
\references{
\insertRef{MaEarn06}{epigrowthfit}
}
