% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R0.R
\name{compute_R0}
\alias{compute_R0}
\title{\loadmathjax
Compute the basic reproduction number}
\usage{
compute_R0(r, breaks, probs)
}
\arguments{
\item{r}{A numeric vector listing values for the initial
growth rate expressed per day.}

\item{breaks}{A numeric vector of length 2 or greater
listing increasing break points, in days, in the support
of the generation interval density.}

\item{probs}{A numeric vector with length \code{length(breaks)-1}.
\code{probs[i]} is the probability that the generation interval
is between \code{break[i]} days and \code{break[i+1]} days. Replaced
with \code{probs / sum(probs)} in the event that \code{sum(probs) != 1}.}
}
\value{
A numeric vector \code{x} of length \code{length(r)}. \code{x[i]} is the
basic reproduction number corresponding to initial growth rate
\code{r[i]}. See Details.
}
\description{
Given an initial growth rate and a generation interval
distribution, compute the basic reproduction number.
}
\details{
Let \mjseqn{t_1 < \cdots < t_\ell} be the break points specified
by \code{breaks}. For \mjseqn{i \in \lbrace 1,\ldots,\ell-1 \rbrace},
let \mjseqn{p_i} be the probability specified by \code{probs[i]} that
the generation interval is in the interval \mjseqn{(t_i,t_{i+1}\rbrack}.
Section 3(d) in \insertCite{WallLips07;textual}{epigrowthfit}
gives the basic reproduction number as a function of initial
growth rate \mjseqn{r}:

\mjsdeqn{\out{\mathcal{R}_0(r) = \left. r \middle/ \bigg\lbrace \sum_{i=1}^{\ell-1} \frac{p_i (e^{-r t_i} - e^{-r t_{i+1}})}{t_{i+1} - t_i} \bigg\rbrace \right.\,.}}
}
\examples{
data(plague_latent_period)
lat <- plague_latent_period$relfreq
m <- length(lat)

data(plague_infectious_period)
inf <- plague_infectious_period$relfreq
n <- length(inf)

r <- seq(0, 1, by = 0.02)
breaks <- 1:(m+n)
probs <- diff(pgi(breaks, lat, inf))

R0 <- compute_R0(r, breaks, probs)
plot(r, R0, las = 1,
  xlab = "initial growth rate, per day",
  ylab = "basic reproduction number"
)

}
\references{
\insertRef{WallLips07}{epigrowthfit}
}
\seealso{
\code{\link[=compute_final_size]{compute_final_size()}} for computing the expected epidemic
final size as a function of the basic reproduction number
}
