% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{confint.egf_fitted}
\alias{confint.egf_fitted}
\title{Confidence intervals on fitted values}
\usage{
\method{confint}{egf_fitted}(object, parm, level = 0.95, link = TRUE, ...)
}
\arguments{
\item{object}{An \code{"egf_fitted"} object returned by \code{\link[=fitted.egf]{fitted.egf()}}.
Must supply standard errors on fitted values.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1). The desired confidence level.}

\item{link}{A logical scalar. If \code{FALSE}, then confidence intervals
on inverse link-transformed fitted values are returned.}

\item{...}{Unused optional arguments.}
}
\value{
If \code{link = TRUE}, then \code{object} but with variable \code{se} replaced
with variables \code{lower} and \code{upper} supplying confidence limits
on link scale fitted values.

Otherwise, the same object but with variables \code{estimate}, \code{lower},
and \code{upper} inverse link transformed and link prefixes stripped
from \code{levels(par)}.

\code{level} is retained as an attribute.
}
\description{
Computes confidence intervals on fitted values of nonlinear
model parameters.
}
\details{
Confidence limits on fitted values (link scale) are computed
as \code{estimate + c(-1, 1) * sqrt(q) * se}, with \code{estimate} and
\code{se} obtained from \code{object} and \code{q = qchisq(level, df = 1)}.
}
