% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.egf_pred}
\alias{confint.egf_pred}
\title{Compute confidence bands on predicted incidence}
\usage{
\method{confint}{egf_pred}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An "egf_pred" object.}

\item{parm}{Unused parameter argument.}

\item{level}{A number in the interval [0,1] indicating a confidence level.}

\item{...}{Unused optional arguments.}
}
\value{
A list of two data frames of the form
\code{data.frame(estimate, lower, upper)},
giving confidence bands on predicted
cumulative and interval incidence, respectively.
}
\description{
A method for obtaining confidence bands on predicted incidence
curves returned by \code{\link[=predict.egf]{predict.egf()}}. Requires that standard errors
were computed via \code{predict.egf(se = TRUE)}.
}
\seealso{
\code{\link[=predict.egf]{predict.egf()}}
}
