% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.egf_predict}
\alias{confint.egf_predict}
\title{Confidence bands on log predicted incidence}
\usage{
\method{confint}{egf_predict}(object, parm, level = 0.95, log = TRUE, ...)
}
\arguments{
\item{object}{An \code{"egf_predict"} object returned by \code{\link[=predict.egf]{predict.egf()}}.
Must supply standard errors on log predicted incidence.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1). The desired confidence level.}

\item{log}{A logical scalar. If \code{FALSE}, then log predicted incidence and
the corresponding confidence bands are inverse log transformed.}

\item{...}{Unused optional arguments.}
}
\value{
If \code{log = TRUE}, then \code{object} but with variable \code{se} in each listed
data frame replaced with two variables \code{lower} and \code{upper} supplying
confidence bands on log predicted incidence.

Otherwise, the same list but with variables \code{estimate}, \code{lower},
and \code{upper} inverse log transformed in each listed data frame.
In this case, the listed data frames are named \code{cum_inc} and
\code{int_inc}, rather than \code{log_cum_inc} and \code{log_int_inc}.
}
\description{
Computes confidence bands on log predicted incidence assuming
asymptotic normality of the estimator.
}
\details{
Confidence bands on log predicted incidence are computed as
\code{estimate + c(-1, 1) * q * se}, where \code{q = qnorm(0.5 * (1 + level))}
and \code{estimate} and \code{se} are log predicted incidence and the
approximate (delta method) standard errors.
}
