% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor2cov}
\alias{cor2cov}
\title{Correlation to covariance matrix conversion}
\usage{
cor2cov(cor, sd)
}
\arguments{
\item{cor}{A square \link{numeric} \link{matrix}.}

\item{sd}{A \link{numeric} vector of length \code{\link{nrow}(cor)}.}
}
\value{
The result of \code{\link{sweep}(sweep(cor, 1L, sd, `*`), 2L, sd, `*`)},
obtained slightly more efficiently, following \code{\link{cov2cor}}.
}
\description{
Perform the inverse of \code{\link{cov2cor}}.
}
\examples{
# X <- replicate(6L, rnorm(10L))
# V <- cov(X, X)
# all.equal(V, cor2cov(cov2cor(V), sqrt(diag(V))))

}
\keyword{internal}
