% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cov2theta}
\alias{cov2theta}
\alias{theta2cov}
\title{Convert between covariance matrix and parameters}
\usage{
cov2theta(S)

theta2cov(theta)
}
\arguments{
\item{S}{A symmetric positive definite numeric matrix.}

\item{theta}{A numeric vector parametrizing a symmetric positive definite
numeric matrix.}
}
\value{
\code{cov2theta(S)} returns a numeric vector of length
\code{choose(nrow(S) + 1, 2)}.
\code{theta2cov(theta)} returns a symmetric positive definite
numeric matrix with dimensions \code{c(n, n)},
where \code{n} is the positive solution of
\code{length(theta) = choose(n + 1, 2)}.
}
\description{
\pkg{TMB} uses \code{choose(n + 1, 2)}-vectors \code{theta}
to parametrize \code{n}-by-\code{n} symmetric positive definite
matrices \code{S}.
The specific parametrization is given by
\code{theta = c(log(sqrt(diag(S))), R1[upper.tri(R1)])},
where
\code{R1 = R \%*\% diag(1 / diag(R))},
\code{R = chol(S)}.
\code{cov2theta} converts from matrix to vector;
\code{theta2cov} performs the inverse operation.
Neither performs any checks on the supplied argument.
}
