% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{covid_generation_interval}
\alias{covid_generation_interval}
\title{Generation interval distribution (COVID-19)}
\format{
A list with 4 elements:

\describe{
\item{\code{shape}}{
Estimated shape parameter. See \code{\link[stats:GammaDist]{dgamma()}}.
}
\item{\code{scale}}{
Estimated scale parameter. See \code{\link[stats:GammaDist]{dgamma()}}.
}
\item{\code{breaks}}{
An integer vector listing numbers of days. Equal to \code{0:20}.
}
\item{\code{probs}}{
A numeric vector of length \code{length(breaks-1)}. \code{probs[i]}
is the probability that the generation interval is between
\code{breaks[i]} and \code{breaks[i+1]} days, conditional on the
fitted generation interval distribution.
Equal to \code{diff(pgamma(breaks, shape = shape, scale = scale))}.
}
}
}
\source{
\code{shape} and \code{scale} were computed from the mean and standard
deviation reported in \insertCite{Gany+20;textual}{epigrowthfit},
Table 4, Scenario 2.
}
\usage{
data(plague_infectious_period)
}
\description{
A gamma distribution of the COVID-19 generation interval
fitted to data from a cluster of 45 cases in Tianjin, China.
}
\examples{
data(covid_generation_interval)
x <- 10^seq(-2, log10(20), length.out = 150)
fx <- with(covid_generation_interval,
  dgamma(x, shape = shape, scale = scale)
)
plot(x, fx, type = "l", ylim = c(0, max(fx[-1])))

}
\references{
\insertRef{Gany+20}{epigrowthfit}
}
\keyword{datasets}
