% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{Daxis}
\alias{Daxis}
\title{Date axis}
\usage{
Daxis(
  origin = .Date(0),
  minor = NULL,
  major = NULL,
  show_minor = TRUE,
  show_major = TRUE
)
}
\arguments{
\item{origin}{A \link{Date}. It is assumed that horizontal user coordinates
measure time as a number of days since the time
\code{\link{unclass}(origin)} days after 1970-01-01 00:00:00.}

\item{minor, major}{Named \link{list}s of arguments to \code{\link{axis}}, affecting
the appearance of the minor (day or month) and major (month or year)
axes, respectively.}

\item{show_minor, show_major}{\link[=logical]{Logical} flags.
If \code{FALSE}, then the corresponding axis is not drawn.}
}
\value{
A \link{list} of \link{numeric} vectors \code{minor} and \code{major}
giving the positions of minor and major axis labels in user coordinates.
}
\description{
Adds an axis to the bottom of the current plot and labels
it with day, month, and year, taking care to ensure that
labels are nicely spaced.
}
\details{
The \link{Date} axis is a calendar consisting of minor
and major axes. The content of these axes depends entirely
on \code{w = \link{diff}(\link{par}("usr"))[1:2]},
the difference between the extreme times in days.

If \code{w <= 210} (7 months), then days are placed on the minor
axis, months are placed on the major axis, and years are not
shown. The spacing on the minor axis in days is the value of
\code{c(1, 2, 4, 7, 14)[w <= c(14, 28, 56, 112, 210)][1]}.
Hence, for example, if `w` is greater than 112 and less than
or equal to 210, then the spacing is 14-daily.

Otherwise, if \code{w <= 3*365} (3 years), then months are placed
on the minor axis, years are placed on the major axis, and days
are not shown. The spacing on the minor axis in months is the
value of \code{c(1, 2, 3)[w <= c(1, 2, 3) * 365][1]}.

Otherwise, if \code{w > 3*365} (3 years), then years are placed on
the minor axis, and days and months are not shown. The spacing on the
minor axis is the value of \code{\link{ceiling}(ceiling(w / 365) / 7)}.
}
\keyword{internal}
