% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{days}
\alias{days}
\title{Get number of days since a date}
\usage{
days(date, since = as.Date("1970-01-01"))
}
\arguments{
\item{date}{A Date vector.}

\item{since}{A Date scalar.}
}
\value{
An integer vector giving the number of days between \code{date}
and \code{since}. Equal to \code{as.integer(date - since)}.
}
\description{
Subtracts a Date scalar from a Date vector and returns the
result as an integer vector instead of a "difftime" object.
}
\keyword{internal}
