% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dceiling}
\alias{dceiling}
\alias{dceiling_m}
\alias{dceiling_y}
\title{Get ceiling of a date}
\usage{
dceiling_m(date)

dceiling_y(date)
}
\arguments{
\item{date}{A Date vector.}
}
\value{
A Date vector of length \code{length(date)}.
For \code{dceiling_m()}, elements are firsts-of-the-month ("YYYY-MM-01").
For \code{dceiling_y()}, elements are firsts-of-the-year ("YYYY-01-01").
}
\description{
Rounds each Date in a Date vector to the next first-of-the-month
or first-of-the-year.
}
\keyword{internal}
