% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ddiff}
\alias{ddiff}
\title{Difference a vector of dates}
\usage{
ddiff(date, lag = 1, differences = 1)
}
\arguments{
\item{date}{A Date vector.}

\item{lag, differences}{Arguments to \code{\link[=diff.Date]{diff.Date()}}.}
}
\value{
An integer vector listing the desired differences in days.
Equal to \code{as.integer(diff(date, lag, differences))}.
}
\description{
Takes differences of a Date vector and returns the
result as an integer vector instead of a "difftime" object.
}
\keyword{internal}
