% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{decontrast_beta}
\alias{decontrast_beta}
\title{Undo zero-sum contrasts}
\usage{
decontrast_beta(beta, fnl)
}
\arguments{
\item{beta}{The segment of a TMB-generated parameter vector with names
\code{"beta"} or, more generally, a numeric vector.}

\item{fnl}{An integer vector giving the length of each segment of \code{beta},
such that \code{sum(fnl) = length(beta)}. See \code{\link[=make_tmb_data]{make_tmb_data()}}.}
}
\value{
A named numeric vector of length \code{length(beta)}.
}
\description{
Multiplies each segment of a vector of fixed effects
coefficients in front by the appropriate matrix of contrasts
to recover within-group means.
}
\examples{
fnl <- 1:4
beta <- sample(5L, sum(fnl), replace = TRUE)
decontrast_beta(beta, fnl)

}
\keyword{internal}
