% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{dfun}
\alias{dfun}
\title{second derivative of lbeta
use Taylor expansion of digamma(a+b) for a>>b
discontinuity in second derivative, but ... probably OK}
\usage{
dfun(x, y, mag = 1e+08)
}
\description{
second derivative of lbeta
use Taylor expansion of digamma(a+b) for a>>b
discontinuity in second derivative, but ... probably OK
}
