% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{do_minus}
\alias{do_minus}
\title{Negate formula terms}
\usage{
do_minus(x)
}
\arguments{
\item{x}{A call, name, or atomic scalar.}
}
\value{
A call, name, or atomic scalar.
}
\description{
Negates formula terms, being careful to resolve \code{-(-x)} to \code{x}.
}
\examples{
# x <- as.name("x")
# minus_x <- call("-", x)
# do_minus(x)
# do_minus(minus_x)

}
\keyword{internal}
