% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_wald}
\alias{do_wald}
\title{Compute Wald confidence intervals}
\usage{
do_wald(estimate, se, level)
}
\arguments{
\item{estimate, se}{\link[=numeric]{Numeric} vectors of equal length supplying maximum
likelihood estimates of parameters and corresponding approximate
standard errors.}

\item{level}{A number in the interval (0,1) indicating a confidence level.}
}
\value{
A \link{numeric} \link{matrix} with \code{length(estimate)} rows
and 2 columns \code{lower} and \code{upper} giving approximate
confidence limits.
}
\description{
Computes approximate confidence intervals around a maximum likelihood
estimate (assuming that the estimator is approximately normally distributed).
}
\details{
Confidence limits are computed as
\code{estimate[i] + c(-1, 1) * sqrt(q) * se[i]},
where \code{q = \link{qchisq}(level, df = 1)}.
See the \href{https://en.wikipedia.org/wiki/Wald_test}{Wald test}.
}
\examples{
## estimate <- rep_len(0, 6L)
## se <- exp(rnorm(6L, sd = 0.1))
## do_wald(estimate = estimate, se = se, level = 0.95)

}
\keyword{internal}
