% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubling_time-methods.R
\name{doubling_time-methods}
\alias{doubling_time-methods}
\alias{print.doubling_time}
\title{Methods for class "doubling_time"}
\usage{
\method{print}{doubling_time}(x, ...)
}
\arguments{
\item{x}{A "doubling_time" object.}

\item{...}{Unused optional arguments.}
}
\value{
The \code{print} method returns \code{x} invisibly.
}
\description{
Methods for "doubling_time" objects returned
by \code{\link[=compute_doubling_time]{compute_doubling_time()}}.
}
\seealso{
\code{\link[=compute_doubling_time]{compute_doubling_time()}}
}
