% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{doublingTime,epigrowthfit-method}
\alias{doublingTime,epigrowthfit-method}
\title{returns the doubling time and its 95\% confidence interval}
\usage{
\S4method{doublingTime}{epigrowthfit}(obj)
}
\value{
a vector
}
\description{
returns the doubling time and its 95\% confidence interval
}
