% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{doublingTime}
\alias{doublingTime}
\title{returns the doubling time and its 95\% confidence interval}
\usage{
doublingTime(obj)
}
\value{
a vector
}
\description{
returns the doubling time and its 95\% confidence interval
}
