% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datetime.R
\name{Dround}
\alias{Dround}
\alias{Dceiling}
\alias{Dfloor}
\title{Round a Date vector}
\usage{
Dceiling(x, to = c("day", "month", "year"))

Dfloor(x, to = c("day", "month", "year"))
}
\arguments{
\item{x}{A \link{Date} vector.}

\item{to}{A \link{character} string.}
}
\value{
\code{x} with elements rounded.
}
\description{
Rounds each element of a \link{Date} vector to the previous or next
midnight (\code{00:00:00}),
midnight on a first-of-the-month (\code{YYYY-MM-01}), or
midnight on a first-of-the-year (\code{YYYY-01-01}).
}
\examples{
## n <- 10L
## x <- .Date(sample.int(10000L, n) + rnorm(n))
## dmy <- c("day", "month", "year")
##
## l <- sapply(dmy, Dceiling, x = x, simplify = FALSE)
## ceiling_x <- data.frame(unclass_x = unclass(x), x = x, l)
##
## l <- sapply(dmy, Dfloor, x = x, simplify = FALSE)
## floor_x <- replace(ceiling_y, dmy, l)

}
\keyword{internal}
