% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf-class}
\alias{egf-class}
\title{Fitted and unfitted egf objects}
\value{
Legitimate \code{"egf"} and \code{"egf_no_fit"} objects are lists
with elements:
\item{model}{
  A copy of the so-named argument.
}
\item{frame}{
  A recursive list of data frames.
}
\item{priors}{
  A recursive list of \code{"\link{egf_prior}"} objects,
  of the form \code{list(top, bottom = list(beta, theta, Sigma))}.
}
\item{control}{
  A copy of the so-named argument.
}
\item{tmb_out}{
  The list output of \code{\link[TMB]{MakeADFun}}.
}
\item{optimizer_out}{
  The \link{list} output of the optimizer specified by
  \code{control$optimizer}.
}
\item{init, best}{

}
\item{random}{

}
\item{value, gradient}{
  Numeric vectors giving the value and gradient of the negative
  log (marginal) likelihood function.
}
\item{hessian}{
  A logical flag indicating whether the Hessian matrix of the negative
  log (marginal) likelihood function, evaluated at \code{best[!random]},
  is positive definite.
  \code{TRUE} means the matrix is positive definite.
  \code{NA} means the matrix has not been computed,
  either because \code{se = FALSE} or because an error
  was thrown during computation by \code{\link{sdreport}}.
}
\item{sdreport}{
  If \code{se = TRUE}, then a list inheriting from class \code{"sdreport"},
  resulting from \code{\link{sdreport}(tmb_out)}. Otherwise, \code{NULL}.
}
\item{effects}{

}
\item{contrasts}{

}
\item{Y0}{

}
\item{call}{
  The \link{call} to \code{\link{egf}},
  enabling updates to the object via \code{\link{update}}.
}
\code{optimizer_out}, \code{best},
\code{value}, \code{gradient}, \code{hessian}, and \code{sdreport}
are \code{NULL} in \code{"egf_no_fit"} objects.
}
\description{
An object returned by \code{\link{egf}},
inheriting from class \code{"egf"} if the nonlinear mixed effects model
that it specifies has been fitted and class \code{"egf_no_fit"} otherwise.
}
