% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf-methods.R
\name{egf-methods}
\alias{egf-methods}
\alias{plot.egf}
\title{Methods for class "egf"}
\usage{
\method{plot}{egf}(x, inc = "interval", tol = 0.025, ...)
}
\arguments{
\item{x}{An "egf" object.}

\item{inc}{One of \code{"interval"} and \code{"cumulative"},
indicating whether to plot \code{x$cases} (interval incidence)
or \code{cumsum(x$cases)} (cumulative incidence).}

\item{tol}{A non-negative number used only if
\code{inc = "interval"}. \code{x$ini$cases[i]} is plotted
in blue if \code{diff(x$init$time)[i] < (1-tol)*m},
in red if \code{diff(x$init$time)[i] > (1+tol)*m},
and in grey otherwise, where \code{m = median(diff(x$init$time))}.
Assign \code{Inf} to ensure that everything is grey.}

\item{...}{Unused optional arguments.}
}
\description{
A method for plotting "egf" objects returned by \code{\link[=egf]{egf()}}.
}
\details{
If \code{x$init$time} is not equally spaced, then \code{inc = "interval"}
should be used with caution. \code{x$init$cases[i]} is the number
of cases observed between \code{x$init$time[i]} and \code{x$init$time[i+1]},
hence \code{x$init$cases} roughly scales with \code{diff(x$init$time)}.
Argument \code{tol} can be used with \code{inc = "interval"} in order to
highlight outliers in \code{diff(x$init$time)}.
}
