% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{egf.egf_model_simulate}
\alias{egf.egf_model_simulate}
\title{Fit simulated incidence time series}
\usage{
\method{egf}{egf_model_simulate}(model, ...)
}
\arguments{
\item{model}{An \code{"\link[=simulate.egf_model]{egf_model_simulate}"} object
supplying simulated time series and details about the generative model
to be estimated.}

\item{...}{Optional arguments passed to \code{\link{egf.egf_model}}.}
}
\description{
Estimates the generative model underlying a set of simulated
incidence time series.
}
\examples{
example("simulate.egf_model", package = "epigrowthfit", local = TRUE, echo = FALSE)
exdata <- system.file("exdata", package = "epigrowthfit", mustWork = TRUE)
model <- readRDS(file.path(exdata, "simulate-egf_model.rds"))

path_to_cache <- file.path(exdata, "egf-egf_model_simulate.rds")
if (file.exists(path_to_cache)) {
  object <- readRDS(path_to_cache)
} else {
  object <- egf(model)
  saveRDS(object, file = path_to_cache)
}
pp <- data.frame(actual = model$actual, fitted = coef(object, full = FALSE))
pp[!object$random, , drop = FALSE]

}
