% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf}
\alias{egf}
\title{Fit phenomenological models of epidemic growth}
\usage{
egf(
  formula,
  formula_par,
  data = parent.frame(),
  data_par = parent.frame(),
  subset = NULL,
  subset_par = NULL,
  na_action = c("fail", "pass"),
  na_action_par = c("fail", "pass"),
  endpoints,
  model = egf_model(),
  control = egf_control(),
  do_fit = TRUE,
  se = FALSE,
  init = NULL,
  origin = .Date(0),
  append = NULL,
  ...
)
}
\arguments{
\item{formula}{A \link{formula} of the form \code{x ~ time} or \code{x ~ time | ts}
specifying one or more incidence time series in long format.
\code{time} must evaluate to a \link{numeric} or \link{Date} vector.
Numeric \code{time} is assumed to measure time as a number of days
since \emph{the end of} Date \code{origin}. Date \code{time} is
coerced to numeric \code{\link{julian}(time, origin)} (see Details).
\code{x} must evaluate to a numeric vector. Within a time series,
\code{x[i]} should specify the number of cases observed from
\code{time[i-1]} to \code{time[i]}. Finally, \code{ts} must evaluate
to a \link{factor} such that
\code{\link{split}(\link{data.frame}(time, x), ts)}
returns a \link{list} of time series. Note that \code{x ~ time}
is equivalent to \code{x ~ time | ts} with \code{ts} set equal
to \code{\link{rep}(\link{factor}(1), \link{length}(x))}.}

\item{formula_par}{A \link{list} of \link{formula}e of the form \code{par ~ terms},
specifying mixed effects models for nonlinear and dispersion
model parameters using \code{\link[lme4:lmer]{lme4}}-like syntax.
\code{\link{deparse}(par)} must be an element of
\code{\link{get_par_names}(model, link = TRUE)}.
\code{~1} is the default for parameters not assigned a formula.
Alternatively, \code{formula_par} may be a formula of the form
\code{~terms}. In this case, the formula is recycled for all
nonlinear and dispersion model parameters. Note that
"individuals" in each mixed effects model are fitting windows,
and \link[=model.frame]{model frames} constructed from
\code{formula_par} and \code{data_par} are expected to
correspond rowwise to \code{endpoints} (see Details).}

\item{data, data_par}{\link[=data.frame]{Data frame}s, \link{list}s, or \link{environment}s.
These are searched prior to \link{formula} environments for variables
used in \code{formula} and \code{formula_par}, respectively.}

\item{subset, subset_par}{Expressions to be evaluated in \code{data} and \code{data_par},
respectively, provided that the latter are data frames. These must
evaluate to \link{logical} vectors of suitable \link{length}
indexing rows to be used to fit the mixed effects model. The default
(\code{\link{NULL}}) is to use all rows. Note that \code{subset}
and \code{subset_par} perform subsetting at different levels of
granularity:
each row of \code{data} corresponds to one time point in a time series,
whereas each row of \code{data_par} corresponds to a fitting window
(and thus multiple time points) in a time series.}

\item{na_action}{A \link{character} string affecting the handling of \code{\link{NA}}
in \code{x} if \code{formula = x ~ time | ts}.
\code{"fail"} is to throw an error.
\code{"pass"} is to ignore \code{NA} when fitting
and replace \code{NA} when predicting.
(TODO: \code{"exclude"} is to ignore \code{NA} when fitting
and preserve \code{NA} when predicting.)
Note that \code{NA} in \code{time} and \code{ts} are an error
regardless of \code{na_action}.}

\item{na_action_par}{A \link{character} string affecting the handling of \code{\link{NA}}
in \code{formula_par} variables.
\code{"fail"} is to throw an error.
\code{"pass"} is to discard fitting windows with incomplete data.}

\item{endpoints}{A \link[=data.frame]{data frame}, \link{list}, or \link{environment}
with variables \code{start} and \code{end}, and any further variables
necessary to evaluate \code{ts} if \code{formula = x ~ time | ts}.
\code{start} and \code{end} must be \link{numeric} or \link{Date}
vectors listing start and end times for all fitting windows. \code{ts}
must evaluate to a \link{factor} indicating the time series in which each
window is found. Within time series, intervals \code{start[i], end[i]]}
must be disjoint and contain at least two time points from \code{time}.}

\item{model}{An \code{"\link{egf_model}"} object specifying a nonlinear model and
an observation model to be estimated.}

\item{control}{An \code{"\link{egf_control}"} object specifying control parameters.}

\item{do_fit}{A \link{logical} scalar. If \code{TRUE}, then \code{egf} returns early
with a \link{list} of optimization inputs.}

\item{se}{A \link{logical} scalar. If \code{TRUE}, then standard errors on mixed
effects model coefficients are computed and stored for later reuse
by methods.}

\item{init}{A full parameter vector for the first likelihood evaluation.
The default (\code{\link{NULL}}) is to accept the internally
generated default. Use \code{do_fit = FALSE} to retrieve this default
and other optimization inputs (in particular \code{Y_init}; see Value),
which can often be used to construct a more informative starting point.}

\item{origin}{A \link{Date} specifying a reference time.}

\item{append}{An expression indicating variables in \code{data_par} to be preserved
in the returned \code{"egf"} object for use by methods. It is evaluated
similarly to the \code{select} argument of \code{\link{subset}}.
The default (\code{\link{NULL}}) is to preserve all variables.
Currently, usage is supported only if \code{data_par} is a
\link[=data.frame]{data frame}.}

\item{...}{Unused optional arguments.}
}
\value{
If \code{do_fit = TRUE}, then a \link{list} inheriting from \link{class}
\code{"egf"}, with elements:
\item{endpoints}{
  A \link[=data.frame]{data frame} with variables \code{ts}, \code{window},
  \code{start}, and \code{end} listing start and end times for all fitting
  windows. Rows are ordered by time series and chronologically within time
  series. \code{origin} is retained as an \link[=attributes]{attribute}.
  Note that \code{start} and \code{end} here are the minimum and maximum of
  the set of time points contained in the supplied intervals. Hence supplied
  and returned interval endpoints may not match exactly.
}
\item{frame}{
  The time series model frame, constructed from \code{formula} and
  \code{data}, with variables \code{ts}, \code{window}, \code{start},
  and \code{end}. (Construction does not use \code{\link{model.frame}},
  as \code{formula} is handled specially.) \code{ts} and \code{window}
  are \link{factor}s that can be used to \link{split} \code{frame} by
  time series and by fitting window, respectively. Rows are ordered by
  time series and chronologically within time series.
  \code{\link{terms}(formula)} and \code{origin} are retained as
  \link{attributes}. \code{\link{as.integer}(window)} indexes rows
  of \code{endpoints}. That is, time series data for the fitting window
  defined in row \code{i} of \code{endpoints} can be found in the rows
  of \code{frame} for which \code{window = \link{levels}(window)[i]}.
}
\item{frame_par}{
  A \link{list} of mixed effects model frames, constructed from
  \code{formula_par} and \code{data_par} using \code{\link{model.frame}}.
  There is one model frame for each nonlinear and dispersion model
  parameter listed in
  \code{\link{get_par_names}(model, link = TRUE)}.
  \code{frame_par[[name]]} retains
  \code{\link{terms}(formula_par[[name]])} as an
  \code{\link[=attributes]{attribute}}.
  Model frames correspond rowwise to \code{endpoints}. That is,
  mixed effects data on the fitting window defined in row \code{i}
  of \code{endpoints} can be found in row \code{i} of each model frame.
}
\item{frame_append}{
  A \link[=data.frame]{data frame} preserving the variables from
  \code{data_par} indicated by \code{append}. Corresponds rowwise
  to \code{endpoints}.
}
\item{model, control}{
  Copies of the so-named arguments.
}
\item{tmb_args}{
  A \link{list} of arguments to \code{\link[TMB]{MakeADFun}}.
  See also \code{\link{make_tmb_args}}.
}
\item{tmb_out}{
  The \link{list} output of \code{\link[TMB]{MakeADFun}}
  (after optimization).
}
\item{optimizer_out}{
  The \link{list} output of the outer optimizer specified
  by \code{control$optimizer}.
}
\item{init}{
  The full parameter vector of the first likelihood evaluation.
}
\item{best}{
  The full parameter vector of the best likelihood evaluation.
}
\item{nonrandom}{
  An \link{integer} vector indexing the non-random segment of
  \code{init} and \code{best}.
}
\item{sdreport}{
  If \code{se = TRUE}, then an \code{"sdreport"} object resulting
  from \code{\link{sdreport}(tmb_out)}. Otherwise, \code{\link{NULL}}.
}
\item{nll}{
  A \link[=double]{numeric} scalar giving the value of the negative
  log Laplace approximation of the marginal likelihood function at
  \code{best[nonrandom]}.
}
\item{call}{
  The \code{\link{call}} to \code{egf}, allowing for updates
  to the \code{"egf"} object via \code{\link{update}}.
}

If \code{do_fit = FALSE}, then a \link{list} containing \code{frame},
\code{frame_par}, \code{endpoints}, \code{tmb_args}, \code{tmb_out}
(\emph{before} optimization), \code{init}, \code{nonrandom}, and
\code{call}, as well as a \link{matrix} \code{Y_init} specifying
a naive estimate of each nonlinear and dispersion model parameter
for each fitting window. \code{Y_init} and \code{endpoints} correspond
rowwise.
}
\description{
Fits nonlinear mixed effects models of epidemic growth to one
or more disease incidence time series.
}
\details{
If \code{formula_ts = x ~ time | ts}, then coercion of \link{Date}
\code{time} to \link{numeric} \code{\link{julian}(time, origin)}
assumes that each element \code{time[i]} can be read as "end of Date
\code{time[i]}", so that observation \code{x[i]} in a given time series
is the number of cases observed from the end of Date \code{time[i-1]}
to the end of Date \code{time[i]}.

To avoid unexpected mismatch between \code{endpoints} and mixed effects
\link[=model.frame]{model frame}s constructed from \code{formula_par}
and \code{data_par}, keep all necessary \code{endpoints} variables in
\code{data_par}, and set \code{endpoints = data_par}.

Day of week effect estimation (see \code{\link{egf_model}}) requires
\code{time} in \code{formula = x ~ time | ts} to evaluate to an
\link{integer} or \link{Date} vector with 1-day spacing in all fitting
windows.
}
