% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_misc.R
\name{egf_combine_frames}
\alias{egf_combine_frames}
\title{Construct the combined model frame}
\usage{
egf_combine_frames(object)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}
}
\value{
A \link[=data.frame]{data frame} combining
(in the sense of \code{\link{cbind}})
all data frames in \link{list} \code{object$frame_parameters}
and data frame \code{object$frame_append}.
}
\description{
Joins in a single \link[=data.frame]{data frame} all mixed effects
\link[=model.frame]{model frames} and further variables originally
indicated by \code{\link{egf}} argument \code{append}.
}
\details{
If a variable name occurs in multiple mixed effects model frames,
then only the first instance is retained. Except in unusual cases
(possible only if model formulae have different formula environments),
all instances of a variable name are identical, and no information is lost.

Since the data frames being combined each correspond rowwise
to \code{object$frame_windows}, so does the result.
}
\examples{
object <- list(
  frame_parameters = list(a = data.frame(x = seq_len(9L)), b = data.frame(y = rnorm(9L))),
  frame_append = data.frame(x = rnorm(9L), z = 0)
)
class(object) <- "egf"
egf_combine_frames(object)

}
