% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils_eval.R
\name{egf_eval}
\alias{egf_eval}
\alias{egf_eval_subset}
\alias{egf_eval_order}
\alias{egf_eval_append}
\alias{egf_eval_label}
\title{Nonstandard evaluation}
\usage{
egf_eval_subset(expr, data, enclos = parent.frame())

egf_eval_order(expr, data, enclos = parent.frame())

egf_eval_append(expr, data, enclos = baseenv())

egf_eval_label(expr, data, enclos = parent.frame())
}
\arguments{
\item{expr}{A \link{language} object to be evaluated in \code{data} or,
in the case of \code{egf_eval_append},
in \code{`\link{names<-}`(\link{as.list}(\link{seq_along}(data)), \link{names}(data))}.
Alternatively, a non-language object to be used in place of
the hypothetical result of evaluation.}

\item{data}{A \link[=data.frame]{data frame}.}

\item{enclos}{An \link{environment} to enclose \code{data}.}
}
\value{
Let \code{res} be the result of evaluating \code{expr}.

\code{egf_eval_subset} returns an increasing \link{integer} vector
indexing rows of \code{data}, obtained after some processing of
\code{res}.

\code{egf_eval_order} returns \code{\link{as.integer}(res)}.

\code{egf_eval_append} returns
\code{\link{match}(\link{names}(data[res]), \link{names}(data), 0L)},
with zeros (if any) deleted.

\code{egf_eval_label} returns \code{\link{as.character}(res)},
repeated to length \code{\link{nrow}(data)}.
}
\description{
Utilities for obtaining index vectors from unevaluated \code{subset},
\code{append}, and \code{order} expressions and character vectors from
unevaluated \code{label} expressions.
}
\details{
\code{subset} must evaluate to a valid index vector
for the rows of \code{data} (see \code{\link{[.data.frame}}).

\code{order} must evaluate to a permutation
of \code{\link{seq_len}(\link{nrow}(data))}.

\code{append} must evaluate to a valid index vector
for the variables in \code{data} (see \code{\link{[.data.frame}}).

\code{label} must evaluate to an \R object coercible
via \code{\link{as.character}} to a \link{character}
vector of length 1 or length \code{\link{nrow}(data)}.
}
\examples{
## year <- 2021L
## data <- data.frame(
##   month = sample(month.abb, 20L, replace = TRUE),
##   day = sample(30L, 20L, replace = TRUE)
## )
##
## subset <- quote(grepl("^J", month) & day < 16L)
## egf_eval_subset(subset, data)
##
## order <- quote(order(month, day))
## egf_eval_order(order, data)
##
## append <- quote(-day)
## egf_eval_append(append, data)
##
## label <- quote(sprintf("\%d-\%02d-\%02d", year, match(month, month.abb, 0L), day))
## egf_eval_label(label, data)

}
\keyword{internal}
