% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_misc.R
\name{egf_has_random}
\alias{egf_has_random}
\title{Check for random effects}
\usage{
egf_has_random(object)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Determines whether an object specifies a random effects model.
}
\examples{
e <- new.env()
e$data <- list()
object <- list(tmb_out = list(env = e))
class(object) <- "egf"

e$data$Z <- matrix(numeric(9L), 3L, 3L) # at least one column
egf_has_random(object)

e$data$Z <- matrix(numeric(0L), 3L, 0L) # zero columns
egf_has_random(object)

}
