% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{egf_make_priors}
\alias{egf_make_priors}
\title{Construct list of prior objects}
\usage{
egf_make_priors(model, priors)
}
\arguments{
\item{model}{An \code{"\link{egf_model}"} object specifying a top level nonlinear model
to be estimated.}

\item{priors}{A \link{list} of \link{formula}e of the form \code{parameter ~ f(...)}
specifying priors on top level nonlinear model parameters.
\code{\link{deparse}(parameter)} must be an element
of \code{\link{get_names_top}(model, link = TRUE)}, and
\code{f(...)} must be a \link{call} to a \link[=egf_prior]{prior function}
with arguments specifying suitable hyperparameters.
Calls are evaluated in the corresponding formula environments.}
}
\value{
A named \link{list} of \code{"\link{egf_prior}"} objects obtained
by evaluating the right hand side of each formula listed in \code{priors}.
}
\description{
Validates and processes \code{\link{egf}} argument \code{priors}.
}
\keyword{internal}
