% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{egf_make_XZ_info}
\alias{egf_make_XZ_info}
\title{Describe design matrix columns}
\usage{
egf_make_XZ_info(x, m)
}
\arguments{
\item{x}{A named \link{list} of \link{formula}e of the form \code{~tt}
or a named list of \link{call}s to binary operator \code{`|`}
of the form \code{(tt | g)}. \code{\link{names}(x)} must indicate
corresponding top level nonlinear model parameters.}

\item{m}{A \link{list} of \code{X} or \code{Z} matrices obtained by
applying \code{\link{egf_make_X}} or \code{\link{egf_make_Z}}
to the elements of \code{x}.}
}
\value{
A \link[=data.frame]{data frame} with rows corresponding to columns
of the combined design matrix \code{\link{do.call}(\link{cbind}, m)},
and variables:
\item{par}{
  Top level nonlinear model parameter.
}
\item{term}{
  \link[=deparse]{Deparse}d term from right hand side of \code{`~`}
  or left hand side of \code{`|`}. Otherwise, \code{"(Intercept)"}.}
\item{group}{
  \link[=deparse]{Deparse}d term from right hand side of \code{`|`}.
  \code{\link{NA}} if not applicable.
}
\item{level}{
  Level of factor or interaction \code{group}.
  \code{\link{NA}} if not applicable.
}
\item{colname}{
  Column name.
}
}
\description{
A utility linking each column of a combined design matrix
to the corresponding nonlinear or dispersion model parameter,
mixed effects formula term, grouping variable, and group.
}
\keyword{internal}
