% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{egf_plot_control}
\alias{egf_plot_control}
\title{Define plot options}
\usage{
egf_plot_control(
  points,
  points_short,
  points_long,
  lines,
  polygon,
  rect,
  rect_bg_panel,
  rect_bg_plab,
  abline,
  segments,
  axis_x_Date_minor,
  axis_x_Date_major,
  axis_x_numeric,
  axis_y,
  box,
  title_main,
  title_sub,
  title_xlab,
  title_ylab,
  title_plab,
  mtext_tdoubling_caption,
  mtext_tdoubling_estimate,
  mtext_tdoubling_ci,
  colorRamp
)
}
\arguments{
\item{points}{A named \link{list} of arguments to \code{\link{points}},
affecting the appearance of observed data.
[\code{type != "rt2"} only.]}

\item{points_short, points_long}{Alternatives to \code{points} used for counts over intervals
shorter or longer than \code{dt} days.
[\code{type = "interval"} only.]}

\item{lines}{A named \link{list} of arguments to \code{\link{lines}},
affecting the appearance of predicted curves.
[\code{type != "rt2"} only.]}

\item{polygon}{A named \link{list} of arguments to \code{\link{polygon}},
affecting the appearance of confidence bands on predicted curves.
[\code{type != "rt2"} only.]}

\item{rect}{A named \link{list} of arguments to \code{\link{rect}},
affecting the appearance of fitting windows.
[\code{type != "rt2"} only.]}

\item{rect_bg_panel, rect_bg_plab}{A named \link{list} of arguments to \code{\link{rect}},
affecting the appearance of panel backgrounds and panel title underlays.
[\code{type = "rt2"} only.]}

\item{abline}{A named \link{list} of arguments to \code{\link{abline}},
affecting the appearance of the line drawn at \code{y = 0}.
[\code{type = "rt1"} only.]}

\item{segments}{A named \link{list} of arguments to \code{\link{segments}},
affecting the appearance of line segments drawn at \code{y = r}
when \code{x$model$curve} is \code{"exponential"}, \code{"logistic"},
or \code{"richards"}.
[\code{type = "rt1"} only.]}

\item{axis_x_Date_minor, axis_x_Date_major, axis_x_numeric, axis_y}{Named \link{list}s of arguments to \code{\link{axis}},
affecting the appearance of plot axes. \code{axis_x_Date_*}
are used for \code{time_as = "Date"}. \code{axis_x_numeric}
is used for \code{time_as = "numeric"}.}

\item{box}{A named \link{list} of arguments to \code{\link{box}},
affecting the appearance of the box drawn around the plot region.
[\code{type != "rt2"} only.]}

\item{title_main, title_sub, title_xlab, title_ylab, title_plab}{Named \link{list}s of arguments to \code{\link{title}},
affecting the appearance of plot, axis, and panel titles.}

\item{mtext_tdoubling_caption, mtext_tdoubling_estimate, mtext_tdoubling_ci}{Named \link{list}s of arguments to \code{\link{mtext}},
affecting the appearance of initial doubling times printed
in the top margin.
[\code{type != "rt2"} only.]}

\item{colorRamp}{A named \link{list} of arguments to \code{\link{colorRamp}},
defining heat map colour palette.
[\code{type = "rt2"} only.]}
}
\value{
A named list.
}
\description{
Sets parameters controlling the graphical output of \code{\link{plot.egf}}.
Here, \code{x}, \code{type}, \code{time_as}, and \code{dt} refer to the
so-named arguments of \code{\link{plot.egf}}.
}
\details{
Unsupported and unmodifiable options are silently discarded.
Modifiable options that are unspecified are assigned a default
value defined internally. A complete list of defaults can be
retrieved with \code{l <- egf_plot_control()}.

Setting an argument to \code{\link{NULL}} has the effect of
suppressing the corresponding plot element. For example,
to suppress observed data, set \code{points} to \code{NULL},
and perhaps also \code{points_short} and \code{points_long}.
}
