% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enum_dupl_str}
\alias{enum_dupl_str}
\alias{enum_dupl_string}
\alias{enum_dupl_names}
\title{Enumerate duplicated strings}
\usage{
enum_dupl_string(x)

enum_dupl_names(x)
}
\arguments{
\item{x}{For \code{enum_dupl_string}, a \link{character} vector.
For \code{enum_dupl_names}, an \R object with a \code{\link{names}}
\link[=attributes]{attribute}.}
}
\value{
\code{enum_dupl_string(x)} returns an enumerated copy of \code{x}.
\code{enum_dupl_names(x)} returns a copy of \code{x} with
\code{\link{names}(x)} replaced by \code{enum_dupl_string(\link{names}(x))}.
}
\description{
Replaces the \code{i}th instance of string \code{s} in a
\link{character} vector with \code{\link{sprintf}("\%s[\%d]", s, i)}.
}
\examples{
# x <- sample(letters[1:3], 10L, replace = TRUE)
# enum_dupl_string(x)
#
# y <- seq_along(x)
# names(y) <- x
# enum_dupl_names(y)

}
\keyword{internal}
