% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\name{epigrowthfit-data}
\alias{epigrowthfit-data}
\title{Data in the \pkg{epigrowthfit} package}
\description{
Epidemic time series to which growth rates can be fit,
and a few other data sets.
}
\details{
Here is a list of available data sets with links to their documentation.

\describe{
\item{\code{\link{canadacovid}}}{
Daily confirmations of COVID-19 in Canadian provinces and territories
from February 14, 2020 to June 21, 2020.
}
\item{\code{\link{husting}}}{
Weekly counts of wills probated in the Court of Husting
during four plague epidemics in 14th century London.
}
\item{\code{\link{canterbury}}}{
Weekly counts of wills probated in the Prerogative Court of Canterbury
during 24 plague epidemics in 16th and 17th century London.
}
\item{\code{\link{londonparish}}}{
Weekly counts of burials listed in extant parish registers
during 24 plague epidemics in 16th and 17th century London.
}
\item{\code{\link{londonbills}}}{
Weekly counts of plague deaths recorded in the London Bills of Mortality
during 24 plague epidemics in 16th and 17th century London.
}
\item{\code{\link{plague_latent_period}}}{
Empirical distribution of the latent period of pneumonic plague.
}
\item{\code{\link{plague_infectious_period}}}{
Empirical distribution of the infectious period of pneumonic plague.
}
}
}
\keyword{internal}
