\name{epigrowthfit-defunct}
\alias{epigrowthfit-defunct}
\title{Defunct Functions in Package \pkg{epigrowthfit}}
\description{
The functions and other objects listed here are no longer part
of \pkg{epigrowthfit} as they are no longer needed.
}
\usage{
## Nothing yet!
}
\details{
These either are stubs reporting that they are defunct or
have been removed completely (apart from being documented here).
}
\seealso{
\code{\link{Deprecated}}, \code{\link{base-deprecated}},
\code{\link{epigrowthfit-deprecated}}, \code{\link{epigrowthfit-notyet}}.
}
