% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfitTMB-package.R
\docType{package}
\name{epigrowthfitTMB-package}
\alias{epigrowthfitTMB-package}
\title{epigrowthfitTMB: Estimation of Epidemic Growth Rates}
\details{
Details about the \pkg{epigrowthfitTMB} package.
}
\references{
\insertRef{Ma+14}{epigrowthfitTMB}

\insertRef{Earn+20}{epigrowthfitTMB}
}
\keyword{internal}
