% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{Eval,Model-method}
\alias{Eval,Model-method}
\title{Evaluate a model}
\usage{
\S4method{Eval}{Model}(obj, data = NULL, par = NULL, ...)
}
\arguments{
\item{data}{a dataframe object holding input values, if NULL, take from ...}

\item{par}{a named vector (or list) of parameter values, if NULL, take from ...}

\item{...}{the input values and parameter values}
}
\value{
numeric
}
\description{
Evaluate a model
}
