% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_inc.R
\name{eval_inc}
\alias{eval_inc}
\title{\loadmathjax
Evaluate expected incidence}
\usage{
eval_inc(time, curve, include_baseline = FALSE, theta)
}
\arguments{
\item{time}{A numeric vector listing (increasing) time points in days
since some initial date.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a phenomenological model for cumulative incidence.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then the
cumulative incidence model will include a linear term \mjseqn{b t}.}

\item{theta}{A named numeric vector listing positive values for all
relevant model parameters:

\describe{
\item{\code{r}}{\mjseqn{\lbrace\,r\,\rbrace}
Initial (exponential) growth rate expressed per day.
}
\item{\code{x0}}{\mjseqn{\lbrace\,x_0\,\rbrace}
Expected cumulative incidence on the initial date.
This is the expectation of the number of cases observed
up to the initial date. Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{\mjseqn{\lbrace\,K\,\rbrace}
Expected epidemic final size. This is the expectation of the
number of cases observed over the full course of the epidemic.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{\mjseqn{\lbrace\,t_\text{half}\,\rbrace}
Expected time at which the epidemic attains half its
final size, expressed as a (possibly non-integer) number
of days since the initial date.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{\mjseqn{\lbrace\,p\,\rbrace}
Richards shape parameter. Used only if \code{curve = "richards"}.
}
\item{\code{b}}{\mjseqn{\lbrace\,b\,\rbrace}
Baseline (linear) growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
}}
}
\value{
A list with elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{cum_inc}}{A numeric vector with length \code{length(time)}
such that \code{cum_inc[i]} is the expected number of cases observed
up to \code{time[i]}.
}
\item{\code{int_inc}}{A numeric vector with length \code{length(time)-1},
such that \code{int_inc[i]} is the expected number of cases observed
between \code{time[i]} and \code{time[i+1]}. Equal to \code{diff(cum_inc)}.
}
}
}
\description{
Evaluates expected cumulative and interval incidence at desired time
points, conditional on a phenomenological model of cumulative incidence.
}
\details{
Equations specifying how expected cumulative incidence is computed
for a given model are presented in Details 1 and 2 of \code{\link[=egf_init]{egf_init()}}.
}
