% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{eval_model}
\alias{eval_model}
\title{\loadmathjax
Evaluate expected cumulative incidence}
\usage{
eval_model(time, curve, include_baseline = FALSE, theta)
}
\arguments{
\item{time}{A numeric vector listing time points in days since
a reference date.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a model of expected cumulative incidence.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then the
cumulative incidence model will include a linear term \mjseqn{b t}.}

\item{theta}{A named numeric vector listing positive values for all
relevant model parameters:

\describe{
\item{\code{r}}{\mjseqn{\lbrace\,r\,\rbrace}
Initial exponential growth rate expressed per day.
}
\item{\code{c0}}{\mjseqn{\lbrace\,c_0\,\rbrace}
Expected cumulative incidence on the reference date.
This is the expected number of cases observed up to
the reference date. Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{\mjseqn{\lbrace\,K\,\rbrace}
Expected epidemic final size. This is expected number
of cases observed over the full course of the epidemic.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{\mjseqn{\lbrace\,t_\text{half}\,\rbrace}
Time at which the epidemic is expected to attain half its
final size, expressed as a (possibly non-integer) number
of days since the reference date.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{\mjseqn{\lbrace\,p\,\rbrace}
Richards shape parameter. Used only if \code{curve = "richards"}.
}
\item{\code{b}}{\mjseqn{\lbrace\,b\,\rbrace}
Baseline linear growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
}}
}
\value{
A numeric vector with length \code{length(time)} such that \code{cum_inc[i]}
is the expected number of cases observed up to \code{time[i]}.
}
\description{
Evaluates expected cumulative incidence at desired time points,
conditional on a model and a vector of model parameters.
}
\details{
A full description of the models of expected cumulative
incidence specified by \code{curve} and \code{include_baseline}
can be found in the package vignette, accessible with
\code{vignette("epigrowthfit-vignette")}.
}
