% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{expand_terms}
\alias{expand_terms}
\title{Expand and simplify formula terms}
\usage{
expand_terms(x)
}
\arguments{
\item{x}{A call, name, or atomic scalar.}
}
\value{
A call, name, or atomic scalar.
}
\description{
Expand and simplify formulae and formula terms
(including those with calls to the \code{|} operator)
using \code{terms(simplify = TRUE)} machinery.
}
\examples{
expand_terms(~x * y)
expand_terms(~x * y + (1 | f/g))
expand_terms(~x * y + (1 | f/g) + (a | f) + (0 + b | f:g))

}
