% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{factor_to_matrix}
\alias{factor_to_matrix}
\title{Construct a design matrix from a factor}
\usage{
factor_to_matrix(x, sparse, intercept)
}
\arguments{
\item{x}{A factor.}

\item{sparse}{A logical scalar. If TRUE, then matrix is returned in sparse
format.}

\item{intercept}{A logical scalar. If TRUE, then zero-sum contrasts are used.
If FALSE, then no contrasts are used. See \code{\link[stats:contrast]{stats::contr.sum()}}.}
}
\value{
A matrix with \code{length(x)-d} rows and \code{nlevels(x)} columns,
where \code{d} is \code{sum(is.na(x))} if \code{NA} is not among \code{levels(x)}
and 0 otherwise.
}
\description{
A wrapper for
\code{\link[stats:model.matrix]{stats::model.matrix()}} and \code{\link[Matrix:sparse.model.matrix]{Matrix::sparse.model.matrix()}}
operating on factors instead of formulae and data frames.
}
\details{
Factors \code{x} with \code{nlevels(x) < 2} are tolerated. This is not true
of \code{\link[stats:model.matrix]{stats::model.matrix()}} or \code{\link[Matrix:sparse.model.matrix]{Matrix::sparse.model.matrix()}}.
}
\examples{
x <- factor(sample(5L, 20L, replace = TRUE))
epigrowthfit:::factor_to_matrix(x, sparse = TRUE, intercept = TRUE)
epigrowthfit:::factor_to_matrix(x, sparse = TRUE, intercept = FALSE)
epigrowthfit:::factor_to_matrix(x, sparse = FALSE, intercept = TRUE)
epigrowthfit:::factor_to_matrix(x, sparse = FALSE, intercept = FALSE)

}
\keyword{internal}
