% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.egf}
\alias{fitted.egf}
\alias{coef.egf}
\title{Extract fitted values}
\usage{
\method{fitted}{egf}(
  object,
  par = get_par_names(object, link = TRUE),
  subset = NULL,
  append = NULL,
  link = TRUE,
  se = FALSE,
  .subset = NULL,
  .append = NULL,
  ...
)

\method{coef}{egf}(
  object,
  par = get_par_names(object, link = TRUE),
  subset = NULL,
  append = NULL,
  link = TRUE,
  se = FALSE,
  .subset = NULL,
  .append = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{par}{A subset of \code{get_par_names(object, link = TRUE)} naming nonlinear
model parameters for which fitted values should be retrieved.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link[=make_combined]{make_combined()}}). Must evaluate to a logical vector
indexing rows of the data frame, and thus fitting windows.
Fitted values are retrieved only for indexed windows.
The default (\code{NULL}) is to consider all windows.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link[=make_combined]{make_combined()}}) to be included with the result.
The default (\code{NULL}) is to append nothing.}

\item{link}{A logical scalar. If \code{FALSE}, then fitted values are inverse
link-transformed.}

\item{se}{A logical scalar. If \code{link = TRUE} and \code{se = TRUE}, then
approximate (delta method) standard errors on fitted values
are reported. Note that standard errors are required for
subsequent use of \code{\link[=confint.egf_fitted]{confint.egf_fitted()}}.}

\item{.subset}{A logical vector, to be used (if non-\code{NULL}) in place
of the result of evaluating \code{subset}.}

\item{.append}{A character vector listing variable names, to be used
(if non-\code{NULL}) in place of the result of evaluating \code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame inheriting from class \code{"egf_fitted"}, with variables:
\item{\code{par}}{
Nonlinear model parameter,
from \code{get_par_names(object, link = link)}.
}
\item{\code{ts}}{
Time series, from \code{levels(object$endpoints$ts)}.
}
\item{\code{window}}{
Fitting window, from \code{levels(object$endpoints$window)}.
}
\item{\code{estimate}}{
Fitted value of parameter \code{par} in fitting window \code{window}.
}
\item{\code{se}}{
(If \code{link = TRUE} and \code{se = TRUE}.)
Approximate (delta method) standard error on \code{value}.
}
}
\description{
Extracts fitted values of nonlinear model parameters.
The fitted value for a given fitting window is obtained
by adding
(i) the population fitted value computed from the
relevant fixed effects coefficients and
(ii) all (if any) random effects, with random effects
coefficients set equal to their conditional modes.
}
\details{
\code{coef.egf()} is currently an alias for \code{fitted.egf()}.

See topic \code{\link{nse}} for details on nonstandard evaluation
of \code{subset} and \code{append}.
}
\seealso{
\code{\link[=confint.egf_fitted]{confint.egf_fitted()}}
}
