% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\name{fitted.epigrowthfit}
\alias{fitted.epigrowthfit}
\title{returns the fitted values}
\usage{
\method{fitted}{epigrowthfit}(object, CI = TRUE, ...)
}
\value{
a dataframe object with columns time and deaths
}
\description{
returns the fitted values
}
