% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef.egf}
\alias{fixef.egf}
\alias{fixef}
\title{Extract fixed effects coefficients}
\usage{
\method{fixef}{egf}(object, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_fixef"}, with one row per coefficient and variables:
\item{name}{
  Coefficient name in the full parameter vector \code{object$best}.
}
\item{par}{
  Top level nonlinear model parameter,
  from \code{\link{get_names_top}(object, link = TRUE)}.
}
\item{term}{
  Term from fixed effects component of mixed effects \link{formula}
  for parameter \code{par}.
}
\item{colname}{
  Corresponding column name in the fixed effects design matrix
  \code{object$tmb_args$data$X}.
}
\item{estimate}{
  Coefficient estimate.
}
}
\description{
Retrieve the coefficients of the fixed effects component
of a mixed effects model.
}
