% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_control_default}
\alias{get_control_default}
\title{Get control default}
\usage{
get_control_default(f = c("plot.egf"), ...)
}
\arguments{
\item{f}{A character string giving the name of a function.}

\item{...}{Optional arguments to the function named \code{f}.}
}
\value{
A named list.
}
\description{
Retrieves the default value of the \code{control} argument of functions
in the \pkg{epigrowthfit} namespace. Currently, the only function
with a \code{control} argument is \code{\link[=plot.egf]{plot.egf()}}.
}
