% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_factor}
\alias{get_factor}
\title{Evaluate a term label in a data frame}
\usage{
get_factor(term_label, frame)
}
\arguments{
\item{term_label}{A character string giving a colon-separated list of names
of factors in \code{frame}.}

\item{frame}{A data frame containing factors.}
}
\value{
A factor of length \code{nrow(frame)} giving the interaction
specified by \code{term_label}. Unused levels are dropped.

In the special case \code{term_label = "(1)"}, the result is
\code{factor(rep("(1)", nrow(frame)))}.
}
\description{
Constructs the interaction specified by a term label
(a character string of the form \code{"f1:...:fn"}) from
factors in a data frame.
}
\examples{
f <- function() factor(sample(5L, 20L, replace = TRUE))
d <- data.frame(x = f(), y = f())
epigrowthfit:::get_factor("x:y", frame = d)
epigrowthfit:::get_factor("(1)", frame = d)

}
\keyword{internal}
