% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_flag}
\alias{get_flag}
\title{Get flags}
\usage{
get_flag(type, enum)
}
\arguments{
\item{type, enum}{Character strings giving the name of an enumerated type
and a name of an enumerator of that type, respectively.}
}
\value{
An integer.
}
\description{
Returns the underlying integer value of an enumerator in the
C++ template.
}
\details{
This function is kept synchronized with the C++ template
using script \code{utils/update_flags.R}.
}
\keyword{internal}
