% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init.R
\name{get_init}
\alias{get_init}
\title{Define fitting windows and initial parameter estimates}
\usage{
get_init(
  time,
  cases,
  model = "exponential",
  error = "poisson",
  theta0 = NULL,
  r_if_leq0 = 0.1,
  min_wlen = switch(model, exponential = 2, logistic = 3, richards = 4) + (error ==
    "nbinom"),
  peak = min_wlen - 1 + which.max(cases[min_wlen:length(cases)]),
  first = NULL,
  first_level = NULL,
  skip_zero = TRUE,
  debug_lm = FALSE
)
}
\arguments{
\item{time}{A numeric vector listing (increasing) time points.}

\item{cases}{A numeric vector with length \code{length(time)}.
\code{cases[i]} is the number of cases observed at time \code{time[i]}.
Here, "cases" can be (reported) infections or (reported) deaths.
Missing values are not tolerated.}

\item{model}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a phenomenological model.}

\item{error}{One of \code{"poisson"} and \code{"nbinom"},
indicating an observation model.}

\item{theta0}{A list with numeric scalar elements specifying
initial estimates of these parameters:

\describe{
\item{\code{r}}{Initial growth rate expressed in reciprocal units of \code{time}.}
\item{\code{x0}}{Expectation of the number of cases observed
at time \code{time[1]}. Used only if \code{model = "exponential"}.
}
\item{\code{K}}{Expectation of the number of cases observed over the full
course of the epidemic (i.e., the expected final epidemic size).
Used only if \code{model \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{Expectation of the time at which the epidemic attains
half of its final size. Used only if
\code{model \%in\% c("logistic", "richards")}
}
\item{\code{p}}{Shape parameter. Used only if \code{model = "richards"}.}
\item{\code{nbdisp}}{Dispersion parameter. Used only if \code{error = "nbinom"}.}
}

\code{theta0} can be \code{NULL} or a list specifying only a subset of
the relevant parameters. In this case, the absent parameters
are set internally. See Value.}

\item{r_if_leq0}{A numeric scalar. Assigned to \code{r}
if \code{!r \%in\% names(theta0)} and \code{beta1 <= 0}.}

\item{min_wlen}{An integer scalar. The minimum number
of observations in the fitting window. Ignored if
less than the number of model parameters. See Details.}

\item{peak}{An integer in \code{seq_along(cases)} indexing
the peak in \code{cases}. The index of the last observation
in the fitting window will be \code{peak} or \code{peak+1}.
See Details.}

\item{first}{An integer in \code{seq_along(cases)} indexing
the first observation in the fitting window, or \code{NULL}.
See Details.}

\item{first_level}{A numeric scalar or \code{NULL}. Can be
used to define \code{first} if \code{first = NULL}. See Details.}

\item{skip_zero}{A logical scalar. If \code{TRUE}, then an
attempt is made when defining \code{first} to ensure that
\code{cases[first] > 0}. See Details.}

\item{debug_lm}{A logical scalar. If \code{TRUE}, then
a diagnostic plot of the linear model (see Value)
is generated in addition to the usual output.}
}
\value{
A list with integer elements \code{first}, \code{last}, \code{peak}
and \code{wlen} and a list element \code{theta0}.

\code{first:last} specifies the chosen fitting window and
\code{wlen = last-first+1} the window length. \code{peak} will
match its value in the function call. \code{last} will be
equal to  \code{peak} if  \code{model = "exponential"} and
equal to \code{max(length(cases), peak + 1)} otherwise.
\code{first} can differ from its value in the function
call depending on the values of other arguments (see Details).

\code{theta0} specifies the chosen initial parameter estimates.
\code{names(theta0)} is the subset of
\code{c("r", "x0", "K", "thalf", "p", "nbdisp")} relevant for
the indicated \code{model} and \code{error}. Parameter values supplied
in the function call are retained. Parameters not specified
in the function call are defined as follows:

\describe{
\item{\code{r}}{The slope of a linear least squares fit to
\code{log(cases + 0.1)} within the first half of the chosen
fitting window. If the slope is negative or zero, then
\code{r} is assigned the value of argument \code{r_if_leq0}.
}
\item{\code{x0}}{The exponential of the intercept of the
linear fit described above.
}
\item{\code{K}}{\code{sum(cases)}}
\item{\code{thalf}}{\code{times[peak]}}
\item{\code{p}}{1.0001}
\item{\code{nbdisp}}{1}
}
}
\description{
Creates a list specifying a fitting window and initial
parameter estimates. Pains are taken to define a window
and estimates that are reasonable in the absence of any
user input.
}
\section{Fitting window selection}{
The fitting window ends at index \code{last = peak}, where \code{last = peak}
if \code{model = "exponential"} and \code{last = max(length(cases), peak + 1)}
otherwise.

The length \code{wlen = last-first+1} of the fitting window is constrained
to be at least \code{m = max(min_wlen, npar)}, where \code{npar} is the number
of parameters being fit:
2 for \code{model = "exponential"},
3 for \code{model = "logistic"}, and
4 for \code{model = "richards"}, plus
0 for \code{error = "poisson"} and
1 for \code{error = "nbinom"}.
This implies two constraints on the arguments: \code{peak >= m}
if \code{model = "exponential"} and \code{peak >= m-1} otherwise,
and \code{first <= peak-m+1}. An error will be thrown
if arguments \code{peak} and \code{first} (unless \code{first = NULL})
do not conform to these constraints.

If \code{first = NULL} and \code{first_level = NULL}, then \code{first}
is set internally to the greatest index \code{i} in \code{1:(peak-m+1)}
for which \code{cases[i] = min(cases[1:(peak-m+1)])}.
If \code{first = NULL} and \code{first_level != NULL}, then \code{first}
is set internally to one plus the greatest index \code{i} in
\code{1:(peak-m)} for which \code{cases[i] < first_level * max(cases)}.
(If no such index exists, then \code{first} is set to 1.)
Finally, if \code{first != NULL}, then the supplied value is
respected (unless \code{skip_zeros = TRUE}) and \code{first_level}
is ignored.

If \code{skip_zeros = TRUE} and \code{first} as defined above
is such that \code{cases[first] = 0} and \code{first < peak-m+1},
then \code{first} is increased by one until \code{cases[first] > 0}
or \code{first = peak-m+1}.
}

\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
get_init(
  time  = ontario$time,
  cases = ontario$new_confirmations,
  model = "richards",
  error = "nbinom"
)

}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
