% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Get nicely formatted tick labels}
\usage{
get_labels(at)
}
\arguments{
\item{at}{A numeric vector listing tick coordinates, probably
generated by \code{\link[graphics:axTicks]{graphics::axTicks()}}.}
}
\value{
An expression vector of length \code{length(at)} listing
tick labels.
}
\description{
Generates "mantissa x 10^power" tick labels for axes.
Intended mainly for count data.
}
\keyword{internal}
