% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_link}
\alias{get_link}
\alias{get_inverse_link}
\title{Get link and inverse link}
\usage{
get_link(s)

get_inverse_link(s)
}
\arguments{
\item{s}{A character string, either \code{"log"} or \code{"logit"}.}
}
\value{
A function.
}
\description{
Retrieve the link or inverse link function associated with a string.
}
\examples{
# get_link("log")
# get_inverse_link("logit")
# get_link("logit")
# get_inverse_link("log")

}
\keyword{internal}
