% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model}
\alias{get_model}
\title{return a model object (deterministic/mean trajectory) for fitting}
\usage{
get_model(model, add_const = FALSE, s_lower = 0.1, link_vals = NULL)
}
\arguments{
\item{model}{base model name ("exp","logistic","richards")}

\item{add_const}{(logical) add a constant incidence term?}

\item{s_lower}{(numeric) for \code{model=="richards"} only
(otherwise ignored): add lower bound above zero for shape (s)
parameter?}

\item{link_vals}{link functions}
}
\description{
return a model object (deterministic/mean trajectory) for fitting
}
