% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{get_names_top}
\alias{get_names_top}
\alias{get_names_top.default}
\alias{get_names_top.egf_model}
\alias{get_names_top.egf}
\alias{get_names_top.tmb_data}
\title{Get top level nonlinear model parameter names}
\usage{
get_names_top(object, ...)

\method{get_names_top}{default}(object, link = TRUE, ...)

\method{get_names_top}{egf_model}(object, link = TRUE, ...)

\method{get_names_top}{egf}(object, link = TRUE, ...)

\method{get_names_top}{tmb_data}(object, link = TRUE, ...)
}
\arguments{
\item{object}{An \code{"\link{egf_model}"}, \code{"\link{egf}"},
or \code{"\link[=egf_make_tmb_data]{tmb_data}"} object
specifying a top level nonlinear model. Otherwise,
\code{\link{NULL}}.}

\item{...}{Unused optional arguments.}

\item{link}{A \link{logical} flag. If \code{TRUE},
then \code{"link(name)"} is returned instead of \code{"name"}.}
}
\value{
A \link{character} vector giving the subset of names relevant to
\code{object}, or the complete set if \code{object = \link{NULL}}.
}
\description{
Retrieves the names used internally for top level nonlinear model
parameters.
}
\examples{
get_names_top(NULL, link = FALSE)
get_names_top(NULL, link = TRUE)

model <- egf_model()
get_names_top(model, link = FALSE)
get_names_top(model, link = TRUE)

}
