% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_par_info}
\alias{get_par_info}
\title{Describe elements of a parameter vector}
\usage{
get_par_info(par, tmb_data)
}
\arguments{
\item{par}{A TMB-generated parameter vector, renamed using \code{\link[=rename_par]{rename_par()}},
so that elements have names of the form
\code{"beta[i]"}, \code{"log_sd_b[j]"}, or \code{"b[k]"}.}

\item{tmb_data}{A \code{"tmb_data"} object returned by \code{\link[=make_tmb_data]{make_tmb_data()}}.}
}
\value{
A data frame with one row per element of \code{beta}, \code{log_sd_b}, \code{b},
and \code{u = decontrast_beta(beta, tmb_data$fnl)} (in that order) and
six variables:
\item{\code{vector}}{
A factor with levels \code{beta}, \code{log_sd_b}, \code{b}, and
\code{.decontrast(beta)}, allowing the data frame to be split by vector.
}
\item{\code{element}}{
A character vector concatenating \code{names(beta)}, \code{names(log_sd_b)},
\code{names(b)}, and \code{names(u)}.
}
\item{\code{estimate}}{
A numeric vector concatenating \code{beta}, \code{log_sd_b}, \code{b}, and \code{u}.
}
\item{\code{response}}{
A factor naming the relevant response variable, i.e., the response
variable of the generalized linear mixed effects model to which
the parameter belongs. For example, when fitting an exponential
model, \code{levels(response)} will include \code{"log_r"} and \code{"log_c0"}.
}
\item{\code{term}}{
A factor naming the relevant term of the mixed effects model
formula.
}
\item{\code{level}}{
A character vector. For elements of \code{u} and \code{b} (within-group
means and within-group random effects), the appropriate element
of \code{levels(term_label)} (i.e., the group label) is given. For
elements of \code{beta} (between-group means and within-group offsets),
\code{".mean(term_label)"} or \code{".offset(group_label)"} is given. For
elements of \code{log_sd_b} (log standard deviations of within-group
random effects), \code{".log_sd_b(term_label)"} is given.
}
}
\description{
Describe elements of a parameter vector
}
\keyword{internal}
