% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_term_labels}
\alias{get_term_labels}
\title{Get term labels from a formula}
\usage{
get_term_labels(f)
}
\arguments{
\item{f}{A formula.}
}
\value{
If \code{f} has no variable terms, as in \code{f = ~1}, then \code{"(1)"}.
Otherwise, the result of \code{labels(terms(f))} \emph{after} replacing
terms of \code{f} of the form \code{(1 | rhs_of_bar)} with \code{rhs_of_bar}.
}
\description{
A wrapper for \code{labels(terms())} with special handling of certain
formulae.
}
\examples{
get_term_labels(~w + x + y:z)
get_term_labels(~1)
get_term_labels(~w*x + y/z) # crosses and nests are expanded
get_term_labels(~w*x + (1 | y/z)) # bars are ignored

}
\keyword{internal}
