% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_yax_cex}
\alias{get_yax_cex}
\title{Find size for y-axis tick labels}
\usage{
get_yax_cex(labels, mex, csi = NULL)
}
\arguments{
\item{labels}{A character or expression vector listing \emph{y}-axis tick labels.}

\item{mex}{A positive number. The (desired) width of the widest tick label
as a number of margin lines. See \code{\link[graphics:par]{graphics::par()}}.}

\item{csi}{A positive number. The width of one margin line in inches.
The default is \code{par("csi")}. See \code{\link[graphics:par]{graphics::par()}}.}
}
\value{
\code{mex * csi / mlw}, where \code{mlw} is the width of the widest tick label
in inches.
}
\description{
Find \code{cex.axis} such that the widest \emph{y}-axis tick label generated by
\code{axis(labels, cex.axis, las = 1)} is exactly \code{mex} margin lines wide.
}
\keyword{internal}
