% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{gof,epigrowthfit-method}
\alias{gof,epigrowthfit-method}
\title{returns the goodness of fit of epigrowthfit object}
\usage{
\S4method{gof}{epigrowthfit}(obj, agg = NULL, debug = FALSE)
}
\value{
a vector
}
\description{
returns the goodness of fit of epigrowthfit object
}
