% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{gof}
\alias{gof}
\title{returns the goodness of fit of epigrowthfit object}
\usage{
gof(obj, ...)
}
\value{
a vector
}
\description{
goodness of fit is defined as \code{cor(predicted,observed)^2}
}
