% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{grad,Model-method}
\alias{grad,Model-method}
\title{compute gradients}
\usage{
\S4method{grad}{Model}(obj, data = NULL, par, ...)
}
\arguments{
\item{data}{a dataframe object holding inut values, if NULL, take from ...}

\item{par}{a named vector (or list) of parameters to compute the derivatives}

\item{...}{the input values and parameter values}
}
\value{
a dataframe with each column as a partial derivative values
}
\description{
compute the gradients of the model negative log-likelihood as a function of the parameters
}
