% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{growthRate,epigrowthfit-method}
\alias{growthRate,epigrowthfit-method}
\title{returns the growth rate and its 95\% confidence interval}
\usage{
\S4method{growthRate}{epigrowthfit}(obj)
}
\value{
a vector
}
\description{
returns the growth rate and its 95\% confidence interval
}
