% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{growthRate}
\alias{growthRate}
\title{returns growth rate and its 95\% confidence interval}
\usage{
growthRate(obj, ...)
}
\value{
a vector
}
\description{
returns growth rate and its 95\% confidence interval
}
