% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formula.R
\name{gsub_bar_plus}
\alias{gsub_bar_plus}
\title{Replace `|` with `+` in formula terms}
\usage{
gsub_bar_plus(x)
}
\arguments{
\item{x}{A \link{formula}.}
}
\value{
\code{x} with \code{`|`} in terms of the form
\code{(expression1 | expression2)} replaced with \code{`+`}.
}
\description{
Replaces the \code{`|`} operator in mixed effects model \link{formula}e
with the \code{`+`} operator, enabling construction of model frames using
\code{\link{model.frame}} machinery.
}
\examples{
# gsub_bar_plus(~x + (1 | f))

}
\keyword{internal}
