% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{has_random}
\alias{has_random}
\title{Check for random effects}
\usage{
has_random(object)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}} or a \code{"tmb_data"} object
returned by \code{\link[=make_tmb_data]{make_tmb_data()}}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Determine whether an object specifies a random effects model.
}
\keyword{internal}
