% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{husting}
\alias{husting}
\title{Husting wills}
\format{
A data frame with 246 rows and 5 variables:

\describe{
\item{\code{date}}{Date formatted as YYYY-MM-DD, weekly within
each level of \code{outbreak} with longer gaps between outbreaks.
}
\item{\code{time}}{Decimal date, equal to year plus fraction
of year. The fraction is 0 on January 1 and either 1/365
or 1/366 on January 2.
}
\item{\code{wills}}{Count of wills written. Within a given level
of \code{outbreak}, \code{wills[i]} is the number of wills written
between \code{date[i-1]+1} and \code{date[i]}.
}
\item{\code{outbreak}}{An ordered factor variable with 4 levels,
partitioning the time series into distinct plague outbreaks.
Levels are named roughly according to the years in which
the outbreaks took place: 1348, 1361, 1368, and 1375.
}
\item{\code{severity}}{An ordered factor variable with levels \code{"minor"}
and \code{"major"}, indicating outbreak severity. All 14th century
outbreaks are classified as major, only because the data are too
sparse to distinguish between minor and major outbreaks
(in contrast with wills probated in the
\link[=canterbury]{Prerogative Court of Canterbury}).
}
}
}
\usage{
data(husting)
}
\description{
Weekly counts of wills probated in the Court of Husting
during 24 plague epidemics in 16th and 17th century London.
}
\examples{
data(husting)
subset(husting, outbreak == 1375)
}
\keyword{datasets}
