% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_place_ragged_apply}
\alias{in_place_ragged_apply}
\title{Apply length-preserving functions to ragged vectors}
\usage{
in_place_ragged_apply(x, index, f)
}
\arguments{
\item{x}{A \link{vector} or \link[=data.frame]{data frame}.}

\item{index}{A \link{factor} (insofar that \code{\link{as.factor}(index)} is a factor)
defining a grouping of the elements or rows of \code{x}.}

\item{f}{A \link{function} or \link{list} of one or more functions to be applied
to the subsets of \code{x} defined by \code{index}. Functions are recycled
to the number of levels of \code{index} (unused levels are not dropped).
Each function must accept an initial \link{vector} argument matching
\code{\link{typeof}(x)} (or \code{typeof(x[[j]])} for all \code{j}
if \code{x} is a \link[=data.frame]{data frame}) and return a vector
of the same length (though not necessarily of the same type). It is
expected that returned vectors have compatible \link{mode}s so that
they can be concatenated in a sensible way.}
}
\value{
If \code{x} is a \link{vector}, then a vector of the same length.
If \code{x} is a \link[=data.frame]{data frame}, then a data frame
with the same dimensions.
}
\description{
Modifies ragged vectors in place by replacing each group of elements
with the result of applying a length-preserving function to that group
of elements.
}
\details{
Let \code{f} be a list of \code{\link{nlevels}(index)} functions,
and let \code{k = \link{split}(\link{seq_along}(index), index)}.
For vectors \code{x}, function \code{f[[i]]} is applied to
\code{x[k[[i]]]}.
For data frames \code{x}, function \code{f[[i]]} is applied to
\code{x[[j]][k[[i]]]} for all \code{j}.
}
\examples{
## x <- 1:10
## in_place_ragged_apply(x, index = gl(2L, 5L), f = list(cumprod, function(x) x - mean(x)))
##
## x <- as.data.frame(replicate(3L, c(exp(rnorm(10L)), qlogis(runif(10L)))))
## in_place_ragged_apply(x, index = gl(2L, 10L), f = list(log, plogis))

}
\keyword{internal}
