% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{is_bar}
\alias{is_bar}
\title{Is an object a call to \code{|}?}
\usage{
is_bar(x)
}
\arguments{
\item{x}{An \R object, typically a formula term.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a call to the \code{|} operator.
}
\examples{
# x <- call("|", as.name("f"), as.name("g"))
# x_plus_x <- call("+", x, x)
# is_bar(x)
# is_bar(x_plus_x)

}
\keyword{internal}
