% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_constant}
\alias{is_constant}
\title{Test whether an atomic vector is "constant"}
\usage{
is_constant(x, na.rm = FALSE, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{An atomic vector or data frame. For data frames \code{x},
\code{is_constant(x)} tests whether all listed vectors are
(individually) constant.}

\item{na.rm}{A logical scalar. If \code{TRUE}, then \code{NA} in vectors \code{x} are ignored.}

\item{tol}{A positive number. Numeric vectors \code{x} are considered constant
if and only if the distance from \code{min(x)} to \code{max(x)} is less
than \code{tol}.}
}
\value{
\code{TRUE}, \code{FALSE}, or \code{NA}.
}
\description{
Tests whether the elements of an atomic vector are equal
(or nearly equal in the case of numeric vectors).
}
\details{
\code{TRUE} is returned for zero-length \code{x}.
}
\examples{
# x <- c(0, 1e-03, NA)
# is_constant(x, na.rm = TRUE, tol = 1e-02)
# is_constant(x, na.rm = TRUE, tol = 1e-04)
# is_constant(x, na.rm = FALSE, tol = 1e-02)
# is_constant(x, na.rm = FALSE, tol = 1e-04)

}
\keyword{internal}
