% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latent.period}
\alias{latent.period}
\title{Estimated latent period for pneumonic plague}
\format{
A data frame containing columns:
\describe{
\item{\code{days}}{days since infection}
\item{\code{frequency}}{frequency distribution of the length of the latent period}
\item{\code{prediction}}{fitted lognormal distribution}
}
}
\source{
These data come from Figure 1A of Gani and Leach (2004).
}
\usage{
latent.period
}
\description{
Estimated latent period for pneumonic plague
}
\references{
\insertRef{GaniLeac04}{epigrowthfit}
}
\keyword{datasets}
