% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locf}
\alias{locf}
\title{Last observation carried forward}
\usage{
locf(x, x0 = NULL)
}
\arguments{
\item{x}{An \link{atomic} vector.}

\item{x0}{An \link{atomic} scalar to replace leading missing values,
for which there are no earlier observations. The default
(\code{\link{NULL}}) is to preserve these missing values.}
}
\value{
\code{x} with missing values replaced.
}
\description{
Replaces missing values (\code{\link{NA}} and, in \link{double}
vectors, \code{\link{NaN}}) with most recent earlier observations.
}
\examples{
# x <- c(NA, NA, 1, NA, 2, 2, 3, NA)
# locf(x)
# locf(x, x0 = 0)

}
\keyword{internal}
