% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfitTMB-data.R
\docType{data}
\name{londonparish}
\alias{londonparish}
\title{Burials in London parish registers}
\format{
A data frame with 1320 rows and 5 variables:

\describe{
\item{\code{date}}{Date formatted as YYYY-MM-DD, weekly within
each level of \code{outbreak} with longer gaps between outbreaks.
}
\item{\code{time}}{Decimal date, equal to year plus fraction
of year. The fraction is 0 on January 1 and either 1/365
or 1/366 on January 2.
}
\item{\code{burials}}{Count of burials. Within a given level
of \code{outbreak}, \code{burials[i]} is the number of burials
registered between \code{date[i-1]+1} and \code{date[i]}.
}
\item{\code{outbreak}}{An ordered factor variable with 24 levels,
partitioning the time series into distinct plague outbreaks.
Levels are named roughly according to the years in which
the outbreaks took place: 1563, 1578, ..., 1647, and 1665.
}
\item{\code{severity}}{An ordered factor variable with levels
\code{"minor"} and \code{"major"}, indicating outbreak severity.
An outbreak is classified as major if and only if plague
deaths per week per 1000 individuals exceeded 5 at least once.
}
}
}
\usage{
data(londonparish)
}
\description{
Weekly counts of burials listed in extant parish registers
during 24 plague epidemics in 16th and 17th century London.
}
\examples{
data(londonparish)
subset(londonparish, outbreak == 1665) # Great Plague of London
}
\keyword{datasets}
