% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{londonparishes}
\alias{londonparishes}
\title{Burials in London parish registers}
\format{
A \link[=data.frame]{data frame} with 14198 rows and 4 variables:
\describe{
\item{\code{Date}}{
  A \link{Date} vector. Weekly, except for 340 instances of 8-day spacing.
}
\item{\code{burials}}{
  An \link{integer} vector. \code{burials[i]} is the number of burials
  registered from the end of \code{Date[i-1]} to the end of \code{Date[i]}.
}
\item{\code{outbreak}}{
  An ordered \link{factor}, \link{split}ting the time series by plague
  outbreak. \link[=levels]{Levels} indicate the years in which outbreaks
  took place: 1563, 1578, ..., 1647, and 1665 (the Great Plague of London).
}
\item{\code{severity}}{
  An ordered \link{factor}, classifying outbreaks as \code{"minor"}
  or \code{"major"}. An outbreak is classified as major if and only if
  plague deaths per week per 1000 individuals exceeded 5 at least once.
}
}
}
\source{
These data were retrieved by Cummins et al. with permission from
\href{https://www.ancestry.com/}{Ancestry.com}.
They have been made publicly available
\href{http://www.neilcummins.com}{here}.
}
\usage{
data(londonparishes)
}
\description{
Weekly counts of burials listed in extant parish registers
during 24 plague epidemics in 16th and 17th century London.
}
\examples{
data(londonparishes)
subset(londonparishes, outbreak == 1665, -outbreak) # Great Plague of London
}
\references{
Cummins N, Kelly M, \'{O} Gr\'{a}da C. Living standards and plague
in London, 1560--1665. Econ Hist Rev. 2016;69(1):3--34.
}
\keyword{datasets}
