% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lpapply}
\alias{lpapply}
\title{Apply length-preserving functions}
\usage{
lpapply(x, index, f)
}
\arguments{
\item{x}{A \link{vector} or \link[=data.frame]{data frame}.}

\item{index}{A \link{factor} defining a partition of the elements or rows of \code{x}.
\code{\link{split}(x, index)} should be a valid expression.}

\item{f}{A \link{function} or \link{list} of one or more functions to be applied
to subsets of \code{x}. These are recycled to match the number of levels
of \code{index} (unused levels are not dropped). Each function must
accept an initial \link{vector} argument matching \code{\link{typeof}(x)}
(or \code{typeof(x[[j]])} for all \code{j} if \code{x} is a
\link[=data.frame]{data frame}) and return a vector of the same length
(but not necessarily of the same type). It is expected that returned
vectors have compatible \link{mode}s: if one has an \link{atomic} mode,
then all should have an atomic mode (likewise for modes \code{"list"}
and \code{"expression"}).}
}
\value{
If \code{x} is a \link{vector}, then a vector of the same length.
If \code{x} is a \link[=data.frame]{data frame}, then a data frame
of the same length, with the same number of rows.
}
\description{
Replaces subsets of ragged vectors and data frames with the results
of length-preserving functions.
}
\details{
Let \code{k = split(seq_len(f), f)}.
For \link{vector}s \code{x}, function \code{f[[i]]} is applied
to \code{x[k[[i]]]}.
For \link[=data.frame]{data frame}s \code{x}, function \code{f[[i]]}
is applied to \code{x[[j]][k[[i]]]} for all \code{j}.
}
\examples{
# x <- 1:10
# lpapply(x, index = gl(2L, 5L), f = list(cumprod, function(x) x - mean(x)))
#
# x <- as.data.frame(replicate(3L, c(exp(rnorm(10L)), qlogis(runif(10L)))))
# lpapply(x, index = gl(2L, 10L), f = list(log, plogis))

}
\keyword{internal}
