% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{make_combined}
\alias{make_combined}
\title{Construct combined model frame}
\usage{
make_combined(object)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}
}
\value{
A \link[=data.frame]{data frame} combining (in the sense
of \code{\link{cbind}}) all data frames in the \link{list}
\code{object$frame_par} and the data frame \code{object$frame_append}.
}
\description{
Joins in a single \link[=data.frame]{data frame} all mixed effects
\link[=model.frame]{model frames} and further variables specified
via \code{\link{egf}} argument \code{append}.
}
\details{
If a variable name occurs in multiple mixed effects model frames,
then only one instance is retained. Except in unusual cases
(possible only if model formulae have different formula environments),
all instances of a variable name are identical, and no information is lost.

Since the data frames being combined each correspond rowwise
to \code{object$endpoints}, so does the result.
}
\keyword{internal}
