% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_frame}
\alias{make_frame}
\title{Construct a model frame}
\usage{
make_frame(
  formula,
  fixed,
  random,
  group_by,
  data,
  index,
  curve,
  distr,
  excess,
  na_action
)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x} specifying one or more incidence
time series in long format, with \code{x} Date and \code{y} numeric.}

\item{fixed}{Named lists of formulae of the form \code{~rhs}, together specifying
a generalized linear mixed effects model (the fixed and random
components) for each parameter of the \emph{incidence} model indicated
by \code{curve}, \code{distr}, and \code{excess}.
A list of valid names can be obtained with
\code{get_par_names(curve, distr, excess, link = TRUE)}.
Alternatively, formulae (rather than lists of formulae) to be
recycled for all parameters.
Syntax is \code{\link[lme4:lmer]{lme4}}-like, but here all variables
must be factors.
\code{fixed} formulae are restricted to one term and so must be \code{~1}
(the default for each missing list element) or have the form
\code{~f1:...:fn}.
\code{random} formulae are restricted to sums of terms of the form
\code{(1 | f1:...:fn)}, \code{(1 | f1/.../fn)}, or \code{(1 | f1 * ... * fn)}.
Use \code{NULL} (the default for each missing list element) instead
of a formula to indicate absence of random effects.}

\item{random}{Named lists of formulae of the form \code{~rhs}, together specifying
a generalized linear mixed effects model (the fixed and random
components) for each parameter of the \emph{incidence} model indicated
by \code{curve}, \code{distr}, and \code{excess}.
A list of valid names can be obtained with
\code{get_par_names(curve, distr, excess, link = TRUE)}.
Alternatively, formulae (rather than lists of formulae) to be
recycled for all parameters.
Syntax is \code{\link[lme4:lmer]{lme4}}-like, but here all variables
must be factors.
\code{fixed} formulae are restricted to one term and so must be \code{~1}
(the default for each missing list element) or have the form
\code{~f1:...:fn}.
\code{random} formulae are restricted to sums of terms of the form
\code{(1 | f1:...:fn)}, \code{(1 | f1/.../fn)}, or \code{(1 | f1 * ... * fn)}.
Use \code{NULL} (the default for each missing list element) instead
of a formula to indicate absence of random effects.}

\item{group_by}{A formula of the form \code{~f1:...:fn},
such that \code{split(data, interaction(data[all.vars(group_by)]))}
splits \code{data} by time series. Use \code{~1} (the default) if there
is only one time series.}

\item{data}{A data frame, list, or environment containing the variables
named in \code{formula}, \code{fixed}, \code{random}, and \code{group_by}. Missing
values in incidence (\code{y} if \code{formula = y ~ x}) are tolerated
only if \code{na_action = "pass"}. Missing values in other variables
are not tolerated.}

\item{index}{A factor of length \code{nrow(data)} such that \code{split(data, index)}
splits \code{data} by fitting window, with \code{is.na(index)} indexing
rows of \code{data} not belonging to a fitting window. \code{NULL}
(the default) is equivalent to \code{rep(factor(0), nrow(data))}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline
mortality rate is estimated. Set to \code{TRUE} if what is
observed (\code{y} if \code{formula = y ~ x}) is multiple causes
mortality rather than disease mortality or disease incidence.}

\item{na_action}{A character string indicating how \code{NA} in the incidence vector
(\code{y} if \code{formula = y ~ x}) are handled. See \code{\link[stats:na.fail]{stats::na.fail()}}.}
}
\value{
A data frame containing the variables named in \code{formula}, \code{fixed},
and \code{random}, with attributes:
\item{\code{extra}}{
A data frame preserving unused data (see Details).
}
\item{\code{index}}{
A factor of length \code{nrow(frame)} such that \code{split(frame, index)}
splits the data frame by fitting window. Not usually identical
to the so-named argument (see Details).
}
\item{\code{group_by}}{
A character vector. The result of applying \code{\link[=all.vars]{all.vars()}} to the
so-named argument.
}
\item{\code{fixed_term_labels}, \code{random_term_labels}}{
Named lists of character vectors naming mixed effects model terms
for each parameter.
}
}
\description{
Constructs a model frame to be used by \code{\link[=egf]{egf()}},
while performing myriad checks on the arguments.
}
\details{
Rows of \code{data} belonging to time series without fitting windows
are discarded. Rows of \code{data[-discarded, ]} belonging to fitting
windows are returned. The remaining rows of \code{data[-discarded, ]}
are preserved in attribute \code{extra}. These are necessary as it
is desirable for \code{\link[=plot.egf]{plot.egf()}} to display entire time series,
not only segments to which a curve has been fit.
}
\examples{
r <- log(2) / 10
c0 <- 10
time <- 0:99
mu <- diff(c0 * exp(r * time))

date <- rep(as.Date(time, origin = "2020-01-01"), 4)
cases <- c(replicate(4, c(NA, rpois(length(mu), mu))))
continent <- factor(rep(c("asia", "europe"), each = 200))
country <- factor(rep(c("china", "japan", "france", "germany"), each = 100))
wave <- factor(rep(rep(c(1, 2), each = 50), 4))

data <- data.frame(date, cases, continent, country, wave)

x <- c(NA, 1, NA, 2, NA,
       NA, 3, NA, 4, NA,
       NA, 5, NA, 6, NA,
       NA, 7, NA, 8, NA)
index <- factor(rep(x, each = 20))

curve <- "exponential"
distr <- "pois"
excess <- FALSE
get_par_names(curve, distr, excess, link = FALSE)
## [1] "r" "c0"

formula <- cases ~ date
fixed <- list(
  r  = ~1,
  c0 = ~wave
)
random <- list(
  r  = ~(1 | continent/country) + (1 | wave),
  c0 = ~(1 | continent/country)
)
group_by <- ~country

frame <- epigrowthfit:::make_frame(
  formula, fixed, random, group_by,
  data, index,
  curve, distr, excess,
  na_action = "pass"
)

}
\keyword{internal}
