% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_help.R
\name{make_index}
\alias{make_index}
\title{Index fitting windows}
\usage{
make_index(date, ts, subset)
}
\arguments{
\item{date}{A Date vector listing time points from one or more time series.}

\item{ts}{A factor of length \code{length(date)} such that \code{split(date, ts)}
splits \code{date} by time series.}

\item{subset}{A named list of lists of integer vectors. \code{names(subset)}
must be a subset of \code{levels(ts)}. Integer vectors listed
in \code{subset[[name]]} must index disjoint segments of the
corresponding time series. That is, they must be disjoint
subsets of \code{seq_len(table(ts)[[name]])}.}
}
\value{
A data frame with \code{length(date)} rows,
listing two factors, \code{index} and \code{wave}.
}
\description{
Constructs a factor splitting time series in long format
by fitting window, to be passed to \code{\link[=egf]{egf()}}.
}
\details{
\code{index} has levels \code{seq_len(sum(lengths(subset)))}.
\code{split(date, index)} splits \code{date} by fitting window,
\code{is.na(index)} indexes points in \code{date} not belonging
to a window.

\code{index} can be assigned to the so-named argument of \code{\link[=egf]{egf()}}.

\code{wave} has levels \verb{c(0, seq_len(max(lengths(subset)))}.
It is a recoding of \code{index}, replacing \code{NA} with \code{0}
and replacing elements in the \code{i}th fitting window of
each time series (chronologically) with \code{i}.

If the \code{i}th fitting window of each time series corresponds
to the \code{i}th epidemic wave in that time series, and if
epidemic wave is being included as a fixed or random effect
in a mixed effects model, then \code{wave} can be passed to
\code{\link[=egf]{egf()}} as a variable in \code{data}.
}
