% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_args}
\alias{make_tmb_args}
\title{Create TMB infrastructure}
\usage{
make_tmb_args(frame, curve, distr, excess, sparse_X, par_init)
}
\arguments{
\item{frame}{A model frame returned by \code{\link[=make_frame]{make_frame()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline
mortality rate is estimated. Set to \code{TRUE} if what is
observed (\code{y} if \code{formula = y ~ x}) is multiple causes
mortality rather than disease mortality or disease incidence.}

\item{sparse_X}{A logical scalar. If \code{TRUE}, then the fixed effects design matrix
will be constructed in sparse format.}

\item{par_init}{A user-specified full parameter vector, or otherwise \code{NULL}.
See \code{\link[=egf]{egf()}}.}
}
\value{
A list with elements \code{data}, \code{parameters}, \code{map}, \code{random},
\code{DLL}, and \code{silent}.
}
\description{
Gathers necessary components of a call to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\examples{
example("make_frame", package = "epigrowthfit")
tmb_args <- epigrowthfit:::make_tmb_args(frame, curve, distr, excess,
                                         sparse_X = FALSE, par_init = NULL)
tmb_out <- do.call(TMB::MakeADFun, tmb_args)

}
\seealso{
\code{\link[=make_tmb_data]{make_tmb_data()}}, \code{\link[=make_tmb_parameters]{make_tmb_parameters()}}
}
\keyword{internal}
