% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{make_tmb_data}
\alias{make_tmb_data}
\title{Construct data objects for C++ template}
\usage{
make_tmb_data(frame, frame_par, model, control, do_fit, init)
}
\arguments{
\item{frame}{Elements of the \link{list} output of \code{\link{make_frames}}.}

\item{frame_par}{Elements of the \link{list} output of \code{\link{make_frames}}.}

\item{model}{An \code{"\link{egf_model}"} object specifying a nonlinear model
and a dispersion model to be estimated.}

\item{control}{An \code{"\link{egf_control}"} object specifying control parameters.}

\item{do_fit}{A \link{logical} flag. If \code{TRUE}, then \code{egf} returns early
with a \link{list} of optimization inputs.}

\item{init}{A \link{numeric} vector to be used as the full parameter vector for
the first likelihood evaluation. The default (\code{\link{NULL}}) is
to accept the internally generated default. Use \code{do_fit = FALSE}
to retrieve this default and other optimization inputs (in particular
\code{Y_init}; see Value), which can often be used to construct a more
informative starting point.}
}
\value{
[Below,
\code{n = \link{sum}(!\link{is.na}(frame$window))}
is the total number of time points belonging a fitting window,
\code{N = \link{nlevels}(frame$window)}
is the number of fitting windows, and
\code{p = \link{length}(frame_par)}
is the number of nonlinear and dispersion model parameters.]

A \link{list} inheriting from \link{class} \code{"tmb_data"},
with elements:
\item{t}{
  A \link[=double]{numeric} vector of length \code{n} giving time
  as a number of days since the earliest time point in the current
  fitting window.
}
\item{t_seg_len}{
  An \link{integer} vector of length \code{N} specifying the
  length of each fitting window as a number of time points.
}
\item{x}{
  A \link[=double]{numeric} vector of length \code{n-N} giving
  incidence in each fitting window. \code{x[i]} in window \code{k}
  is the number of cases observed from time \code{t[k+i-1]}
  to time \code{t[k+i]}.
}
\item{day_of_week_on_day0}{
  If \code{model$day_of_week > 0}, then an \link{integer} vector
  of length \code{N} indicating the first day of week in each
  fitting window, with value \code{i} in \code{0:6} mapping to
  the day of week \code{i} days after the reference day specified
  by \code{model$day_of_week}. Otherwise, an integer vector of
  the same length filled with \code{-1}.
}
\item{Yo}{
  The \link[=model.offset]{offset} matrix in dense format,
  with \code{N} rows and \code{p} columns.
}
\item{X}{
  The fixed effects design matrix in \link[Matrix:sparseMatrix]{sparse}
  or dense format, depending on \code{control$sparse_X}, with \code{N} rows.
}
\item{Z}{
  The random effects design matrix in \link[Matrix:sparseMatrix]{sparse}
  format, with \code{N} rows. If there are no random effects, then \code{Z}
  is an empty sparse matrix.
}
\item{Xs, Xd}{
  If \code{control$sparse_X = TRUE}, then \code{Xs = X} and \code{Xd}
  is an empty dense matrix. Otherwise, \code{Xd = X} and \code{Xs} is
  an empty sparse matrix.
}
\item{X_info, Z_info}{
  \link[=data.frame]{Data frame}s with \code{\link{ncol}(X)}
  and \code{\link{ncol}(Z)} rows, respectively, constructed
  by \code{\link{make_XZ_info}}. Row \code{j} describes the
  coefficient associated with column \code{j} of \code{X}
  or \code{Z}. \code{Z_info} has additional \link{factor}s
  \code{cor} and \code{vec} \link{split}ting coefficients
  by relation to a common covariance matrix and random vector,
  respectively.
}
\item{beta_seg_len, b_seg_len}{
  \link[=integer]{Integer} vectors of length \code{p} counting the
  columns of \code{X} and \code{Z}, respectively, that relate to
  a common nonlinear or dispersion model parameter.
}
\item{beta_seg_index, b_seg_index}{
  \link[=integer]{Integer} vectors of length \code{\link{ncol}(X)}
  and \code{\link{ncol}(Z)}, respectively, with values in \code{0:(p-1)}.
  These split the columns of \code{X} and \code{Z} by relation to
  a common nonlinear or dispersion model parameter.
}
\item{block_rows, block_cols}{
  \link[=integer]{Integer} vectors together giving the dimensions
  of each block of the random effects matrix.
}
\item{curve_flag, excess_flag, family_flag, day_of_week_flag, trace_flag, sparse_X_flag}{
  \link[=integer]{Integer} flags referencing so-named elements
  of \code{model} and \code{control}.
}
\item{predict_flag}{
  An \link{integer} flag set equal to 0 so that prediction code is not run.
}
Additional \link{integer} elements of the form \code{j_link_parameter}
(e.g., \code{j_log_r}) give the 0-index of nonlinear and dispersion
model parameter names (e.g., \code{"log(r)"}) in \code{names(frame_par)}.
The value \code{-1} is used for parameters not belonging to the model
being fit.
}
\description{
Gathers in a \link{list} data objects to be passed to
the package's C++ template via \pkg{TMB}'s \code{DATA_*} macros.
}
\seealso{
\code{\link[TMB]{MakeADFun}}
}
\keyword{internal}
