% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_parameters}
\alias{make_tmb_parameters}
\title{Construct parameter objects for C++ template}
\usage{
make_tmb_parameters(tmb_data, frame_ts, frame_par, curve, do_fit, init)
}
\arguments{
\item{tmb_data}{A \code{"tmb_data"} object returned by \code{\link[=make_tmb_data]{make_tmb_data()}}.}

\item{frame_ts}{Elements of the list output of \code{\link[=make_frames]{make_frames()}}.}

\item{frame_par}{Elements of the list output of \code{\link[=make_frames]{make_frames()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{do_fit}{A logical scalar used for debugging. If \code{TRUE}, then
\code{egf()} returns early with a list of optimization inputs.}

\item{init}{A full parameter vector for the first likelihood evaluation.
The default (\code{NULL}) is to accept the internally generated
default. Use \code{do_fit = FALSE} to retrieve this default and
other optimization inputs (in particular \code{Y_init}; see Value),
which can often be used to construct a more informative
starting point.}
}
\value{
A list with elements:
\item{\code{beta}}{
A numeric vector of length \code{sum(tmb_data$beta_seg_len)}
listing initial values for the fixed effects coefficients.
}
\item{\code{b}}{
If there are random effects, then a numeric vector of
length \code{sum(tmb_data$b_seg_len)} listing initial values
for the (unit variance scale) random effects coefficients.
If there are no random effects, then \code{b = NA_real_}.
}
\item{\code{theta}}{
If there are random effects, then a numeric vector of length
\code{sum(f(tmb_data$block_rows))}, where \code{f(n) = n*(n+1)/2},
specifying initial covariance matrices for all random vectors.
Let \code{u} be a segment of \code{theta} of length \code{f(n)},
corresponding to a random vector \code{x} of length \code{n}.
Then \code{u[1:n]} lists log standard deviations of the elements
of \code{x}, and \code{u[-(1:n)]} lists (in row-major order) the
subdiagonal elements of the unit lower triangular matrix \code{L}
in the Cholesky factorization
\code{S = A \%*\% t(A)},
\code{A = D^-0.5 \%*\% L},
\code{D = diag(diag(L \%*\% t(L)))}
of the correlation matrix \code{S} of \code{x}.
If there are no random effects, then \code{theta = NA_real_}.
}
}
\description{
Gathers in a list parameter objects to be passed to the C++ template
via TMB's \code{PARAMETER_} macros during the first likelihood evaluation.
See \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\details{
When \code{init = NULL}, naive estimates of nonlinear model parameters
are obtained for each fitting window as follows:
\describe{
\item{\code{r}}{
The slope of a linear model fit to \verb{log1p(cumsum(x)))}.
}
\item{\code{alpha}}{
\code{r*c0^(1-p)} if \code{curve = "subexponential"},
\code{r/log(K/c0)} if \code{curve = "gompertz"}.
These are the values obtained by setting the per capita growth
rate at time 0 in the subexponential and Gompertz models equal
to \code{r}, substituting the naive estimates of \code{r}, \code{c0}, \code{K}, and
\code{p}, and solving for \code{alpha}.
}
\item{\code{c0}}{
\code{exp(log_c0)}, where \code{log_c0} is the intercept of a linear model
fit to \code{log1p(cumsum(x))}.
}
\item{\code{tinfl}}{
\code{max(t)}, assuming that the fitting window ends near the time
of a peak in interval incidence.
}
\item{\code{K}}{
\code{2*sum(x)}, assuming that the fitting window ends near the time
of a peak in interval incidence \emph{and} that interval incidence is
roughly symmetric about the peak.
}
\item{\code{p}}{0.8}
\item{\code{a}, \code{b}, \code{nbdisp}, \code{w[123456]}}{1}
}
The naive estimates are log- or logit-transformed (all but \code{p}
use a log link), and, for each nonlinear model parameter, the
initial value of the \code{"(Intercept)"} coefficient in its fixed
effects model (if one exists) is taken to be the mean of the
link scale estimates across fitting windows.
The remaining elements of \code{beta} take initial value 0.
All elements of \code{b} and \code{theta} take initial value 0,
so that, initially, all random vectors follow a standard
normal distribution.
}
\keyword{internal}
