% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_window}
\alias{make_window}
\alias{make_wave}
\title{Index fitting windows}
\usage{
make_window(date, ts, endpoints)

make_wave(window, ts)
}
\arguments{
\item{date}{A Date vector listing time points from one or more time series
in long format. Must be increasing in each level of \code{ts}.}

\item{ts}{A factor of length \code{length(date)} such that \code{split(date, ts)}
splits \code{date} by time series.}

\item{endpoints}{A data frame with variables \code{ts} (factor), \code{start} (Date), and
\code{end} (Date) and one row per fitting window. Window \code{i} runs
from the end of date \code{start[i]} to the end of date \code{end[i]} in
time series \code{ts[i]}.}

\item{window}{A factor generated by \code{make_window(, ts, )}.}
}
\value{
A factor of length \code{length(date)}. For \code{make_window()}, there are
as many levels as windows. For \code{make_wave()}, the number of levels
is equal to the maximum number of windows found in one time series.
}
\description{
Constructs a factor suitable for \code{\link[=egf]{egf()}} argument \code{window}
from a data frame listing fitting window endpoints.
}
\details{
\code{make_window(date, ts, endpoints)} constructs a factor \code{window}
such that \code{split(date, window)} splits the time points in \code{date}
by fitting window.

\code{make_wave(window, ts)} recodes \code{window}, assigning level \code{i}
to the \code{i}th window in all time series. The resulting factor
is useful if windows within time series correspond to epidemic
waves and one wants to include epidemic wave as a variable in
a mixed effects model.

Missing values in \code{date} and \code{ts} are an error.
Rows of \code{endpoints} containing missing values are discarded silently.
Window endpoints not enclosing at least two time points are an error.
}
