% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_X}
\alias{make_X}
\alias{make_Z}
\title{Construct design matrices}
\usage{
make_X(x, frame, sparse)

make_Z(x, frame)
}
\arguments{
\item{x}{For \code{make_X()}, a fixed effects formula of the form \code{~terms}.
For \code{make_Z()}, a random effects term of the form \code{(terms | group)}.}

\item{frame}{A model frame (a data frame with an appropriate \code{terms} attribute)
defining the variables used in \code{x}.}

\item{sparse}{A logical scalar. If \code{TRUE}, then the design matrix is returned
in sparse format.}
}
\value{
A matrix with attributes:
\item{\code{contrasts}, \code{assign}}{
See \code{\link[stats:model.matrix]{stats::model.matrix()}}.
\code{contrasts} will be absent if \code{foo} in \code{x = ~foo} or
\code{x = (foo | group)} does not use factors. \code{assign} indexes
\code{labels(terms(~foo))} for \code{x = ~foo} and \code{x = (foo | group)}.
}
\item{\code{group}}{
(\code{make_Z()} only.) For \code{x = (foo | group)}, a factor of length
\code{ncol(Z)} with levels \code{levels(group)}, useful for splitting \code{Z}
into group-specific submatrices.
}
}
\description{
Construct the \code{X} matrix associated with a fixed effects
model formula or the \code{Z} matrix associated with a random
effects term.
}
\details{
\code{make_X(x, frame, sparse)} is equivalent
to \code{stats::model.matrix(x, frame)} or
\code{Matrix::sparse.model.matrix(x, frame)}
(depending on \code{sparse}), except that
columns containing only zeros are dropped
from the result.

\code{make_Z()} follows the steps outlined in
\code{vignette("lmer", "lme4")} to construct a
\code{Z} matrix for \code{x}, using \code{make_X()} to
construct the so-called raw model matrix
from \code{x[[2L]]}. The result is always sparse.
}
\keyword{internal}
