% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{match_link}
\alias{match_link}
\title{Get link and inverse link functions}
\usage{
match_link(f, inverse = FALSE)
}
\arguments{
\item{f}{A character string naming a link function,
either \code{"log"} or \code{"logit"}.}

\item{inverse}{A logical scalar. If \code{TRUE}, then the inverse is returned.}
}
\value{
A \link{function}.
}
\description{
Retrieve the link function named by a string, or its inverse.
}
\examples{
# match_link("log")
# match_link("log", inverse = TRUE)
# match_link("logit")
# match_link("logit", inverse = TRUE)

}
\keyword{internal}
