% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formula.R
\name{negate}
\alias{negate}
\title{Negate formula terms}
\usage{
negate(x)
}
\arguments{
\item{x}{A \link{call}, \link{name}, or \link{atomic} scalar.}
}
\value{
If \code{x} is a \link{call} to the \code{`-`} operator with
one argument \code{y}, then \code{y}. Otherwise, a call to
the \code{`-`} operator with \code{x} as its only argument.
}
\description{
Negates formula terms, taking some care to simplify double negatives.
}
\examples{
# x <- quote(x)
# minus_x <- call("-", x)
# identical(negate(x), minus_x)
# identical(negate(minus_x), x)

}
\keyword{internal}
