% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nse.R
\name{nse}
\alias{nse}
\alias{eval_subset}
\alias{eval_append}
\alias{eval_order}
\alias{eval_label}
\title{Nonstandard evaluation}
\usage{
eval_subset(subset, envir, enclos, .subset = NULL)

eval_append(append, envir, enclos, .append = NULL)

eval_order(order, envir, enclos, .order = NULL)

eval_label(label, envir, enclos, .label = NULL)
}
\arguments{
\item{subset, order, label}{Expressions to be evaluated in \code{envir}.}

\item{envir, enclos}{See \code{\link{eval}}. \code{\link{is.list}(envir)} must be \code{TRUE}
for \code{eval_append}.}

\item{.subset, .order, .append, .label}{\link[=atomic]{Atomic} vectors to be used (if non-\code{\link{NULL}})
in place of the result of evaluating the undotted argument.}

\item{append}{An expression to be evaluated in
\code{`\link{names<-}`(\link{as.list}(\link{seq_along}(envir)), \link{names}(envir))}.}
}
\value{
Let \code{r} be the result of evaluating the supplied expression.

\code{eval_subset} returns \code{r},
with \code{\link{NA}} (if any) coerced to \code{FALSE}.

\code{eval_order} returns \code{r} as is.

\code{eval_append} returns
\code{\link{match}(\link{names}(envir[r]), \link{names}(envir), 0L)},
with zeros (if any) deleted.

\code{eval_label} returns \code{\link{as.character}(r)},
repeated to length \code{\link{nrow}(envir)}
if \code{envir} is a \link[=data.frame]{data frame}.
}
\description{
Utilities for obtaining index vectors from unevaluated \code{subset},
\code{append}, and \code{order} expressions and character vectors from
unevaluated \code{label} expressions.
}
\details{
\code{subset} must evaluate to a \link{logical} vector.
The result must have length \code{\link{nrow}(envir)}
if \code{envir} is a \link[=data.frame]{data frame}.

\code{order} must evaluate to a permutation of \code{\link{seq_len}(n)}
for some \code{n}. \code{n} must be \code{nrow(envir)} if \code{envir}
is a data frame.

\code{append} must evaluate to an \link{atomic} vector
indexing \code{envir}.

\code{label} must evaluate to an \link{atomic} vector.
The result must have length of length 1 or length \code{nrow(envir)}
if \code{envir} is a data frame.
}
\section{Note}{

\code{subset} and \code{append} are processed similarly to arguments
\code{subset} and \code{select} of function \code{\link{subset}}.
See \code{\link{subset}} for additional usage examples.
}

\section{Warning}{

Nonstandard evaluation of \code{subset}, \code{order}, \code{append},
and \code{label} is intended only to make interactive use more convenient.
To avoid unexpected behaviour, especially when programming, use the dotted
arguments.
}

\examples{
year <- 2021L
data <- data.frame(
  month = sample(month.abb, 20L, replace = TRUE),
  day = sample(30L, 20L, replace = TRUE)
)

subset <- quote(grepl("^J", month) & day < 16L)
order <- quote(order(month, day))
append <- quote(-day)
label <- quote(sprintf("\%d-\%02d-\%02d", year, match(month, month.abb, 0L), day))

## eval_subset(subset, data, parent.frame())
## eval_order(order, data, parent.frame())
## eval_append(append, data, parent.frame())
## eval_label(label, data, parent.frame())

}
\keyword{internal}
