% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{optim_tmb_out}
\alias{optim_tmb_out}
\title{Optimize a TMB model object}
\usage{
optim_tmb_out(tmb_out, method_outer, ...)
}
\arguments{
\item{tmb_out}{The list output of \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}

\item{method_outer}{Character strings specifying optimizers to be used
for the outer and inner optimizations, respectively.
The outer optimization permits \code{\link[stats:nlminb]{stats::nlminb()}},
\code{\link[stats:nlm]{stats::nlm()}}, and certain optimizers available
through \code{\link[stats:optim]{stats::optim()}}. The inner optimization
permits \code{\link[TMB:newton]{TMB::newton()}} and the same \code{\link[stats:optim]{stats::optim()}}
methods. \code{method_inner} may have length greater than 1;
in this case, the listed optimizers will be tried
in turn until one succeeds.}

\item{...}{Unused optional arguments}
}
\value{
The list output of \code{\link[stats:nlminb]{stats::nlminb()}}, \code{\link[stats:nlm]{stats::nlm()}},
or \code{\link[stats:optim]{stats::optim()}}, depending on \code{method_outer}.
}
\description{
Evaluates a call to an optimizer constructed from the components
of a TMB model object.
}
\keyword{internal}
