% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{patch_fn}
\alias{patch_fn}
\alias{patch_gr}
\title{Patch TMB-generated functions}
\usage{
patch_fn(fn, inner_optimizer)

patch_gr(gr, inner_optimizer)
}
\arguments{
\item{fn, gr}{\link[=function]{Function}s from a \code{\link[TMB]{MakeADFun}}-generated
\link{list} object to be patched.}

\item{inner_optimizer}{A \link{list} of \code{"\link{egf_inner_optimizer}"} objects
specifying inner optimization methods to be tried in turn.}
}
\value{
A patched version of \link{function} \code{fn} or \code{gr}.
}
\description{
Define a wrapper function on top of \code{\link[TMB]{MakeADFun}}-generated
functions \code{fn} and \code{gr}, so that function and gradient evaluations
can retry inner optimization using fallback methods in the event that the
default method (usually \code{\link[TMB]{newton}}) fails.
}
\keyword{internal}
