% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{plague_infectious_period}
\alias{plague_infectious_period}
\title{Infectious period distribution (pneumonic plague)}
\format{
A data frame with 12 rows and 2 variables:

\describe{
\item{\code{days}}{[integer] Infectious period in days, from 1 to 12.}
\item{\code{relfreq}}{[numeric] Observed relative frequency
out of 225 cases.
}
}
}
\source{
Obtained from Figure 1B in \insertCite{GaniLeac04;textual}{epigrowthfit}.
}
\usage{
data(plague_infectious_period)
}
\description{
Empirical distribution of the infectious period of pneumonic plague.
}
\examples{
data(plague_infectious_period)
plot(relfreq ~ days, data = plague_infectious_period)

}
\references{
\insertRef{GaniLeac04}{epigrowthfit}
}
\keyword{datasets}
