% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{plague_latent_period}
\alias{plague_latent_period}
\title{Latent period distribution (pneumonic plague)}
\format{
A data frame with 13 rows and 3 variables:

\describe{
\item{\code{days}}{[integer] Latent period in days, from 0 to 12.}
\item{\code{observed}}{[numeric] Observed relative frequency
out of 224 cases.
}
\item{\code{fitted}}{[numeric] Probability from a fitted lognormal
distribution.
}
}
}
\source{
Obtained from Figure 1A in \insertCite{GaniLeac04;textual}{epigrowthfit}.
}
\usage{
data(plague_latent_period)
}
\description{
Distributions (observed and fitted) of the latent period
of pneumonic plague.
}
\references{
\insertRef{GaniLeac04}{epigrowthfit}
}
\keyword{datasets}
