% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{plot,Model,missing-method}
\alias{plot,Model,missing-method}
\title{plot a model used for fitting exponential growth rates}
\usage{
\S4method{plot}{Model,missing}(x, par = NULL, times = NULL, add = FALSE, cumulative = TRUE, ...)
}
\arguments{
\item{x}{a \code{Model} object}

\item{par}{a named vector or list of parameter values associated
with the model.  Default values will be chosen for any that
are not specified.}

\item{times}{vector of times at which to plot the model (currently,
only \code{range(times)} affects the resulting plot)}

\item{add}{logical: if \code{TRUE} then add to existing plot}

\item{cumulative}{logical: if \code{TRUE} show cumulative curve}

\item{...}{further arguments to be passed to \code{\link{plot}}
and \code{\link{curve}}}
}
\description{
plot a model used for fitting exponential growth rates
}
