% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\title{Plot nonlinear mixed effects models of epidemic growth}
\usage{
\method{plot}{egf}(
  x,
  type = c("interval", "cumulative", "rt1", "rt2"),
  subset = NULL,
  order = NULL,
  cache = NULL,
  do_plot = TRUE,
  time_as = c("Date", "numeric"),
  dt = 1,
  log = TRUE,
  show_predict = TRUE,
  show_tdoubling = TRUE,
  show_legend = FALSE,
  level = 0.95,
  per_plot = 6L,
  ips = 0.25,
  control = egf_plot_control(),
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  ylab_outer = NULL,
  plab = NULL,
  align_right_tdoubling_caption = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{"\link{egf}"} object.}

\item{type}{A \link{character} string indicating a type of plot. The options are:
interval incidence (\code{"interval"}),
cumulative incidence (\code{"cumulative"}),
per capita growth rate as curve (\code{"rt1"}), and
per capita growth rate as heat map (\code{"rt2"}).
\code{type != "rt2"} displays one time series per plot.
\code{type = "rt2"} displays \code{per_plot} time series per plot.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{make_combined}}). Must evaluate to
a \link{logical} vector indexing rows of the data frame,
and thus fitting windows. Only indexed fitting windows are plotted.
The default (\code{\link{NULL}}) is to plot all fitting windows.}

\item{order}{An expression to be evaluated in the combined model frame
(see \code{\link{make_combined}}), typically a call to
\code{\link{order}}, determining the order in which time series
are plotted. The default (\code{\link{NULL}}) is equivalent to
\code{\link{seq_len}(\link{nrow}(combined))}.}

\item{cache}{An \code{"egf_plot_cache"} object returned by a previous evaluation
of \code{plot.egf(x)}. Predicted values and standard errors stored
in \code{cache} will be reused (rather than recomputed) if possible.}

\item{do_plot}{A \link{logical} flag. If \code{FALSE}, then nothing is plotted.
Useful when only the returned \code{"egf_plot_cache"} object is desired.}

\item{time_as}{A \link{character} string indicating how time is displayed on the
bottom axis. The options are: as a calendar (\code{"Date"}) and
as a number of days since the earliest time point (\code{"numeric"}).}

\item{dt}{A positive number indicating an observation interval in days.
Predicted curves are evaluated on grids with this spacing.
When \code{type = "interval"}, counts observed over a shorter
or longer interval \code{dt0} are scaled by a factor of
\code{dt / dt0} so that their scale matches that of the curves.
These points can be highlighted via \code{control}.}

\item{log}{A \link{logical} flag. If \code{TRUE}, then the dependent variable
is plotted on a logarithmic scale.
[\code{type != "rt1"} only.]}

\item{show_predict}{An integer flag: 2 is to draw predicted curves with confidence bands,
1 is draw predicted curves only, 0 is to draw neither.
\link[=logical]{Logical} values are coerced to integer.
[\code{type != "rt2"} only.]}

\item{show_tdoubling}{An integer flag: 2 is to print initial doubling time estimates in
the top margin with confidence intervals, 1 is to print estimates
only, 0 is to print neither. \link[=logical]{Logical} values are
coerced to integer. Supported only if \code{x$model$curve} is
\code{"exponential"}, \code{"logistic"}, or \code{"richards"}.
[\code{type != "rt2"} only.]}

\item{show_legend}{A \link{logical} flag. If \code{TRUE}, then a legend is displayed
in the right margin.
[\code{type = "interval"} only.]}

\item{level}{A number in the interval (0,1). This is the confidence level used
when \code{show_predict = 2} or \code{show_tdoubling = 2}.
[\code{type != "rt2"} only.]}

\item{per_plot}{A positive integer. This is the number of panels (time series)
displayed in one plot.
[\code{type = "rt2"} only.]}

\item{ips}{(\code{type = "rt2"} only.)
A non-negative number specifying the space between panels
("interpanel space") in multipanel plots as a number of margin lines.
[\code{type = "rt2"} only.]}

\item{control}{An \code{"\link{egf_plot_control}"} object controlling the appearance
of almost all plot elements.}

\item{xlim, ylim}{\link[=numeric]{Numeric} vectors of length 2 specifying axis limits,
which are recycled for all plots. If \code{time_as = "Date"}, then
\code{xlim} can instead be a \link{Date} vector or a \link{character}
vector coercible to Date via \code{\link{as.Date}(xlim)}. \code{ylim}
is unused by \code{type = "rt2"}.}

\item{main, sub, xlab, ylab, ylab_outer, plab}{\link[=character]{Character} strings or expressions used to generate
plot (\code{main}, \code{sub}), axis (\code{xlab}, \code{ylab},
\code{ylab_outer}), and panel (\code{plab}) titles. \code{main},
\code{xlab}, and \code{ylab} are supported for all values of \code{type}.
\code{sub} is unused by \code{type = "rt2"}.
\code{plab} is used by \code{type = "rt2"} only.
\code{ylab_outer} is used by \code{type = "rt[12]"} only.
When \code{type != "rt2"}, \code{main} and \code{sub} are evaluated
in the combined model frame (see \code{\link{make_combined}})
in order to generate unique (sub)titles for each plot.
When \code{type = "rt2"}, \code{plab} is evaluated similarly
in order to generate unique titles for each panel.
\code{\link{plotmath}} expressions are not supported
for \code{main}, \code{sub}, and \code{plab} in these cases.}

\item{align_right_tdoubling_caption}{A \link{logical} flag, used if \code{show_tdoubling > 0}
to determine whether the caption explaining doubling time
annotation should be aligned left or right in the top margin.
[\code{type != "rt2"} only.]}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data.frame} inheriting from \link{class}
\code{"egf_plot_cache"}. If argument \code{cache} was supplied
in the function call, then this data frame is the result of
augmenting \code{cache} with new computations.

This object is returned \emph{in spite of} errors thrown
during plotting, to avoid waste of computation time.
}
\description{
Methods for plotting \code{"\link{egf}"} objects.
}
\details{
Caching functionality must be used with care, as mismatch between \code{x}
and \code{cache} will \emph{not} be detected. Constructions such as
\code{plot(x2, cache = plot(x1))} should \emph{not} be expected to produce
correct results.

See topic \code{\link{nse}} for details on nonstandard evaluation
of \code{subset} and \code{order}.
}
