% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{plot.egf_confint}
\alias{plot.egf_confint}
\title{Plot confidence intervals}
\usage{
\method{plot}{egf_confint}(x, group_by = ~1, subset = NULL, ci_per_panel = 8L, ...)
}
\arguments{
\item{x}{An \code{"egf_confint"} object returned by \code{\link[=confint.egf]{confint.egf()}}.}

\item{group_by}{A formula of the form \code{~f1:...:fn} specifying an interaction of
the factors in \code{x} other than \code{par}. \code{group_by} determines the
order in which confidence intervals are plotted, with \code{f1} varying
slowest. Use \code{~1} (the default) to preserve the order given in \code{x}
or if \code{x} has no factors other than \code{par}.}

\item{subset}{A named list of atomic vectors with elements specifying
levels of factors in \code{x}. Only the relevant subset of
confidence intervals is plotted. Use \code{NULL} (the default)
to plot all confidence intervals.}

\item{ci_per_panel}{A positive integer. One panel will display at most this many
confidence intervals.}

\item{...}{Unused optional arguments.}
}
\value{
\code{NULL} (invisibly).
}
\description{
A method for graphically comparing confidence intervals
on fitted values across groups.
}
\details{
If an endpoint of a confidence interval is \code{NA}, then a dashed
line is drawn from the point estimate to the left or right boundary
of the plotting region (depending on the missing endpoint).
}
