% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{plot.egf_profile}
\alias{plot.egf_profile}
\title{Plot likelihood profiles}
\usage{
\method{plot}{egf_profile}(x, subset = NULL, level = 0.95, sqrt = TRUE, ...)
}
\arguments{
\item{x}{An \code{"egf_profile"} object returned by \code{\link[=profile.egf]{profile.egf()}}.}

\item{subset}{A subset of \code{levels(x$name)} specifying profiles to be plotted,
or otherwise \code{NULL}, in which case all profiles are plotted.}

\item{level}{A numeric vector with elements in (0,1), or otherwise \code{NULL}.
If non-\code{NULL}, then horizontal lines are drawn at
\code{qchisq(level, df = 1)} or its square root, depending on \code{sqrt}.}

\item{sqrt}{A logical scalar. If \code{TRUE}, then square root-transformed
deviance is plotted.}

\item{...}{Optional arguments to \code{plot()}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
A method for quickly inspecting computed likelihood profiles.
}
