% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{plot,epigrowthfit,missing-method}
\alias{plot,epigrowthfit,missing-method}
\title{plot method for epigrowthfit objects}
\usage{
\S4method{plot}{epigrowthfit,missing}(
  x,
  window.only = FALSE,
  cumulative = FALSE,
  show.data = TRUE,
  show.pred = TRUE,
  show.legend = TRUE,
  show.gof = TRUE,
  show.totals = TRUE,
  xlab = "",
  ylab = NULL,
  col = "black",
  col.bg = "red",
  col.pred = "black",
  lwd.pred = 3,
  col.year = "blue",
  confint = FALSE,
  col.ci = adjustcolor("black", alpha.f = 0.3),
  log = "",
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{an \code{\link{epigrowthfit}} object}

\item{window.only}{logical flag: if \code{TRUE} then restrict plot to fitting window}

\item{cumulative}{logical flag: if \code{TRUE} then plot cumulative data}

\item{show.data}{logical flag: if \code{TRUE} then display the observed data}

\item{show.pred}{logical flag: if \code{TRUE} then display the fitted model}

\item{show.legend}{logical flag: if \code{TRUE} then display trajectory model and noise model in a legend}

\item{show.gof}{logical flag: if \code{TRUE} then display goodness of fit in a legend}

\item{show.totals}{logical flag: if \code{TRUE} then display total deaths in all and in the fitting window in a legend}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{col}{colour of lines and points for observed data}

\item{col.bg}{fill colour for observed data points}

\item{col.pred}{colour of fitted model}

\item{lwd.pred}{line width of fitted model}

\item{col.year}{colour of year labels on x axis (in multi-year plots)}

\item{confint}{include confidence interval ribbon in plot?}

\item{col.ci}{colour of confidence interval ribbon}

\item{log}{log scale ("", "x", "y", or "xy", as in \code{\link{plot.default}})}

\item{add}{logical flag: if \code{TRUE} then add to existing plot}

\item{...}{further arguments passed to \code{\link{plot}} and \code{\link{points}}}
}
\value{
nothing at the moment
}
\description{
simple plotting method to show observed data and fitted model
}
