% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotPlague}
\alias{plotPlague}
\title{plot plague epidemics}
\usage{
plotPlague(
  epidemic = NULL,
  data = epigrowthfitPNAS::plague,
  name.data = "plague.deaths",
  toprightlabel = NULL,
  cex.toprightlabel = 2,
  col.toprightlabel = "black",
  col.data = "red",
  cex.data = 1,
  show.place = TRUE,
  cumulative = FALSE,
  show.acm = FALSE,
  show.legend = show.acm,
  col.acm = "grey",
  cex.acm = 0.5,
  plot.type = "timeplot",
  col.year = "blue",
  add = FALSE,
  log = "",
  ...
)
}
\arguments{
\item{epidemic}{epidemic year(s); all available years by default}

\item{data}{data frame containing epidemic time series}

\item{name.data}{the column name of the variable to be plotted (\code{"plague.deaths"} by default)}

\item{toprightlabel}{epidemic name by default}

\item{cex.toprightlabel}{character expansion factor for \code{toprightlabel}}

\item{col.toprightlabel}{font colour for \code{toprightlabel}}

\item{col.data}{fill colour for points in \code{\link{plot}} or
bars in \code{\link{barplot}}}

\item{cex.data}{character expansion factor for plague data points}

\item{show.place}{logical: if \code{TRUE} then indicate place in y axis label}

\item{cumulative}{logical: if \code{TRUE} then show cumulative rather than raw data}

\item{show.acm}{logical: if \code{TRUE} then show all cause mortality}

\item{show.legend}{logical: if \code{TRUE} then show a legend indicating which curve is which (relevant only if \code{show.acm==TRUE})}

\item{col.acm}{colour of all cause mortality data if shown}

\item{cex.acm}{character expansion factor for all cause mortality points}

\item{plot.type}{character string: either "barplot" or "timeplot"}

\item{col.year}{colour of year labels on x axis (for multi-year plots)}

\item{add}{(logical) add to an existing plot?}

\item{log}{log scale ("", "x", "y", or "xy", as in \code{\link{plot.default}})}

\item{...}{further parameters passed to \code{\link{plot}}, \code{\link{barplot}}, or \code{\link{points}}}
}
\value{
A \code{\link{data.frame}} that is the subset of the \code{\link{plague}} data
frame used to make the plot, with an additional \code{date} column of class \code{\link{Date}}.
}
\description{
A plot of one or more plague epidemics available
in the \code{\link{plague}} data frame.
}
\details{
If multiple epidemic years are given then all will be plotted.

WARNINGS:
\itemize{
\item With \code{plot.type="barplot"}, temporal gaps between
      distinct epidemics will not appear.
\item If multiple epidemics are plotted and they include
      different levels of aggregation (e.g., weekly, monthly)
      the wrong impression will be given for the magnitude
      of more highly aggregated epidemics.
}
}
\examples{
## plot the Black Death in London
plotPlague(1348)
plotPlague(1348,plot.type="barplot")
## plot a sequence minor plague epidemics in London
plotPlague(1578:1582)
plotPlague(1578:1582,plot.type="barplot")
## plot the Great Plague of London
plotPlague(1665)
}
