% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\name{predict,epigrowthfit-method}
\alias{predict,epigrowthfit-method}
\title{predict method}
\usage{
\S4method{predict}{epigrowthfit}(
  object,
  tvec = NULL,
  newpars = NULL,
  confint = FALSE,
  seed = 101,
  nsim = 500,
  level = 0.95
)
}
\arguments{
\item{object}{\code{\link{epigrowthfit}} object}

\item{tvec}{time vector}

\item{newpars}{modified model parameters for prediction}

\item{confint}{return confidence intervals?}

\item{seed}{random-number seed (for importance sampling CIs)}

\item{nsim}{number of importance samples for CIs}

\item{level}{confidence interval}
}
\value{
a data frame with columns \code{time} and \code{deaths}
and (if confidence intervals requested) \code{lwr} and \code{upr}
}
\description{
predict method
}
