% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.egf}
\alias{predict.egf}
\title{Compute predicted values}
\usage{
\method{predict}{egf}(
  object,
  time,
  varname = c("log_int_inc", "log_cum_inc", "log_rt"),
  subset = NULL,
  se = FALSE,
  ...
)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{time}{A numeric vector listing increasing time points in days since
the start of the fitting window specified by \code{subset}.}

\item{varname}{A character vector listing one or more variables for which
predicted values are desired.}

\item{subset}{A named list of atomic vectors of length 1 specifying exactly
one level for each factor in \code{object$frame} (and thus a unique
fitting window). Use the default (\code{NULL}) if and only if
\code{object$frame} has no factors.}

\item{se}{A logical scalar. If \code{TRUE}, then standard errors on predicted
values are also reported.}

\item{...}{Unused optional arguments.}
}
\value{
A list inheriting from class \code{"egf_predict"},
containing data frames (one for each \code{varname})
with numeric variables \code{time}, \code{estimate}, and
\code{se} (if \code{se = TRUE}).

If \code{"log_int_inc" \%in\% varname}, then, for \code{i > 1},
\code{log_int_inc$estimate[i]} stores log incidence
during the interval from \code{time[i-1]} to \code{time[i]}.
}
\description{
Computes predicted values of log cumulative incidence, log
interval incidence, and log instantaneous exponential growth
rate given user-specified time points.
}
\details{
Elements of \code{subset} (if non-\code{NULL}) must have the form
\code{factor_name = level_name}, where \code{factor_name} is
the name of a factor in \code{object$frame} and \code{level_name}
is an element of \code{levels(object$frame$factor_name)}.
}
\seealso{
\code{\link[=confint.egf_predict]{confint.egf_predict()}}
}
