% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.doubling_time}
\alias{print.doubling_time}
\title{Print doubling times}
\usage{
\method{print}{doubling_time}(x, ...)
}
\arguments{
\item{x}{A "doubling_time" object.}

\item{...}{Unused optional arguments.}
}
\value{
\code{x} (invisibly).
}
\description{
A method for printing objects of class "doubling_time".
}
\seealso{
\code{\link[=compute_doubling_time]{compute_doubling_time()}}
}
