% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.egf}
\alias{print.egf}
\alias{print.egf_init}
\title{Print models of epidemic growth}
\usage{
\method{print}{egf_init}(x, ...)

\method{print}{egf}(x, ...)
}
\arguments{
\item{x}{An "egf_init" or "egf" object.}

\item{...}{Unused optional arguments.}
}
\value{
\code{x} (invisibly).
}
\description{
Methods for printing objects of class "egf_init" or "egf".
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \code{\link[=egf]{egf()}}
}
