% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.egf_sim}
\alias{print.egf_sim}
\title{Print simulations}
\usage{
\method{print}{egf_sim}(x, ...)
}
\arguments{
\item{x}{An "egf_sim" object.}

\item{...}{Unused optional arguments.}
}
\value{
\code{x} (invisibly).
}
\description{
A method for printing objects of class "egf_sim".
}
\seealso{
\code{\link[=simulate.egf]{simulate.egf()}}
}
