% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.egf}
\alias{profile.egf}
\title{Compute likelihood profiles}
\usage{
\method{profile}{egf}(
  fitted,
  which = NULL,
  A = NULL,
  parm = get_par_names(fitted, link = TRUE),
  max_level = 0.99,
  grid_len = 12L,
  trace = TRUE,
  parallel = c("serial", "multicore", "snow"),
  cores = getOption("egf.cores", 2L),
  outfile = NULL,
  cl = NULL,
  ...
)
}
\arguments{
\item{fitted}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{which}{An integer vector indexing coefficients in \code{fitted$par}
to be profiled. Must be a subset of \code{fitted$nonrandom}.}

\item{A}{A numeric matrix with \code{length(fitted$nonrandom)} columns.
Each row specifies a linear combination of the elements
of \code{fitted$par[fitted$nonrandom]} to be profiled. Ignored
if \code{which} is non-\code{NULL}.}

\item{parm}{A character vector listing nonlinear model parameters whose
population fitted values should be profiled. Must be a subset
of \code{get_par_names(fitted, link = TRUE)}. Ignored if one of
\code{which} and \code{lin_comb} is non-\code{NULL}. (Here, "population
fitted value" refers to the fixed effects component of the
individual or overall fitted value, which accounts for both
fixed and random effects.)}

\item{max_level}{A number in the interval (0,1). Profiles will be computed
up to a deviance of \code{qchisq(max_level, df = 1)}.}

\item{grid_len}{A positive integer. Step sizes chosen adaptively by
\code{\link[TMB:tmbprofile]{TMB::tmbprofile()}} will generate approximately this
many points on each side of a profile's minimum point.}

\item{trace}{A logical scalar. If \code{TRUE}, then basic tracing messages
are printed.}

\item{parallel}{A character string indicating a library for parallel computation.
\code{"serial"} indicates no parallelization. \code{"multicore"} forks and
is intended for use from a terminal rather than, e.g., RStudio.
On Windows, it is equivalent to \code{"serial"}. \code{"snow"} creates
socket clusters and parallelizes on both Unix-alikes and Windows.}

\item{cores}{A positive integer. The number of worker processes spawned when
\code{parallel != "serial"}. See also \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{outfile}{A character string containing a file path or otherwise \code{NULL}.
If a file path, then console output is diverted there. If \code{NULL},
then there is no diversion. When \code{parallel = "snow"}, diversion
may be necessary to view output.}

\item{cl}{A optional socket cluster created by
\code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}}, to be used
when \code{parallel = "snow"}. If non-\code{NULL},
then \code{cores} and \code{outfile} are ignored.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame inheriting from class \code{"egf_profile"}, with variables:
\item{\code{index}}{
Row index of linear combination, from \code{seq_len(nrow(A))}.
}
\item{\code{value}}{
Value of linear combination being profiled.
}
\item{\code{deviance}}{
Deviance of the restricted model that assumes \code{value}
for the linear combination being profiled.
}
\item{\code{par}}{
(\code{parm}-based calls only.)
Nonlinear model parameter, from \code{parm}.
}
\item{\code{window}}{
(\code{parm}-based calls only.)
Fitting window, from \code{levels(fitted$frame_ts$.window)}.
}
Matrix \code{A} and vector \code{par = fitted$par[fitted$nonrandom]}
are retained as attributes.
}
\description{
Computes the univariate likelihood profile of fixed effects
coefficients, log standard deviations of random effects
coefficients, and linear combinations thereof.
}
\details{
\code{which} is mapped to an \code{A} matrix composed of unit row vectors,
with 1 at array index \verb{[i, which[i]]} for all \code{i}.
\code{A \%*\% fitted$par[fitted$nonrandom]} reproduces \code{fitted$par[which]}.

\code{parm} is mapped to an \code{A} matrix composed of blocks that
are precisely the fixed effects design matrices for nonlinear
model parameters \code{parm}. \code{A \%*\% fitted$par[fitted$nonrandom]}
is a vector listing each parameters' population fitted values
(one per fitting window per parameter).
}
\seealso{
\code{\link[=confint.egf_profile]{confint.egf_profile()}}, \code{\link[=plot.egf_profile]{plot.egf_profile()}}
}
