% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{R0,numeric-method}
\alias{R0,numeric-method}
\title{compute R0 given r and generation interval distribution}
\usage{
\S4method{R0}{numeric}(obj, generation.interval = generation.interval.plague())
}
\arguments{
\item{obj}{instantaneous exponential growth rate}

\item{generation.interval}{the generation interval distribution (days)}
}
\description{
compute R0 given r and generation interval distribution
}
