% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{R0}
\alias{R0}
\title{returns R0 and its 95\% confidence interval}
\usage{
R0(obj, ...)
}
\arguments{
\item{obj}{a fitted model object}
}
\value{
a vector
}
\description{
uses the exponential growth rate from the expgrowthrate
object that is passed, and the serial interval distribution available
in the package
}
\references{
\insertRef{WallLips07}{epigrowthfitPNAS}
}
