% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{rename_par}
\alias{rename_par}
\title{Enumerate TMB-generated parameter names}
\usage{
rename_par(par)
}
\arguments{
\item{par}{A TMB-generated parameter vector or, more generally,
a named numeric vector such that
\code{all(names(par) \%in\% c("beta", "log_sd_b", "b"))}.}
}
\value{
\code{par} with \code{names(par)} replaced with names of the form
\code{"beta[i]"}, \code{"log_sd_b[j]"}, or \code{"b[k]"} (see Examples).
}
\description{
Enumerates the names of a parameter vector generated by
\code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}, making them unique.
}
\examples{
s <- c("beta", "log_sd_b", "b")
n <- 10L
par <- numeric(n)
names(par) <- sample(s, n, replace = TRUE)
rename_par(par)

}
\keyword{internal}
