% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rle_literal}
\alias{rle_literal}
\title{Literal run length encoding}
\usage{
rle_literal(x)
}
\arguments{
\item{x}{An \link{atomic} vector.}
}
\value{
A \link{list} similar to \code{\link{rle}(x)},
though not inheriting from \link{class} \code{"rle"}.
}
\description{
A replacement for \code{\link{rle}} that regards \code{\link{NA}}
as equal to previous \code{NA}. For \link{double} vectors,
it regards \code{\link{NaN}} as equal to previous \code{NaN}
but unequal to previous \code{NA}.
}
\details{
The result is compatible with \code{\link{inverse.rle}}.
More precisely,
\code{function(x) inverse.rle(rle_literal(x))}
is an identity function for atomic \code{x}, and
\code{function(x) rle_literal(inverse.rle(x))}
is an identity function for lists \code{x} generated
by \code{rle_literal}.
}
\examples{
## x <- rep.int(c(0, NA, NaN, 1), 1:4)
## rle_x <- rle_literal(x)
## identical(x, inverse.rle(rle_x))

}
\keyword{internal}
