% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\name{safe_date}
\alias{safe_date}
\title{epidemic growth fit}
\usage{
safe_date(tvec)
}
\arguments{
\item{tvec}{vector of julian dates (YYYY.fff)}
}
\description{
convert a julian date to a date, being careful about truncation (\code{\link{as.Date}} truncates values instead of rounding)
}
