% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.egf_model}
\alias{simulate.egf_model}
\title{Simulate incidence time series}
\usage{
\method{simulate}{egf_model}(
  object,
  nsim = 1L,
  seed = NULL,
  mu,
  Sigma = NULL,
  tol = 1e-06,
  tmax = 100,
  cstart = 0,
  origin = .Date(0L),
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf_model}"} specifying a phenomenological
epidemic model to be simulated.}

\item{nsim}{A positive integer indicating a number of simulated time series.}

\item{seed}{An integer used to set the \link{RNG} state before simulation.
The default (\code{\link{NULL}}) is to use the state at the
time of the function call. The RNG state is preserved as an
\link[=attributes]{attribute} of the result (see Value).}

\item{mu}{A \link{numeric} vector listing means across time series
of nonlinear and dispersion model parameters (link scale).
It is assumed that elements are ordered as in
\code{\link{get_par_names}(object, link = TRUE)}.}

\item{Sigma}{A symmetric positive definite \link{numeric} \link{matrix}
to be used as the covariance matrix corresponding to \code{mu}.
The default (\code{\link{NULL}}) is equivalent (conceptually)
to a zero matrix and is handled specially.}

\item{tol}{A non-negative number indicating a tolerance for lack of positive
definiteness of \code{Sigma}. Negative eigenvalues of \code{Sigma}
must not be less than \code{-tol * rho}, where \code{rho} is the
spectral radius of \code{Sigma}. (However, regardless of \code{tol},
\code{\link{diag}(Sigma)} must be positive, as standard deviations
are processed on the log scale.)}

\item{tmax}{A positive number. Simulated time series run from 0 days to 1 day
after the inflection time of the underlying cumulative incidence
function indicated by \code{object$curve}. However, if there is no
inflection time, as in the exponential and subexponential models
of cumulative incidence, then they run from 0 days to \code{tmax}
days.}

\item{cstart}{A number indicating a threshold value of cumulative incidence.
Left endpoints of suggested fitting windows are those times
when cumulative incidence first exceeds this threshold.}

\item{origin}{A \link{Date} specifying a reference time.}

\item{...}{Unused optional arguments.}
}
\value{
A \link{list} inheriting from \link{class} \code{"egf_model_simulate"},
with elements:
\item{model, mu, Sigma, origin}{
  Copies of the so-named arguments.
  (\code{model} is a copy of argument \code{object}.)
}
\item{Y}{
  A \link[=double]{numeric} \link{matrix} with \code{nsim} rows and
  \code{length(mu)} columns listing the nonlinear and dispersion model
  parameter values underlying each time series.
  If \code{Sigma} is \code{\link{NULL}}, then the row vectors of \code{Y}
  are all \code{mu}. Otherwise, \code{Y} is (conceptually) the result of
  \code{MASS::mvrnorm(nsim, mu, Sigma, tol)}.
}
\item{formula}{
  A \link{formula}, \code{x ~ time | ts}, expressing how simulated
  time series are stored in \code{data}.
}
\item{formula_par}{
  A \link{formula} specifying the generative model.
  If \code{Sigma = \link{NULL}}, then the formula is
  \code{~1} if \code{nsim = 1} and \code{~ts} if \code{nsim > 1}.
  Otherwise, it is \code{~(1 | ts)}.
}
\item{data}{
  A \link[=data.frame]{data frame} with variables \code{ts}, \code{time},
  and \code{x} storing \code{nsim} simulated time series in long format.
  \code{ts} is a \link{factor} with \code{nsim} \link{levels} grouping
  the rows by time series. \code{time} is a \link[=double]{numeric}
  vector listing time points as numbers of days since \emph{the end of}
  \link{Date} \code{origin}. \code{x} is an \link{integer} vector such that,
  within time series, \code{x[i]} is the number of cases observed between
  \code{time[i-1]} and \code{time[i]}.
}
\item{data_par}{
  A \link[=data.frame]{data frame} with \code{nsim} rows defining
  variables used in \code{formula_par}. It is either empty or has
  one variable \code{ts = \link{gl}(nsim, 1L)}.
}
\item{endpoints}{
  A \link[=data.frame]{data frame} with \code{nsim} rows and variables
  \code{ts}, \code{start}, and \code{end} suggesting a fitting window
  for each simulated time series. Start times are determined according
  to \code{cstart}. End times are precisely the last time point in the
  corresponding time series.
}
\item{actual}{
  A \link[=double]{numeric} vector giving the full parameter vector
  corresponding to the generative model. When estimating this model
  from \code{data}, \code{\link{egf}} output should be compared against
  \code{actual}. More precisely, if \code{m} is the \code{"\link{egf}"}
  object, then \code{m$best} estimates \code{actual}.
}
\item{nonrandom}{
  An \link{integer} vector indexing the elements of \code{actual}
  that are not random effects.
}
\item{call}{
  The \link{call} to \code{simulate.egf_model}, allowing for updates
  to the \code{"egf_model_simulate"} object via \code{\link{update}}.
}
If \code{seed = \link{NULL}}, then \link[=attributes]{attribute}
\code{RNGstate} is the value of \code{\link{.Random.seed}} at the
time of simulation.
Otherwise, it is a \link{list} of arguments to \code{\link{set.seed}}
such that \code{\link{do.call}(\link{set.seed}, RNGstate)} restores
the value of \code{\link{.Random.seed}} at the time of simulation.
}
\description{
Simulates daily incidence time series according to a specified
phenomenological epidemic model. Nonlinear and dispersion model
parameters vary between time series according to a fixed intercept
model \code{~ts} or random intercept model \code{~(1 | ts)}.
}
\examples{
model <- egf_model(curve = "logistic", family = "nbinom")

r <- log(2) / 20
tinfl <- 160
K <- 25000
nbdisp <- 50

mu <- log(c(r, tinfl, K, nbdisp))
Sigma <- diag(rep_len(0.5^2, length(mu)))

sim <- simulate(model,
  nsim = 20L,
  seed = 202737L,
  mu = mu,
  Sigma = Sigma,
  cstart = 10
)

}
\seealso{
\code{\link{egf.egf_model_simulate}} for estimating the generative model
from simulated time series,
\code{\link{simulate.egf}} for simulating time series from \emph{fitted}
models
}
