% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cases.R
\name{smooth_cases}
\alias{smooth_cases}
\title{Fit cubic splines to interval incidence}
\usage{
smooth_cases(date, cases, log = TRUE, spar = seq(0.55, 0.9, by = 0.05), ...)
}
\arguments{
\item{date}{A Date vector listing increasing time points.
Should start at or before the date of the first observed case
in an epidemic wave.}

\item{cases}{A numeric vector of length \code{length(date)-1}.
\code{cases[i]} is the number of cases observed between \code{date[i]}
and \code{date[i+1]}. Here, "cases" can mean infections,
reported infections, or reported deaths (either disease deaths or
deaths due to multiple causes including the disease of interest).
Missing values are not tolerated.}

\item{log}{A logical scalar. If \code{TRUE}, then \code{log10(1+cases)} is
smoothed instead of \code{cases}.}

\item{spar}{A numeric vector with elements in the interval (0,1].
Degree of smoothing increases with \code{spar} (smoothing parameter).
See \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}

\item{...}{Optional arguments to \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}
}
\value{
A "smooth_cases" object. A list containing copies of arguments
\code{date}, \code{cases}, \code{log}, and \code{spar}, with these additional elements:

\describe{
\item{\code{time}}{Time as a number of days since \code{date[1]}.
Equal to \code{as.numeric(date - date[1])}.
}
\item{\code{ss}}{A list of \code{length(spar)} "smoothing.spline" objects,
giving the list output of \code{\link[stats:smooth.spline]{stats::smooth.spline()}} for each
element of \code{spar}.
}
\item{\code{peaks}}{An integer vector indexing peaks in \code{cases} or
\code{log10(1+cases)}.
}
\item{\code{troughs}}{An integer vector indexing troughs in \code{cases} or
\code{log10(1+cases)}.
}
}
}
\description{
Fits cubic splines of varying smoothness to an interval incidence
time series using \code{\link[stats:smooth.spline]{stats::smooth.spline()}}, and looks for local
extrema by checking where the first derivative changes sign.
Intended for use in conjunction with \code{\link[=plot.smooth_cases]{plot.smooth_cases()}}
and \code{\link[=egf_init]{egf_init()}}, to make the process of selecting fitting windows
more robust to noise.
}
\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
x <- smooth_cases(
  date = ontario$date,
  cases = ontario$new_confirmations[-1],
  log = TRUE,
  spar = seq(0.55, 0.9, by = 0.05)
)
plot(x)

}
\seealso{
\code{\link[=plot.smooth_cases]{plot.smooth_cases()}}
}
