% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{split_effects}
\alias{split_effects}
\title{Split fixed and random effects terms}
\usage{
split_effects(x)
}
\arguments{
\item{x}{A formula.}
}
\value{
A list with elements:
\item{\code{fixed}}{
A fixed effects formula.
}
\item{\code{random}}{
A list of random effects terms (calls to the \code{|} operator).
}
}
\description{
Retrieve from a mixed effects formula the corresponding
fixed effects formula and a list of all random effects terms.
}
\examples{
split_effects(y ~ 0 + x + (1 | f) + (a | g))

}
