% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{split_interaction}
\alias{split_interaction}
\title{Split an interaction term}
\usage{
split_interaction(x)
}
\arguments{
\item{x}{A call, name, or atomic scalar.}
}
\value{
If \code{x} is a (possibly nested) call to the \code{:} operator, then
a list of the arguments to \code{:} that are not themselves calls
to \code{:}. Otherwise, \code{list(x)}.
}
\description{
From an interaction term, recursively construct a list
of the interacted variables.
}
\examples{
# x <- str2lang("f:g:I(a:b):log(x)")
# split_interaction(x)

}
\keyword{internal}
