% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{split_sdreport}
\alias{split_sdreport}
\title{Split \code{ADREPORT()}ed variables}
\usage{
split_sdreport(sdreport)
}
\arguments{
\item{sdreport}{An \code{"sdreport"} object returned by \code{\link[TMB:sdreport]{TMB::sdreport()}}.}
}
\value{
A named list of data frames with variables \code{estimate} and \code{se}
giving estimates and standard errors.
}
\description{
Extracts reported variables from an \code{"sdreport"} object
and returns them in a list.
}
\details{
When reconstructing reported matrices, which are returned as vectors,
assume column-major order.
}
\keyword{internal}
