% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{split_terms}
\alias{split_terms}
\alias{unsplit_terms}
\title{Split and unsplit formula terms}
\usage{
split_terms(x)

unsplit_terms(l)
}
\arguments{
\item{x}{A call, name, or atomic scalar.}

\item{l}{A list of calls, names, and atomic scalars.}
}
\value{
A list of calls, names, and atomic scalars.
}
\description{
Split formula terms related by \code{+} and \code{-}
operators with \code{split_terms()}. Perform the inverse
operation with \code{unsplit_terms()}.
}
\details{
If \code{x} is a formula, then \code{x[[length(x)]]} is split.
Hence \code{unsplit_terms(split_terms(x))} reproduces
\code{x[[length(x)]]} but not \code{x}.

Extraneous \code{(} calls, as in the expression
\code{(x) + (y)} are stripped when splitting and
not replaced when unsplitting. In these cases,
\code{identical(x, unsplit_terms(split_terms(x)))}
may return \code{FALSE}.
}
\examples{
# x <- str2lang("1 + a + b * c - c + (d | e) + (0 + f * g | h)")
# l <- split_terms(x)
# y <- unsplit_terms(l)
# identical(x, y)
# # [1] FALSE

}
\keyword{internal}
