% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{string_get_link}
\alias{string_get_link}
\alias{string_add_link}
\alias{string_remove_link}
\alias{string_extract_link}
\title{Manipulate model parameter names}
\usage{
string_get_link(s)

string_add_link(s)

string_remove_link(fs)

string_extract_link(fs)
}
\arguments{
\item{s, fs}{\link[=character]{Character} vectors.}
}
\value{
For strings \code{s}, \code{f}, and \code{fs} as described above,
\code{string_get_link(s)} returns \code{f}, \code{string_add_link(s)}
returns \code{fs}, \code{string_remove_link(fs)} returns \code{s},
and \code{string_extract_link(fs)} returns \code{f}.

All functions are vectorized. \code{\link{NA_character_}}
is returned for invalid input (see Examples).
}
\description{
Utilities for modifying strings \code{s} and
\code{fs = \link{sprintf}("\%s(\%s)", f, s)},
where \code{s} is the name used internally for
a nonlinear or dispersion model parameter and
\code{f} is the name (either \code{"log"} or \code{"logit"})
of the corresponding link function.
}
\examples{
# string_get_link("r")
# string_add_link("r")
# string_remove_link("log(r)")
# string_extract_link("log(r)")
# string_extract_link("invalid", "input", "log(r)", "r")

}
\keyword{internal}
