% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.fitList}
\alias{summary.fitList}
\title{summarize nested lists of fits}
\usage{
\method{summary}{fitList}(
  object,
  level = 1,
  keys = c("outbreak.year", "source"),
  type = "sum",
  gof_agg = "7 days",
  verbose = FALSE,
  confint_pred = TRUE,
  add_NBdisp_coef = TRUE,
  ...
)
}
\arguments{
\item{object}{a (possibly nested) list of epidemic fits}

\item{level}{number of levels to collapse}

\item{keys}{names of levels}

\item{type}{"sum": summary; "pred"; predicted values; "deaths"; original data used to fit}

\item{gof_agg}{aggregation unit for goodness of fit calculations}

\item{verbose}{print messages?}

\item{confint_pred}{??}

\item{add_NBdisp_coef}{add NA value for NB disp if necessary?}

\item{...}{extra arguments (ignored: for generic consistency)}
}
\description{
level=0 returns the summary from a single fit
level=1 returns the summary from a list of fits
level=2 returns the summary from a nested list of fits ...
}
