% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{TransformedModel-class}
\alias{TransformedModel-class}
\alias{TransformedModel}
\title{a model with transformed parameters}
\description{
a model with transformed parameters
}
\section{Slots}{

\describe{
\item{\code{orig_expr}}{original expression}

\item{\code{transforms}}{the parameter transforms}

\item{\code{inverses}}{the inverse transforms}
}}

