% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{transformPar,Model-method}
\alias{transformPar,Model-method}
\title{transform parameter values}
\usage{
\S4method{transformPar}{Model}(model, par, inverse = FALSE)
}
\arguments{
\item{model}{a Model object}

\item{par}{a named vector (or list) of parameters to be transformed}

\item{inverse}{logical, if TRUE, map from the TransformedModel parameters to the original model parameters; if FALSE, map from the original model parameters to the TransformedModel parameters}
}
\value{
a named list of transformed parameter values
}
\description{
return the same parameter values as input, but in a named list
}
