% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{transformPar}
\alias{transformPar}
\title{transform parameter values}
\usage{
transformPar(model, ...)
}
\arguments{
\item{model}{a list of Model objects defining parameter transformations}
}
\value{
a named list of transformed parameter values
}
\description{
Map parameter values of a transformed model to/from the parameter values of the original model.
}
