% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.egf}
\alias{update.egf}
\alias{update.egf_no_fit}
\title{Update a fitted model}
\usage{
\method{update}{egf}(object, ..., evaluate = TRUE)

\method{update}{egf_no_fit}(object, ..., evaluate = TRUE)
}
\arguments{
\item{object}{An \code{"\link{egf}"} or \code{"\link[=egf]{egf_no_fit}"} object.}

\item{...}{Arguments for the updated \link{call} (see \code{\link{update}}).}

\item{evaluate}{A \link{logical} flag, indicating whether the updated \link{call}
should be evaluated.}
}
\value{
The updated \link{call} or the result of its evaluation,
depending on \code{evaluate}.
}
\description{
When \code{\link{match.call}} is called from a method for a generic function,
a \link{call} to the method is returned, rather than a call to the
generic function. Typically, the method is not exported, so the call
cannot be evaluated by the default method for \code{\link{update}}.
These methods for \code{\link{update}} circumvent this issue for objects
returned by \code{\link{egf}}.
}
