% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{window.epigrowthfit}
\alias{window.epigrowthfit}
\title{returns the fitting window}
\usage{
\method{window}{epigrowthfit}(x, ...)
}
\arguments{
\item{x}{an \code{epigrowthfit} object}

\item{...}{extra arguments (ignored)}
}
\value{
a vector
}
\description{
returns the fitting window
}
