% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap}
\alias{wrap}
\title{Concatenate, wrap, collapse}
\usage{
wrap(..., width = 0.9 * getOption("width"))
}
\arguments{
\item{...}{Zero or more \R objects (typically \link{atomic} vectors of length 1),
to be coerced to \link{character} and concatenated with no separator.}

\item{width}{A positive integer passed to \code{\link{strwrap}},
indicating a target column width.}
}
\value{
A \link{character} vector of length 1.
}
\description{
Concatenates \R objects, formats the resulting string in paragraphs
using \code{\link{strwrap}}, and collapses paragraph lines with newlines.
Under default settings, the result is a string that prints in the console
as one or more nicely wrapped paragraphs.
}
\keyword{internal}
