% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ymd}
\alias{ymd}
\title{Decompose a vector of dates}
\usage{
ymd(date, which = 1:3)
}
\arguments{
\item{date}{A Date vector.}

\item{which}{A subset of \code{c(1, 2, 3)} indicating which of
year, month, and day should be returned.}
}
\value{
\code{imat[, which]}, where \code{imat} is an integer matrix with
\code{length(date)} rows and 3 columns listing year, month,
and day, respectively, for each Date in \code{date}.
}
\description{
Extracts year, month, and day from a Date vector.
}
\keyword{internal}
