% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{baxis}
\alias{baxis}
\title{Better axis}
\usage{
baxis(side, a = NULL, b = NULL, at = NULL, labels = TRUE, ...)
}
\arguments{
\item{side, at, labels, ...}{Arguments to \code{\link{axis}}.}

\item{a, b}{Lower and upper limits of the axis line in user coordinates.
The default (\code{\link{NULL}}) is to take the limit from
\code{\link{par}("usr")}.}
}
\value{
A \link{numeric} vector.
This is the value of \code{\link{axis}(side, at, labels, ...)}.
}
\description{
A replacement for \code{\link{axis}} allowing the user to define
the extent of the axis line.
}
