% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{baxis}
\alias{baxis}
\title{Better axis}
\usage{
baxis(side, a = NULL, b = NULL, at = NULL, labels = TRUE, ...)
}
\arguments{
\item{side, at, labels, ...}{Arguments to \code{\link[graphics:axis]{graphics::axis()}}.}

\item{a, b}{Lower and upper limits of the axis line in user coordinates.
The default (\code{NULL}) is to use values from \code{par("usr")}.}
}
\value{
A numeric vector. The value of \code{axis(side, at, labels, ...)}.
}
\description{
A replacement for \code{\link[graphics:axis]{graphics::axis()}} allowing the user
to define the extent of the axis line.
}
