% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_par}
\alias{boot_par}
\title{Parametric bootstrap}
\usage{
boot_par(
  object,
  n = 6L,
  method = c("nlminb", "nlm", "Nelder-Mead", "BFGS"),
  trace = TRUE,
  parallel = c("serial", "multicore", "snow"),
  cores = getOption("egf.cores", 2L),
  outfile = NULL,
  cl = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{n}{A positive integer. The desired number of samples.}

\item{method}{A character string specifying an optimizer available through
\code{\link[stats:nlminb]{stats::nlminb()}}, \code{\link[stats:nlm]{stats::nlm()}}, or \code{\link[stats:optim]{stats::optim()}}.}

\item{trace}{A logical scalar. If \code{TRUE}, then basic tracing information is
printed.}

\item{parallel}{A character string indicating a library for parallel computation.
\code{"serial"} indicates no parallelization. \code{"multicore"} forks and
is intended for use from a terminal rather than, e.g., RStudio.
On Windows, it is equivalent to \code{"serial"}. \code{"snow"} creates
socket clusters and parallelizes on both Unix-alikes and Windows.}

\item{cores}{A positive integer. The number of worker processes spawned when
\code{parallel != "serial"}. See also \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{outfile}{A character string containing a file path or otherwise \code{NULL}.
If a file path, then console output is diverted there. If \code{NULL},
then there is no diversion. When \code{parallel = "snow"}, diversion
may be necessary to view output.}

\item{cl}{A optional socket cluster created by
\code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}}, to be used
when \code{parallel = "snow"}. If non-\code{NULL},
then \code{cores} and \code{outfile} are ignored.}

\item{...}{Unused optional arguments.}
}
\value{
A numeric matrix inheriting from class \code{"egf_boot"},
with \code{length(object$par)} rows and \code{n} columns.
}
\description{
Sample from the conditional distribution of the full parameter vector
given the fitted epidemic model.
}
