% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_par}
\alias{boot_par}
\title{Parametric bootstrap}
\usage{
boot_par(
  object,
  n = 6L,
  control = object$control,
  parallel = egf_parallel(),
  trace = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{n}{A positive integer. The desired number of samples.}

\item{control}{An \code{"\link{egf_control}"} object specifying control parameters.}

\item{parallel}{An \code{"\link{egf_parallel}"} object defining a parallelization method.}

\item{trace}{A \link{logical} flag. If \code{TRUE}, then basic tracing messages
indicating progress are printed. Depending on \code{control$trace},
these messages may be mixed with optimization output.}

\item{...}{Unused optional arguments.}
}
\value{
A \link{numeric} \link{matrix} inheriting from \link{class}
\code{"egf_boot"}, with \code{\link{length}(object$best)} rows
and \code{n} columns.
}
\description{
Sample from the conditional distribution of the
full parameter vector given its fitted value.
}
