% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{canadacovid}
\alias{canadacovid}
\title{COVID-19 in Canadian provinces and territories}
\format{
A data frame with 5451 rows and 4 variables:
\describe{
\item{\code{province}}{
A factor. Canadian province or territory (postal abbreviation).
}
\item{\code{Date}}{
A Date vector. Daily within each level of \code{province},
except prior to invasion, when less frequent reports were common.
}
\item{\code{cases_tot}}{
An integer vector. \code{cases_tot[i]} is the number of cases confirmed
up to the end of \code{Date[i]} in \code{province[i]}.
}
\item{\code{cases_new}}{
An integer vector. \code{cases_new[i]} is the number of cases confirmed
from the end of \code{Date[i-1]} to the end of \code{Date[i]} in \code{province[i]}.
}
}
}
\source{
Raw data were downloaded from Michael Li's public
\href{https://github.com/wzmli}{Github repository}.
Up-to-date data can be downloaded
\href{https://wzmli.github.io/COVID19-Canada/COVID19_Canada.csv}{here}.
}
\usage{
data(canadacovid)
}
\description{
Daily confirmations of COVID-19 in Canadian provinces
and territories, from first confirmation to May 8, 2021.
}
\examples{
data(canadacovid)
subset(canadacovid, province == "ON", -province)

}
\keyword{datasets}
