% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{canadacovid}
\alias{canadacovid}
\title{COVID-19 in Canadian provinces and territories}
\format{
A data frame with 1677 rows and 3 variables:

\describe{
\item{\code{date}}{[Date] Date, daily from February 14, 2020
to June 21, 2020 within each level of \code{province}.
}
\item{\code{time}}{[numeric] Decimal date, equal to year plus
fraction of year. The fraction is 0 on January 1, 1/365
or 1/366 on January 2, and so on.
}
\item{\code{province}}{[factor, 13 levels] A Canadian province
or territory (postal abbreviation).
}
\item{\code{new_confirmations}}{[numeric] Number of new confirmations
of COVID-19. There are 37 missing values, at least one for each
level of \code{province}.
}
}
}
\usage{
data(canadacovid)
}
\description{
Daily confirmations of COVID-19 in Canadian provinces and territories
from February 14, 2020 to June 21, 2020.
}
\examples{
data(canadacovid)
subset(canadacovid, province == "ON")

}
\keyword{datasets}
