% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{coef.egf}
\alias{coef.egf}
\alias{coef.egf_no_fit}
\title{Extract coefficients and covariance parameters}
\usage{
\method{coef}{egf}(object, full = FALSE, list = FALSE, ...)

\method{coef}{egf_no_fit}(object, full = FALSE, list = FALSE, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} or \code{"\link[=egf]{egf_no_fit}"} object.}

\item{full}{A logical flag. If \code{FALSE}, then parameter vectors are returned
in the condensed format used by \pkg{TMB}, which excludes mapped elements.
(See \code{\link{egf}} argument \code{map}.)}

\item{list}{A logical flag.
If \code{FALSE}, then the parameter vectors are concatenated.}

\item{...}{Unused optional arguments.}
}
\value{
If \code{list = FALSE}, then a numeric vector \code{x = c(beta, theta, b)},
with \code{names(x)} providing the grouping.

If \code{list = TRUE}, then a named list \code{list(beta, theta, b)}.
Each parameter vector has an attribute \code{map}, an index vector such
that the full vector \code{y} and condensed vector \code{x} are related
by \code{y = x[map]}, with the exception that \code{map[i]} is \code{NA}
if \code{y[i]} is mapped to an initial value
(hence \code{x[map[i]]} is equal to \code{NA}, not \code{y[i]}).

In both cases, the result inherits from class \code{"egf_coef"}.
}
\description{
Extracts bottom level parameter vectors \code{beta}, \code{theta},
and \code{b}. These store
linear fixed effects coefficients,
random effect covariance parameters, and
linear random effects coefficients, respectively.
}
\details{
If \code{object} inherits from class \code{"egf_no_fit"},
then the result lists initial parameter values to be used
in the first likelihood evaluation.
}
\seealso{
Other extractors: 
\code{\link{df.residual.egf}()},
\code{\link{fitted.egf}()},
\code{\link{fixef.egf}()},
\code{\link{formula.egf}()},
\code{\link{getCall.egf}()},
\code{\link{logLik.egf}()},
\code{\link{model.frame.egf}()},
\code{\link{model.matrix.egf}()},
\code{\link{nobs.egf}()},
\code{\link{ranef.egf}()},
\code{\link{terms.egf}()},
\code{\link{vcov.egf}()}
}
\concept{extractors}
