% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.egf}
\alias{coef.egf}
\alias{coef.egf_init}
\title{Extract model parameter estimates}
\usage{
\method{coef}{egf_init}(object, log = FALSE, ...)

\method{coef}{egf}(object, log = FALSE, ...)
}
\arguments{
\item{object}{An "egf_init" or "egf" object.}

\item{log}{A logical scalar. If \code{TRUE}, then parameter estimates are
log-transformed.}

\item{...}{Unused optional arguments.}
}
\value{
The method for class "egf_init" returns
\code{object$theta_init} if \code{log = FALSE} and
\code{object$log_theta_init} if \code{log = TRUE}.

The method for class "egf" returns
\code{object$theta_fit} if \code{log = FALSE} and
\code{object$log_theta_fit} if \code{log = TRUE}.
}
\description{
Methods for extracting estimates of model parameters from objects
of class "egf_init" or "egf".
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \code{\link[=egf]{egf()}}
}
