% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R0.R
\name{compute_R0}
\alias{compute_R0}
\alias{compute_R0.default}
\alias{compute_R0.egf}
\title{\loadmathjax
Compute the basic reproduction number}
\usage{
compute_R0(r, breaks, probs, ...)

\method{compute_R0}{default}(r, breaks, probs, ...)

\method{compute_R0}{egf}(r, breaks, probs, subset, ...)
}
\arguments{
\item{r}{A numeric vector with non-negative elements listing initial
exponential growth rates \mjseqn{r}. Alternatively, an \code{"egf"}
object returned by \code{\link[=egf]{egf()}}.}

\item{breaks}{A numeric vector of length 2 or greater listing increasing
break points \mjseqn{t_i} in the support of the generation
interval distribution, in reciprocal units of \code{r} (days if
\code{r} is an \code{"egf"} object).}

\item{probs}{A numeric vector of length \code{length(breaks)-1}. \code{probs[i]}
is the probability \mjseqn{p_i} that the generation interval
is between \code{breaks[i]} and \code{breaks[i+1]}. If \code{sum(probs) != 1},
then \code{probs} is replaced with \code{probs / sum(probs)}.}

\item{...}{Arguments passed to methods.}

\item{subset}{A named list passed to \code{\link[=fitted.egf]{fitted.egf()}}.}
}
\value{
The default method returns a numeric vector of length \code{length(r)}
listing basic reproduction numbers \mjseqn{\mathcal{R}USCORE0}
(see Details).

The method for class \code{"egf"} constructs the data frame
\code{d = fitted(r, subset, link = TRUE)}, appends
\code{R0 = compute_R0.default(r = exp(d$log_r), breaks, probs)},
and returns the result omitting extraneous variables.
}
\description{
Calculates basic reproduction numbers given initial exponential
growth rates and a binned generation interval distribution.
}
\details{
Let \mjseqn{t_0 < \cdots < t_m} be the break points specified by
\code{breaks}. For \mjseqn{i \in \lbrace 1,\ldots,m \rbrace}, let
\mjseqn{p_i} be the probability specified by \code{probs[i]}. This is
the probability that the generation interval is in the interval
\mjseqn{\lbrack t_{i-1},t_i)}.
Section 3d in \insertCite{WallLips07;textual}{epigrowthfit}
gives the basic reproduction number \mjseqn{\mathcal{R}USCORE0}
as a function of the initial exponential growth rate \mjseqn{r}:

\mjsdeqn{\mathcal{R}USCORE0(r) = \left. r \middle/ \bigg\lbrace \sum_{i=1}^{m} \frac{p_i (e^{-r t_{i-1}} - e^{-r t_i})}{t_i - t_{i-1}} \bigg\rbrace \right.\,.}
}
\references{
\insertRef{WallLips07}{epigrowthfit}
}
\seealso{
\code{\link[=compute_final_size]{compute_final_size()}}
}
