% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R0.R
\name{compute_R0}
\alias{compute_R0}
\alias{compute_R0.numeric}
\alias{compute_R0.egf_init}
\alias{compute_R0.egf}
\title{\loadmathjax
Compute the basic reproduction number}
\usage{
compute_R0(x, breaks, probs)

\method{compute_R0}{numeric}(x, breaks, probs)

\method{compute_R0}{egf_init}(x, breaks, probs)

\method{compute_R0}{egf}(x, breaks, probs)
}
\arguments{
\item{x}{A numeric vector listing values for the initial exponential
growth rate expressed per day. Alternatively, an "egf_init"
or "egf" object.}

\item{breaks}{A numeric vector of length 2 or greater listing increasing
break points \mjseqn{t_i}, in days, in the support of the
generation interval distribution.}

\item{probs}{A numeric vector with length \code{length(breaks)-1}. \code{probs[i]}
is the probability that the generation interval is between
\code{break[i]} days and \code{break[i+1]} days. Replaced with
\code{probs / sum(probs)} in the event that \code{sum(probs) != 1}.}
}
\value{
The method for class "numeric" returns a numeric vector
of length \code{length(x)}, whose \code{i}th element is the basic
reproduction number corresponding to initial exponential
growth rate \code{x[i]}. See Details.

The method for class "egf_init" applies the method for
class "numeric" to \code{x$theta_init[["r"]]}.

The method for class "egf" applies the method for
class "numeric" to \code{x$theta_fit[["r"]]}.
}
\description{
Compute the basic reproduction number corresponding
to an initial exponential growth rate and a binned
generation interval distribution.
}
\details{
Let \mjseqn{t_0 < \cdots < t_m} be the break points specified
by \code{breaks}. For \mjseqn{i \in \lbrace 1,\ldots,m \rbrace},
let \mjseqn{p_i} be the probability specified by \code{probs[i]}
that the generation interval is in the interval
\mjseqn{\lbrace t_{i-1},t_i \rbrace}.
Section 3d in \insertCite{WallLips07;textual}{epigrowthfit}
gives the basic reproduction number \mjseqn{\mathcal{R}_0}
as a function of the initial exponential growth rate \mjseqn{r}:

\mjsdeqn{\mathcal{R}_0(r) = \left. r \middle/ \bigg\lbrace \sum_{i=1}^{m} \frac{p_i (e^{-r t_{i-1}} - e^{-r t_i})}{t_i - t_{i-1}} \bigg\rbrace \right.\,.}
}
\examples{
data(plague_latent_period)
lat <- plague_latent_period$relfreq
m <- length(lat)

data(plague_infectious_period)
inf <- plague_infectious_period$relfreq
n <- length(inf)

r <- seq(0, 1, by = 0.02)
breaks <- 1:(m+n)
probs <- diff(pgi(breaks, lat, inf))

R0 <- compute_R0(r, breaks, probs)
plot(r, R0, las = 1,
  xlab = "initial exponential growth rate, per day",
  ylab = "basic reproduction number"
)

}
\references{
\insertRef{WallLips07}{epigrowthfit}
}
\seealso{
\code{\link[=compute_final_size]{compute_final_size()}} for computing the expected epidemic
final size as a function of the basic reproduction number
}
