% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_R0}
\alias{compute_R0}
\title{Compute the basic reproduction number}
\usage{
compute_R0(r, breaks, probs)
}
\arguments{
\item{r}{A non-negative \link{numeric} vector listing initial exponential
growth rates.}

\item{breaks}{An increasing \link{numeric} vector of length 2 or greater listing
break points in the support of the generation interval distribution,
in reciprocal units of \code{r}.}

\item{probs}{A \link{numeric} vector of length \code{\link{length}(breaks)-1}.
\code{probs[i]} is the probability that the generation interval
is between \code{breaks[i]} and \code{breaks[i+1]}.
If \code{\link{sum}(probs) != 1}, then \code{probs} is replaced
with \code{probs / \link{sum}(probs)}.}
}
\value{
A numeric vector of length \code{\link{length}(r)} listing
basic reproduction numbers.
}
\description{
Computes the basic reproduction number as a function of the
initial exponential growth rate, conditional on a binned
generation interval distribution.
}
\section{Computation}{

For an initial exponential growth rate \code{r},
the basic reproduction number \code{R0} is computed as

\code{R0(r) = r / sum(probs * (exp(-r * breaks[-n]) - exp(-r * breaks[-1L])) / (breaks[-1L] - breaks[-n]))},

where \code{n = \link{length}(breaks)}.
}

\examples{
r <- seq(0, 1, by = 0.02)
breaks <- 0:20
probs <- diff(pgamma(breaks, shape = 1, scale = 2.5))
R0 <- compute_R0(r, breaks, probs)

plot(r, R0, las = 1,
  xlab = "initial exponential growth rate",
  ylab = "basic reproduction number"
)

}
\references{
Wallinga J, Lipsitch M. How generation intervals shape the relationship
between growth rates and reproductive numbers. Proc R Soc Lond B Biol Sci.
2007;274:599--604.
}
\seealso{
Other epidemic parameters: 
\code{\link{compute_final_size}()},
\code{\link{compute_tdoubling}()}
}
\concept{epidemic parameters}
