% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R0.R
\name{compute_R0}
\alias{compute_R0}
\title{\loadmathjax
Compute the basic reproduction number}
\usage{
compute_R0(r, breaks, probs)
}
\arguments{
\item{r}{A numeric vector listing values for the initial
growth rate expressed per day.}

\item{breaks}{A numeric vector of length 2 or greater
listing increasing break points, in days, in the support
of the generation interval density.}

\item{probs}{A numeric vector with length \code{length(breaks)-1}.
\code{probs[i]} is the probability that the generation interval
is between \code{break[i]} days and \code{break[i+1]} days. Replaced
with \code{probs / sum(probs)} in the event that \code{sum(probs) != 1}.}
}
\value{
A numeric vector of length \code{length(r)} listing values
for the basic reproduction number. The \code{i}th element is
computed using initial growth rate \code{r[i]}. See Details.
}
\description{
Given an initial growth rate and a generation interval
distribution, compute the basic reproduction number.
}
\details{
Let \mjseqn{a_1 < \cdots < a_\ell} be the break points
specified by \code{breaks}.
For \mjseqn{i \in \lbrace 1,\ldots,\ell-1 \rbrace}, let
\mjseqn{z_i} be the probability specified by \code{probs[i]}
that the generation interval \mjseqn{Z} is in the interval
\mjseqn{(a_i,a_{i+1}\rbrack}.

Section 3(d) in \insertCite{WallLips07;textual}{epigrowthfit}
states that the basic reproduction number is given by

\mjsdeqn{\out{\mathcal{R}_0 = \left. r \middle/ \bigg\lbrace \sum_{i=1}^{\ell-1} \frac{z_i (e^{-r a_i} - e^{-r a_{i+1}})}{a_{i+1} - a_i} \bigg\rbrace \right.\,,}}

where \mjseqn{r} is the initial growth rate. \code{compute_R0()}
evaluates the right-hand side using each value for \mjseqn{r}
specified in argument \code{r}.
}
\examples{
data(plague_latent_period)
latent <- plague_latent_period$relfreq
m <- length(latent)

data(plague_infectious_period)
infectious <- plague_latent_period$relfreq
n <- length(infectious)

r <- seq(0, 1, by = 0.02)
breaks <- 1:(m+n)
probs <- diff(pgi(breaks, latent, infectious))

R0 <- compute_R0(r, breaks, probs)
plot(r, R0, las = 1,
  xlab = "initial growth rate, per day",
  ylab = "basic reproduction number"
)

}
\references{
\insertRef{WallLips07}{epigrowthfit}
}
\seealso{
\code{\link[=dgi]{dgi()}}, \code{\link[=pgi]{pgi()}} for evaluating generation interval
density and distribution functions assuming discrete latent
and infectious period distributions
}
