% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_tdoubling}
\alias{compute_tdoubling}
\title{Compute doubling time}
\usage{
compute_tdoubling(r, per = NULL)
}
\arguments{
\item{r}{A non-negative \link{numeric} vector.}

\item{per}{A positive integer indicating that \code{r} is a rate per
\code{per} days, in which case the result is printed with units.
Use the default (\code{\link{NULL}}) if \code{r} is unitless.}
}
\value{
\code{\link{log}(2) / r} with \code{"tdoubling"} prepended
to its \link{class}.
\code{per} is retained as an \link[=attributes]{attribute}
for use by \code{print.tdoubling}.
}
\description{
Computes doubling times corresponding to exponential growth rates.
}
\examples{
r <- 10^(-2:0)
tdoubling <- compute_tdoubling(r)
all.equal(tdoubling, log(2) / r)

## 'per' affects printing
for (i in c(1, 2, 7, 8, 365, 366)) {
  attr(tdoubling, "per") <- i
  print(tdoubling)
}

}
\seealso{
Other epidemic parameters: 
\code{\link{compute_R0}()},
\code{\link{compute_final_size}()}
}
\concept{epidemic parameters}
