% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.egf}
\alias{confint.egf}
\title{Confidence intervals on fitted values}
\usage{
\method{confint}{egf}(
  object,
  parm,
  level = 0.95,
  top = egf_get_names_top(object, link = TRUE),
  subset = NULL,
  append = NULL,
  link = TRUE,
  method = c("wald", "profile", "uniroot"),
  grid_len = 12,
  trace = TRUE,
  parallel = egf_parallel(),
  max_width = 7,
  breaks = NULL,
  probs = NULL,
  .subset = NULL,
  .append = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1) indicating a confidence level.}

\item{top}{A subset of \code{\link{egf_get_names_top}(object, link = TRUE)}
naming top level nonlinear model parameters for which confidence
intervals should be computed. If \code{object$model$curve} is
\code{"exponential"}, \code{"logistic"}, or \code{"richards"},
then \code{top} may also contain \code{"tdoubling"} and \code{"R0"}.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{egf_combine_frames}}). It must evaluate
to a valid index vector for the rows of the data frame
(see \code{\link{[.data.frame}}), and thus fitting windows.
Confidence intervals are computed only for indexed windows.
The default (\code{\link{NULL}}) is to consider all windows.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link{egf_combine_frames}}) to be included with the
result. The default (\code{\link{NULL}}) is to append nothing.}

\item{link}{A \link{logical} flag. If \code{FALSE}, then confidence intervals
on inverse link-transformed fitted values are computed.}

\item{method}{A \link{character} string indicating how confidence intervals
should be computed.}

\item{grid_len}{(For \code{method = "profile"}.)
A positive integer. Step sizes chosen adaptively by
\code{\link[TMB]{tmbprofile}} will generate approximately
this many points on each side of a profile's minimum point.}

\item{trace}{(For \code{method != "wald"}.)
A \link{logical} flag.
If \code{TRUE}, then basic tracing messages indicating progress
are printed.
Depending on \code{object$control$trace}, these may be mixed with
optimizer output.}

\item{parallel}{(For \code{method != "wald"}.)
An \code{"\link{egf_parallel}"} object defining options for \R level
parallelization.}

\item{max_width}{(For \code{method = "uniroot"}.)
A positive number. \code{\link[TMB]{tmbroot}} will search for roots
in the interval from \code{x-max_width} to \code{x+max_width}, where
\code{x} is the fitted value (link scale).}

\item{breaks, probs}{(For \code{top = "R0"}.)
Arguments to \code{\link{compute_R0}}.}

\item{.subset, .append}{Index vectors to be used (if non-\code{\link{NULL}}) in place of
the result of evaluating \code{subset} and \code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_confint"}, with variables:
\item{top}{
  Top level nonlinear model parameter,
  from \code{\link{egf_get_names_top}(object, link = TRUE)}.
}
\item{ts}{
  Time series, from \code{\link{levels}(object$frame_windows$ts)}.
}
\item{window}{
  Fitting window, from \code{\link{levels}(object$frame_windows$window)}.
}
\item{estimate, lower, upper}{
  Fitted value and approximate lower and upper confidence limits.
}
\code{level} and \code{object$frame_windows} are retained
as \link{attributes}.
}
\description{
Computes confidence intervals on \link[=fitted.egf]{fitted values}
of top level nonlinear model parameters and, where appropriate,
initial doubling times and basic reproduction numbers.
}
\details{
Three methods are provided for calculating confidence intervals:
\describe{
\item{\code{wald}}{
  See \code{\link{confint.egf_fitted}}.
}
\item{\code{profile}}{
  See \code{\link{confint.egf_profile}}.
}
\item{\code{uniroot}}{
  Similar to \code{"profile"}, except that the two solutions
  of \code{deviance(value) = \link{qchisq}(level, df = 1)} are
  approximated by root-finding using \code{\link[TMB]{tmbroot}}
  (\code{\link{uniroot}} internally).
}
}
For top level parameters following random effects models,
\code{"wald"} returns confidence intervals on individual fitted values,
while \code{"profile"} and \code{"uniroot"} return confidence intervals
on population fitted values, which are the fixed effects components of
individual fitted values.

\code{"wald"} assumes e.g., asymptotic normality of the maximum likelihood
estimator. \code{"profile"} and \code{"uniroot"} avoid these issues but are
typically more expensive, requiring estimation of many restricted models.
They are parallelized at the C++ level when there is OpenMP support and
\code{object$control$omp_num_threads} is set to an integer greater than 1.
If there is no OpenMP support, then computation can still be parallelized
at the \R level with appropriate setting of \code{parallel}.

See topic \code{\link{egf_eval}} for details on nonstandard evaluation
of \code{subset} and \code{append}.
}
\examples{
example("egf", "epigrowthfit")
zz <- confint(object)
str(zz)

}
\seealso{
\code{plot.egf_confint}
}
