% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.egf}
\alias{confint.egf}
\title{Confidence intervals on point estimates}
\usage{
\method{confint}{egf}(object, parm = "r", level = 0.95, method = "linear", ...)
}
\arguments{
\item{object}{An "egf" object.}

\item{parm}{An element of \code{names(object$theta_fit)} indicating
a parameter for which to compute a confidence interval.}

\item{level}{A number in the interval [0,1] indicating a
confidence level.}

\item{method}{One of \code{"linear"}, \code{"uniroot"}, and \code{"wald"}
indicating a method with which to compute the confidence interval.}

\item{...}{Unused optional arguments.}
}
\value{
A numeric vector of the form \code{c(lower = a, upper = b)},
where \code{a} and \code{b} are the left and right endpoints of
the confidence interval specified by \code{parm}, \code{level},
and \code{method}.
}
\description{
A method for obtaining confidence intervals on point estimates
of model parameters specified by "egf" objects.
}
\seealso{
\code{\link[=egf]{egf()}}, \code{\link[TMB:tmbprofile]{TMB::tmbprofile()}}, \code{\link[TMB:tmbroot]{TMB::tmbroot()}}
}
