% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.egf}
\alias{confint.egf}
\title{Confidence intervals on fitted values}
\usage{
\method{confint}{egf}(
  object,
  parm,
  level = 0.95,
  par = get_names_top(object, link = TRUE),
  subset = NULL,
  append = NULL,
  link = TRUE,
  method = c("wald", "profile", "uniroot"),
  grid_len = 12,
  parallel = egf_parallel(),
  trace = TRUE,
  max_width = 7,
  breaks = NULL,
  probs = NULL,
  .subset = NULL,
  .append = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1) indicating a confidence level.}

\item{par}{A subset of \code{\link{get_names_top}(object, link = TRUE)}
naming top level nonlinear model parameters for which confidence
intervals should be computed. If \code{object$model$curve} is
\code{"exponential"}, \code{"logistic"}, or \code{"richards"}),
then \code{par} may also contain \code{"tdoubling"} and \code{"R0"}.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{egf_make_combined}}). Must evaluate to
a \link{logical} vector indexing rows of the data frame,
and thus fitting windows. Confidence intervals are computed
only for indexed windows. The default (\code{\link{NULL}})
is to consider all windows.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link{egf_make_combined}}) to be included with the
result. The default (\code{\link{NULL}}) is to append nothing.}

\item{link}{A \link{logical} flag. If \code{FALSE}, then confidence intervals
on inverse link-transformed fitted values are computed.}

\item{method}{A \link{character} string indicating how confidence intervals
should be computed.}

\item{grid_len}{(For \code{method = "profile"}.)
A positive integer. Step sizes chosen adaptively by
\code{\link[TMB]{tmbprofile}} will generate approximately
this many points on each side of a profile's minimum point.}

\item{parallel}{(For \code{method = "profile"}.)
An \code{"\link{egf_parallel}"} object defining parallelization options.}

\item{trace}{(For \code{method != "wald"}.)
A \link{logical} flag. If \code{TRUE}, then basic tracing messages
are printed.}

\item{max_width}{(For \code{method = "uniroot"}.)
A positive number. \code{\link[TMB]{tmbroot}} will search for roots
in the interval from \code{x-max_width} to \code{x+max_width}, where
\code{x} is the fitted value (link scale).}

\item{breaks, probs}{(For \code{par = "R0"}.)
Arguments to \code{\link{compute_R0}}.}

\item{.subset}{A \link{logical} vector to be used (if non-\code{\link{NULL}})
in place of the result of evaluating \code{subset}.}

\item{.append}{A \link{character} vector listing variable names to be used
(if non-\code{\link{NULL}}) in place of the result of evaluating
\code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_confint"}, with variables:
\item{par}{
  Top level nonlinear model parameter,
  from \code{\link{get_names_top}(object, link = TRUE)}.
}
\item{ts}{
  Time series, from \code{\link{levels}(object$frame_windows$ts)}.
}
\item{window}{
  Fitting window, from \code{\link{levels}(object$frame_windows$window)}.
}
\item{estimate, lower, upper}{
  Fitted value and approximate lower and upper confidence limits.
}
\code{level} and \code{object$frame_windows} are retained
as \link{attributes}.
}
\description{
Computes confidence intervals on \link[=fitted.egf]{fitted values}
of top level nonlinear model parameters and, where appropriate,
initial doubling times (in days) and basic reproduction numbers.
}
\details{
Three methods are provided for calculating confidence intervals:
\describe{
\item{\code{wald}}{
  See \code{\link{fitted.egf}} and \code{\link{confint.egf_fitted}}.
}
\item{\code{profile}}{
  See \code{\link{profile.egf}} and \code{\link{confint.egf_profile}}.
}
\item{\code{uniroot}}{
  Similar to \code{"profile"}, except that the two solutions
  of \code{deviance(value) = \link{qchisq}(level, df = 1)} are
  approximated by root-finding using \code{\link[TMB]{tmbroot}}
  (\code{\link{uniroot}} internally).
}
}
For top level parameters following random effects models, \code{"wald"}
returns confidence intervals on individual fitted values,
whereas \code{"profile"} and \code{"uniroot"} return confidence intervals
on the fixed effects components only, namely the population fitted values.

\code{"wald"} requires minimal computation time but assumes,
e.g., asymptotic normality of the maximum likelihood estimator.
A further limitation of \code{"wald"} is functional non-invariance.
\code{"profile"} and \code{"uniroot"} avoid these issues but are
much slower, requiring estimation of restricted models. Of the two,
\code{"profile"} is more robust.

See topic \code{\link{nse}} for details on nonstandard evaluation
of \code{subset} and \code{append}.
}
\seealso{
\code{plot.egf_confint}
}
