% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.egf_profile}
\alias{confint.egf_profile}
\title{Confidence intervals from likelihood profiles}
\usage{
\method{confint}{egf_profile}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An \code{"egf_profile"} object returned by \code{\link[=profile.egf]{profile.egf()}}.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1). The desired confidence level.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame with character variable \code{name} equal to
\code{levels(object$name)} and numeric variables \code{estimate},
\code{lower}, and \code{upper} supplying estimates and confidence
intervals.
}
\description{
Computes confidence intervals on fixed effect coefficients,
log standard deviations of random effects, and linear combinations
thereof from their (univariate) likelihood profiles.
}
\details{
For each likelihood profile (level of \code{object$name}),
\code{\link[TMB:confint.tmbprofile]{TMB::confint.tmbprofile()}} is called to approximate
(via linear interpolation) the two solutions of
\code{deviance(x) = qchisq(level, df = 1)}. These are
the lower and upper confidence limits of interest.
}
