% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{confint.egf_profile}
\alias{confint.egf_profile}
\title{Confidence intervals from likelihood profiles}
\usage{
\method{confint}{egf_profile}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An \code{"egf_profile"} object returned by \code{\link[=profile.egf]{profile.egf()}}.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1), indicating a confidence level.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame with variables:
\item{\code{index}}{
Row index of linear combination, from \code{levels(object$index)}.
}
\item{\code{estimate}, \code{lower}, \code{upper}}{
Estimate and (approximate) lower and upper confidence limits
of linear combination.
}
\item{\code{par}}{
(If \code{object$par} is non-\code{NULL}.)
Nonlinear model parameter, from \code{levels(object$par)}.
}
\item{\code{window}}{
(If \code{object$window} is non-\code{NULL}.)
Fitting window, from \code{levels(object$window)}.
}
\code{level} and attributes \code{A} and \code{par} of \code{object} are retained
as attributes.
}
\description{
Computes confidence intervals on fixed effects coefficients,
log standard deviations of random effects coefficients, and
linear combinations thereof from their univariate likelihood
profiles.
}
\details{
For each supplied likelihood profile (level of \code{object$index}),
\code{\link[stats:approxfun]{stats::approx()}} is called to approximate (via linear interpolation)
the two solutions of \code{deviance(value) = qchisq(level, df = 1)},
which provide the lower and upper confidence limits of interest.
}
