% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{confint.egf_profile}
\alias{confint.egf_profile}
\title{Confidence intervals from likelihood profiles}
\usage{
\method{confint}{egf_profile}(object, parm, level = 0.95, link = FALSE, ...)
}
\arguments{
\item{object}{An \code{"egf_profile"} object returned by \code{\link[=profile.egf]{profile.egf()}}.}

\item{parm}{Unused argument included for generic consistency.}

\item{level}{A number in the interval (0,1). The desired confidence level.}

\item{link}{A logical scalar. If \code{FALSE} and \code{object} supplies
likelihood profiles of population fitted values of
nonlinear model parameters, then confidence intervals
on inverse link-transformed fitted values are returned.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame with variables:
\item{\code{linear_combination}}{
Row index of linear combination,
from \code{seq_len(nrow(attr(object, "A")))}.
}
\item{\code{estimate}, \code{lower}, \code{upper}}{
Estimate of linear combination and approximate lower
and upper confidence limits, inverse-link transformed
if \code{link = FALSE} and linear combinations represent
population fitted values of nonlinear model parameter.
}
\code{attr(object, "A")}, \code{attr(object, "x")}, and \code{level}
are retained as attributes.
}
\description{
Computes confidence intervals on fixed effects coefficients,
log standard deviations of random effects coefficients, and
linear combinations thereof from their univariate likelihood
profiles.
}
\details{
Each supplied likelihood profile
(level of \code{object$linear_combination}),
is linearly interpolated to approximate the two solutions of
\code{deviance(value) = qchisq(level, df = 1)}.
These provide the lower and upper confidence limits of interest
(see Wilks' theorem).
}
