% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor2cov}
\alias{cor2cov}
\title{Correlation to covariance matrix conversion}
\usage{
cor2cov(cor, sd)
}
\arguments{
\item{cor}{A symmetric numeric matrix. Diagonal elements should be 1 and
off-diagonal elements should be numbers in the interval \verb{[-1,1]}.}

\item{sd}{A numeric vector of length \code{nrow(cor)}.}
}
\value{
The result of \verb{sweep(sweep(cor, 1L, sd, ```*```), 2L, sd, ```*```)},
obtained slightly more efficiently, following \code{\link[stats:cor]{stats::cov2cor()}}.
}
\description{
Perform the inverse of \code{\link[stats:cor]{stats::cov2cor()}}.
}
\examples{
# X <- replicate(6L, rnorm(10L))
# V <- cov(X, X)
# all.equal(V, cor2cov(cov2cor(V), sqrt(diag(V))))

}
\keyword{internal}
