% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{covid_generation_interval}
\alias{covid_generation_interval}
\title{Generation interval distribution (COVID-19)}
\format{
A \link{list} with 4 elements:
\describe{
\item{\code{shape}}{
  Estimated shape parameter. See \code{\link{dgamma}}.
}
\item{\code{scale}}{
  Estimated scale parameter. See \code{\link{dgamma}}.
}
\item{\code{breaks}}{
  An \link{integer} vector listing numbers of days. Equal to `0:20`.
}
\item{\code{probs}}{
  A \link[=double]{numeric} vector of length \code{\link{length}(breaks-1)}.
  \code{probs[i]} is the probability that the generation interval
  is between \code{breaks[i]} and \code{breaks[i+1]} days, conditional
  on \code{shape} and \code{scale}. Equal to
  \code{\link{diff}(\link{pgamma}(breaks, shape = shape, scale = scale))}.
}
}
}
\source{
\code{shape} and \code{scale} were computed from the mean and standard
deviation reported in Ganyani et al. (2020), Table 4, Scenario 2.
}
\usage{
data(plague_infectious_period)
}
\description{
Gamma distribution of the COVID-19 generation interval fitted
to data from a cluster of 45 cases in Tianjin, China.
}
\examples{
data(covid_generation_interval)
x <- 10^seq(-2, log10(20), length.out = 150L)
fx <- with(covid_generation_interval,
  dgamma(x, shape = shape, scale = scale)
)
plot(x, fx, type = "l", ylim = c(0, max(fx[-1L])))

}
\references{
Ganyani T, Kremer C, Chen D, Torneri A, Faes C, Wallinga J, et al.
Estimating the generation interval for coronavirus disease (COVID-19)
based on symptom onset data, March 2020. Eurosurveillance.
2020;25(17):2000257.
}
\keyword{datasets}
