% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daxis}
\alias{daxis}
\title{Annotate a date axis}
\usage{
daxis(
  left,
  right,
  refdate,
  tcl = -0.2,
  mgp2 = c(0.05, 1),
  col.axis = "black",
  cex.axis = c(0.7, 0.85)
)
}
\arguments{
\item{left}{Left endpoint of the bottom axis in user coordinates.}

\item{right}{Right endpoint of the bottom axis in user coordinates.}

\item{refdate}{A Date scalar. \code{t0} and \code{t1} represent a number of
days since this date.}

\item{tcl}{A numeric scalar. Passed to \code{\link[graphics:axis]{graphics::axis()}} argument
\code{tcl} when creating the minor axis. Defines tick length.
See \code{\link[graphics:par]{graphics::par()}}.}

\item{mgp2}{A numeric vector of length 2. Elements are passed
to \code{\link[graphics:axis]{graphics::axis()}} argument \code{mgp} (the second component)
when creating the minor and major axes, respectively. Defines
the distance between the axis and the top of tick labels.
See \code{\link[graphics:par]{graphics::par()}}.}

\item{col.axis}{A numeric or character vector of length 2.
Elements are passed to \code{\link[graphics:axis]{graphics::axis()}} argument \code{col.axis}
when creating the minor and major axes, respectively. Defines
the colour of tick labels. See \code{\link[graphics:par]{graphics::par()}}.}

\item{cex.axis}{A numeric vector of length 2. Elements are
passed to \code{\link[graphics:axis]{graphics::axis()}} argument \code{cex.axis} when creating
the minor and major axes, respectively. Defines the size of
tick labels. See \code{\link[graphics:par]{graphics::par()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Labels day, month, and year on a bottom axis, taking care to ensure
that labels are nicely spaced.
}
\details{
\code{daxis()} assumes that horizontal user coordinates measure
a number of days since \code{refdate}.

The date axis consists of minor and major axes. The content
of these axes depends entirely on \code{w = round(right-left)},
the rounded difference between the extreme times in days.

If \code{w <= 210} (7 months), then days are placed on the minor
axis, months are placed on the major axis, and years are not
shown. The spacing of the minor axis in days is the value of
\code{c(1, 2, 4, 7, 14)[w <= c(14, 28, 56, 112, 210)][1]}. Hence,
for example, if \code{w} is greater than 112 and less than or equal
to 210, then the spacing is 14-daily.

Otherwise, if \code{w <= 3*365} (3 years), then months are placed
on the minor axis, years are placed on the major axis, and days
are not shown. The spacing of the minor axis in months is the
value of \code{c(1, 2, 3)[w <= c(1, 2, 3) * 365][1]}.

Otherwise (i.e., if \code{w > 3*365}), years are placed on the minor
axis, and days and months are not shown. The spacing of the
minor axis is the value of \code{ceiling(ceiling(w / 365) / 7)}.

All nonzero lengths of \code{mgp2}, \code{cex.axis}, and \code{col.axis}
are tolerated. Vectors of length 1 are recycled. Only the
first two elements of vectors of length 3 or greater are used.
}
\keyword{internal}
