% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daxis.R
\name{daxis}
\alias{daxis}
\title{Annotate a date axis}
\usage{
daxis(
  t0 = par("usr")[1],
  t1 = par("usr")[2],
  refdate,
  tcl = -0.2,
  line = c(0.05, 1),
  col.axis = c("black", "black"),
  cex.axis = c(0.7, 0.85)
)
}
\arguments{
\item{t0}{Left endpoint of the axis in user coordinates.}

\item{t1}{Right endpoint of the axis in user coordinates.}

\item{refdate}{A Date scalar. \code{t0} and \code{t1} represent a number of
days since this date.}

\item{tcl}{A numeric scalar. Passed to \code{\link[graphics:axis]{axis()}}
argument \code{tcl} when creating the minor axis. Defines tick length.
See \code{\link[graphics:par]{par()}}.}

\item{line}{A numeric vector of length 2. Elements are passed
to \code{\link[graphics:axis]{axis()}} argument \code{mgp} (the second
component) when creating the minor and major axes,
respectively. Defines the distance between the axis and
the top of tick labels. See \code{\link[graphics:par]{par()}}.}

\item{col.axis}{A numeric or character vector of length 2.
Elements are passed to \code{\link[graphics:axis]{axis()}} argument
\code{col.axis} when creating the minor and major axes, respectively.
Defines the colour of tick labels. See \code{\link[graphics:par]{par()}}.}

\item{cex.axis}{A numeric vector of length 2. Elements are
passed to \code{\link[graphics:axis]{axis()}} argument \code{cex.axis}
when creating the minor and major axes, respectively.
Defines the size of tick labels. See \code{\link[graphics:par]{par()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Labels day, month, and year on a horizontal time axis, taking care
to ensure that labels are nicely spaced. Intended only for internal
use within
\code{\link[=egf_init-methods]{plot.egf_init()}} and
\code{\link[=egf_init-methods]{plot.egf()}}.
}
\details{
The date axis consists of minor and major axes.
Let \code{w} be the difference between the first and last date in days,
If \code{w <= 210} (7 months), then days are placed on the minor axis,
months are placed on the major axis, and years are not shown.
Otherwise, if \code{w <= 3*365} (3 years), then months are placed on
the minor axis, years are placed on the major axis, and days are
not shown. Otherwise (i.e., if \code{w > 3*365}), then years are placed
on the minor axis, and days and months are not shown.

The spacing of days, months, and years in each of these cases
varies with \code{w} to prevent overcrowding. How the spacing varies
with \code{w} is defined internally.

All nonzero lengths of \code{line}, \code{cex.axis}, and \code{col.axis}
are tolerated. Vectors of length 1 are recycled. Only the
first two elements of vectors of length 3 or greater are used.
}
\keyword{internal}
