% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{daxis}
\alias{daxis}
\title{Annotate a date axis}
\usage{
daxis(left, right, origin = .Date(0), plot = TRUE, minor = NULL, major = NULL)
}
\arguments{
\item{left, right}{Left and right endpoints of the bottom axis in user coordinates.
Taken from \code{par("usr")} if not supplied.}

\item{origin}{A Date scalar. It is assumed that horizontal user coordinates
measure time as a number of days since time 00:00:00 on date
\code{origin}.}

\item{plot}{A logical scalar. If \code{FALSE}, then an axis is not drawn.
Useful when the return value is desired but graphical
output is not.}

\item{minor, major}{Named lists of arguments to \code{\link[graphics:axis]{graphics::axis()}}, affecting the
appearance of the minor (day or month) and major (month or year)
axes, respectively.}
}
\value{
A list of numeric vectors \code{minor} and \code{major} giving
the positions of minor and major axis labels in user
coordinates.
}
\description{
Adds an axis to the bottom of the current plot and labels
it with day, month, and year, taking care to ensure that
labels are nicely spaced.
}
\details{
The date axis consists of minor and major axes. The content
of these axes depends entirely on \code{w = floor(right)-ceiling(left)},
the difference between the extreme times in days.

If \code{w <= 210} (7 months), then days are placed on the minor
axis, months are placed on the major axis, and years are not
shown. The spacing of the minor axis in days is the value of
\code{c(1, 2, 4, 7, 14)[w <= c(14, 28, 56, 112, 210)][1]}. Hence,
for example, if \code{w} is greater than 112 and less than or equal
to 210, then the spacing is 14-daily.

Otherwise, if \code{w <= 3*365} (3 years), then months are placed
on the minor axis, years are placed on the major axis, and days
are not shown. The spacing of the minor axis in months is the
value of \code{c(1, 2, 3)[w <= c(1, 2, 3) * 365][1]}.

Otherwise, if \code{w > 3*365} (3 years), then years are placed on
the minor axis, and days and months are not shown. The spacing
of the minor axis is the value of \code{ceiling(ceiling(w / 365) / 7)}.
}
\keyword{internal}
