% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daxis.R
\name{daxis}
\alias{daxis}
\title{Annotate a date axis}
\usage{
daxis(
  left,
  right,
  refdate,
  tcl = -0.2,
  line = c(0.05, 1),
  col.axis = c("black", "black"),
  cex.axis = c(0.7, 0.85)
)
}
\arguments{
\item{left}{Left endpoint of the axis in user coordinates.}

\item{right}{Right endpoint of the axis in user coordinates.}

\item{refdate}{A Date scalar. \code{t0} and \code{t1} represent a number of
days since this date.}

\item{tcl}{A numeric scalar. Passed to \code{\link[graphics:axis]{axis()}}
argument \code{tcl} when creating the minor axis. Defines tick length.
See \code{\link[graphics:par]{par()}}.}

\item{line}{A numeric vector of length 2. Elements are passed
to \code{\link[graphics:axis]{axis()}} argument \code{mgp} (the second
component) when creating the minor and major axes,
respectively. Defines the distance between the axis and
the top of tick labels. See \code{\link[graphics:par]{par()}}.}

\item{col.axis}{A numeric or character vector of length 2.
Elements are passed to \code{\link[graphics:axis]{axis()}} argument
\code{col.axis} when creating the minor and major axes, respectively.
Defines the colour of tick labels. See \code{\link[graphics:par]{par()}}.}

\item{cex.axis}{A numeric vector of length 2. Elements are
passed to \code{\link[graphics:axis]{axis()}} argument \code{cex.axis}
when creating the minor and major axes, respectively.
Defines the size of tick labels. See \code{\link[graphics:par]{par()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Labels day, month, and year on a horizontal time axis, taking care
to ensure that labels are nicely spaced. Intended mainly for internal
use within \code{\link[=plot.egf_init]{plot.egf_init()}} and \code{\link[=plot.egf]{plot.egf()}}.
}
\details{
The date axis consists of minor and major axes. The content
of these axes depends greatly on \code{w = round(right-left)},
roughly the difference between the first and last date in days.

If \code{w <= 210} (7 months), then days are placed on the minor
axis, months are placed on the major axis, and years are not
shown. The spacing of the minor axis in days is the value of
\code{c(1, 2, 4, 7, 14)[w <= c(14, 28, 56, 112, 210)][1]}. Hence,
for example, if \code{w} is greater than 112 and less than or equal
to 210, then the spacing is 14-daily.

Otherwise, if \code{w <= 3*365} (3 years), then months are placed
on the minor axis, years are placed on the major axis, and days
are not shown. The spacing of the minor axis in months is the
value of \code{c(1, 2, 3)[w <= c(1, 2, 3) * 365][1]}.

Otherwise (i.e., if \code{w > 3*365}), years are placed on the minor
axis, and days and months are not shown. The spacing of the
minor axis is the value of \code{ceiling(ceiling(w / 365) / 7)}.

All nonzero lengths of \code{line}, \code{cex.axis}, and \code{col.axis}
are tolerated. Vectors of length 1 are recycled. Only the
first two elements of vectors of length 3 or greater are used.
}
\keyword{internal}
