% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dceiling}
\alias{dceiling}
\title{Get ceiling of a date}
\usage{
dceiling(x, to = c("month", "year"))
}
\arguments{
\item{x}{A Date vector.}

\item{to}{A character string.}
}
\value{
\code{x} with elements replaced by firsts-of-the-month (\code{"YYYY-MM-01"})
or firsts-of-the-year (\code{"YYYY-01-01"}), depending on \code{to}.
}
\description{
Rounds each Date in a Date vector to the next first-of-the-month
or first-of-the-year.
}
\keyword{internal}
