% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf-methods.R
\name{egf-methods}
\alias{egf-methods}
\alias{print.egf}
\alias{coef.egf}
\alias{predict.egf}
\alias{simulate.egf}
\alias{plot.egf}
\title{Methods for class "egf"}
\usage{
\method{print}{egf}(x, ...)

\method{coef}{egf}(object, log = FALSE, ...)

\method{predict}{egf}(object, time = object$init$time, ...)

\method{simulate}{egf}(object, nsim = 1, seed = NULL, time = object$init$time, ...)

\method{plot}{egf}(x, inc = "cumulative", tol = 0.025, ...)
}
\arguments{
\item{x, object}{An "egf" object.}

\item{...}{Unused optional arguments.}

\item{log}{A logical scalar. If \code{TRUE}, then parameter values are
log-transformed.}

\item{time}{A numeric vector listing increasing time points in days
since \code{object$init$date[1]}. Missing values are not tolerated.
The simulate method requires \code{length(time) >= 2}.}

\item{nsim}{A positive integer specifying a number of simulations.}

\item{seed}{An integer specifying a seed for RNG, otherwise \code{NULL}.}

\item{inc}{One of \code{"interval"} and \code{"cumulative"},
indicating a type of incidence to plot.}

\item{tol}{A non-negative number used only if \code{inc = "interval"}.
\code{x$init$cases[i]} is highlighted if
\code{diff(x$init$time)[i] < (1-tol)*m} or
\code{diff(x$init$time)[i] > (1+tol)*m}, where
\code{m = median(diff(x$init$time))}.
Assign 0 to ensure that all deviations from \code{m} are highlighted.
Assign \code{Inf} to disable highlighting.}
}
\value{
The \code{print} method returns \code{x} invisibly.

The \code{coef} method returns \code{object$theta_hat} if \code{log = FALSE}
and \code{object$log_theta_hat} if \code{log = TRUE}.

The \code{predict} method returns a list with numeric elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{cum_inc}}{Expected cumulative incidence at time points
\code{time}, conditional on parameter vector \code{object$theta_hat}.
Equal to \code{object$cum_inc(time)}.
}
\item{\code{int_inc}}{Expected interval incidence given interval
endpoints \code{time}, conditional on parameter vector
\code{object$theta_hat}. Equal to \code{diff(object$cum_inc(time))}
if \code{length(time) >= 2} and omitted otherwise.
}
}

The \code{simulate} method returns a list with numeric elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{int_inc}}{A matrix with \code{length(time)-1} rows and \code{nsim}
columns, such that \code{int_inc[i, j]} is the number of cases
observed between \code{time[i]} and \code{time[i+1]} in simulation \code{j}
of \code{nsim}. Row vector \code{int_inc[i, ]} is sampled from a
Poisson or negative binomial distribution (depending on
\code{object$init$distr}) with mean \code{predict(object, time)$int_inc[i]}.
The negative binomial dispersion parameter is taken from
\code{object$theta_hat[["nbdisp"]]}.
}
\item{\code{cum_inc}}{A matrix with \code{length(time)} rows and \code{nsim}
columns, such that \code{cum_inc[i, j]} is the number of cases
observed up to \code{time[i]} in simulation \code{j}. Column vector
\code{cum_inc[, j]} is computed as \code{c0 + cumsum(c(0, int_inc[, j]))},
where \code{c0 = predict(object, time)$cum_inc[1]} is the expected
value of cumulative incidence at \code{time[1]} conditional on
parameter vector \code{object$theta_hat}.
}
}

The \code{plot} method returns \code{NULL} invisibly.
}
\description{
Methods for "egf" objects returned by \code{\link[=egf]{egf()}}.
}
\details{
\subsection{Plot elements}{

The bottom axis measures the number of days since \code{x$init$date[1]}.
The left axis measures interval or cumulative incidence
(depending on \code{inc}) on a log scale. Zeros are plotted as if
they were 10^(-0.2), and are therefore distinguished from
nonzero counts, which are always at least 1.

Observed data, specified by \code{x$init$time} and either \code{x$init$cases}
or \code{cumsum(x$init$cases)} (depending on \code{inc}), are plotted as
points. \code{cases[i]} gives the number of cases observed between
\code{time[i]} and \code{time[i+1]}, and \code{cumsum(cases)[i]} the number
observed between \code{time[1]} and \code{time[i+1]}. Both are plotted at
\code{time[i+1]}.

The left and right endpoints of the fitting window, specified by
indices \code{x$init$first} and \code{x$init$last}, are displayed as vertical
lines at \code{time[first+1]} and \code{time[last+1]} (adding one since \code{first}
and \code{last} index \code{cases}, and \code{length(time) = length(cases)+1}).

The incidence curve predicted by fitted parameter estimates
\code{x$theta_hat} is plotted as a teal line on grid points
\code{wgrid = seq(time[first+1], time[last+1], by = m)},
where \code{m = 1} for cumulative incidence and
where \code{m = median(diff(time))} for interval incidence
(to ensure that the curve has the correct scale; see below).

Careful interpretation of observed interval incidence is required
if the plotted time series is not equally spaced, as \code{cases}
roughly scales with \code{diff(time)}. That is, certain observations
may vary from the predicted curve not due to chance, but because
they represent a count over fewer or more days than the typical
observation interval, namely \code{median(diff(time))}. Observations
for which \code{diff(time)} differs from \code{median(diff(time))} are
highlighted according to argument \code{tol}.
}
}
\seealso{
\code{\link[=egf]{egf()}}
}
