% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf-methods.R
\name{egf-methods}
\alias{egf-methods}
\alias{print.egf}
\alias{coef.egf}
\alias{predict.egf}
\alias{simulate.egf}
\alias{plot.egf}
\title{Methods for class "egf"}
\usage{
\method{print}{egf}(x, ...)

\method{coef}{egf}(object, log = FALSE, ...)

\method{predict}{egf}(object, time = object$init$time, ...)

\method{simulate}{egf}(object, nsim = 1, seed = NULL, time = object$init$time, ...)

\method{plot}{egf}(
  x,
  inc = "interval",
  xty = "Date",
  log = FALSE,
  add = FALSE,
  annotate = TRUE,
  tol = 0,
  date_style = list(tcl = -0.2, line = c(0.05, 1), col.axis = c("black", "black"),
    cex.axis = c(0.7, 0.85)),
  window_style = list(col = "#DDCC7740", border = NA),
  line_style = list(lty = 1, lwd = 3, col = "#44AA99"),
  point_style_main = list(pch = 21, col = "#BBBBBB", bg = "#DDDDDD", cex = 1),
  point_style_short = list(pch = 1, col = "#882255", bg = NA, cex = 1),
  point_style_long = list(pch = 16, col = "#882255", bg = NA, cex = 1),
  text_style = list(pos = 3, offset = 0.3, col = "#BBBBBB", cex = 0.7, font = 2),
  ...
)
}
\arguments{
\item{x, object}{An "egf" object.}

\item{...}{Optional arguments. Used only by the \code{plot} method
to specify graphical parameters. Currently, only \code{xlim},
\code{ylim}, \code{xlab}, \code{ylab}, and \code{main} are implemented. Further
arguments are ignored.
See \code{\link[graphics:plot.default]{plot()}} and \code{\link[graphics:par]{par()}}
for a catalogue of graphical parameters.}

\item{log}{A logical scalar. For the \code{coef} method, if \code{TRUE},
then parameter estimates are log-transformed. For the plot
method, if \code{TRUE}, then incidence is displayed on a
logarithmic scale.}

\item{time}{A numeric vector listing increasing time points in days
since \code{object$init$date[1]}. Missing values are not tolerated.
The simulate method requires \code{length(time) >= 2}.}

\item{nsim}{A positive integer specifying a number of simulations.}

\item{seed}{An integer specifying a seed for RNG, otherwise \code{NULL}.}

\item{inc}{One of \code{"cumulative"} and \code{"interval"},
indicating a type of incidence to plot.}

\item{xty}{One of \code{"Date"} and \code{"numeric"}, indicating
a type of horizontal axis.}

\item{add}{A logical scalar. If \code{TRUE}, then the fitting window
and predicted curve are added to the current graphics device,
presumably started by a previous invocation of the plot method.}

\item{annotate}{A logical scalar. If \code{TRUE}, then a legend and
a list of parameter estimates are displayed in the right margin.}

\item{tol}{A non-negative number, used only if \code{inc = "interval"}.
\code{cases[i]} is highlighted according to \code{point_style_short} if
\code{diff(time)[i] < (1-tol)*m} and according to \code{point_style_long}
if \code{diff(time)[i] > (1+tol)*m}, where \code{m = median(diff(time))}.
In both cases, the value of \code{diff(time)[i]} is printed above
the point. Assign 0 to ensure that all deviations from \code{m} are
highlighted. Assign \code{Inf} to disable highlighting.}

\item{date_style}{A named list of arguments to \code{\link[=daxis]{daxis()}}
(a subset of \code{tcl}, \code{line}, \code{col.axis}, and \code{cex.axis}),
affecting the appearance of the bottom axis if \code{xty = "date"}.}

\item{window_style}{A named list of arguments to
\code{\link[graphics:polygon]{polygon()}}, affecting the appearance
of the fitting window. Currently, only \code{col} and \code{border} are
implemented.}

\item{line_style}{A named list of arguments to
\code{\link[graphics:lines]{lines()}}, affecting the appearance
of the predicted curve. Currently, only \code{lty}, \code{lwd}, and \code{col}
are implemented.}

\item{point_style_main}{A named list of arguments to
\code{\link[graphics:points]{points()}}, affecting the appearance
of the observed data. Currently, only \code{pch}, \code{col}, \code{bg} and \code{cex}
are implemented.}

\item{point_style_short}{Alternatives to \code{point_style_main} used to highlight
certain points when \code{inc = "interval"}. See argument \code{tol}.}

\item{point_style_long}{Alternatives to \code{point_style_main} used to highlight
certain points when \code{inc = "interval"}. See argument \code{tol}.}

\item{text_style}{A named list of arguments to
\code{\link[graphics:text]{text()}}, affecting the appearance
of text printed above highlighted points. See argument \code{tol}.
Currently, only \code{pos}, \code{offset}, \code{col}, \code{cex}, and \code{font}
are implemented.}
}
\value{
The \code{print} method returns \code{x} invisibly.

The \code{coef} method returns \code{object$theta_hat} if \code{log = FALSE}
and \code{object$log_theta_hat} if \code{log = TRUE}.

The \code{predict} method returns a list with numeric elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{refdate}}{Matches \code{object$init$date[1]}.}
\item{\code{cum_inc}}{Expected cumulative incidence at time points
\code{time}, conditional on parameter vector \code{object$theta_hat}.
Equal to \code{object$eval_cum_inc(time)}.
}
\item{\code{int_inc}}{Expected interval incidence given interval
endpoints \code{time}, conditional on parameter vector
\code{object$theta_hat}. Equal to \code{diff(object$eval_cum_inc(time))}
if \code{length(time) >= 2} and omitted otherwise.
}
}

The \code{simulate} method returns an "egf_sim" object,
which is a list with elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{refdate}}{Matches \code{object$init$date[1]}.}
\item{\code{int_inc}}{A matrix with \code{length(time)-1} rows and \code{nsim}
columns, such that \code{int_inc[i, j]} is the number of cases
observed between \code{time[i]} and \code{time[i+1]} in simulation \code{j}
of \code{nsim}. Row vector \code{int_inc[i, ]} is sampled from a
Poisson or negative binomial distribution (depending on
\code{object$init$distr}) with mean \code{predict(object, time)$int_inc[i]}.
The negative binomial dispersion parameter is taken from
\code{object$theta_hat[["nbdisp"]]}.
}
\item{\code{cum_inc}}{A matrix with \code{length(time)} rows and \code{nsim}
columns, such that \code{cum_inc[i, j]} is the number of cases
observed up to \code{time[i]} in simulation \code{j}. Column vector
\code{cum_inc[, j]} is computed as \code{c0 + cumsum(c(0, int_inc[, j]))},
where \code{c0 = predict(object, time)$cum_inc[1]} is the expected
value of cumulative incidence at \code{time[1]} conditional on
parameter vector \code{object$theta_hat}.
}
\item{\code{object}}{Matches argument.}
}

The \code{plot} method returns \code{NULL} invisibly.
}
\description{
Methods for "egf" objects returned by \code{\link[=egf]{egf()}}.
}
\details{
\subsection{Plot elements}{

If \code{xty = "Date"}, then the bottom axis displays the dates specified
by \code{x$init$date}. If \code{xty = "numeric"}, then the bottom axis displays
the number of days since \code{x$init$date[1]}. Regardless of \code{xty},
numeric coordinates are used, hence the left and right boundaries of
the plot region are specified by \code{range(x$init$time)}.

The left axis measures interval or cumulative incidence (depending
on \code{inc}). When incidence is displayed on a logarithmic scale, zeros
are plotted as a positive number less than 1. They are therefore
distinguished from nonzero counts, which are always at least 1.

Observed data, specified by \code{x$init$date} and either \code{x$init$cases}
or \code{cumsum(x$init$cases)} (depending on \code{inc}), are plotted as
points. \code{cases[i]} gives the number of cases observed between
\code{date[i]} and \code{date[i+1]}, and \code{cumsum(cases)[i]} the number
observed between \code{date[1]} and \code{date[i+1]}. Both are plotted at
\code{time[i+1]}.

The fitting window, specified by indices \code{x$init$first} and
\code{x$init$last}, is displayed as a shaded rectangle behind the other
plot elements. The left and right boundaries occur at \code{time[first]}
and \code{time[last+1]}. (\code{cases[first]} is a count from \code{time[first]}
to \code{time[first+1]}, and \code{cases[last]} is a count from \code{time[last]}
to \code{time[last+1]}. Hence the fitting window starts at \code{time[first]}
and ends at \code{time[last+1]}.)

The incidence curve predicted by fitted parameter estimates
\code{x$theta_hat} is displayed as a line supported on grid points
\code{wgrid = seq(time[first], time[last+1], by)}, where \code{by = 1}
for cumulative incidence and \code{by = median(diff(time))} for
interval incidence (ensuring that the interval incidence
curve has the correct scale; see below). The predicted curve
is obtained with \code{predict(x, wgrid)}.

Careful interpretation of interval incidence is required if
the plotted time series is not equally spaced, because \code{cases}
roughly scales with \code{diff(time)}. That is, certain observations
may vary from the predicted curve not due to chance, but because
they represent a count over fewer or more days than the typical
observation interval, namely \code{median(diff(time))}. Observations
for which \code{diff(time)} differs from \code{median(diff(time))} are
highlighted according to argument \code{tol} and labeled with the
value of \code{diff(time)}.

If \code{annotate = TRUE} and \code{add = FALSE}, then a legend and the
initial parameter estimates \code{x$theta_hat} are displayed in the
right margin.
}
}
\seealso{
\code{\link[=egf]{egf()}}, \link[=egf_sim-methods]{methods for class "egf_sim"}
}
