% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{egf.egf_model_simulate}
\alias{egf.egf_model_simulate}
\title{Fit simulated incidence time series}
\usage{
\method{egf}{egf_model_simulate}(object, ...)
}
\arguments{
\item{object}{An \code{"\link[=simulate.egf_model]{egf_model_simulate}"} object
supplying simulated time series and details about the generative
model to be estimated.}

\item{...}{Optional arguments passed to \code{\link{egf.egf_model}},
such as \code{priors} and \code{control}.}
}
\description{
Estimates the generative model underlying a set of simulated
incidence time series.
}
\examples{
example("simulate.egf_model", "epigrowthfit")
fit <- egf(sim)
pp <- data.frame(actual = sim$actual, fitted = fit$best)
pp[fit$nonrandom, , drop = FALSE]

}
