% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{egf.egf_model_simulate}
\alias{egf.egf_model_simulate}
\title{Fit simulated incidence time series}
\usage{
\method{egf}{egf_model_simulate}(model, ...)
}
\arguments{
\item{model}{An \code{"\link[=simulate.egf_model]{egf_model_simulate}"} object
supplying simulated time series and details about the generative model
to be estimated.}

\item{...}{Optional arguments passed to \code{\link{egf.egf_model}}.}
}
\description{
Estimates the generative model underlying a set of simulated
incidence time series.
}
\examples{
example("simulate.egf_model", package = "epigrowthfit", local = TRUE, echo = FALSE)
model <- readRDS(system.file("exdata", "egf_model_simulate.rds",
                             package = "epigrowthfit", mustWork = TRUE))

zz <- egf(model)
pp <- data.frame(actual = model$actual, fitted = zz$best)
pp[zz$nonrandom, , drop = FALSE]

}
