% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf}
\alias{egf}
\title{\loadmathjax
Fit models of epidemic growth}
\usage{
egf(x, method = "BFGS", ...)
}
\arguments{
\item{x}{An "egf_init" object, specifying epidemic data,
a fitting window, and initial parameter estimates.
See \code{\link[=egf_init]{egf_init()}}.}

\item{method}{One of \code{"Nelder-Mead"}, \code{"BFGS"}, \code{"CG"},
\code{"L-BFGS-S"}, and \code{"SANN"}, indicating a method used
to minimize the negative log likelihood function.
Passed directly to \code{\link[stats:optim]{stats::optim()}}.}

\item{...}{Additional arguments to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
An "egf" object.
}
\description{
A description of function \code{egf()}.
}
\details{
\subsection{Phenomenological model}{

Let \mjseqn{x(t)} be the expected number of cases observed
up to time \mjseqn{t}, i.e., expected cumulative incidence.
\mjseqn{x(t)} is modeled by an exponential, logistic, or
Richards curve, plus a linear term representing baseline
mortality, relevant if \mjseqn{x(t)} counts all causes deaths
instead of disease deaths (and otherwise excluded).
\subsection{Exponential}{

\mjsdeqn{x(t) = b t + x_0 e^{r t}}
}

\subsection{Logistic}{

\mjsdeqn{x(t) = b t + \frac{K}{1 + \big(\frac{K}{x(0)} - 1\big) e^{-r t}}}

reparametrized as

\mjsdeqn{x(t) = b t + \frac{K}{1 + e^{-r (t - t_\text{half}}}}

where \mjseqn{t_\text{half}} satisfies
\mjseqn{x(t_\text{half}) - b t_\text{half} = \frac{K}{2}}.
}

\subsection{Richards}{

\mjsdeqn{x(t) = b t + \frac{K}{\big\lbrack 1 + \big(\big(\frac{K}{x(0)}\big)^p - 1\big) e^{-r p t} \big\rbrack^{1/p}}}

reparametrized as

\mjsdeqn{x(t) = b t + \frac{K}{\big\lbrack 1 + (2^p - 1) * e^{-r p (t - t_\text{half}} \big\rbrack}^{1/p}}

where \mjseqn{t_\text{half}} satisfies
\mjseqn{x(t_\text{half}) - b t_\text{half} = \frac{K}{2}}.
}

}

\subsection{Observation model}{

Let \mjseqn{Y(t_1,t_2)} be the number of cases observed between times
\mjseqn{t_1} and \mjseqn{t_2 > t_1}, i.e., observed interval incidence.
\mjseqn{Y(t_1,t_2)} is modeled as either a Poisson-distributed random
variable with mean \mjseqn{x(t_2) - x(t_1)},

\mjsdeqn{Y(t_1,t_2) \sim \mathrm{Poisson}\big(x(t_2) - x(t_1)\big)\,,}

or as negative binomial-distributed random variable with mean
\mjseqn{x(t_2) - x(t_1)} and dispersion \mjseqn{k},

\mjsdeqn{Y(t_1,t_2) \sim \mathrm{NegativeBinomial}\big(x(t_2) - x(t_1), k\big)\,.}
}
}
\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
init <- egf_init(
  date = ontario$date,
  cases = ontario$new_confirmations[-1],
  curve = "richards",
  distr = "nbinom"
x <- egf(init)
plot(x, inc = "interval")
plot(x, inc = "cumulative")

}
