% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf}
\alias{egf}
\title{Fit models of epidemic growth}
\usage{
egf(
  formula,
  fixed = ~1,
  random = NULL,
  data = parent.frame(),
  index = NULL,
  curve = c("logistic", "richards", "exponential", "subexponential", "gompertz"),
  distr = c("nbinom", "pois"),
  excess = FALSE,
  method = c("nlminb", "nlm", "Nelder-Mead", "BFGS"),
  na_action = c("pass", "fail"),
  sparse_X = FALSE,
  date_format = "\%Y-\%m-\%d",
  debug = FALSE,
  par_init = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x} specifying an incidence time series
or multiple incidence time series in long format. \code{x} and \code{y} must
be vectors of dates (see \code{date_format}) and integers, respectively.}

\item{fixed, random}{Named lists of formulae of the form \code{~rhs}, together specifying
a generalized linear mixed effects model (the fixed and random
components) for each parameter of the \emph{incidence} model indicated
by \code{curve}, \code{distr}, and \code{excess}. A list of valid names can be
obtained with \code{get_par_names(curve, distr, excess, link = TRUE)}.
Alternatively, formulae (rather than lists of formulae) to be
recycled for all parameters. Syntax is \code{\link[lme4:lmer]{lme4}}-like,
but here all variables must be factors. \code{fixed} formulae are
restricted to one term and so must have the form \code{~f1:...:fn}
or \code{~1} (the default for each missing list element). \code{random}
formulae must give sums of terms of the form \code{(1 | f1:...:fn)},
\code{(1 | f1/.../fn)}, or \code{(1 | f1 * ... * fn)}. Use \code{NULL} (the
default for each missing list element) instead of a formula to
indicate absence of random effects.}

\item{data}{A data frame, list, or environment containing the variables
named in \code{formula}, \code{fixed}, and \code{random}. Missing values in
incidence (\code{y} if \code{formula = y ~ x}) are not tolerated unless
\code{na_action = "pass"}. Missing values in other variables are
never tolerated.}

\item{index}{A factor of length \code{nrow(data)} such that \code{split(data, index)}
splits \code{data} by fitting window and \code{is.na(index)} indexes
the rows of \code{data} not belonging to a fitting window. \code{NULL}
(the default) is equivalent to \code{rep(factor("0"), nrow(data))}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline
mortality rate is estimated. Set to \code{TRUE} if what is
observed (\code{y} if \code{formula = y ~ x}) is multiple causes
mortality rather than disease mortality or disease incidence.}

\item{method}{A character string specifying an optimizer available through
\code{\link[stats:nlminb]{stats::nlminb()}}, \code{\link[stats:nlm]{stats::nlm()}}, or \code{\link[stats:optim]{stats::optim()}}.}

\item{na_action}{A character string indicating how \code{NA} in the incidence vector
(\code{y} if \code{formula = y ~ x}) are handled. See \code{\link[stats:na.fail]{stats::na.fail()}}.}

\item{sparse_X}{A logical scalar. If \code{TRUE}, then the fixed effects design matrix
will be constructed in sparse format.}

\item{date_format}{A character vector passed to \code{\link[=as.Date]{as.Date()}} argument \code{tryFormats},
used to convert dates (\code{x} if \code{formula = y ~ x}) from character
to Date, if necessary. See \code{\link[base:strptime]{base::strptime()}} for a list of
conversion specifications.}

\item{debug}{A logical scalar used for debugging. If \code{TRUE}, then \code{egf()}
returns early with a list of optimization inputs.}

\item{par_init}{A full parameter vector for the first likelihood evaluation.
Set to \code{NULL} to accept the internally generated default. If
the default causes errors, then use \code{debug = TRUE} to retrieve
its value, which will have the correct structure, and modify as
desired.}

\item{...}{Optional arguments to the optimizer specified by \code{method}.}
}
\value{
If \code{debug = FALSE}, then a list inheriting from class \code{"egf"},
with elements:
\item{\code{frame}}{
The model frame. This is a data frame containing only the
variables named in \code{formula}, \code{fixed}, and \code{random}. Data
not belonging to a fitting window are omitted (but stored
as an attribute). See \code{\link[=make_frame]{make_frame()}}.
}
\item{\code{index}}{
A factor of length \code{nrow(data)} such that \code{split(frame, index)}
splits the model frame by fitting window. Not necessarily
identical to the so-named argument, as \code{frame} is constructed
by permuting and subsetting the rows of \code{data}.
}
\item{\code{curve}, \code{distr}, \code{excess}, \code{method}}{
Copies of the so-named arguments (after matching).
}
\item{\code{tmb_args}}{
A list of arguments to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}. See \code{\link[=make_tmb_args]{make_tmb_args()}}.
}
\item{\code{tmb_out}}{
The list output of \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} (after optimization).
}
\item{\code{optim_out}}{
The list output of the optimizer.
}
\item{\code{par_init}}{
The full parameter vector of the first likelihood evaluation.
Matches the so-named argument if given.
}
\item{\code{par}}{
The full parameter vector of the best likelihood evaluation.
}
\item{\code{nonrandom}}{
An integer vector indexing the nonrandom segment of \code{par}.
}
\item{\code{report}}{
A list containing all variables \code{REPORT()}ed and \code{ADREPORT()}ed
in the C++ template, with an additional element \code{cov} giving the
covariance matrix corresponding to \code{par[nonrandom]}.
}
\item{\code{nll}}{
A numeric scalar giving the value of the negative log
Laplace approximation of the marginal likelihood function
at \code{par[nonrandom]}.
}
\item{\code{nll_func}, \code{nll_grad}}{
Closures taking a numeric vector \code{x} as an argument and returning
the value or gradient of the negative log Laplace approximation of
the marginal likelihood function at \code{x[nonrandom]}.
}
\item{\code{call}}{
The call to \code{egf()}, allowing for updates to the \code{"egf"} object
via \code{\link[stats:update]{stats::update()}}.
}
If \code{debug = TRUE}, then a list with only the elements \code{frame},
\code{tmb_args}, and \code{par_init}.
}
\description{
Fits nonlinear mixed effects models of epidemic growth to one
or more disease incidence time series. First and second order
derivatives of the negative log likelihood (more precisely,
the negative log Laplace approximation of the marginal likelihood)
are calculated by automatic differentiation using R package \strong{TMB}.
}
