% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf}
\alias{egf}
\title{Fit phenomenological models of epidemic growth}
\usage{
egf(
  model = egf_model(),
  formula,
  formula_par = list(),
  data = parent.frame(),
  data_par = parent.frame(),
  subset = NULL,
  subset_par = NULL,
  na_action = c("fail", "pass"),
  na_action_par = c("fail", "omit"),
  endpoints,
  priors = list(),
  control = egf_control(),
  do_fit = TRUE,
  se = FALSE,
  init = NULL,
  map = NULL,
  origin = .Date(0L),
  append = NULL,
  ...
)
}
\arguments{
\item{model}{An \code{"\link{egf_model}"} object specifying a nonlinear model
and a dispersion model to be estimated.}

\item{formula}{A \link{formula} of the form \code{x ~ time} or \code{x ~ time | ts}
specifying one or more incidence time series in long format.
\code{time} must evaluate to a \link{numeric} or \link{Date} vector.
Numeric \code{time} is expected to measure time as a number of days
since \emph{the end of} Date \code{origin}. Date \code{time} is
coerced to numeric \code{\link{julian}(time, origin)}. \code{x}
must evaluate to a numeric vector. Within time series, \code{x[i]}
should specify the number of cases observed from \code{time[i-1]}
to \code{time[i]}. Finally, \code{ts} must evaluate to a \link{factor}
such that \code{\link{split}(\link{data.frame}(time, x), ts)}
returns a \link{list} of time series. Formulae \code{x ~ time}
are equivalent to \code{x ~ time | ts} with \code{ts} set equal
to \code{\link{rep}(\link{factor}(1), \link{length}(x))}.}

\item{formula_par}{A \link{list} of \link{formula}e of the form \code{par ~ terms},
specifying mixed effects models for nonlinear and dispersion model
parameters using \code{\link[lme4:lmer]{lme4}}-like syntax.
Alternatively, a formula of the form \code{~terms} to be recycled
for all parameters. A list of parameters for which formulae may
be specified can be retrieved with \code{\link{get_par_names}}.
Specifically, \code{\link{deparse}(par)} must be an element of
\code{\link{get_par_names}(model, link = TRUE)}. The default for
parameters not assigned a formula is \code{~1}.}

\item{data, data_par}{\link[=data.frame]{Data frame}s, \link{list}s, or \link{environment}s.
These are searched prior to \link{formula} environments for variables
used in \code{formula} and \code{formula_par}, respectively.}

\item{subset, subset_par}{Expressions to be evaluated in \code{data} and \code{data_par},
respectively, provided the latter are data frames. The values,
which must be \link{logical} vectors of suitable length,
index rows of data to be used when fitting (see Details).
The default (\code{\link{NULL}}) is to use all rows.}

\item{na_action}{A \link{character} string affecting the handling of \code{\link{NA}}
in \code{x} for \code{formula = x ~ time | ts}.
\code{"fail"} is to throw an error.
\code{"pass"} is to ignore \code{NA} when fitting and replace \code{NA}
when predicting.
Note that \code{NA} in \code{time} and \code{ts} are an error regardless
of \code{na_action}.}

\item{na_action_par}{A \link{character} string affecting the handling of \code{\link{NA}}
in \code{formula_par} variables.
\code{"fail"} is to throw an error.
\code{"omit"} is to discard fitting windows with incomplete data.}

\item{endpoints}{A \link[=data.frame]{data frame}, \link{list}, or \link{environment}
with variables \code{start} and \code{end}, and any further variables
necessary to evaluate \code{ts} when \code{formula = x ~ time | ts}.
\code{start} and \code{end} must be \link{numeric} or \link{Date}
vectors listing start and end times for all fitting windows. \code{ts}
must evaluate to a \link{factor} indicating the time series in which each
window is found. Within time series, intervals \code{[start[i], end[i]]}
must be disjoint and contain at least two time points from \code{time}.}

\item{priors}{An \link{list} of \link{formula}e of the form \code{par ~ f(...)},
specifying priors on nonlinear and dispersion model parameters.
\code{\link{deparse}(par)} must be an element of
\code{\link{get_par_names}(model, link = TRUE)}.
\code{f(...)} must be a \link{call} to a \link[=egf_prior]{prior function}
with arguments specifying suitable hyperparameters.
This call is evaluated in the corresponding formula environment.
(Currently, the only implemented prior is Gaussian. As a result,
formulae must have the form \code{par ~ \link{Normal}(mu, sigma)}.)}

\item{control}{An \code{"\link{egf_control}"} object specifying control parameters.}

\item{do_fit}{A \link{logical} flag. If \code{TRUE}, then \code{egf} returns early
with a \link{list} of optimization inputs.}

\item{se}{A \link{logical} flag. If \code{TRUE}, then standard errors on mixed
effects model coefficients are computed and stored for later reuse
by methods.}

\item{init}{A \link{numeric} vector to be used as the full parameter vector for
the first likelihood evaluation. The default (\code{\link{NULL}}) is
to accept the internally generated default. Use \code{do_fit = FALSE}
to retrieve this default and other optimization inputs (in particular
\code{Y_init}; see Value), which can often be used to construct a more
informative starting point.}

\item{map}{A \link{factor} of length \code{\link{length}(init)}. Elements
of the full parameter vector corresponding to \code{\link{NA}}
in \code{map} are fixed at their initial values, rather than
estimated. Elements corresponding to a common factor level
are constrained to have a common value during estimation.
\link[=logical]{Logical} values of \code{map} are accepted
and coerced to factor internally with
\code{map <- \link{replace}(\link{gl}(\link{length}(map), 1L), map, NA)}.}

\item{origin}{A \link{Date} specifying a reference time.}

\item{append}{An expression indicating variables in \code{data_par} to be preserved
in the returned object for use by methods. It is evaluated similarly
to the \code{select} argument of \code{\link{subset}}. The default
(\code{\link{NULL}}) is to preserve all variables. Currently, usage
is supported only if \code{data_par} is a \link[=data.frame]{data frame}.}

\item{...}{Unused optional arguments.}
}
\value{
If \code{do_fit = TRUE}, then a \link{list} inheriting from \link{class}
\code{"egf"}, with elements:
\item{endpoints}{
  A \link[=data.frame]{data frame} with variables \code{ts}, \code{window},
  \code{start}, and \code{end} listing start and end times for all fitting
  windows. Rows are ordered by time series and chronologically within time
  series. \code{origin} is retained as an \link[=attributes]{attribute}.
  Note that \code{start} and \code{end} here are the minimum and maximum of
  the set of time points contained in the supplied intervals. Hence supplied
  and returned interval endpoints may not match exactly.
}
\item{frame}{
  The time series model frame, constructed from \code{formula} and
  \code{data}, with variables \code{ts}, \code{window}, \code{start},
  and \code{end}. \code{ts} and \code{window} are \link{factor}s
  that can be used to \link{split} \code{frame} by time series and
  by fitting window, respectively. Rows are ordered by time series
  and chronologically within time series. \code{\link{terms}(formula)}
  and \code{origin} are retained as \link{attributes}.
  \code{\link{as.integer}(window)} indexes rows of \code{endpoints}.
  That is, time series data for the fitting window defined in row
  \code{i} of \code{endpoints} can be found in the rows of \code{frame}
  for which \code{window = \link{levels}(window)[i]}.
}
\item{frame_par}{
  A \link{list} of mixed effects model frames, constructed from
  \code{formula_par} and \code{data_par} using \code{\link{model.frame}}.
  There is one model frame for each nonlinear and dispersion model
  parameter. \code{frame_par[[name]]} retains
  \code{\link{terms}(formula_par[[name]])} as
  an \link[=attributes]{attribute}.
  Model frames correspond rowwise to \code{endpoints}. That is,
  mixed effects data on the fitting window defined in row \code{i}
  of \code{endpoints} can be found in row \code{i} of each model frame.
}
\item{frame_append}{
  A \link[=data.frame]{data frame} preserving the variables from
  \code{data_par} indicated by \code{append}. Corresponds rowwise
  to \code{endpoints}.
}
\item{model}{
  A copy of the so-named argument.
}
\item{priors}{
  A \link{list} of \code{"\link{egf_prior}"} objects,
  obtained by evaluating the right hand side of each
  \link{formula} listed in the so-named argument.
}
\item{control}{
  A copy of the so-named argument.
}
\item{tmb_args}{
  A \link{list} of arguments to \code{\link[TMB]{MakeADFun}}
  constructed by \code{\link{make_tmb_args}}.
}
\item{tmb_out}{
  The \link{list} output of \code{\link[TMB]{MakeADFun}}
  (after optimization).
}
\item{optimizer_out}{
  The \link{list} output of the outer optimizer specified
  by \code{control$optimizer}.
}
\item{init, best}{
  Numeric vectors. These are the full parameter vectors
  of the first and best likelihood evaluations.
}
\item{nonrandom}{
  An \link{integer} vector indexing components \code{beta} and \code{theta}
  of \code{init} and \code{best}.
}
\item{sdreport}{
  If \code{se = TRUE}, then an \code{"sdreport"} object resulting
  from \code{\link{sdreport}(tmb_out)}. Otherwise, \code{\link{NULL}}.
}
\item{nll}{
  A \link[=double]{numeric} scalar giving the value of the negative
  log Laplace approximation of the marginal likelihood function at
  \code{best[nonrandom]}.
}
\item{call}{
  The \code{\link{call}} to \code{egf}, allowing for updates
  to the \code{"egf"} object via \code{\link{update}}.
}
If \code{do_fit = FALSE}, then a \link{list} containing \code{frame},
\code{frame_par}, \code{endpoints}, \code{tmb_args}, \code{tmb_out}
(\emph{before} optimization), \code{init}, \code{nonrandom}, and
\code{call}, as well as a \link{matrix} \code{Y_init} specifying
a naive estimate of each nonlinear and dispersion model parameter
for each fitting window, corresponding rowwise to \code{endpoints}.
}
\description{
Fits nonlinear mixed effects models of epidemic growth to one
or more disease incidence time series.
}
\details{
\code{endpoints} and \link[=model.frame]{model frame}s constructed from
\code{formula_par} and \code{data_par} are expected to correspond rowwise.

Day of week effect estimation (see \code{\link{egf_model}}) requires
\code{time} in \code{formula = x ~ time | ts} to evaluate to
an \link{integer} or \link{Date} vector with 1-day spacing in
all fitting windows.
}
