% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf}
\alias{egf}
\title{\loadmathjax
Fit a model of epidemic growth}
\usage{
egf(init, method = "nlminb", ...)
}
\arguments{
\item{init}{An "egf_init" object specifying epidemic data,
a fitting window, and initial parameter estimates.
See \code{\link[=egf_init]{egf_init()}}.}

\item{method}{One of \code{"nlminb"}, \code{"nlm"}, \code{"Nelder-Mead"},
\code{"BFGS"}, \code{"L-BFGS-S"}, and \code{"CG"},
indicating an optimization algorithm.}

\item{...}{Additional arguments to \code{\link[stats:nlminb]{nlminb()}},
\code{\link[stats:nlm]{nlm()}}, or \code{\link[stats:optim]{optim()}}.}
}
\value{
An "egf" object. A list containing copies of arguments
\code{init} and \code{method}, with these additional elements:

\describe{
\item{\code{theta_hat}}{A named numeric vector giving the optimizer's
approximation of the maximum likelihood parameter vector.
}
\item{\code{log_theta_hat}}{Log-transformed \code{theta_hat}. Identical
to \code{log(theta_hat)} but with \code{"log_"} prepended to the names.
}
\item{\code{nll}}{A numeric scalar giving the negative log likelihood
of \code{log_theta_hat}.
}
\item{\code{nll_func}}{A closure with a numeric argument \code{log_theta}
specifying a log-transformed parameter vector (default is
\code{log_theta_hat}). Returns the negative log likelihood function
evaluated at \code{log_theta}.
}
\item{\code{nll_grad}}{A closure with a numeric argument \code{log_theta}
specifying a log-transformed parameter vector (default is
\code{log_theta_hat}). Returns the gradient of the negative log
likelihood with respect to log-transformed parameters
evaluated at \code{log_theta}.
}
\item{\code{eval_cum_inc}}{A closure with numeric arguments \code{time}
and \code{theta} (default is \code{theta_hat}), evaluating expected
cumulative incidence at \code{time} days using parameter
vector \code{theta}. Elements must be named as in \code{theta_hat}.
}
\item{\code{madf_out}}{The list output of \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}
\item{\code{optim_out}}{The list output of
\code{\link[stats:nlminb]{nlminb()}},
\code{\link[stats:nlm]{nlm()}}, or
\code{\link[stats:optim]{optim()}},
depending on \code{method}.
}
\item{\code{call}}{The call to \code{egf()}, allowing the output to
be updated using \code{\link[stats:update]{update()}}.
}
}
}
\description{
Minimizes the negative log likelihood of an epidemic growth model
using \code{\link[stats:nlminb]{nlminb()}}, \code{\link[stats:nlm]{nlm()}}, or
one of the routines provided through \code{\link[stats:optim]{optim()}}.
The negative log likelihood function is written as a C++ template,
and its gradient with respect to log-transformed parameters is
defined by automatic differentiation using package \pkg{TMB}.
}
\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
init <- egf_init(
  date = ontario$date,
  cases = ontario$new_confirmations[-1],
  curve = "logistic",
  distr = "nbinom"
)
x <- egf(init)
print(x)
coef(x, log = FALSE)
coef(x, log = TRUE)
plot(x, inc = "interval")
plot(x, inc = "cumulative")
time_obs <- init$time
time_pred <- seq(min(time_obs), max(time_obs), by = median(diff(time_obs)))
pred <- predict(x, time = time_pred)
sim <- simulate(x, nsim = 6, time = time_obs)
plot(sim, inc = "interval")
plot(sim, inc = "cumulative")

}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \link[=egf-methods]{methods for class "egf"}
}
