% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf}
\alias{egf}
\title{Fit models of epidemic growth}
\usage{
egf(
  formula_ts,
  formula_par,
  data = parent.frame(),
  window,
  origin = .Date(0),
  curve = c("logistic", "richards", "exponential", "subexponential", "gompertz"),
  excess = FALSE,
  distr = c("nbinom", "pois"),
  weekday = FALSE,
  method = c("nlminb", "nlm", "Nelder-Mead", "BFGS"),
  na_action = c("pass", "fail"),
  sparse_X = FALSE,
  debug = FALSE,
  init = NULL,
  ...
)
}
\arguments{
\item{formula_ts}{A formula of the form \code{x ~ time} or \code{x ~ time | ts} specifying one
or more incidence time series in long format. \code{time} must evaluate
to a numeric or Date vector. Numeric \code{time} is assumed to measure
time as a number of days since \emph{the end of} date \code{origin}. Date
\code{time} is coerced to numeric \code{julian(time, origin)} (see Details).
\code{x} must evaluate to a numeric vector. Within a time series,
\code{x[i]} should specify the number of cases observed from \code{time[i-1]}
to \code{time[i]}. Finally, \code{ts} must evaluate to a factor, such that
\code{split(data.frame(time, x), ts)} returns a list of time series.
Note that \code{x ~ time} is equivalent to \code{x ~ time | ts} with \code{ts}
set equal to \code{rep(factor(1), length(x))}.}

\item{formula_par}{A named list of formulae of the form \code{~terms}, specifying
mixed effects models (\code{\link[lme4:lmer]{lme4}}-like syntax)
for nonlinear model parameters. \code{names(formula_par)} must
be a subset of
\code{get_par_names(curve, distr, excess, weekday, link = TRUE)}.
\code{~1} is the default for parameters not assigned a formula.
Alternatively, \code{formula_par} may itself be a formula of
the form \code{~terms}. In this case, the formula is recycled
for all nonlinear model parameters. Mixed effects variables
must match the length of the time series variables and be
constant in each level of \code{window}.}

\item{data}{A data frame, list, or environment. \code{egf()} looks here
for variables used in \code{formula_ts} and \code{formula_par},
before looking in formula environments.}

\item{window}{A factor of length \code{nrow(data)} such that \code{split(data, window)}
splits \code{data} by fitting window, with \code{is.na(window)} indexing
rows of \code{data} not belonging to a fitting window. Note that
such a factor can be constructed from a table of fitting window
endpoints using \code{\link[=make_window]{make_window()}}.)}

\item{origin}{A Date specifying a reference time (see \code{formula_ts}).}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline mortality
rate is estimated. Set to \code{TRUE} if what is observed is
multiple causes mortality rather than disease mortality
or disease incidence.}

\item{distr}{A character string specifying an observation model.}

\item{weekday}{An integer or logical scalar. If \code{weekday > 0}, then weekday
effects are estimated as offsets relative to the indicated day
(Sunday if \code{weekday = 1}, Monday if \code{weekday = 2}, and so on).
Currently, weekday effect estimation requires time (that is,
the value of \code{foo} in \code{formula_ts = x ~ foo | ts}) to be an
integer (in the sense of \code{all.equal(foo, round(foo))}) or Date
vector with 1-day spacing in all fitting windows.}

\item{method}{A character string specifying an optimizer available through
\code{\link[stats:nlminb]{stats::nlminb()}}, \code{\link[stats:nlm]{stats::nlm()}}, or \code{\link[stats:optim]{stats::optim()}}.}

\item{na_action}{A character string indicating how \code{NA} in incidence (\code{x}
if \code{formula_ts = x ~ time | ts}) are handled. Note that \code{NA}
in other \code{formula_ts} variables are an error, and \code{NA} in
\code{formula_par} variables within fitting windows are an error.}

\item{sparse_X}{A logical scalar. If \code{TRUE}, then the fixed effects design
matrix is constructed in sparse format.}

\item{debug}{A logical scalar used for debugging. If \code{TRUE}, then
\code{egf()} returns early with a list of optimization inputs.}

\item{init}{A full parameter vector for the first likelihood evaluation.
Set to \code{NULL} to accept the internally generated default.
Use \code{debug = TRUE} to retrieve this default, which is useful
as a template.}

\item{...}{Optional arguments to the optimizer specified by \code{method}.}
}
\value{
If \code{debug = FALSE}, then a list inheriting from class \code{"egf"},
with elements:
\item{\code{frame_ts}}{
The time series model frame, constructed from \code{formula_ts}.
\code{origin} is retained as an attribute. See \code{\link[=make_frames]{make_frames()}}.
}
\item{\code{frame_par}}{
A list of mixed effects model frames (one per nonlinear model
parameter), constructed from \code{formula_par} (after completion).
See \code{\link[=make_frames]{make_frames()}}.
}
\item{\code{endpoints}}{
A data frame with variables \code{ts}, \code{window}, \code{start}, and \code{end}
listing fitting window endpoints as numbers of days since time
23:59:59 on a reference date, which is stored as an attribute.
}
\item{\code{curve}, \code{excess}, \code{distr}, \code{weekday}, \code{method}}{
Copies of the so-named arguments (after matching).
}
\item{\code{tmb_args}}{
A list of arguments to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}. See \code{\link[=make_tmb_args]{make_tmb_args()}}.
}
\item{\code{tmb_out}}{
The list output of \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} (after optimization).
}
\item{\code{optim_out}}{
The list output of the optimizer specified by \code{method}.
}
\item{\code{init}}{
The full parameter vector of the first likelihood evaluation.
}
\item{\code{best}}{
The full parameter vector of the best likelihood evaluation.
}
\item{\code{nonrandom}}{
An integer vector indexing the nonrandom segment of \code{best}.
}
\item{\code{report}}{
A list containing all variables \code{REPORT()}ed and \code{ADREPORT()}ed
in the C++ template, with an additional element \code{cov} giving the
covariance matrix corresponding to \code{par[nonrandom]}.
}
\item{\code{nll}}{
A numeric scalar giving the value of the negative log
Laplace approximation of the marginal likelihood function
at \code{best[nonrandom]}.
}
\item{\code{nll_func}, \code{nll_grad}}{
Closures taking a numeric vector \code{x} as an argument and returning
the value or gradient of the negative log Laplace approximation of
the marginal likelihood function at \code{x[nonrandom]}.
}
\item{\code{call}}{
The call to \code{egf()}, allowing for updates to the \code{"egf"} object
via \code{\link[stats:update]{stats::update()}}.
}
If \code{debug = TRUE}, then a list containing only
\code{frame_ts}, \code{frame_par}, \code{init}, \code{tmb_args}, and \code{call}.
}
\description{
Fits nonlinear mixed effects models of epidemic growth to one
or more disease incidence time series.
}
\details{
Coercion of Date \code{time} to numeric \code{julian(time, origin)} assumes
that each element \code{time[i]} can be read as "end of date \code{time[i]}",
so that observation \code{x[i]} in a given time series is the number of
cases observed from the end of date \code{time[i-1]} to the end of date
\code{time[i]}.
}
