\name{egf_has_random}
\alias{egf_has_random}
\title{Check for random effects}
\usage{
egf_has_random(object)
}
\arguments{
\item{object}{An \code{"\link{egf}"} or \code{"egf_no_fit"} object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Determines whether an object specifies a random effects model.
}
