% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_misc.R
\name{egf_has_random}
\alias{egf_has_random}
\alias{egf_has_random.egf}
\title{Check for random effects}
\usage{
egf_has_random(object)

\method{egf_has_random}{egf}(object)
}
\arguments{
\item{object}{An \R object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Determines whether an object specifies a random effects model.
}
