% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_init-methods.R
\name{egf_init-methods}
\alias{egf_init-methods}
\alias{print.egf_init}
\alias{plot.egf_init}
\title{Methods for class "epigrowthfit_init"}
\usage{
\method{print}{egf_init}(x, ...)

\method{plot}{egf_init}(x, ...)
}
\arguments{
\item{x}{An "egf_init" object.}

\item{...}{Unused optional arguments.}
}
\description{
Methods for printing and plotting "egf_init" objects
returned by \code{\link[=egf_init]{egf_init()}}.
}
