% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_init-methods.R
\name{egf_init-methods}
\alias{egf_init-methods}
\alias{print.egf_init}
\alias{coef.egf_init}
\alias{predict.egf_init}
\alias{plot.egf_init}
\title{Methods for class "egf_init"}
\usage{
\method{print}{egf_init}(x, ...)

\method{coef}{egf_init}(object, log = FALSE, ...)

\method{predict}{egf_init}(object, time = object$time, ...)

\method{plot}{egf_init}(x, inc = "cumulative", tol = 0.025, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x, object}{An "egf_init" object.}

\item{...}{Unused optional arguments.}

\item{log}{A logical scalar. If \code{TRUE}, then parameter values are
log-transformed.}

\item{time}{A numeric vector listing increasing time points in days
since \code{object$date[1]}. Missing values are not tolerated.}

\item{inc}{One of \code{"interval"} and \code{"cumulative"},
indicating a type of incidence to plot.}

\item{tol}{A non-negative number used only if \code{inc = "interval"}.
\code{x$cases[i]} is highlighted if
\code{diff(x$time)[i] < (1-tol)*m}, or
\code{diff(x$time)[i] > (1+tol)*m}, where
\code{m = median(diff(x$time))}.
Assign 0 to ensure that all deviations from \code{m} are highlighted.
Assign \code{Inf} to disable highlighting.}
}
\value{
The \code{print} method returns \code{x} invisibly.

The \code{coef} method returns \code{object$theta0} if \code{log = FALSE}
and \code{object$log_theta0} if \code{log = TRUE}.

The \code{predict} method returns a list with numeric elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{cum_inc}}{Expected cumulative incidence at time points
\code{time}, conditional on parameter vector \code{object$theta0}.
Equal to \code{object$cum_inc(time)}.
}
\item{\code{int_inc}}{Expected interval incidence given interval
endpoints \code{time}, conditional on parameter vector
\code{object$theta0}. Equal to \code{diff(object$cum_inc(time))}
if \code{length(time) >= 2} and omitted otherwise.
}
}

The \code{plot} method returns \code{NULL} invisibly.
}
\description{
Methods for "egf_init" objects returned by \code{\link[=egf_init]{egf_init()}}.
}
\details{
\subsection{Plot elements}{

The bottom axis measures the number of days since \code{x$date[1]}.
The left axis measures interval or cumulative incidence
(depending on \code{inc}) on a log scale. Zeros are plotted as if
they were 10^(-0.2), and are therefore distinguished from
nonzero counts, which are always at least 1.

Observed data, specified by \code{x$time} and either \code{x$cases} or
\code{cumsum(x$cases)} (depending on \code{inc}), are plotted as points.
\code{cases[i]} gives the number of cases observed between \code{time[i]}
and \code{time[i+1]}, and \code{cumsum(cases)[i]} the number observed
between \code{time[1]} and \code{time[i+1]}. Both are plotted at \code{time[i+1]}.

The left and right endpoints of the fitting window, specified by
indices \code{x$first} and \code{x$last}, are displayed as vertical lines
at \code{time[first+1]} and \code{time[last+1]} (adding one since \code{first}
and \code{last} index \code{cases}, and \code{length(time) = length(cases)+1}).

The incidence curve predicted by initial parameter estimates
\code{x$theta0} is plotted as a teal line on grid points
\code{wgrid = seq(time[first+1], time[last+1], by = m)},
where \code{m = 1} for cumulative incidence and
where \code{m = median(diff(time))} for interval incidence
(to ensure that the curve has the correct scale; see below).

Careful interpretation of observed interval incidence is required
if the plotted time series is not equally spaced, as \code{cases}
roughly scales with \code{diff(time)}. That is, certain observations
may vary from the predicted curve not due to chance, but because
they represent a count over fewer or more days than the typical
observation interval, namely \code{median(diff(time))}. Observations
for which \code{diff(time)} differs from \code{median(diff(time))} are
highlighted according to argument \code{tol}.
}
}
\seealso{
\code{\link[=egf_init]{egf_init()}}
}
