% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_init.R
\name{egf_init}
\alias{egf_init}
\title{Define a fitting window and initial parameter estimates}
\usage{
egf_init(
  formula = cases ~ date,
  data = data.frame(date, cases),
  date,
  cases,
  curve = "logistic",
  distr = "nbinom",
  include_baseline = FALSE,
  theta_init = NULL,
  peak = NULL,
  last = NULL,
  first = NULL,
  min_first = NULL,
  max_first = NULL,
  dfmt = "\%Y-\%m-\%d"
)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x} used to locate an interval incidence
time series in \code{data}.}

\item{data}{A data frame, list, or environment containing the variables in
\code{formula}. For \code{formula} of the form \code{y ~ x}, variables \code{x} and \code{y}
must conform to the constraints on arguments \code{date} and \code{cases},
respectively.}

\item{date}{A Date vector listing increasing time points, starting at or
before the date of the first observed case in an epidemic wave.
Alternatively, a character vector coercible to such a Date
vector with \code{as.Date(date, tryFormats = dfmt)}. Ignored if
\code{data} is set explicitly and mandatory otherwise.}

\item{cases}{A numeric vector of length \code{length(date)}. For \code{i > 1}, \code{cases[i]}
must specify the number of cases observed between \code{date[i-1]} and
\code{date[i]}. \code{cases[1]} is ignored. Missing values are tolerated.
Ignored if \code{data} is set explicitly and mandatory otherwise.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"}, indicating
a model of expected cumulative incidence.}

\item{distr}{One of \code{"pois"} and \code{"nbinom"}, indicating a model of observed
interval incidence.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then the model of expected cumulative
incidence will include a linear baseline. Assign \code{TRUE} if \code{cases}
counts deaths due to multiple causes and \code{FALSE} otherwise.}

\item{theta_init}{A named numeric vector specifying positive initial estimates of
relevant model parameters:

\describe{
\item{\code{r}}{Initial exponential growth rate expressed per day.}
\item{\code{c0}}{
Expected cumulative incidence on \code{date[first]}, where
"cumulative incidence" refers to the number of cases observed
since \code{date[first]}. Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{
Expected epidemic final size. This is the expected number of
cases observed over the full course of the epidemic wave.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{
Time at which the epidemic wave is expected to attain half its
final size expressed as a number of days since \code{date[first]}.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{
Richards shape parameter.
Used only if \code{curve = "richards"}.
}
\item{\code{b}}{
Baseline linear growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
\item{\code{nbdisp}}{
Negative binomial dispersion parameter.
Used only if \code{distr = "nbinom"}.
}
}

\code{theta_init} can be \code{NULL} or a vector specifying a subset of the
relevant parameters. Unspecified parameters are set internally.}

\item{peak, first, last}{Integers in \code{seq_along(date)} indexing the time of a peak
in \code{cases} (\code{peak}) or endpoints of the fitting window
(\code{first}, \code{last}). Alternatively, Dates between \code{min(date)}
and \code{max(date)} or characters so coercible with
\code{as.Date(x, tryFormats = dfmt)}. In this case, coercion
from Date to index is done by \code{which.min(abs(date - x))}.
If \code{NULL} (default), then an index is chosen internally.}

\item{min_first, max_first}{Bounds on \code{first} used only if \code{first} is chosen internally.
Like \code{first}, can be integer, Date, or character. If \code{NULL}
(default), then the least strict bounds are used internally.}

\item{dfmt}{A character vector listing possible date formats using
the conversion specifications outlined under \code{\link[base:strptime]{base::strptime()}},
e.g., \code{"\%Y-\%m-\%d"} (default).}
}
\value{
An "egf_init" object. A list containing copies of arguments
\code{curve}, \code{distr}, and \code{include_baseline} and these additional
elements:

\describe{
\item{\code{data}}{
A data frame with variables
\code{date},
\code{time = as.integer(date - date[first])}, and
\code{cases}.
\code{date} and \code{cases} are obtained from \code{data} using \code{formula}.
Names in \code{formula} are discarded.
}
\item{\code{window}}{
An integer vector indexing the elements of \code{date} in the fitting
window. Equal to \code{first:last}.
}
\item{\code{first}, \code{last}}{
Integers in \code{seq_along(date)}. The endpoints of the fitting
window are \code{date[first]} and \code{date[last]}.
}
\item{\code{theta_init}}{
A named numeric vector whose elements are the subset of \code{r},
\code{c0}, \code{K}, \code{thalf}, \code{p}, \code{b}, and \code{nbdisp} relevant to \code{curve},
\code{distr}, and \code{include_baseline}. Values from the argument are
retained if they are positive numbers and discarded otherwise.
Parameters not specified in the argument are assigned their
default value:
\describe{
\item{\code{r}, \code{c0}}{
\code{beta1} and \code{exp(beta0)}, respectively,
where \code{beta1} and \code{beta0} are the slope and intercept
of a linear model fit to \code{x = time[first+(1:h)]} and
\code{y = log1p(cumsum(cases[first+(1:h)]))}, where
\code{h = max(2, floor((last-first)/2))}. If either
\code{cases[first+1]} or \code{cases[first+2]} is \code{NA},
then fitting a linear model to \code{x} and \code{y} is
impossible, and \code{r} and \code{c0} default to 0.1 and 1.
}
\item{\code{K}}{\code{2 * sum(cases[(first+1):peak])}}
\item{\code{thalf}}{\code{time[peak]}}
\item{\code{p}}{1}
\item{\code{nbdisp}}{1}
\item{\code{b}}{1}
}
Can be extracted with \code{coef(object, log = FALSE)}.
}
\item{\code{log_theta_init}}{
Log-transformed \code{theta_init}. Identical to \code{log(theta_init)},
but with \code{"log_"} prepended to the names. Can be extracted
with \code{coef(object, log = TRUE)}.
}
\item{\code{eval_cum_inc}}{
A closure with numeric arguments \code{time} and \code{theta}
(default is \code{theta_init}) evaluating expected cumulative
incidence at \code{time} days since \code{date[first]} conditional
on parameter vector \code{theta}. Elements of \code{theta} must be
named as in \code{theta_init}. \code{predict(object, time)} should
be used instead of \code{object$eval_cum_inc(time)}.
}
\item{\code{call}}{
The call to \code{egf_init()}, allowing the output to be updated
using \code{\link[stats:update]{stats::update()}}.
}
}
}
\description{
Defines a fitting window and initial estimates of model parameters
given an interval incidence time series. Used to initialize \code{\link[=egf]{egf()}}.
}
\details{
\subsection{1. Models}{

A full description of the models specified by arguments
\code{curve}, \code{distr}, and \code{include_baseline} can be found in
the package vignette, accessible with
\code{vignette("epigrowthfit-vignette")}.

The number \code{npar} of model parameters is given by
\code{switch(curve, exponential = 2, logistic = 3, richards = 4)},
plus one if \code{distr = "nbinom"},
plus one if \code{include_baseline = TRUE}.
}

\subsection{2. Fitting window selection}{

The "fitting window" is the time interval with endpoints
\code{date[first]} and \code{date[last]}. The cases observed during
this interval are counted in \code{cases[(first+1):last]}.
Only these \code{last-first} elements of \code{cases} are used when
fitting model parameters to the data, hence \code{last-first}
is constrained to be at least \code{npar}.

If \code{cases} spans multiple epidemic waves, then the fitting
window must not include data from more than one wave. The
fitting window should start when \code{cases} (restricted to the
focal wave), begins growing roughly exponentially (linearly
on a logarithmic scale). When it should end depends on the
model being fit to the data. If \code{curve = "exponential"},
then the window should end at or before the time when when
\code{cases} (restricted to the focal wave) stops growing
exponentially. If \code{curve \%in\% c("logistic", "richards")},
then it should end between that time and the time of the
peak in \code{cases} (during the focal wave).
}

\subsection{3. Default behaviour}{

\code{egf_init()} tries to make fitting window selection with
\code{curve \%in\% c("logistic", "richards")} as simple as possible
by reducing the need for user input. However, behaviour with
\code{peak}, \code{last}, \code{first}, \code{min_first}, and \code{max_first} all
set to \code{NULL} (see below) relies on several assumptions.
It can be expected to define a reasonable fitting window if
(i) \code{cases} gives data for exactly one epidemic wave,
(ii) \code{cases} is close to equally spaced, and
(iii) \code{cases} is close to smooth.
If any of these conditions fail to hold, then consider
setting \code{peak}, \code{last}, and \code{first} explicitly.
\code{\link[=smooth_cases]{smooth_cases()}} simplifies this task by fitting a cubic spline
to the data and providing the times of peaks in the fitted curve
as indices of \code{date}.

If \code{peak = NULL}, then \code{peak} is set to \code{which.max(cases)}
internally.

If \code{last = NULL}, then \code{last} is set to \code{peak} internally,
unless \code{peak < npar+1},
in which case \code{last} is set to \code{npar+which.max(cases[-(1:npar)])}
in order to enforce \code{last >= npar+1}.

If \code{first = NULL}, then \code{first} is set to
\code{max_first-which.min(cases[max_first:min_first])+1} internally.
This is the greatest index \code{i} in \code{min_first:max_first} satisfying
\code{cases[i] = min(cases[min_first:max_first])}. This default assumes
that \code{min_first} points to the peak of the wave before the focal
wave or to the start of the epidemic if the focal wave is the first
wave. In this case, the chosen index \code{first} will point to the
"base" of the focal wave, a crude approximation of the time when
\code{cases} begins to grow exponentially.

If \code{min_first = NULL}, then \code{min_first} is set to 1 internally.
If \code{max_first = NULL}, then \code{max_first} is set to \code{last-npar}
internally. These are the least strict bounds on \code{first} in that
they allow the fitting window to contain as few as \code{npar} and as
many as \code{last} elements of \code{cases}. The default \code{min_first = 1}
should be accepted only if the focal wave is the first wave and
otherwise set explicitly (see above paragraph).
}

\subsection{4. Missing data}{

Missing values in \code{cases} are tolerated insofar that they do not
prevent fitting of a model to the data by \code{egf(object)} unless
\code{sum(!is.na(cases[(first+1):last])) < npar}. However, they do
prevent calculation of cumulative incidence since \code{date[1]} and
so cause \code{plot(object, inc = "cumulative")} to throw an error.
}
}
\examples{
data(canadacovid)
ontario <- subset(canadacovid, province == "ON")
sc <- smooth_cases(
  date = ontario$date,
  cases = ontario$new_confirmed,
  log = TRUE,
  spar = 0.7
)
plot(sc)
v <- c("2020-03-01", "2020-03-28", "2020-09-01", "2020-09-26")
dline(v, lty = 2, col = "#CCCCCC")
x1 <- egf_init(new_confirmed ~ date,
  data = ontario,
  curve = "logistic",
  distr = "nbinom",
  first = 27,
  last = 77,
  peak = 77
)
x2 <- update(x1, first = 211, last = 236, peak = "2020-10-10")
print(x1)
coef(x1, log = FALSE)
coef(x1, log = TRUE)
predict(x1)
plot(x1, inc = "interval")
plot(x2, inc = "interval", add = TRUE)
plot(x1, inc = "cumulative")
plot(x2, inc = "cumulative", add = TRUE)

}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
\seealso{
\code{\link[=egf]{egf()}}, \code{\link[=smooth_cases]{smooth_cases()}}, \code{\link[=plot.egf_init]{plot.egf_init()}}
}
