% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_init.R
\name{egf_init}
\alias{egf_init}
\title{\loadmathjax
Define a fitting window and initial parameter estimates}
\usage{
egf_init(
  date,
  cases,
  curve = "logistic",
  distr = "nbinom",
  include_baseline = FALSE,
  theta_init = NULL,
  peak = min_wlen - 1 + which.max(cases[min_wlen:length(cases)]),
  last = min(length(cases), peak + 1),
  first = NULL,
  wlen = NULL,
  min_wlen = 6,
  max_wlen = last
)
}
\arguments{
\item{date}{A Date vector listing increasing time points.
Should start at or before the date of the first observed case
in an epidemic.}

\item{cases}{A numeric vector with length \code{length(date)-1}.
\code{cases[i]} is the number of cases observed between \code{date[i]}
and \code{date[i+1]}. Here, "cases" can mean infections,
reported infections, or reported deaths (either disease deaths or
deaths due to multiple causes including the disease of interest).
Missing values are not tolerated.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a model of expected cumulative incidence. See Details 1.}

\item{distr}{One of \code{"pois"} and \code{"nbinom"}, indicating a
model of observed interval incidence. See Details 1.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then
the model of expected cumulative incidence will include
a linear baseline. Assign \code{TRUE} if \code{cases} counts deaths
due to multiple causes and \code{FALSE} otherwise. See Details 1.}

\item{theta_init}{A named numeric vector specifying positive
initial estimates of relevant model parameters:

\describe{
\item{\code{r}}{\mjseqn{\lbrace\,r\,\rbrace}
Initial exponential growth rate expressed per day.
}
\item{\code{c0}}{\mjseqn{\lbrace\,c_0\,\rbrace}
Expected cumulative incidence on \code{date[first]}.
Here, "cumulative incidence" refers to the number of cases
observed since the start of the epidemic wave (i.e., since
\code{date[first]}), which is not necessarily the number of
cases observed since the start of the epidemic (i.e., since
\code{date[1]}). Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{\mjseqn{\lbrace\,K\,\rbrace}
Expected epidemic final size. This is the expected number
of cases observed over the full course of the epidemic wave.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{\mjseqn{\lbrace\,t_\text{half}\,\rbrace}
Time at which the epidemic wave is expected to attain
half its final size, expressed as a number of days since
\code{date[first]}.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{\mjseqn{\lbrace\,p\,\rbrace}
Richards shape parameter. Used only if \code{curve = "richards"}.
}
\item{\code{b}}{\mjseqn{\lbrace\,b\,\rbrace}
Baseline linear growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
\item{\code{nbdisp}}{\mjseqn{\lbrace\,k\,\rbrace}
Negative binomial dispersion parameter.
Used only if \code{distr = "nbinom"}.
}
}

\code{theta_init} can be \code{NULL} or a vector specifying a subset
of the relevant parameters. Unspecified parameters are
set internally. See Value.}

\item{peak}{An integer in \code{seq_along(cases)} indexing a
peak in \code{cases}. If \code{cases} describes just one epidemic
wave, then the default value is typically acceptable.
If \code{cases} spans multiple waves, then \code{peak} should be
set explicitly, as its default value may not locate the
peak in the focal wave.}

\item{last}{An integer in \code{seq_along(cases)}. The last
observation in the fitting window will be \code{cases[last]}.
See Details 2.}

\item{first}{An integer in \code{seq_along(cases)}, or \code{NULL}.
If non-\code{NULL}, then the first observation in the fitting
window will be \code{cases[first]}. If \code{NULL}, then \code{first}
is chosen internally according to \code{wlen}, and otherwise
according to \code{min_wlen} and \code{max_wlen}. See Details 2.}

\item{wlen}{An integer indicating the number of observations
in the fitting window. Must be at least the number of
parameters in the model being fit. See Details 2.}

\item{min_wlen, max_wlen}{Integers indicating the minimum
and maximum number of observations in the fitting window.
Along with \code{last}, these determine the range of integers
considered for \code{first} if \code{first} and \code{wlen} are \code{NULL}.
If \code{cases} describes just one epidemic wave, then the
default values are typically acceptable. If \code{cases} spans
multiple epidemic waves, then \code{max_wlen} should be set
explicitly. See Details 2.}
}
\value{
An "egf_init" object. A list containing copies of arguments
\code{date}, \code{cases}, \code{last}, \code{curve}, \code{distr}, and \code{include_baseline},
with these additional elements:

\describe{
\item{\code{time}}{Time as a number of days since \code{date[1]}.
Equal to \code{as.numeric(date - date[1])}.
}
\item{\code{first}}{A copy of the argument if set explicitly.
Otherwise, the result of an internal selection algorithm.
See Details 2.
}
\item{\code{theta_init}}{A named numeric vector whose elements are
the subset of \code{r}, \code{c0}, \code{K}, \code{thalf}, \code{p}, \code{b}, and
\code{nbdisp} relevant to \code{curve}, \code{distr}, and \code{include_baseline}.
Values from the argument are retained if they are positive
numbers and discarded otherwise. Parameters not specified
in the argument are assigned their default value:
\describe{
\item{\code{r}, \code{c0}}{\code{beta1} and \code{exp(beta0)}, respectively,
where \code{beta1} and \code{beta0} are the slope and intercept
of a linear model fit to the first half of
\code{log(cumsum(cases[first:last]) + 0.1)}, where
the corresponding time points are taken to be
\code{time[(first:last)+1] - time[first]}.
}
\item{\code{K}}{\code{2 * sum(cases[first:peak])}}
\item{\code{thalf}}{\code{time[peak+1] - time[first]}}
\item{\code{p}}{1}
\item{\code{nbdisp}}{1}
\item{\code{b}}{1}
}
Can be extracted with \code{coef(object, log = FALSE)}.
}
\item{\code{log_theta_init}}{Log-transformed \code{theta_init}. Identical to
\code{log(theta_init)}, but with \code{"log_"} prepended to the names.
Can be extracted with \code{coef(object, log = TRUE)}.
}
\item{\code{eval_cum_inc}}{A closure with numeric arguments
\code{time} and \code{theta} (default is \code{theta_init}) evaluating
expected cumulative incidence at \code{time} days using
parameter vector \code{theta}. Elements of \code{theta} must be
named as in \code{theta_init}. \code{predict(object, time)} wraps
\code{eval_cum_inc(time, theta = theta_init)} and provides
additional useful information.
}
\item{\code{call}}{The call to \code{egf_init()}, allowing the output
to be updated using \code{\link[stats:update]{update()}}.
}
}
}
\description{
Given an interval incidence time series, constructs an
"egf_init" object specifying a fitting window and initial
estimates of model parameters. Attempts are made to define
a window and estimates that are reasonable in the absence
of any user input, but the default behaviour is not robust,
and it is likely that some optional arguments must be set
explicitly.
}
\details{
\subsection{1. Models}{

A full description of the models of expected cumulative
incidence and observed interval incidence specified by
arguments \code{curve}, \code{distr}, and \code{include_baseline}, can
be found in the package vignette, accessible with
\code{vignette("epigrowthfit-vignette")}.
}

\subsection{2. Fitting window selection}{

The "fitting window" is the subset of \code{cases} used
when fitting model parameters to the data, starting
at index \code{first} and ending at index \code{last}. If
\code{cases} contains data from multiple epidemic waves,
then this subset should not include data from more
than one wave. Regardless of the model being fit,
the fitting window should start at the point in the
focal wave at which \code{cases} begins growing roughly
exponentially. This is precisely the point at which
\code{log(cases)} becomes roughly linear. Where the
fitting window should end depends on the cumulative
incidence curve being fit to the data.
If \code{curve = "exponential"}, then the window should
end at the point in the focal wave at which \code{cases}
stops growing exponentially. This is the point at
which \code{log(cases)} \emph{stops} being linear.
If \code{curve = "logistic"} or \code{curve = "richards"}, then
the fitting window should end near the peak in \code{cases}
during the focal wave. If the wave is incomplete and
the peak has not yet occurred (e.g., when fitting in
real time, during a growing epidemic), then the window
should simply end at \code{cases[length(cases)]}.

The default behaviour of \code{egf_init()} tries to make
usage with \code{curve = "logistic"} and \code{curve = "richards"}
(which should be preferred in practice over
\code{curve = "exponential"}) as simple as possible
by reducing the need for user input. Specifically, if
\code{last} is not set explicitly, then it is set by default
to \code{min(length(cases), peak+1)}, ensuring that the
fitting window ends near the peak in \code{cases} during
the focal wave. If \code{first} is not set explicitly but
\code{wlen} is, then \code{first} is set to \code{max(1, last-wlen+1)}.
It neither \code{first} nor \code{wlen} is set explicitly, then
\code{egf_init()} attempts to choose \code{first} such that
\code{cases[first]} occurs at the "base" of the focal wave,
an approximation of the point at which \code{cases} begins
growing exponentially. This selection is made as follows:

First, constraints on \code{first} are determined according
to \code{last}, \code{min_wlen} and \code{max_wlen}. The length
\code{wlen = last-first+1} of the fitting window is
constrained to be at least \code{min_wlen} at most \code{max_wlen}.
This means that \code{first} must satisfy
\code{first >= last-max_wlen+1} and \code{first <= last-min_wlen+1}.
The default values of \code{min_wlen} and \code{max_wlen} make the
range of valid integers as large as possible. This is
typically acceptable when \code{cases} describes just one
epidemic wave. However, when \code{cases} spans multiple waves,
it is necessary to set \code{max_wlen} so that the minimum
allowed value of \code{first} indexes a point before exponential
growth begins in the focal wave, but after any previous waves.

Second, \code{m = min(cases[(last-max_wlen+1):(last-min_wlen+1)])}
is computed. Provided \code{max_wlen} is set appropriately, this
value can be described as the minimum of \code{cases} in the trough
between the focal wave and the previous wave (or the start of
the time series, if the focal wave is the first or only wave).

Finally, \code{first} is set to the greatest integer \code{i} in
\code{(last-max_wlen+1):(last-min_wlen+1)} such that \code{cases[i] = m}.

If \code{m = 0}, which typically arises when the focal wave is the
first wave in the epidemic, then the selected index \code{first}
is such that \code{cases[first] = 0}. In this situation, \code{egf_init()}
enforces \code{cases[first] > 0} by assigning \code{first <- first + 1},
provided that this assignment does not cause \code{first} to exceed
the maximum allowed value, namely \code{last-min_wlen+1}.
}
}
\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
x <- egf_init(
  date = ontario$date,
  cases = ontario$new_confirmations[-1],
  curve = "logistic",
  distr = "nbinom"
)
print(x)
coef(x, log = FALSE)
coef(x, log = TRUE)
time_obs <- x$time
time_pred <- seq(min(time_obs), max(time_obs), by = median(diff(time_obs)))
pred <- predict(x, time = time_pred)
plot(x, inc = "interval")
plot(x, inc = "cumulative")

}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
\seealso{
\code{\link[=egf]{egf()}}, \link[=egf_init-methods]{methods for class "egf_init"}
}
