% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_init.R
\name{egf_init}
\alias{egf_init}
\title{\loadmathjax
Define a fitting window and initial parameter estimates}
\usage{
egf_init(
  date,
  cases,
  curve = "logistic",
  distr = "nbinom",
  include_baseline = FALSE,
  theta0 = NULL,
  min_wlen = 6,
  peak = min_wlen - 1 + which.max(cases[min_wlen:length(cases)]),
  first = NULL,
  first_level = NULL,
  skip_zero = TRUE
)
}
\arguments{
\item{date}{A Date vector listing increasing time points.
Should start at or before the date of the first observed case.}

\item{cases}{A numeric vector with length \code{length(date)-1}.
\code{cases[i]} is the number of cases observed between \code{date[i]}
and \code{date[i+1]}. Here, "cases" can mean infections, reported
infections, or reported deaths (either disease deaths or deaths
due to multiple causes including the disease of interest).
Missing values are not tolerated.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a model of expected cumulative incidence. See Details 2.}

\item{distr}{One of \code{"pois"} and \code{"nbinom"}, indicating a
model of observed interval incidence. See Details 2.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then
the model of expected cumulative incidence will include
a linear baseline \mjseqn{b t}. Assign \code{TRUE} if \code{cases}
counts deaths due to multiple causes. See Details 2.}

\item{theta0}{A named numeric vector specifying positive
initial estimates of relevant model parameters:

\describe{
\item{\code{r}}{\mjseqn{\lbrace\,r\,\rbrace}
Initial exponential growth rate expressed per day.
}
\item{\code{c0}}{\mjseqn{\lbrace\,c_0\,\rbrace}
Expected cumulative incidence on \code{date[1]}. This is
the expected number of cases observed up to \code{date[1]}.
Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{\mjseqn{\lbrace\,K\,\rbrace}
Expected epidemic final size. This is the expected number
of cases observed over the full course of the epidemic.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{\mjseqn{\lbrace\,t_\text{half}\,\rbrace}
Time at which the epidemic is expected to attain half its
final size, expressed as a (possibly non-integer) number
of days since \code{date[1]}.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{\mjseqn{\lbrace\,p\,\rbrace}
Richards shape parameter. Used only if \code{curve = "richards"}.
}
\item{\code{b}}{\mjseqn{\lbrace\,b\,\rbrace}
Baseline linear growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
\item{\code{nbdisp}}{\mjseqn{\lbrace\,k\,\rbrace}
Negative binomial dispersion parameter.
Used only if \code{distr = "nbinom"}.
}
}

\code{theta0} can be \code{NULL} or a vector specifying a subset
of the relevant parameters. Unspecified parameters are
set internally. See Value.}

\item{min_wlen}{An integer scalar. The minimum number
of observations in the fitting window. Must be at
least the number of parameters being fit.}

\item{peak}{An integer in \code{seq_along(cases)} indexing
the peak in \code{cases}. The index of the last observation
in the fitting window will be \code{peak} or \code{peak+1}.
The default is like \code{which.max(cases)} but careful
to prevent \code{peak < min_wlen}. See Details 1.}

\item{first}{An integer in \code{seq_along(cases)} indexing
the first observation in the fitting window, or \code{NULL}.
Non-\code{NULL} values may not be respected, depending on
other arguments. See Details 1.}

\item{first_level}{A numeric scalar or \code{NULL}. Can be
used to define \code{first} if \code{first = NULL}. See Details 1.}

\item{skip_zero}{A logical scalar. If \code{TRUE}, then an
attempt is made when defining \code{first} to ensure that
\code{cases[first] > 0}. See Details 1.}
}
\value{
An "egf_init" object. A list with elements:

\describe{
\item{\code{date}}{Matches argument.}
\item{\code{time}}{Time as a number of days since \code{date[1]}.
Equal to \code{as.numeric(date - date[1])}.
}
\item{\code{cases}}{Matches argument.}
\item{\code{first}}{An integer indexing the start of the fitting window,
so that the first observation is \code{cases[first]}. May not match
the argument of the same name. See Details 1.
}
\item{\code{last}}{An integer indexing the end of the fitting window,
so that the last observation is \code{cases[last]}. Equal to \code{peak}
if \code{curve = "exponential"} and \code{max(length(cases), peak + 1)}
otherwise.
}
\item{\code{curve}}{Matches argument.}
\item{\code{distr}}{Matches argument.}
\item{\code{include_baseline}}{Matches argument.}
\item{\code{theta0}}{A named numeric vector whose elements
are the subset of \code{r}, \code{c0}, \code{K}, \code{thalf}, \code{p}, \code{b},
and \code{nbdisp} relevant to \code{curve}, \code{include_baseline},
and \code{distr}. Values from the argument of the same
name are retained if they are positive and discarded
otherwise. Parameters not specified in the argument
are assigned their default value:
\describe{
\item{\code{r}, \code{c0}}{\code{beta1} and \code{exp(beta0)}, respectively,
where \code{beta1} and \code{beta0} are the slope and intercept
of a linear model fit to \code{log(cumsum(cases) + 0.1)}
within the first half of the fitting window.
}
\item{\code{K}}{\code{sum(cases)}}
\item{\code{thalf}}{\code{time[peak+1]}}
\item{\code{p}}{1}
\item{\code{nbdisp}}{1}
\item{\code{b}}{1}
}
}
\item{\code{log_theta0}}{Log-transformed \code{theta0}. Identical to
\code{log(theta0)}, but with \code{"log_"} prepended to the names.
}
\item{\code{cum_inc}}{A closure with numeric arguments \code{time}
and \code{theta} (default is \code{theta0}), evaluating expected
cumulative incidence at \code{time} days using parameter
vector \code{theta}. Elements must be named as in \code{theta0}.
}
\item{\code{call}}{The call to \code{egf_init()}, making the output
reproducible with \code{eval(call)}.
}
}
}
\description{
Given an interval incidence time series, constructs an
"egf_init" object specifying a fitting window and initial
estimates of model parameters. Attempts are made to define
a window and estimates that are reasonable in the absence
of any user input.
}
\details{
\subsection{1. Fitting window selection}{

The "fitting window" is the subset of \code{cases} used
when fitting model parameters to the data. It ends
at index \code{last}, where
\code{last = peak} if \code{curve = "exponential"} and
\code{last = max(length(cases), peak + 1)} otherwise.

The length \code{wlen = last-first+1} of the fitting window
is constrained to be at least \code{min_wlen}. This implies
two constraints on the arguments: \code{peak >= min_wlen} if
\code{curve = "exponential"} and \code{peak >= min_wlen-1} otherwise,
and \code{first <= peak-min_wlen+1}. An error will be thrown
if arguments \code{peak} and \code{first} (unless \code{first = NULL})
do not conform to these constraints.

If \code{first = NULL} and \code{first_level = NULL}, then \code{first}
is set internally to the greatest index \code{i} in \code{1:(peak-min_wlen+1)}
for which \code{cases[i] = min(cases[1:(peak-min_wlen+1)])}.
If \code{first = NULL} and \code{first_level != NULL}, then \code{first}
is set internally to one plus the greatest index \code{i} in
\code{1:(peak-min_wlen)} for which \code{cases[i] < first_level * max(cases)}.
If no such index exists, then \code{first} is set to 1.
Finally, if \code{first != NULL}, then the supplied value is
respected (unless \code{skip_zeros = TRUE}) and \code{first_level}
is ignored.

If \code{skip_zeros = TRUE} and \code{first} (as defined above) is
such that \code{cases[first] = 0} and \code{first < peak-min_wlen+1},
then \code{first} is increased by one until \code{cases[first] > 0}
or \code{first = peak-min_wlen+1}.
}

\subsection{2. Models}{

A full description of the models of expected cumulative
incidence and observed interval incidence, specified by
\code{curve}, \code{distr}, and \code{include_baseline}, can be found
in the package vignette, accessible with
\code{vignette("epigrowthfit-vignette")}.
}
}
\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
x <- egf_init(
  date = ontario$date,
  cases = ontario$new_confirmations[-1],
  curve = "richards",
  distr = "nbinom"
)
print(x)
coef(x, log = FALSE)
coef(x, log = TRUE)
time_obs <- x$time
time_pred <- seq(min(time_obs), max(time_obs), by = median(diff(time_obs)))
predict(x, time = time_pred)
plot(x, inc = "cumulative")
plot(x, inc = "interval")

}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
\seealso{
\code{\link[=egf]{egf()}}, \link[=egf_init-methods]{methods for class "egf_init"}
}
