% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_init.R
\name{egf_init}
\alias{egf_init}
\title{\loadmathjax
Define fitting windows and initial parameter estimates}
\usage{
egf_init(
  date,
  cases,
  curve = "logistic",
  include_baseline = FALSE,
  distr = "nbinom",
  theta0 = NULL,
  min_wlen = 6,
  peak = min_wlen - 1 + which.max(cases[min_wlen:length(cases)]),
  first = NULL,
  first_level = NULL,
  skip_zero = TRUE
)
}
\arguments{
\item{date}{A Date vector listing increasing time points.}

\item{cases}{A numeric vector with length \code{length(date)-1}.
\code{cases[i]} is the number of cases observed between
\code{date[i]} and \code{date[i+1]}. Here, "cases" can mean
(reported) infections or (reported) deaths. Missing
values are not tolerated.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a phenomenological model for cumulative incidence.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then the
cumulative incidence model will include a linear term \mjseqn{b t}.
Assign \code{TRUE} only if \code{cases} counts deaths due to all causes and
you want to model growth above a baseline mortality rate \mjseqn{b}.}

\item{distr}{One of \code{"pois"} and \code{"nbinom"}, indicating an
observation model for interval incidence.}

\item{theta0}{A named numeric vector specifying positive
initial estimates of model parameters:

\describe{
\item{\code{r}}{\mjseqn{\lbrace\,r\,\rbrace}
Initial (exponential) growth rate expressed per day.
}
\item{\code{x0}}{\mjseqn{\lbrace\,x_0\,\rbrace}
Expected cumulative incidence on \code{date[first]} (see \code{first} below).
This is the expectation of the number of cases observed
up to \code{date[first]}. Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{\mjseqn{\lbrace\,K\,\rbrace}
Expected epidemic final size. This is the expectation of the
number of cases observed over the full course of the epidemic.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{\mjseqn{\lbrace\,t_\text{half}\,\rbrace}
Expected time at which the epidemic attains half its
final size, expressed as a (possibly non-integer) number
of days since \code{date[1]}.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{\mjseqn{\lbrace\,p\,\rbrace}
Richards shape parameter. Used only if \code{curve = "richards"}.
}
\item{\code{b}}{\mjseqn{\lbrace\,b\,\rbrace}
Baseline (linear) growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
\item{\code{nbdisp}}{\mjseqn{\lbrace\,k\,\rbrace}
Negative binomial dispersion parameter.
Used only if \code{distr = "nbinom"}.
}
}

\code{theta0} can be \code{NULL} or a vector specifying a subset
of the relevant parameters. Absent parameters are set
internally. See Value.}

\item{min_wlen}{An integer scalar. The minimum number
of observations in the fitting window. Must be at
least the number of parameters being fit.}

\item{peak}{An integer in \code{seq_along(cases)} indexing
the peak in \code{cases}. The index of the last observation
in the fitting window will be \code{peak} or \code{peak+1}.
The default is like \code{which.max(cases)} but careful
to prevent \code{peak < min_wlen}. See Details 4.}

\item{first}{An integer in \code{seq_along(cases)} indexing
the first observation in the fitting window, or \code{NULL}.
Non-\code{NULL} values may not be respected, depending on
other arguments. See Details 4.}

\item{first_level}{A numeric scalar or \code{NULL}. Can be
used to define \code{first} if \code{first = NULL}. See Details 4.}

\item{skip_zero}{A logical scalar. If \code{TRUE}, then an
attempt is made when defining \code{first} to ensure that
\code{cases[first] > 0}. See Details 4.}
}
\value{
An "egf_init" object. A list with elements:

\describe{
\item{\code{date}}{Matches argument.}
\item{\code{time}}{Time as a number of days since \code{date[1]}.
Equal to \code{as.numeric(date - date[1])}.
}
\item{\code{cases}}{Matches argument.}
\item{\code{first}}{An integer indexing the start of the fitting window,
so that the first observation is \code{cases[first]}. May not match
the argument of the same name. See Details 4.
}
\item{\code{last}}{An integer indexing the end of the fitting window,
so that the last observation is \code{cases[last]}. Equal to \code{peak}
if \code{curve = "exponential"} and \code{max(length(cases), peak + 1)}
otherwise.
}
\item{\code{curve}}{Matches argument.}
\item{\code{include_baseline}}{Matches argument.}
\item{\code{distr}}{Matches argument.}
\item{\code{theta0}}{A named numeric vector whose elements
are the subset of \code{r}, \code{x0}, \code{K}, \code{thalf}, \code{p}, \code{b},
and \code{nbdisp} relevant to \code{curve}, \code{include_baseline},
and \code{distr}. Values from the argument of the same
name are retained if they are positive and discarded
otherwise. Parameters not specified in the argument
are assigned their default value below:\preformatted{\\begin\{describe\}\{
  \\item\{`r`, `x0`\}\{`beta1` and `exp(beta0)`, respectively,
    where `beta1` and `beta0` are the slope and intercept
    of a linear fit to `log(cumsum(cases) + 0.1)` within
    the first half of the fitting window. Here, "intercept"
    means the value of the fit at `time[first]`.
  \}
  \\item\{`K`\}\{`sum(cases)`\}
  \\item\{`thalf`\}\{`time[peak+1]`\}
  \\item\{`p`\}\{1\}
  \\item\{`b`\}\{1\}
  \\item\{`nbdisp`\}\{1\}
\}
}

}
\item{\code{log_theta0}}{A named numeric vector, identical to
\code{log(theta0)} with \code{"log_"} prepended to the names.
}
\item{\code{call}}{The call to \code{egf_init()}, making the output
reproducible with \code{eval(call)}.
}
}
}
\description{
Given an incidence or mortality time series, constructs an
"egf_init" object specifying a fitting window and initial
estimates of model parameters. Pains are taken to define a
window and estimates that are reasonable in the absence of
any user input. For model details, see \code{\link[=egf]{egf()}}.
}
\details{
\subsection{1. Phenomenological models}{

Let \mjseqn{x(t)} be the expected number of cases observed
up to time \mjseqn{t} (i.e., expected cumulative incidence),
and let \mjseqn{x_0 = x(0) > 0}. Ignoring any baseline growth
(see Details 2), \mjseqn{x(t)} is modeled as an exponential,
logistic, or Richards curve.
\subsection{Exponential model}{

If \mjseqn{x(t)} follows

\mjsdeqn{x'(t) = r x(t)\,,\qquad r > 0\,,}

then \mjseqn{x(t)} grows exponentially as

\mjsdeqn{x(t) = x_0 e^{r t}\,.}

Hence the exponential model for \mjseqn{x(t)} requires
fitting two parameters to observed data: the exponential
growth rate \mjseqn{r} and initial value \mjseqn{x_0}.

The exponential model ignores depletion of susceptible
individuals and implies continuous exponential growth
in \mjseqn{x(t)}. Hence it will only agree with epidemic
data during the (typically short) initial exponential
growth phase. \insertCite{Ma+14;textual}{epigrowthfit}
show that estimates of \mjseqn{r} are sensitive to the
choice of fitting window, and that more robust fits to
the data are likely to obtained with the logistic and
Richards models, at negligible cost.
}

\subsection{Logistic model}{

If \mjseqn{x(t)} follows

\mjsdeqn{x'(t) = r x(t) \bigg(1 - \frac{x(t)}{K}\bigg)\,,\qquad r, K > 0\,,}

and if \mjseqn{x_0 \in (0,K)}, then \mjseqn{x(t)} grows as

\mjsdeqn{x(t) = \frac{K}{1 + \big(\frac{K}{x_0} - 1\big) e^{-r t}}}

and increases to \mjseqn{K} as \mjseqn{t \to \infty}.
The logistic model can be reparametrized as

\mjsdeqn{x(t) = \frac{K}{1 + e^{-r (t - t_\text{half})}}\,,}

where \mjseqn{t_\text{half}} is the time at which
cumulative incidence attains half its final size,
satisfying \mjseqn{x(t_\text{half}) = \frac{K}{2}}.

The reparametrized logistic model requires fitting
\mjseqn{r}, \mjseqn{K}, and \mjseqn{t_\text{half}}
to observed data.
}

\subsection{Richards}{

If \mjsdeqn{x(t)} follows

\mjsdeqn{x'(t) = r x(t) \bigg(1 - \bigg(\frac{x(t)}{K}\bigg)^p\bigg)\,,\qquad r, K, p > 0\,,}

and if \mjseqn{x_0 \in (0,K)}, then \mjseqn{x(t)} grows as

\mjsdeqn{x(t) = \frac{K}{\big\lbrack 1 + \big(\big(\frac{K}{x_0}\big)^p - 1\big) e^{-r p t} \big\rbrack^{1/p}}}

and increases to \mjseqn{K} as \mjseqn{t \to \infty}.
The Richards model can be reparametrized as

\mjsdeqn{x(t) = \frac{K}{\big\lbrack 1 + (2^p - 1) * e^{-r p (t - t_\text{half}}) \big\rbrack^{1/p}}\,,}

where, as with the logistic model, \mjseqn{t_\text{half}}
satisfies \mjseqn{x(t_\text{half}) = \frac{K}{2}}.

The reparametrized logistic model requires fitting
\mjseqn{r}, \mjseqn{K}, \mjseqn{t_\text{half}}, and \mjseqn{p}
to observed data.
}

}

\subsection{2. Baseline growth}{

For many historical epidemics, the observed data are counts
deaths due to all causes, not only the disease of interest.
Growth in disease mortality over time can still be understood
phenomenologically, provided that baseline mortality (deaths
unrelated to the epidemic) and disease mortality are modeled
separately. To account for baseline mortality, it is assumed
that deaths occur at a constant rate \mjseqn{b > 0} in the
absence of an epidemic. Then, for example, the logistic model
(see Details 1) becomes

\mjsdeqn{x(t) = b t + \frac{K}{1 + \big(\frac{K}{x_0} - 1\big) e^{-r t}}\,,}

where \mjseqn{x(t)} is to be interpreted as expected
cumulative mortality instead of expected cumulative incidence.
Accounting for baseline mortality requires that \mjseqn{b} is
fit in addition to the other model parameters.
}

\subsection{3. Observation model}{

Let \mjseqn{Y(t_1,t_2)} be the number of cases observed between times
\mjseqn{t_1} and \mjseqn{t_2 > t_1} (i.e., observed interval incidence).
\mjseqn{Y(t_1,t_2)} is modeled as either a Poisson-distributed random
variable with mean \mjseqn{x(t_2) - x(t_1)},

\mjsdeqn{Y(t_1,t_2) \sim \mathrm{Poisson}\big(x(t_2) - x(t_1)\big)\,,}

or as negative binomial-distributed random variable with mean
\mjseqn{x(t_2) - x(t_1)} and dispersion \mjseqn{k > 0},

\mjsdeqn{Y(t_1,t_2) \sim \mathrm{NegativeBinomial}\big(x(t_2) - x(t_1), k\big)\,.}

The negative binomial observation model requires that \mjseqn{k} is fit
in addition to the other model parameters.
}

\subsection{4. Fitting window selection}{

The "fitting window" is the subset of \code{cases} used
when fitting model parameters to the data. It ends
at index \code{last}, where
\code{last = peak} if \code{curve = "exponential"} and
\code{last = max(length(cases), peak + 1)} otherwise.

The length \code{wlen = last-first+1} of the fitting window
is constrained to be at least \code{min_wlen}. This implies
two constraints on the arguments: \code{peak >= min_wlen} if
\code{curve = "exponential"} and \code{peak >= min_wlen-1} otherwise,
and \code{first <= peak-min_wlen+1}. An error will be thrown
if arguments \code{peak} and \code{first} (unless \code{first = NULL})
do not conform to these constraints.

If \code{first = NULL} and \code{first_level = NULL}, then \code{first}
is set internally to the greatest index \code{i} in \code{1:(peak-min_wlen+1)}
for which \code{cases[i] = min(cases[1:(peak-min_wlen+1)])}.
If \code{first = NULL} and \code{first_level != NULL}, then \code{first}
is set internally to one plus the greatest index \code{i} in
\code{1:(peak-min_wlen)} for which \code{cases[i] < first_level * max(cases)}.
If no such index exists, then \code{first} is set to 1.
Finally, if \code{first != NULL}, then the supplied value is
respected (unless \code{skip_zeros = TRUE}) and \code{first_level}
is ignored.

If \code{skip_zeros = TRUE} and \code{first} (as defined above) is
such that \code{cases[first] = 0} and \code{first < peak-min_wlen+1},
then \code{first} is increased by one until \code{cases[first] > 0}
or \code{first = peak-min_wlen+1}.
}
}
\examples{
data(canadacovid)
ontario <- na.omit(subset(canadacovid, province == "ON"))
x <- egf_init(
  date = ontario$date,
  cases = ontario$new_confirmations[-1],
  curve = "richards",
  distr = "nbinom"
)
print(x)
plot(x)
plot(x, inc = "cumulative")

}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
\seealso{
\link[=egf_init-methods]{methods for class "egf_init"}, \code{\link[=egf]{egf()}}
}
