% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf_model}
\alias{egf_model}
\title{Define an epidemic model}
\usage{
egf_model(
  curve = c("logistic", "richards", "exponential", "subexponential", "gompertz"),
  excess = FALSE,
  family = c("nbinom", "pois"),
  day_of_week = FALSE
)
}
\arguments{
\item{curve}{A \link{character} string specifying a model for expected cumulative
incidence as a function of time.}

\item{excess}{A \link{logical} scalar. If \code{TRUE}, then a constant baseline
mortality rate is estimated. Set to \code{TRUE} if what is observed
is multiple causes mortality rather than disease mortality or disease
incidence.}

\item{family}{A \link{character} string specifying a model for observed interval
incidence (i.e., an error distribution). Note that a log link is used
regardless of \code{family}.}

\item{day_of_week}{An integer. If positive, then day of week effects are estimated as
offsets relative to the indicated day of week
(Sunday if \code{day_of_week = 1}, Monday if \code{day_of_week = 2},
and so on). \link[=logical]{Logical} values are coerced to integer.}
}
\value{
A \link{list} inheriting from \link{class} \code{"egf_model"}
containing the arguments (after possible matching and coercion).
}
\description{
Sets flags defining the phenomenological model of epidemic growth
to be estimated by \code{\link{egf}}.
}
