% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf.R
\name{egf_parallel}
\alias{egf_parallel}
\title{Define a parallelization method}
\usage{
egf_parallel(
  method = c("serial", "multicore", "snow"),
  outfile = "",
  cores = getOption("egf.cores", 1L),
  options = list(),
  cl = NULL
)
}
\arguments{
\item{method}{A \link{character} string indicating a method of parallelization.
\code{"\link[=lapply]{serial}"} indicates no parallelization.
\code{"\link[parallel:mclapply]{multicore}"} indicates forking.
It is intended for use from a terminal rather than a GUI.
On Windows, it is equivalent to \code{"serial"}.
\code{"\link[parallel:parLapply]{snow}"} indicates socket clusters.
It supports parallelization on both Unix-alikes and Windows.}

\item{outfile}{A \link{character} string indicating a file path where
console output should be diverted. \code{\link{NULL}} is
equivalent to \code{\link{nullfile}()}. \code{""} means
no diversion. If \code{method = "snow"}, then diversion
may be necessary to view output.}

\item{cores}{A positive integer indicating a number of worker processes to spawn
when \code{parallel != "serial"}. The effective maximum is typically
\code{\link[parallel]{detectCores}(TRUE, FALSE)}.}

\item{options}{A \link{list} of optional arguments to
\code{\link[parallel]{mclapply}} (\code{method = "multicore"}) or
\code{\link[parallel]{makePSOCKcluster}} (\code{method = "snow"}).}

\item{cl}{An optional \link[parallel:makePSOCKcluster]{socket cluster}
to be used when \code{parallel = "snow"}. If non-\code{\link{NULL}},
then \code{outfile}, \code{cores}, and \code{options} are ignored.}
}
\value{
A \link{list} inheriting from \link{class} \code{"egf_parallel"}
containing the arguments (after possible matching and substitution).
}
\description{
Defines instructions for parallelization by linking a method with options.
}
