% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_sim-methods.R
\name{egf_sim-methods}
\alias{egf_sim-methods}
\alias{plot.egf_sim}
\title{Methods for class "egf_sim"}
\usage{
\method{plot}{egf_sim}(x, inc = "cumulative", alpha = 0.5, ...)
}
\arguments{
\item{x}{An "egf_sim" object.}

\item{inc}{One of \code{"cumulative"} and \code{"interval"},
indicating a type of incidence to plot.}

\item{alpha}{A number in [0,1] indicating an alpha level.
Determines the transparency of plotted simulations
(0=invisible, 1=opaque). More simulations tend to require
more transparency.}

\item{...}{Optional arguments. Used only by the \code{plot} method
to specify graphical parameters. See \code{\link[base:plot]{plot()}}
and \code{\link[graphics:par]{par()}}. Currently, only \code{xlim} and
\code{ylim} are implemented. Any additional parameters will be
ignored.}
}
\value{
The \code{plot} method returns \code{NULL} invisibly.
}
\description{
Methods for "egf_sim" objects returned by \code{\link[=simulate.egf]{simulate.egf()}}.
}
\details{
\subsection{Plot elements}{

The bottom axis measures the number of days since
\code{x$object$init$date[1]}. The left axis measures
interval or cumulative incidence (depending on \code{inc}).

Simulations are obtained as the columns of \code{x$cum_inc}
or \code{x$int_inc} (depending on \code{inc}). They are plotted
in grey with transparency at the time points given by
\code{x$time} or \code{x$time[-1]}, respectively.

The predicted curve is obtained as \code{pred$cum_inc}
or \code{pred$int_inc} (depending on \code{inc}),
where \code{pred = predict(x$object, time = x$time)}.
It is plotted in teal without transparency.
}
}
\seealso{
\code{\link[=simulate.egf]{simulate.egf()}}
}
