% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_sim-methods.R
\name{egf_sim-methods}
\alias{egf_sim-methods}
\alias{plot.egf_sim}
\title{Methods for class "egf_sim"}
\usage{
\method{plot}{egf_sim}(x, inc = "cumulative", col_pred = "#44AA99", col_sim = "#BBBBBB66", ...)
}
\arguments{
\item{x}{An "egf_sim" object.}

\item{inc}{One of \code{"cumulative"} and \code{"interval"},
indicating a type of incidence to plot.}

\item{col_sim, col_pred}{Character or numeric scalars specifying
colours for simulated incidence time series and predicted incidence
curves, respectively.}

\item{...}{Optional arguments. Used only by the \code{plot} method
to specify graphical parameters. Currently, only \code{xlim},
\code{ylim}, \code{xlab}, \code{ylab}, and \code{main} are implemented. Further
arguments are ignored.
See \code{\link[graphics:plot.default]{plot()}} and \code{\link[graphics:par]{par()}}
for a catalogue of graphical parameters.}
}
\value{
The \code{plot} method returns \code{NULL} invisibly.
}
\description{
Methods for "egf_sim" objects returned by \code{\link[=simulate.egf]{simulate.egf()}}.
}
\details{
\subsection{Plot elements}{

The bottom axis measures the number of days since
\code{x$object$init$date[1]}. The left axis measures interval
or cumulative incidence (depending on \code{inc}). The predicted
incidence curve is obtained as \code{pred$cum_inc} or \code{pred$int_inc},
where \code{pred = predict(x$object, time = x$time)}. The simulated
incidence time series are obtained as the columns of \code{x$cum_inc}
or \code{x$int_inc}. These are plotted together behind the predicted
incidence curve. To help visualize the distribution of simulated
incidence at a given time, assign \code{col_sim} a sufficiently
transparent colour.
}
}
\seealso{
\code{\link[=simulate.egf]{simulate.egf()}}
}
