% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enum_dupl_str}
\alias{enum_dupl_str}
\alias{enum_dupl_names}
\title{Enumerate duplicated strings}
\usage{
enum_dupl_str(x)

enum_dupl_names(x)
}
\arguments{
\item{x}{For \code{enum_dupl_str()}, a character vector.
For \code{enum_dupl_names()}, an \R object with a \code{names} attribute.}
}
\value{
\code{enum_dupl_str(x)} returns an enumerated copy of \code{x}.
\code{enum_dupl_names(x)} returns a copy of \code{x} with \code{names(x)}
replaced by \code{enum_dupl_str(names(x))}.
}
\description{
Replaces the \code{i}th instance of strings \code{s} in a character vector
with \code{sprintf("\%s[\%d]", s, i)}.
}
\examples{
# x <- sample(letters[1:3], 10L, replace = TRUE)
# enum_dupl_str(x)
#
# y <- seq_along(x)
# names(y) <- x
# enum_dupl_names(y)

}
\keyword{internal}
