\name{epigrowthfit-deprecated}
\alias{epigrowthfit-deprecated}
\title{Deprecated Functions in Package \pkg{epigrowthfit}}
\description{
The functions and other objects listed here are provided only
for compatibility with older versions of \pkg{epigrowthfit} and
may become defunct as soon as the next release.
}
\usage{
## Nothing yet!
}
\seealso{
\code{\link{Defunct}}, \code{\link{base-defunct}},
\code{\link{epigrowthfit-defunct}}, \code{\link{epigrowthfit-notyet}}.
}
