\name{epigrowthfit-package}
\alias{epigrowthfit-package}
\docType{package}
\title{\R{} Package \pkg{epigrowthfit}}
\description{
An \R{} package for estimating nonlinear mixed effects models of
epidemic growth.
}
\details{
The \dQuote{main} model estimating function is \code{\link{egf}}.

To render a list of available help topics, use
\code{\link{help}(package = "epigrowthfit")}.
Many of these document methods for the class of objects returned
by \code{\link{egf}}.

To report a bug or request a change, use
\code{\link{bug.report}(package = "epigrowthfit")}.
}
\author{
Mikael Jagan
\email{jaganmn@mcmaster.ca}
}
