% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-package.R
\docType{package}
\name{epigrowthfit-package}
\alias{epigrowthfit-package}
\title{epigrowthfit: Nonlinear Mixed Effects Models of Epidemic Growth}
\description{
A description of the \pkg{epigrowthfit} package.
}
\details{
Details about the \pkg{epigrowthfit} package.
}
\references{
\insertRef{Ma+14}{epigrowthfit}

\insertRef{Earn+20}{epigrowthfit}
}
\keyword{internal}
