% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{eval_model}
\alias{eval_model}
\title{\loadmathjax
Evaluate expected cumulative incidence}
\usage{
eval_model(time, curve, include_baseline = FALSE, theta)
}
\arguments{
\item{time}{A numeric vector listing time points in days.}

\item{curve}{One of \code{"exponential"}, \code{"logistic"}, and \code{"richards"},
indicating a model of expected cumulative incidence.}

\item{include_baseline}{A logical scalar. If \code{TRUE}, then the model of expected
cumulative incidence will include a linear baseline.}

\item{theta}{A named numeric vector listing positive values for all
relevant model parameters:

\describe{
\item{\code{r}}{Initial exponential growth rate, expressed per day.}
\item{\code{c0}}{
Expected cumulative incidence at 0 days.
Used only if \code{curve = "exponential"}.
}
\item{\code{K}}{
Expected epidemic final size.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{thalf}}{
Time in days at which the epidemic is expected
to attain half its final size.
Used only if \code{curve \%in\% c("logistic", "richards")}.
}
\item{\code{p}}{Richards shape parameter.
Used only if \code{curve = "richards"}.
}
\item{\code{b}}{Baseline linear growth rate expressed per day.
Used only if \code{include_baseline = TRUE}.
}
}}
}
\value{
A numeric vector with length \code{length(time)} whose \code{i}th
element is expected cumulative incidence at \code{time[i]}.
}
\description{
Evaluates expected cumulative incidence at desired time points
conditional on a vector of model parameters.
}
\details{
A full description of the models specified by arguments
\code{curve} and \code{include_baseline} can be found in
the package vignette, accessible with
\code{vignette("epigrowthfit-vignette")}.
}
