% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.egf}
\alias{fitted.egf}
\alias{coef.egf}
\title{Extract fitted values}
\usage{
\method{fitted}{egf}(
  object,
  subset = NULL,
  select = NULL,
  link = TRUE,
  append_frame = TRUE,
  ...
)

\method{coef}{egf}(
  object,
  subset = NULL,
  select = NULL,
  link = TRUE,
  append_frame = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{subset}{A list of logical expressions to be evaluated in the combined
model frame \code{do.call(cbind, object$frame_glmm)}. Fitted values
are returned only for the indexed fitting windows. Use the
default (\code{NULL}) to retrieve fitted values for all fitting
windows.}

\item{select}{A subset of \code{get_par_names(object, link = TRUE)} naming
nonlinear model parameters for which fitted values should
be returned. Use the default (\code{NULL}) to retrieve fitted
values for all nonlinear model parameters.}

\item{link}{A logical scalar. If \code{FALSE}, then fitted values are inverse
link-transformed.}

\item{append_frame}{A logical scalar. If \code{TRUE}, then observed values of covariates
are reported alongside fitted values.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame inheriting from class \code{"egf_fitted"}, with variables:
\item{\code{.par}}{
Nonlinear model parameter, from \code{get_par_names(object, link)}.
}
\item{\code{.window}}{
Fitting window, from \code{levels(object$frame_ts$.window)}.
}
\item{\code{.value}}{
Fitted value of parameter \code{.par} in fitting window \code{.window}.
}

If \code{append_frame = TRUE}, then observed values of covariates
from the combined model frame \code{do.call(cbind, object$frame_glmm)}
are also reported. (If a variable occurs more than once there,
because it appears in multiple model frames, then only the
earliest instance is retained. Except in very unusual cases,
all instances of the variable are identical, so no information
is lost.)
}
\description{
Extracts the fitted values of nonlinear model parameters.
The fitted value for a given fitting window is obtained
by adding
(i) the population fitted value computed from the relevant
fixed effects coefficients and
(ii) all random effects terms (if any), with random effects
coefficients set equal to their conditional modes.
}
\details{
\code{coef.egf()} is currently an alias for \code{fitted.egf()}.
}
