% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.egf}
\alias{fitted.egf}
\alias{coef.egf}
\title{Extract fitted values}
\usage{
\method{fitted}{egf}(object, link = TRUE, ...)

\method{coef}{egf}(object, link = TRUE, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{link}{A logical scalar. If \code{FALSE}, then inverse link-transformed
fitted values are returned.}

\item{...}{Atomic vectors, optionally specifying fitting windows
for which fitted values are desired. See Details.}
}
\value{
A data frame inheriting from class \code{"egf_fitted"}, with one row
per fitting window. The first \code{length(object$frame)-2} variables
specify levels for each factor in \code{object$frame[-(1:2)]}
(and thus fitting windows). The remaining variables list fitted
values for each element of \code{get_par_names(object, link)}.
}
\description{
Extracts the fitted values of incidence model parameters.
The fitted value for a given fitting window is obtained
by adding
(i) the population fitted value computed from the relevant
fixed effect coefficients and
(ii) the random effect conditional modes relevant to the
window.
}
\details{
Elements of \code{list(...)} must be named and have the form
\code{factor_name = c(level_names)}, where \code{factor_name} must
be the name of a factor in \code{object$frame} and \code{level_names}
must be a subset of \code{levels(object$frame[[factor_name]])}.
}
