% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.egf}
\alias{fitted.egf}
\alias{coef.egf}
\title{Extract fitted values}
\usage{
\method{fitted}{egf}(object, subset = NULL, link = TRUE, ...)

\method{coef}{egf}(object, subset = NULL, link = TRUE, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{subset}{A named list of atomic vectors with elements specifying levels
of factors in \code{object$frame} (and thus fitting windows). Use
the default (\code{NULL}) to retrieve fitted values for every fitting
window or if \code{object$frame} has no factors.}

\item{link}{A logical scalar. If \code{FALSE}, then fitted values
are inverse link-transformed.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame inheriting from class \code{"egf_fitted"}, with one
row per fitting window. The first \code{length(object$frame)-2}
variables specify levels for each factor in \code{object$frame}
(and thus fitting windows). The remaining variables list
fitted values for each element of \code{get_par_names(object, link)}.
}
\description{
Extracts the fitted values of incidence model parameters.
The fitted value for a given fitting window is obtained
by adding
(i) the population fitted value computed from the relevant
fixed effect coefficients and
(ii) the random effect conditional modes relevant to the
window (if any).
}
\details{
Elements of \code{subset} (if non-\code{NULL}) must have the form
\code{factor_name = level_names}, where \code{factor_name} is the
name of a factor in \code{object$frame} and \code{level_names} is
a subset of \code{levels(object$frame$factor_name)}.

Currently, \code{coef.egf()} is an alias for \code{fitted.egf()}.
}
