% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.egf}
\alias{fitted.egf}
\alias{coef.egf}
\title{Extract fitted values}
\usage{
\method{fitted}{egf}(
  object,
  top = egf_get_names_top(object, link = TRUE),
  subset = NULL,
  append = NULL,
  link = TRUE,
  se = FALSE,
  .subset = NULL,
  .append = NULL,
  ...
)

\method{coef}{egf}(
  object,
  top = egf_get_names_top(object, link = TRUE),
  subset = NULL,
  append = NULL,
  link = TRUE,
  se = FALSE,
  .subset = NULL,
  .append = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{top}{A subset of \code{\link{egf_get_names_top}(object, link = TRUE)}
naming top level nonlinear model parameters for which fitted
values should be retrieved.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{egf_combine_frames}}). It must evaluate
to a valid index vector for the rows of the data frame
(see \code{\link{[.data.frame}}), and thus fitting windows.
Fitted values are retrieved only for indexed windows.
The default (\code{\link{NULL}}) is to consider all windows.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link{egf_combine_frames}}) to be included with the
result. The default (\code{\link{NULL}}) is to append nothing.}

\item{link}{A \link{logical} flag. If \code{FALSE}, then fitted values are
inverse link-transformed.}

\item{se}{A \link{logical} flag. If \code{link = TRUE} and \code{se = TRUE},
then approximate delta method standard errors on fitted values
are reported. Standard errors are required for subsequent use of
\code{\link{confint.egf_fitted}}.}

\item{.subset, .append}{Index vectors to be used (if non-\code{\link{NULL}}) in place of
the result of evaluating \code{subset} and \code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_fitted"}, with variables:
\item{top}{
  Top level nonlinear model parameter,
  from \code{\link{egf_get_names_top}(object, link = link)}.
}
\item{ts}{
  Time series, from \code{\link{levels}(object$frame_windows$ts)}.
}
\item{window}{
  Fitting window, from \code{\link{levels}(object$frame_windows$window)}.
}
\item{estimate}{
  Fitted value of parameter \code{top} in fitting window \code{window}.
}
\item{se}{
  (If \code{link = TRUE} and \code{se = TRUE}.)
  Approximate delta method standard error on \code{estimate}.
}
}
\description{
Computes fitted values of top level nonlinear model parameters.
The fitted value of a given parameter for a given fitting window
is obtained by adding
(i) the population fitted value computed as a linear combination
of fixed effects coefficients and
(ii) all applicable random effects, with random effects coefficients
set equal to their conditional modes.
}
\details{
\code{coef.egf} is currently an alias for \code{fitted.egf}.

See topic \code{\link{egf_eval}} for details on nonstandard evaluation
of \code{subset} and \code{append}.
}
\examples{
example("egf", "epigrowthfit")
zz <- fitted(object, se = TRUE)
str(zz)

}
\seealso{
\code{\link{confint.egf_fitted}}

Other coefficient extractors: 
\code{\link{fixef.egf}()},
\code{\link{ranef.egf}()}
}
\concept{coefficient extractors}
