% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef.egf}
\alias{fixef.egf}
\alias{fixef}
\title{Extract fixed effects coefficients}
\usage{
\method{fixef}{egf}(object, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame inheriting from class \code{"egf_fixef"},
with one row per coefficient and variables:
\item{\code{name}}{
Coefficient name in the full parameter vector \code{object$par}.
}
\item{\code{par}}{
Nonlinear model parameter, from
\code{get_par_names(object, link = TRUE)}.
}
\item{\code{term}}{
Term from fixed effects component of mixed effects formula.
}
\item{\code{colname}}{
Corresponding column name in the fixed effects design matrix
\code{object$tmb_args$data$X}.
}
\item{\code{estimate}}{
Coefficient estimate.
}
}
\description{
Retrieve the coefficients of the fixed effects component
of a mixed effects model.
}
