% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef.egf}
\alias{fixef.egf}
\alias{fixef}
\title{Extract fixed effect coefficients}
\usage{
\method{fixef}{egf}(object, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_fixef"}, with one row per coefficient and variables:
\item{bottom}{
  Name of a bottom level mixed effects model parameter
  (a fixed effect coefficient).
}
\item{top}{
  Name of the top level nonlinear model parameter whose fitted value
  is a function of \code{bottom},
  from \code{\link{egf_get_names_top}(object, link = TRUE)}.
}
\item{term}{
  Term from fixed effects component of mixed effects model \link{formula}
  for parameter \code{top}.
}
\item{colname}{
  Column name in the fixed effects design matrix
  \code{object$tmb_out$env$data$X}.
}
\item{estimate}{
  Coefficient estimate.
}
}
\description{
Retrieve the coefficients of the fixed effects component
of a mixed effects model.
}
\examples{
example("egf", package = "epigrowthfit", local = TRUE, echo = FALSE)
object <- readRDS(system.file("exdata", "egf.rds", package = "epigrowthfit", mustWork = TRUE))

zz <- fixef(object)
str(zz)

}
\seealso{
Other coefficient extractors: 
\code{\link{fitted.egf}()},
\code{\link{ranef.egf}()}
}
\concept{coefficient extractors}
