% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_link_string}
\alias{get_link_string}
\alias{add_link_string}
\alias{remove_link_string}
\title{Manipulate nonlinear model parameter names}
\usage{
add_link_string(s)

remove_link_string(s)

get_link_string(s)
}
\arguments{
\item{s}{A character vector.}
}
\value{
\code{add_link_string(s)} replaces valid \code{s[i]} with
\code{paste(get_link_string(s[i]), s[i], sep = "_")}.

\code{remove_link_string(s)} replaces valid \code{s[i]} with
\code{sub("^(log|logit)_.*$", "\\\\1", s[i])}.

\code{get_link_string(s)} replaces valid \code{s[i]} with
\code{"log"} or \code{"logit"} according to the link function
used internally for \code{s[i]}.

In all cases, invalid \code{s[i]} are replaced with
\code{NA_character_}.
}
\description{
Prepend, strip, and retrieve link prefixes
from nonlinear model parameter names.
}
\examples{
# get_link_string("r")
# add_link_string("r")
# remove_link_string("log_r")

}
\keyword{internal}
