% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_par_names}
\alias{get_par_names}
\alias{add_link_string}
\alias{remove_link_string}
\title{Get parameter names}
\usage{
get_par_names(curve = NULL, distr = NULL, excess = NULL, link = TRUE)

add_link_string(s)

remove_link_string(s)
}
\arguments{
\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline mortality
rate will be estimated. Set to \code{TRUE} if what is observed
(\code{y} if \code{formula = y ~ x}) is multiple causes mortality
rather than disease mortality or incidence.}

\item{link}{A logical scalar. If \code{TRUE}, then a prefix indicating a link
function (either \code{"log_"} or \code{"logit_"}) is prepended to each
parameter name.}

\item{s}{A character vector listing parameter names with or without
prefixes.}
}
\value{
\code{get_par_names(link = FALSE)} returns the subset of
\code{c("r", "alpha", "c0", "tinfl", "K", "p", "a", "b", "nbdisp")}
relevant to \code{curve}, \code{distr}, and \code{excess}.
}
\description{
Returns the names used internally for parameters of the specified
incidence model.
}
\details{
\code{add_link_string()} adds prefixes to parameter names.
\code{remove_link_string()} is its inverse.
}
\examples{
a <- list(curve = "exponential", distr = "pois", excess = FALSE)
pn <- do.call(get_par_names, c(link = FALSE, a))
link_pn <- do.call(get_par_names, c(link = TRUE, a))
identical(link_pn, add_link_string(pn))
identical(pn, remove_link_string(link_pn))

}
\keyword{internal}
