% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{get_par_names}
\alias{get_par_names}
\alias{get_par_names.default}
\alias{get_par_names.egf_model}
\alias{get_par_names.egf}
\alias{get_par_names.tmb_data}
\title{Get model parameter names}
\usage{
get_par_names(object, ...)

\method{get_par_names}{default}(object, link = TRUE, ...)

\method{get_par_names}{egf_model}(object, link = TRUE, ...)

\method{get_par_names}{egf}(object, link = TRUE, ...)

\method{get_par_names}{tmb_data}(object, link = TRUE, ...)
}
\arguments{
\item{object}{An \code{"\link{egf_model}"}, \code{"\link{egf}"},
or \code{"\link[=make_tmb_data]{tmb_data}"} object
specifying a model. Otherwise, \code{\link{NULL}}.}

\item{...}{Unused optional arguments.}

\item{link}{A \link{logical} flag. If \code{TRUE}, then \code{"name"}
is replaced with \code{"log(name)"} or \code{"logit(name)"}
depending on the link function used internally for the
parameter.}
}
\value{
A \link{character} vector giving a subset of \code{"r"}, \code{"alpha"},
\code{"c0"}, \code{"tinfl"}, \code{"K"}, \code{"p"}, \code{"a"},
\code{"b"}, \code{"nbdisp"}, and \code{\link{paste0}("w", 1:6)}
(modified if \code{link = TRUE}).

If \code{object = \link{NULL}}, then the complete set is returned.
}
\description{
Retrieves the names used internally for nonlinear and dispersion
model parameters.
}
\examples{
get_par_names(NULL, link = FALSE)
get_par_names(NULL, link = TRUE)

model <- egf_model(
  curve = "logistic",
  excess = FALSE,
  family = "nbinom",
  day_of_week = 0L
)
get_par_names(model, link = FALSE)

}
