% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_par_names}
\alias{get_par_names}
\title{Get parameter names}
\usage{
get_par_names(curve = NULL, distr = NULL, excess = NULL, link = TRUE)
}
\arguments{
\item{curve, distr, excess}{Character strings specifying an incidence model. See \code{\link[=egf]{egf()}}.
Alternatively, \code{curve} can be an \code{"egf"} object returned by
\code{\link[=egf]{egf()}} or a \code{"tmb_data"} object returned by \code{\link[=make_tmb_data]{make_tmb_data()}},
in which case \code{distr} and \code{excess} are ignored.}

\item{link}{A logical scalar. If \code{TRUE}, then a prefix indicating the
link function used internally (either \code{"log_"} or \code{"logit_"})
is prepended to each parameter name.}
}
\value{
\code{get_par_names(curve, distr, excess, link = FALSE)} returns the
subset of
\code{c("r", "alpha", "c0", "tinfl", "K", "p", "a", "b", "nbdisp")}
relevant to \code{curve}, \code{distr}, and \code{excess}.
}
\description{
Returns the names used internally for parameters of the specified
incidence model.
}
\examples{
pn  <- get_par_names(curve = "exponential", distr = "pois",
                     excess = FALSE, link = FALSE)
lpn <- get_par_names(curve = "exponential", distr = "pois",
                     excess = FALSE, link = TRUE)
identical(pn, sub("^(log|logit)_", "", lpn))

}
