% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{get_par_names}
\alias{get_par_names}
\alias{get_par_names.default}
\alias{get_par_names.tmb_data}
\alias{get_par_names.egf}
\title{Get nonlinear model parameter names}
\usage{
get_par_names(curve, ...)

\method{get_par_names}{default}(
  curve = NULL,
  excess = NULL,
  distr = NULL,
  weekday = NULL,
  link = TRUE,
  ...
)

\method{get_par_names}{tmb_data}(curve, link = TRUE, ...)

\method{get_par_names}{egf}(curve, link = TRUE, ...)
}
\arguments{
\item{curve, excess, distr, weekday}{For the default method, atomic scalar flags specifying
an incidence model; see \code{\link[=egf]{egf()}}. Set to \code{NULL} or \code{NA}
to retrieve or suppress, respectively, all names associated
with an argument. For other methods, \code{curve} is an object
returned by \code{\link[=egf]{egf()}} or \code{\link[=make_tmb_data]{make_tmb_data()}}, and the remaining
arguments are ignored.}

\item{...}{Arguments passed to methods by the generic function.}

\item{link}{A logical scalar. If \code{TRUE}, then \code{"name"} is replaced
with \code{"log(name)"} or \code{"logit(name)"} depending on the
link function used internally for the parameter.}
}
\value{
If \code{link = FALSE}, then a subset of \code{"r"}, \code{"alpha"},
\code{"c0"}, \code{"tinfl"}, \code{"K"}, \code{"p"}, \code{"a"}, \code{"b"}, \code{"nbdisp"},
and \code{paste0("w", 1:6)}.
}
\description{
Retrieves the names used internally for nonlinear model parameters.
}
\examples{
## For a specific model
get_par_names(
  curve = "logistic",
  excess = FALSE,
  distr = "nbinom",
  weekday = 0L,
  link = FALSE
)

## Across all possible models
get_par_names(NULL, NULL, NULL, NULL, link = FALSE)

## With link
get_par_names(NULL, NULL, NULL, NULL, link = TRUE)

}
