% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{get_yax_cex}
\alias{get_yax_cex}
\title{Find size for y-axis tick labels}
\usage{
get_yax_cex(labels, mex, ...)
}
\arguments{
\item{labels}{A \link{character} or \link{expression} vector listing \eqn{y}-axis
tick labels.}

\item{mex}{A positive number. The desired maximum label width as a number
of margin lines.}

\item{...}{Graphical parameters passed to \code{\link{strwidth}}.}
}
\value{
A positive number.
}
\description{
Find \code{cex.axis} such that the widest \eqn{y}-axis tick label
generated by \code{\link{axis}(labels, cex.axis, las = 1)} spans
a desired number of margin lines.
}
\keyword{internal}
