% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{get_yax_cex}
\alias{get_yax_cex}
\title{Find size for y-axis tick labels}
\usage{
get_yax_cex(labels, mex, cex = NULL, font = NULL, csi = NULL)
}
\arguments{
\item{labels}{A \link{character} or \link{expression} vector listing \eqn{y}-axis
tick labels.}

\item{mex}{A positive number. The (desired) width of the widest tick label
as a number of margin lines.}

\item{cex, font, csi}{Graphical parameters affecting the width of a string as a number
of margin lines. The default (\code{\link{NULL}}) is to obtain values
using \code{\link{par}}.}
}
\value{
\code{mex * csi / mlw}, where \code{mlw} is the width
of the widest tick label in inches given \code{cex} and \code{font}.
}
\description{
Find \code{cex.axis} such that the widest \eqn{y}-axis tick label
generated by \code{\link{axis}(labels, cex.axis, las = 1)} is exactly
\code{mex} margin lines wide.
}
\keyword{internal}
