% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.R
\name{gi}
\alias{gi}
\alias{dgi}
\alias{pgi}
\alias{qgi}
\alias{rgi}
\title{\loadmathjax
Generation interval distribution}
\usage{
dgi(x, lat, inf)

pgi(q, lat, inf)

qgi(p, lat, inf)

rgi(n, lat, inf)
}
\arguments{
\item{x, q}{A numeric vector listing generation intervals.}

\item{lat}{A numeric vector listing the probability \mjseqn{p_i} that the
latent period is \mjseqn{i} (or the corresponding probability
weight) for all \mjseqn{i \in \lbrace 1,\ldots,m \rbrace}.}

\item{inf}{A numeric vector listing the probability \mjseqn{q_i} that the
infectious period is \mjseqn{i} (or the corresponding probability
weight) for all \mjseqn{i \in \lbrace 1,\ldots,n \rbrace}.}

\item{p}{A numeric vector listing probabilities.}

\item{n}{A non-negative integer. The number of samples generated.}
}
\value{
A numeric vector with length equal to the length of the
first argument, or with length \code{n} in the case of \code{rgi()}.

\code{dgi()} evaluates the density function \mjseqn{f_\textrm{gen}}.
\code{pgi()} evaluates the distribution function \mjseqn{F_\textrm{gen}}.
\code{qgi()} evaluates the quantile function, which is defined as the
left-continuous generalized inverse of \code{pgi()}.
\code{rgi()} samples from the distribution.
}
\description{
Generation interval density, distribution, and quantile functions.
Assumes:
\itemize{
\item that the latent and infectious periods are integer-valued,
\item that the latent period and infectious waiting time are independent, and
\item that infectiousness is constant over the infectious period.
}
}
\details{
Let \mjseqn{\tau_\textrm{lat}} and \mjseqn{\tau_\textrm{inf}} be the
latent and infectious periods. Let \mjseqn{f_\textrm{lat}} the density
function of \mjseqn{\tau_\textrm{lat}}, and let \mjseqn{F_\textrm{inf}}
be the distribution function of \mjseqn{\tau_\textrm{inf}}.

Suppose that the distributions of \mjseqn{\tau_\textrm{lat}} and
\mjseqn{\tau_\textrm{inf}} are supported on
\mjseqn{\lbrace 1,\ldots,m \rbrace} and
\mjseqn{\lbrace 1,\ldots,n \rbrace}, respectively, so that

\mjtdeqn{\begin{array}{r@{}ll} f_\textrm{lat}(t) &= \sum_{i=1}^{m} p_i \delta(t - i)\,, \cr F_\textrm{inf}(t) & {}= \mathrm{1}_{\lbrack 1,\infty)} \sum_{i=1}^{\min\lbrace n,\lfloor t \rfloor\rbrace} q_i\,, \end{array}}{\begin{align} f_\textrm{lat}(t) & {}= \sum_{i=1}^{m} p_i \delta(t - i)\,, \cr F_\textrm{inf}(t) &= \unicode{x1D7D9}_{\lbrack 1,\infty)} \sum_{i=1}^{\min\lbrace n,\lfloor t \rfloor\rbrace} q_i\,, \end{align}}{LaTeX}

where \mjseqn{p_i,q_i \in \lbrack 0,1 \rbrack},
\mjseqn{\sum_{i=1}^{m} p_i = \sum_{i=1}^{n} q_i = 1}
and \mjseqn{\delta} is the Dirac delta.

Now let \mjseqn{\tau_\textrm{wait}} be the infectious waiting time
and \mjseqn{f_\textrm{wait}} its density. The generation interval is
\mjseqn{\tau_\textrm{gen} = \tau_\textrm{lat} + \tau_\textrm{wait}} and
has density

\mjsdeqn{f_\textrm{gen}(t) = (f_\textrm{lat} * f_\textrm{wait})(t) = \sum_{i=1}^{m} p_i f_\textrm{wait}(t - i)\,,}

assuming independence of \mjseqn{\tau_\textrm{lat}} and
\mjseqn{\tau_\textrm{wait}}.

Equation 5.7 of \insertCite{Sven07;textual}{epigrowthfit}
gives an expression for \mjseqn{f_\textrm{wait}} for the case
in which infectiousness is constant over the infectious period:

\mjtdeqn{f_\textrm{wait} = \mathrm{1}_{(0,\infty)} \frac{1 - F_\textrm{inf}}{\mathrm{E}\lbrack \tau_\textrm{inf} \rbrack}\,.}{f_\textrm{wait} = \unicode{x1D7D9}_{(0,\infty)} \frac{1 - F_\textrm{inf}}{\unicode{x1D53C}\lbrack \tau_\textrm{inf} \rbrack}\,.}{LaTeX}

In this case, \mjseqn{f_\textrm{gen}} is supported on the interval
\mjseqn{\lbrack 1,m+n)} and is constant on the unit intervals
\mjseqn{\lbrack i,i+1)}. Hence evaluation of \mjseqn{f_\textrm{gen}}
(as well as the corresponding distribution and quantile functions)
is straightforward and fast.
}
\examples{
data(plague_latent_period)
lat <- plague_latent_period$relfreq
m <- length(lat)

data(plague_infectious_period)
inf <- plague_infectious_period$relfreq
n <- length(inf)

## Histogram of samples
y <- rgi(1e06, lat, inf)
hist(y, breaks = seq(0, m + n + 1), freq = FALSE, las = 1,
  ylab = "relative frequency",
  main = ""
)

## Density and distribution functions
x <- seq(0, m + n + 1, by = 0.02)
fx <- dgi(x, lat, inf)
Fx <- pgi(x, lat, inf)
plot(x, fx, type = "l", las = 1, # consistent with histogram
  xlab = "generation interval",
  ylab = "density function"
)
plot(x, Fx, type = "l", las = 1,
  xlab = "generation interval",
  ylab = "distribution function"
)

## Quantile function
p <- seq(0, 1, by = 0.001)
qp <- qgi(p, lat, inf)
plot(p, qp, type = "l", las = 1,
  xlab = "probability",
  ylab = "quantile function"
)

}
\references{
\insertRef{Sven07}{epigrowthfit}
}
