% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{gsub_bar_plus}
\alias{gsub_bar_plus}
\title{Replace \code{|} with \code{+} in formula terms}
\usage{
gsub_bar_plus(x)
}
\arguments{
\item{x}{A call, name, or atomic scalar.}
}
\value{
A call, name, or atomic scalar.
}
\description{
Replaces the \code{|} operator in formulae and formula terms with
the \code{+} operator, enabling construction of model frames from
mixed effects formulae using \code{\link[stats:model.frame]{stats::model.frame()}}.
}
\examples{
# gsub_bar_plus(~x + (1 | f))

}
\keyword{internal}
