% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{has_random}
\alias{has_random}
\title{Check for random effects}
\usage{
has_random(object)
}
\arguments{
\item{object}{An \code{"\link{egf}"} or \code{"\link[=make_tmb_data]{tmb_data}"} object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Determine whether an object specifies a random effects model.
}
\keyword{internal}
