% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{husting}
\alias{husting}
\title{Husting wills}
\format{
A \link[=data.frame]{data frame} with 723 rows and 4 variables:
\describe{
\item{\code{Date}}{
  A \link{Date} vector. Spacing varies, as only nonzero counts of wills
  are included.
}
\item{\code{wills}}{
  An \link{integer} vector. \code{wills[i]} is the number of wills
  written on \code{Date[i]}.
}
\item{\code{outbreak}}{
  An ordered \link{factor}, \link{split}ting the time series by plague
  outbreak. \link[=levels]{Levels} indicate the years in which outbreaks
  took place: 1348, 1361, 1368, and 1375.
}
\item{\code{severity}}{
  An ordered \link{factor}, classifying outbreaks as \code{"minor"}
  or \code{"major"}. All 14th century outbreaks are classified as major,
  as the data are too sparse to distinguish between minor and major
  outbreaks, in contrast with wills probated in the
  \link[=canterbury]{Prerogative Court of Canterbury}.
}
}
}
\source{
These data were transcribed from Sharpe (1889).
}
\usage{
data(husting)
}
\description{
Counts of wills probated in the Court of Husting
during four plague epidemics in 14th century London.
}
\examples{
data(husting)
subset(husting, outbreak == 1375, -outbreak)
}
\references{
Sharpe RR. Calendar of wills proved and enrolled in the Court of Husting,
London, A.D. 1258--A.D. 1688. London: JC Francis; 1889.
}
\keyword{datasets}
