% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_constant}
\alias{is_constant}
\title{Test whether an atomic vector is "constant"}
\usage{
is_constant(x, na.rm = FALSE, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{An \link{atomic} vector or \link[=data.frame]{data frame}.}

\item{na.rm}{A \link{logical} scalar. If \code{TRUE}, then \code{\link{NA}}
are ignored.}

\item{tol}{A positive number. \link[=double]{Double} vectors \code{x}
are considered constant if and only if the distance from
\code{\link{min}(x)} to \code{\link{max}(x)} is less than
\code{tol}.}
}
\value{
\code{TRUE}, \code{FALSE}, or \code{\link{NA}}.
}
\description{
Tests whether the elements of an \link{atomic} vector are equal
(or nearly equal in the case of \link{double} vectors).
}
\details{
\code{TRUE} is returned if \code{\link{length}(x) = 0}.

For data frames \code{x}, \code{is_constant(x)} tests
whether all listed vectors are constant.
}
\examples{
## x <- c(0, 1e-03, NA)
## is_constant(x, na.rm = TRUE, tol = 1e-02)
## is_constant(x, na.rm = TRUE, tol = 1e-04)
## is_constant(x, na.rm = FALSE, tol = 1e-02)
## is_constant(x, na.rm = FALSE, tol = 1e-04)

}
\keyword{internal}
