% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_combined}
\alias{make_combined}
\title{Construct combined model frame}
\usage{
make_combined(object)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}
}
\value{
A data frame combining (in the sense of \code{\link[=cbind]{cbind()}}) all data frames
listed in \code{object$frame_par} and the data frame \code{object$frame_append}.
}
\description{
Joins in a single data frame all mixed effects model frames
and further variables specified via \code{\link[=egf]{egf()}} argument \code{append}.
}
\details{
If, for example, a variable name occurs in multiple mixed effects
model frames, then only one instance is retained. Except in
unusual cases, all instances of a variable name are identical,
and no information is lost.

Since the data frames being combined each correspond rowwise to
\code{object$endpoints}, so does the result.
}
\keyword{internal}
