% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_frames}
\alias{make_frames}
\title{Construct model frames}
\usage{
make_frames(
  formula_ts,
  formula_par,
  data_ts,
  data_par,
  endpoints,
  origin,
  curve,
  excess,
  distr,
  weekday,
  na_action_ts,
  na_action_par,
  init,
  append
)
}
\arguments{
\item{formula_ts}{A formula of the form \code{x ~ time} or \code{x ~ time | ts} specifying
one or more incidence time series in long format. \code{time} must
evaluate to a numeric or Date vector. Numeric \code{time} is assumed
to measure time as a number of days since \emph{the end of} date
\code{origin}. Date \code{time} is coerced to numeric \code{julian(time, origin)}
(see Details). \code{x} must evaluate to a numeric vector. Within a
time series, \code{x[i]} should specify the number of cases observed
from \code{time[i-1]} to \code{time[i]}. Finally, \code{ts} must evaluate to
a factor such that \code{split(data.frame(time, x), ts)} returns
a list of time series. Note that \code{x ~ time} is equivalent to
\code{x ~ time | ts} with \code{ts} set equal to \code{rep(factor(1), length(x))}.}

\item{formula_par}{A list of formulae of the form \code{par ~ terms} specifying
mixed effects models (\code{\link[lme4:lmer]{lme4}}-like syntax)
for nonlinear model parameters. \code{deparse(par)} must be an
element of
\code{get_par_names(curve, distr, excess, weekday, link = TRUE)}.
\code{~1} is the default for parameters not assigned a formula.
Alternatively, \code{formula_par} may be a formula of the form
\code{~terms}. In this case, the formula is recycled for all
nonlinear model parameters. Note that "individuals" in each
model are fitting windows, hence model frames constructed
from \code{formula_par} and \code{data_par} are expected to correspond
rowwise to \code{endpoints} (see Details).}

\item{data_ts}{Data frames, lists, or environments. These are searched prior
to formula environments for variables used in \code{formula_ts} and
\code{formula_par}, respectively.}

\item{data_par}{Data frames, lists, or environments. These are searched prior
to formula environments for variables used in \code{formula_ts} and
\code{formula_par}, respectively.}

\item{endpoints}{A data frame, list, or environment with variables \code{start}
and \code{end}, and any further variables necessary to evaluate
\code{ts} if \code{formula_ts = x ~ time | ts}. \code{start} and \code{end}
must be numeric or Date vectors listing start and end times
for all fitting windows. \code{ts} must evaluate to a factor
indicating the time series in which each window is found.
Within time series, intervals \verb{[start[i], end[i]]} must be
disjoint and contain at least two time points from \code{time}.}

\item{origin}{A Date specifying a reference time.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline mortality
rate is estimated. Set to \code{TRUE} if what is observed is multiple
causes mortality rather than disease mortality or disease incidence.}

\item{distr}{A character string specifying an observation model.}

\item{weekday}{An integer scalar. If \code{weekday > 0}, then weekday effects are
estimated as offsets relative to the indicated day
(Sunday if \code{weekday = 1}, Monday if \code{weekday = 2}, and so on).
Currently, weekday effect estimation requires \code{time} in
\code{formula_ts = x ~ time | ts} to evaluate to an integer
(in the sense of \code{all.equal(time, round(time))}) or Date vector
with 1-day spacing in all fitting windows.
Logical \code{weekday} is equivalent to \code{as.integer(weekday)}.}

\item{na_action_ts}{A character string affecting the handling of \code{NA}
in \code{x} if \code{formula_ts = x ~ time | ts}.
\code{"fail"} is to throw an error.
\code{"pass"} is to ignore \code{NA} when fitting and replace \code{NA}
when predicting.
(Not yet implemented:
\code{"exclude"} is to ignore \code{NA} when fitting and preserve \code{NA}
when predicting.)
Note that \code{NA} in \code{time} and \code{ts} are an error regardless
of \code{na_action_ts}.}

\item{na_action_par}{A character string affecting the handling of \code{NA}
in \code{formula_par} variables.
\code{"fail"} is to throw an error.
\code{"pass"} is to discard fitting windows with incomplete data.}

\item{init}{A full parameter vector for the first likelihood evaluation.
The default (\code{NULL}) is to accept the internally generated
default. Use \code{do_fit = FALSE} to retrieve this default and
other optimization inputs (in particular \code{Y_init}; see Value),
which can often be used to construct a more informative
starting point.}

\item{append}{An expression indicating variables in \code{data_par} to be preserved
in the returned \code{"egf"} object for use by methods.
It is evaluated similarly to the \code{select} argument of \code{\link[=subset]{subset()}}.
The default (\code{NULL}) is to preserve all variables.
Currently, usage is supported only if \code{data_par} is a data frame.}
}
\value{
A list with elements \code{endpoints}, \code{frame_ts}, \code{frame_par},
and \code{frame_append}. See descriptions in \code{\link[=egf]{egf()}}.
}
\description{
Constructs time series and mixed effects model frames
to be used by \code{\link[=egf]{egf()}}, while completing a battery of
checks on the supplied formulae and data.
}
