% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{make_tmb_args}
\alias{make_tmb_args}
\title{Create TMB infrastructure}
\usage{
make_tmb_args(
  model,
  frame,
  frame_parameters,
  priors,
  control,
  do_fit,
  init,
  map
)
}
\arguments{
\item{model}{An \code{"\link{egf_model}"} object specifying a top level nonlinear model
to be estimated.}

\item{frame, frame_parameters}{Elements of the \link{list} output of \code{\link{make_frames}}.}

\item{priors}{A \link{list} of \link{formula}e of the form \code{parameter ~ f(...)}
specifying priors on top level nonlinear model parameters.
\code{\link{deparse}(parameter)} must be an element
of \code{\link{get_names_top}(model, link = TRUE)}, and
\code{f(...)} must be a \link{call} to a \link[=egf_prior]{prior function}
with arguments specifying suitable hyperparameters.
Calls are evaluated in the corresponding formula environments.}

\item{control}{An \code{"\link{egf_control}"} object specifying control parameters.}

\item{do_fit}{A \link{logical} flag. If \code{TRUE}, then \code{egf} returns early
with a \link{list} of optimization inputs.}

\item{init}{A \link{numeric} vector to be used as the full parameter vector for
the first likelihood evaluation. The default (\code{\link{NULL}}) is
to accept the internally generated default. Use \code{do_fit = FALSE}
to retrieve this default and other optimization inputs (in particular
\code{Y_init}; see Value), which can often be used to construct a more
informative starting point.}

\item{map}{A \link{factor} of length \code{\link{length}(init)}.
Elements of the full parameter vector corresponding to \code{\link{NA}}
in \code{map} are fixed at their initial values, rather than estimated.
Elements corresponding to a common factor level are constrained to have
a common value during estimation. \link[=logical]{Logical} values
of \code{map} are accepted and coerced to factor internally with
\code{map <- \link{replace}(\link{gl}(\link{length}(map), 1), map, NA)}.}
}
\value{
A \link{list} with elements \code{data}, \code{parameters}, \code{map},
\code{random}, \code{profile}, \code{DLL}, and \code{silent}.
}
\description{
Gathers necessary components of a call to \code{\link[TMB]{MakeADFun}}.
}
\seealso{
\code{\link{make_tmb_data}}, \code{\link{make_tmb_parameters}}
}
\keyword{internal}
