% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_args}
\alias{make_tmb_args}
\title{Create TMB infrastructure}
\usage{
make_tmb_args(
  frame_ts,
  frame_par,
  curve,
  distr,
  excess,
  weekday,
  weekday_ref,
  sparse_X,
  init
)
}
\arguments{
\item{frame_ts, frame_par}{Elements of the list output of \code{\link[=make_frames]{make_frames()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline
mortality rate is estimated. Set to \code{TRUE} if what is
observed (\code{x} if \code{formula_ts = x ~ time | ts}) is
multiple causes mortality rather than disease mortality
or disease incidence.}

\item{weekday}{A logical scalar. If \code{TRUE}, then weekday effects are estimated,
though this currently requires daily observation of incidence in
all fitting windows.}

\item{weekday_ref}{An integer indicating a weekday, with 1 mapping to Sunday,
2 mapping to Monday, and so on. Weekday effects are estimated
as offsets relative to the indicated day.}

\item{sparse_X}{A logical scalar. If \code{TRUE}, then the fixed effects design
matrix is constructed in sparse format.}

\item{init}{A full parameter vector for the first likelihood evaluation.
Set to \code{NULL} to accept the internally generated default.
Use \code{debug = TRUE} to retrieve this default, which is useful
as a template.}
}
\value{
A list with elements \code{data}, \code{parameters}, \code{map}, \code{random},
\code{DLL}, and \code{silent}.
}
\description{
Gathers necessary components of a call to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\seealso{
\code{\link[=make_tmb_data]{make_tmb_data()}}, \code{\link[=make_tmb_parameters]{make_tmb_parameters()}}
}
\keyword{internal}
