% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_args}
\alias{make_tmb_args}
\title{Create TMB infrastructure}
\usage{
make_tmb_args(
  frame_ts,
  frame_par,
  curve,
  distr,
  excess,
  weekday,
  sparse_X,
  init,
  debug
)
}
\arguments{
\item{frame_ts, frame_par}{Elements of the list output of \code{\link[=make_frames]{make_frames()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline mortality
rate is estimated. Set to \code{TRUE} if what is observed is multiple
causes mortality rather than disease mortality or disease incidence.}

\item{weekday}{An integer or logical scalar. If \code{weekday > 0}, then weekday
effects are estimated as offsets relative to the indicated day
(Sunday if \code{weekday = 1}, Monday if \code{weekday = 2}, and so on).
Currently, weekday effect estimation requires \code{time} in
\code{formula_ts = x ~ time | ts} to evaluate to an integer
(in the sense of \code{all.equal(time, round(time))}) or Date vector
with 1-day spacing in all fitting windows.}

\item{sparse_X}{A logical scalar. If \code{TRUE}, then the fixed effects design
matrix is constructed in sparse format.}

\item{init}{A full parameter vector for the first likelihood evaluation.
Set to \code{NULL} to accept the internally generated default.
Use \code{debug = TRUE} to retrieve this default, which may be
useful as a template.}

\item{debug}{A logical scalar used for debugging. If \code{TRUE}, then
\code{egf()} returns early with a list of optimization inputs.}
}
\value{
A list with elements \code{data}, \code{parameters}, \code{map}, \code{random},
\code{DLL}, and \code{silent}.
}
\description{
Gathers necessary components of a call to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\seealso{
\code{\link[=make_tmb_data]{make_tmb_data()}}, \code{\link[=make_tmb_parameters]{make_tmb_parameters()}}
}
\keyword{internal}
