% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_data}
\alias{make_tmb_data}
\title{Construct data objects for C++ template}
\usage{
make_tmb_data(frame, curve, distr, excess, sparse_X)
}
\arguments{
\item{frame}{A model frame returned by \code{\link[=make_frame]{make_frame()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline mortality
rate will be estimated. Set to \code{TRUE} if what is observed
(\code{y} if \code{formula = y ~ x}) is multiple causes mortality
rather than disease mortality or disease incidence.}

\item{sparse_X}{A logical scalar. If \code{TRUE}, then the fixed effects design matrix
will be constructed in sparse format.}
}
\value{
A \code{"tmb_data"} object, which is a list with elements:
\item{\code{index}}{
A copy of \code{attr(frame, "index")}. See \code{\link[=make_frame]{make_frame()}}.
}
\item{\code{t}}{
An integer vector of length \code{nrow(frame)} giving time as a number
of days since the earliest date in the fitting window.
}
\item{\code{x}}{
An integer vector of length \code{nrow(frame)}. Within each fitting
window, the first element is \code{NA} and the remaining elements
are such that \code{x[i]} is the number of cases observed from \code{t[i-1]}
to \code{t[i]}.
}
\item{\code{wl}}{
An integer vector of length \code{nlevels(index)} giving the length
of each fitting window.
}
\item{\code{Xs}, \code{Xd}}{
The fixed effects design matrix in sparse and dense format.
If \code{sparse_X = TRUE}, then \code{Xd} is empty. If \code{sparse_X = FALSE},
then \code{Xs} is empty. Uses zero-sum contrasts.
}
\item{\code{Z}}{
The random effects design matrix in sparse format, without
contrasts.
}
\item{\code{ffr}, \code{rfr}}{
Data frames containing the factors referenced by the elements of
\code{attr(frame, "fixed_term_labels")} and
\code{attr(frame, "random_term_labels")}. See \code{\link[=make_frame]{make_frame()}}.
}
\item{\code{fnl}, \code{rnl}}{
Integer vectors listing the number of levels of each factor in
\code{ffr} and \code{rfr}.
}
\item{\code{fid}, \code{rid}}{
Indicator matrices with term labels for row names and parameter
names (with prefixes) for column names. Element \verb{[i,j]} is equal
to 1 if and only if term \code{i} appears in the mixed effects model
(fixed/random component) for parameter \code{j}.
}
\item{\code{curve_flag}, \code{distr_flag}, \code{excess_flag}, \code{sparse_X_flag}}{
Integer flags referencing \code{curve}, \code{distr}, \code{excess}, and
\code{sparse_X}.
}
\item{\code{predict_flag}}{
An integer flag set equal to 0 so that prediction code is not run.
}
Additional integer elements of the form \code{j_link_parameter} (e.g.,
\code{j_log_r}) give the index of parameter names in \code{colnames(rid)}.
Indexing starts at 0, and the value -1 indicates that the parameter
does not appear in the incidence model being fit.
}
\description{
Gathers in a list data objects to be passed to the C++ template
via TMB's \code{DATA_} macros. See \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\examples{
example("make_frame", package = "epigrowthfit")
tmb_data <-
  epigrowthfit:::make_tmb_data(frame, curve, distr, excess,
                               sparse_X = FALSE)

}
\keyword{internal}
