% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_parameters}
\alias{make_tmb_parameters}
\title{Construct parameter objects for C++ template}
\usage{
make_tmb_parameters(tmb_data, frame_ts, frame_glmm, curve, par_init)
}
\arguments{
\item{tmb_data}{A \code{"tmb_data"} object returned by \code{\link[=make_tmb_data]{make_tmb_data()}}.}

\item{frame_ts}{Elements of the list output of \code{\link[=make_frames]{make_frames()}}.}

\item{frame_glmm}{Elements of the list output of \code{\link[=make_frames]{make_frames()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{par_init}{A full parameter vector for the first likelihood evaluation.
Set to \code{NULL} to accept the internally generated default. If
the default causes errors, then use \code{debug = TRUE} to retrieve
its value, which will have the correct length, and modify as
necessary.}
}
\value{
A list with elements:
\item{\code{beta}}{
A numeric vector of length \code{sum(tmb_data$beta_seg_len)}
listing initial values for the fixed effects coefficients.
}
\item{\code{b}}{
If there are random effects, then a numeric vector of
length \code{sum(tmb_data$b_seg_len)} listing initial values
for the (unit variance scale) random effects coefficients.
Otherwise, \code{NA_real_}.
}
\item{\code{log_sd_b}}{
If there are random effects, then a numeric vector of
length \code{sum(tmb_data$block_rows)} listing initial values
for the log standard deviations of the random effects
coefficients. Otherwise, \code{NA_real_}.
}
}
\description{
Gathers in a list parameter objects to be passed to the C++ template
via TMB's \code{PARAMETER_} macros during the first likelihood evaluation.
See \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\details{
When \code{par_init = NULL}, naive estimates of nonlinear model
parameters are obtained for each fitting window as follows:
\describe{
\item{\code{r}}{
The slope of a linear model fit to \verb{log1p(cumsum(x)))}.
}
\item{\code{alpha}}{
\code{r*c0^(1-p)} if \code{curve = "subexponential"},
\code{r/log(K/c0)} if \code{curve = "gompertz"}.
These are the values obtained by setting the instantaneous
exponential growth rate at time 0 in the subexponential and
Gompertz models equal to \code{r}, substituting the naive estimates
of \code{r}, \code{c0}, \code{K}, and \code{p}, and solving for \code{alpha}.
}
\item{\code{c0}}{
\code{exp(log_c0)}, where \code{log_c0} is the intercept of a linear model
fit to \code{log1p(cumsum(x))}.
}
\item{\code{tinfl}}{
\code{max(t)}, assuming that the fitting window ends near the time
of a peak in interval incidence.
}
\item{\code{K}}{
\code{2*sum(x)}, assuming that the fitting window ends near the time
of a peak in interval incidence \emph{and} that interval incidence is
roughly symmetric about the peak.
}
\item{\code{p}}{0.8}
\item{\code{a}, \code{b}, \code{nbdisp}, \code{w[123456]}}{1}
}
The naive estimates are log- or logit-transformed (all but \code{p}
use a log link), and, for each nonlinear model parameter, the
initial value of the \code{"(Intercept)"} coefficient in its fixed
effects model (if any) is taken to be the mean of the link scale
estimates across fitting windows. All other mixed effects model
parameters take initial value 0.
}
\keyword{internal}
