% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_tmb_parameters}
\alias{make_tmb_parameters}
\title{Construct parameter objects for C++ template}
\usage{
make_tmb_parameters(tmb_data, curve, distr, excess, par_init)
}
\arguments{
\item{tmb_data}{A \code{"tmb_data"} object returned by \code{\link[=make_tmb_data]{make_tmb_data()}}.}

\item{curve}{A character string specifying a cumulative incidence model.}

\item{distr}{A character string specifying an observation model.}

\item{excess}{A logical scalar. If \code{TRUE}, then a constant baseline mortality
rate will be estimated. Set to \code{TRUE} if what is observed
(\code{y} if \code{formula = y ~ x}) is multiple causes mortality
rather than disease mortality or incidence.}

\item{par_init}{A user-specified full parameter vector, or otherwise \code{NULL}.
See \code{\link[=egf]{egf()}}.}
}
\value{
A list with elements:
\item{\code{beta}}{
A numeric vector of length \code{sum(tmb_data$fnl)} listing initial
values for the fixed effects coefficients, namely between-group
means and within-group offsets.
}
\item{\code{log_sd_b}}{
If there are random effects, then a numeric vector of length
\code{sum(tmb_data$rid)} listing initial values for the log standard
deviations of the within-group random effects. Otherwise, \code{NA_real_}.
}
\item{\code{b}}{
If there are random effects, then a numeric vector of length
\code{sum(tmb_data$rnl * rowSums(tmb_data$rid))} listing initial
values for the within-group random effects. Otherwise, \code{NA_real_}.
}
}
\description{
Gathers in a list parameter objects to be passed to the C++ template
via TMB's \code{PARAMETER_} macros during the first likelihood evaluation.
See \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\details{
When \code{par_init = NULL}, naive estimates of incidence model parameters
are obtained for each fitting window as follows:
\describe{
\item{\code{r}}{
The slope of a linear model fit to \code{log1p(c(0, cumsum(x[-1])))}.
}
\item{\code{alpha}}{
\code{r*c0^(1-p)} if \code{curve = "subexponential"},
\code{r/log(K/c0)} if \code{curve = "gompertz"}.
These are the values obtained by setting the instantaneous
exponential growth rate at time 0 in the subexponential and
Gompertz models equal to \code{r}, substituting the naive estimates
of \code{r}, \code{c0}, \code{K}, and \code{p}, and solving for \code{alpha}.
}
\item{\code{c0}}{
\code{exp(log_c0)}, where \code{log_c0} is the intercept of a linear model
fit to \code{log1p(c(0, cumsum(x[-1])))}.
}
\item{\code{tinfl}}{
\code{max(t)}, assuming that the fitting window ends near the time
of a peak in interval incidence.
}
\item{\code{K}}{
\code{2*sum(x[-1])}, assuming that the fitting window ends near the
time of a peak in interval incidence \emph{and} that interval incidence
is roughly symmetric about the peak.
}
\item{\code{p}}{0.8}
\item{\code{a}}{1}
\item{\code{b}}{1}
\item{\code{nbdisp}}{1}
}
The naive estimates are log-transformed (all but \code{p}) or
logit-transformed (\code{p} only), and coefficients under zero-sum
contrasts are obtained from the within-group means to form
segments of parameter object \code{beta}.

When \code{par_init = NULL} and there are random effects, \code{log_sd_b}
and \code{b} are for simplicity initialized as zero vectors.
}
\examples{
example("make_tmb_data", package = "epigrowthfit")
tmb_parameters <-
  epigrowthfit:::make_tmb_parameters(tmb_data, curve, distr, excess,
                                     par_init = NULL)

}
\keyword{internal}
