% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{make_tmb_parameters}
\alias{make_tmb_parameters}
\title{Construct parameter objects for C++ template}
\usage{
make_tmb_parameters(tmb_data, model, frame, frame_par, do_fit, init)
}
\arguments{
\item{tmb_data}{A \code{"\link[=make_tmb_data]{tmb_data}"} object.}

\item{frame}{Elements of the \link{list} output of \code{\link{make_frames}}.}

\item{frame_par}{Elements of the \link{list} output of \code{\link{make_frames}}.}

\item{do_fit}{A \link{logical} flag. If \code{TRUE}, then \code{egf} returns early
with a \link{list} of optimization inputs.}

\item{init}{A \link{numeric} vector to be used as the full parameter vector for
the first likelihood evaluation. The default (\code{\link{NULL}}) is
to accept the internally generated default. Use \code{do_fit = FALSE}
to retrieve this default and other optimization inputs (in particular
\code{Y_init}; see Value), which can often be used to construct a more
informative starting point.}
}
\value{
A \link{list} with elements:
\item{beta}{
  A \link[=double]{numeric} vector of length
  \code{\link{sum}(tmb_data$beta_seg_len)} listing
  initial values for fixed effects coefficients.
}
\item{b}{
  A \link[=double]{numeric} vector of length
  \code{\link{sum}(tmb_data$b_seg_len)} listing initial values
  for random effects coefficients (unit variance scale).
}
\item{theta}{
  A \link[=double]{numeric} vector of length
  \code{\link{sum}(f(tmb_data$block_rows))}, where \code{f(n) = n*(n+1)/2},
  specifying initial covariance matrices for all random vectors.
  Let \code{x} be a segment of \code{theta} of length \code{f(n)},
  corresponding to a random vector \code{u} of length \code{n}.
  Then \code{x[1:n]} lists log standard deviations of the elements
  of \code{u}, and \code{x[-(1:n)]} lists (in row-major order)
  the subdiagonal elements of the unit lower triangular matrix
  \code{L} in the Cholesky factorization of the correlation matrix
  \code{S} of \code{u}:\cr
  \code{S = A \link[=matmult]{\%*\%} \link{t}(A)}\cr
  \code{A = D^-0.5 \link[=matmult]{\%*\%} L}\cr
  \code{D = \link{diag}(diag(L \link[=matmult]{\%*\%} \link{t}(L)))}
}
}
\description{
Gathers in a \link{list} parameter objects to be passed to
the package's C++ template via \pkg{TMB}'s \code{PARAMETER_*}
macros during the first likelihood evaluation.
See also \code{\link[TMB]{MakeADFun}}.
}
\details{
When \code{init = NULL}, naive estimates of nonlinear and dispersion
model parameters are obtained for each fitting window as follows:
\describe{
\item{\code{r}}{
  The slope of a linear model fit to \code{\link{log1p}(\link{cumsum}(x)))}.
}
\item{\code{alpha}}{
  \code{r*c0^(1-p)} if \code{curve = "subexponential"},
  \code{r/log(K/c0)} if \code{curve = "gompertz"}.
  These are the values obtained by setting the per capita growth
  rate at time 0 in the subexponential and Gompertz models equal
  to \code{r}, substituting the naive estimates of \code{r},
  \code{c0}, \code{K}, and \code{p}, and solving for \code{alpha}.
}
\item{\code{c0}}{
  \code{\link{exp}(log_c0)}, where \code{log_c0} is the intercept
  of a linear model fit to \code{\link{log1p}(\link{cumsum}(x))}.
}
\item{\code{tinfl}}{
  \code{\link{max}(t)}, assuming that the fitting window ends
  near the time of a peak in interval incidence.
}
\item{\code{K}}{
  \code{2*\link{sum}(x)}, assuming that the fitting window
  ends near the time of a peak in interval incidence _and_
  that interval incidence is roughly symmetric about the peak.
}
\item{\code{p}}{0.8}
\item{\code{a, b, nbdisp, w[123456]}}{1}
}
The naive estimates are log- or logit-transformed (all but
\code{p} use a log link), and, for each nonlinear or dispersion
model parameter, the initial value of the \code{"(Intercept)"}
coefficient in its fixed effects model (if one exists) is taken
to be the mean of the link scale estimates across fitting windows.
The remaining elements of \code{beta} take initial value 0.
All elements of \code{b} and \code{theta} take initial value 0,
so that, initially, all random vectors follow a standard normal
distribution.
}
\keyword{internal}
