% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{make_XZ_info}
\alias{make_XZ_info}
\title{Describe design matrix columns}
\usage{
make_XZ_info(xl, ml)
}
\arguments{
\item{xl}{A named list of fixed effects formulae of the form \code{~terms}, or
a named list of random effects terms of the form \code{(terms | group)}.
\code{names(xl)} must specify the respective nonlinear model parameters.}

\item{ml}{A list of \code{X} or \code{Z} matrices obtained by applying \code{\link[=make_X]{make_X()}} or
\code{\link[=make_Z]{make_Z()}} to the to the elements of \code{xl}.}
}
\value{
A data frame with \code{sum(vapply(ml, ncol, 0L))} with one row
per column of \code{do.call(cbind, ml)} and variables:
\item{\code{par}}{
Nonlinear model parameter.
}
\item{\code{term}}{
Deparsed term from right hand side of \code{~}
or left hand side of \code{|}, or \code{"(Intercept)"}.}
\item{\code{group}}{
Deparsed term from right hand side of \code{|}.
\code{NA} if not applicable.
}
\item{\code{level}}{
Level of the interaction specified by \code{group}.
\code{NA} if not applicable.
}
\item{\code{colname}}{
Column name.
}
}
\description{
A utility allowing combined design matrices and parameter vectors to
be split by nonlinear model parameter, term, group, or interactions
thereof.
}
\keyword{internal}
