% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{match_link}
\alias{match_link}
\title{Get link and inverse link functions}
\usage{
match_link(f, inverse = FALSE)
}
\arguments{
\item{f}{A \link{character} string naming a link function.}

\item{inverse}{A \link{logical} flag. If \code{TRUE}, then the inverse is returned.}
}
\value{
A \link{function}.
}
\description{
Retrieve the link function named by a string, or its inverse.
}
\examples{
## match_link("log")
## match_link("log", inverse = TRUE)

}
\keyword{internal}
