% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nse}
\alias{nse}
\alias{subset_to_index}
\alias{append_to_index}
\alias{order_to_index}
\alias{label_to_character}
\title{Utilities for nonstandard evaluation}
\usage{
subset_to_index(subset, data, enclos, .subset = NULL)

append_to_index(append, data, enclos, .append = NULL)

order_to_index(order, data, enclos, .order = NULL)

label_to_character(label, data, enclos, .label = NULL)
}
\arguments{
\item{subset, order, label}{Expressions to be evaluated in \code{data}.}

\item{data}{A data frame.}

\item{enclos}{An environment to enclose \code{data}.}

\item{.subset, .order, .append, `.label`}{Atomic vectors to be used (if non-\code{NULL}) in place of
the result of evaluating the undotted argument.}

\item{append}{An expression to be evaluated in
````names<-```(as.list(seq_along(data)), names(data))`.}
}
\value{
Let \code{r} be the result of evaluating the supplied expression
or, otherwise, the dotted argument.

\code{subset_to_index()} returns \code{which(r)}.

\code{order_to_index()} returns \code{r} as is.

\code{append_to_index()} returns
\code{match(names(data[r]), names(data), 0L)}, without zeros.

\code{label_to_character()} returns
\code{rep_len(as.character(r), nrow(data))}.
}
\description{
Utilities for obtaining index vectors from unevaluated \code{subset},
\code{append}, and \code{order} expressions and character vectors from
unevaluated \code{label} expressions (\code{xlab}, \code{ylab}, etc.). Used by
several methods for class \code{"egf"}.
}
\details{
\code{subset} must evaluate to a logical vector of length
\code{n = nrow(data)}.
\code{NULL} is equivalent to \code{rep_len(TRUE, n)}.

\code{order} must evaluate to a permutation of \code{seq_len(n)}.
\code{NULL} is equivalent to \code{seq_len(n)}.

\code{append} must evaluate to an atomic vector indexing \code{data}.
\code{NULL} is equivalent to \code{integer(0L)}.

\code{label} must evaluate to an atomic vector of length 1 or \code{n}.
\code{NULL} is a no-op.

Note that \code{subset} and \code{append} are processed similarly to
\code{\link[=subset]{subset()}} arguments \code{subset} and \code{select}. See \code{\link[=subset]{subset()}}
for additional usage examples.
}
\section{Warning}{
Nonstandard evaluation of \code{subset}, \code{order}, \code{append}, and \code{label}
is intended only to make interactive use more convenient. To avoid
unexpected behaviour, especially when programming, use the dotted
versions.
}

\examples{
year <- 2021L
data <- data.frame(
  month = sample(month.abb, 20L, replace = TRUE),
  day = sample(30L, 20L, replace = TRUE)
)

subset <- substitute(grepl("^J", month) & day < 16L)
order <- substitute(order(month, day))
append <- substitute(-day)
label <- substitute(sprintf("\%d-\%02d-\%02d", year, match(month, month.abb, 0L), day))

## Not run
\dontrun{
subset_to_index(subset, data, parent.frame())
order_to_index(order, data, parent.frame())
append_to_index(append, data, parent.frame())
label_to_character(label, data, parent.frame())
}

}
\keyword{internal}
