% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nse.R
\name{nse}
\alias{nse}
\alias{subset_to_index}
\alias{append_to_index}
\alias{order_to_index}
\alias{label_to_character}
\title{Nonstandard evaluation}
\usage{
subset_to_index(subset, data, enclos = parent.frame(), .subset = NULL)

append_to_index(append, data, enclos = baseenv(), .append = NULL)

order_to_index(order, data, enclos = parent.frame(), .order = NULL)

label_to_character(label, data, enclos = parent.frame(), .label = NULL)
}
\arguments{
\item{subset, order, label}{Expressions to be evaluated in \code{data}.}

\item{data}{A \link[=data.frame]{data frame}.}

\item{enclos}{An \link{environment} to enclose \code{data}.}

\item{.subset, .order, .append, .label}{\link[=atomic]{Atomic} vectors to be used (if non-\code{\link{NULL}})
in place of the result of evaluating the undotted argument.}

\item{append}{An expression to be evaluated in
\code{`\link{names}<-`(\link{as.list}(\link{seq_along}(data)), names(data))}.}
}
\value{
Let \code{r} be the result of evaluating the supplied expression or,
otherwise, the dotted argument.

\code{subset_to_index} returns \code{r} with \code{FALSE} replacing
\code{\link{NA}} (if any).

\code{order_to_index} returns \code{r} as is.

\code{append_to_index} returns
\code{\link{match}(\link{names}(data[r]), names(data), 0L)},
with zeros (if any) deleted.

\code{label_to_character} returns
\code{\link{rep_len}(\link{as.character}(r), \link{nrow}(data))}.
}
\description{
Utilities for obtaining index vectors from unevaluated \code{subset},
\code{append}, and \code{order} expressions and character vectors from
unevaluated \code{label} expressions (\code{xlab}, \code{ylab}, etc.).
}
\details{
\code{subset} must evaluate to a logical vector of length
\code{n = \link{nrow}(data)}.
\code{\link{NULL}} is equivalent to \code{\link{rep_len}(TRUE, n)}.

\code{order} must evaluate to a permutation of \code{\link{seq_len}(n)}.
\code{\link{NULL}} is equivalent to \code{\link{seq_len}(n)}.

\code{append} must evaluate to an \link{atomic} vector indexing \code{data}.
\code{\link{NULL}} is equivalent to \code{\link{integer}(0L)}.

\code{label} must evaluate to an \link{atomic} vector of length 1
or \code{n}.
\code{\link{NULL}} is a no-op.
}
\section{Note}{

Note that \code{subset} and \code{append} are processed similarly to
arguments \code{subset} and \code{select} of function \code{\link{subset}}.
See \code{\link{subset}} for additional usage examples.
}

\section{Warning}{

Nonstandard evaluation of \code{subset}, \code{order}, \code{append},
and \code{label} is intended only to make interactive use more convenient.
To avoid unexpected behaviour, especially when programming, use the dotted
versions.
}

\examples{
year <- 2021L
data <- data.frame(
  month = sample(month.abb, 20L, replace = TRUE),
  day = sample(30L, 20L, replace = TRUE)
)

subset <- quote(grepl("^J", month) & day < 16L)
order <- quote(order(month, day))
append <- quote(-day)
label <- quote(sprintf("\%d-\%02d-\%02d", year, match(month, month.abb, 0L), day))

# subset_to_index(subset, data, parent.frame())
# order_to_index(order, data, parent.frame())
# append_to_index(append, data, parent.frame())
# label_to_character(label, data, parent.frame())

}
\keyword{internal}
