% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egf_utils.R
\name{optim_tmb_out}
\alias{optim_tmb_out}
\title{Optimize a TMB model object}
\usage{
optim_tmb_out(tmb_out, method, ...)
}
\arguments{
\item{tmb_out}{The list output of \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}

\item{method}{A character string specifying an optimizer available through
\code{\link[stats:nlminb]{stats::nlminb()}}, \code{\link[stats:nlm]{stats::nlm()}}, or \code{\link[stats:optim]{stats::optim()}}.}

\item{...}{Optional arguments to the optimizer specified by \code{method}.}
}
\value{
The list output of \code{\link[stats:nlminb]{stats::nlminb()}}, \code{\link[stats:nlm]{stats::nlm()}},
or \code{\link[stats:optim]{stats::optim()}}, depending on \code{method}.
}
\description{
Evaluates a call to an optimizer constructed from the components
of a TMB model object.
}
\examples{
example("make_tmb_args", package = "epigrowthfit")
m <- c("nlminb", "nlm", "Nelder-Mead", "BFGS")
s <- c("par", "estimate", "par", "par")
f <- function(m, s) epigrowthfit:::optim_tmb_out(tmb_out, method = m)[[s]]
## Not run
\dontrun{
pp <- mapply(f, m = m, s = s) # needs 1-2 min
}

}
\keyword{internal}
