% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egf.R
\name{patch_fn_gr}
\alias{patch_fn_gr}
\title{Patch TMB-generated functions}
\usage{
patch_fn_gr(f, order, inner_optimizer)
}
\arguments{
\item{f}{A \link{function} to be patched, either \code{fn} or \code{gr}
from a \code{\link[TMB]{MakeADFun}}-generated \link{list} object.}

\item{order}{An integer identifying \code{f}, either 0 (\code{fn}) or 1 (\code{gr}).}

\item{inner_optimizer}{A \link{list} of \code{"\link{egf_inner_optimizer}"} objects
specifying inner optimization methods to be tried in turn.}
}
\value{
A patched version of \link{function} \code{f}.
}
\description{
Defines a wrapper function on top of \code{\link[TMB]{MakeADFun}}-generated
functions \code{fn} and \code{gr}, so that function and gradient evaluations
can retry inner optimization using fallback methods in the event of failure
using the default method (usually \code{\link[TMB]{newton}}).
}
\keyword{internal}
