% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{plague_infectious_period}
\alias{plague_infectious_period}
\title{Infectious period distribution (pneumonic plague)}
\format{
A data frame with 13 rows and 3 variables:

\describe{
\item{\code{days}}{[integer] Infectious period in days, from 0 to 12.}
\item{\code{observed}}{[numeric] Observed relative frequency
out of 225 cases.
}
\item{\code{fitted}}{[numeric] Probability from a fitted lognormal
distribution.
}
}
}
\source{
Obtained from Figure 1B in \insertCite{GaniLeac04;textual}{epigrowthfit}.
}
\usage{
data(plague_infectious_period)
}
\description{
Distributions (observed and fitted) of the infectious period
of pneumonic plague.
}
\references{
\insertRef{GaniLeac04}{epigrowthfit}
}
\keyword{datasets}
