% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{plague_latent_period}
\alias{plague_latent_period}
\title{Latent period distribution (pneumonic plague)}
\format{
A data frame with 12 rows and 2 variables:

\describe{
\item{\code{days}}{[integer] Latent period in days, from 1 to 12.}
\item{\code{relfreq}}{[numeric] Observed relative frequency
out of 224 cases.
}
}
}
\source{
Obtained from Figure 1A in \insertCite{GaniLeac04;textual}{epigrowthfit}.
}
\usage{
data(plague_latent_period)
}
\description{
Empirical distribution of the latent period of pneumonic plague.
}
\examples{
data(plague_latent_period)
plot(relfreq ~ days, data = plague_latent_period)

}
\references{
\insertRef{GaniLeac04}{epigrowthfit}
}
\keyword{datasets}
