% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit-data.R
\docType{data}
\name{plague_latent_period}
\alias{plague_latent_period}
\title{Latent period distribution (pneumonic plague)}
\format{
A \link[=data.frame]{data frame} with 12 rows and 2 variables:
\describe{
\item{\code{days}}{Latent period in days, from 1 to 12.}
\item{\code{relfreq}}{Observed relative frequency out of 224 cases.}
}
}
\source{
Obtained from Figure 1A in Gani and Leach (2004).
}
\usage{
data(plague_latent_period)
}
\description{
Empirical distribution of the latent period of pneumonic plague.
}
\examples{
data(plague_latent_period)
plot(relfreq ~ days, data = plague_latent_period)

}
\references{
Gani R, Leach S. Epidemiological determinants for modeling
pneumonic plague outbreaks. Emerg Infect Dis. 2004;10:608--14.
}
\keyword{datasets}
