% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\title{Plot models of epidemic growth}
\usage{
\method{plot}{egf}(
  x,
  type = c("interval", "cumulative", "rt1", "rt2"),
  subset = NULL,
  log = TRUE,
  tol = 0,
  legend = FALSE,
  level = 0.95,
  bands = FALSE,
  control = NULL,
  per_plot = 6L,
  bw_panels = 0.25,
  ...
)
}
\arguments{
\item{x}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{type}{A character string indicating a type of plot. Options are
\code{"interval"} (interval incidence),
\code{"cumulative"} (cumulative incidence),
\code{"rt1"}, and \code{"rt2"} (instantaneous exponential growth rate).}

\item{subset}{A named list of atomic vectors with elements specifying levels of
factors in \code{x$frame}. Only the subset of fitting windows belonging
to these levels is plotted. Use \code{NULL} (the default) to plot all
fitting windows or if \code{x$frame} has no factors.}

\item{log}{A logical scalar. If \code{TRUE}, then the dependent variable is plotted
on a logarithmic scale. Unused by \code{type \%in\% c("rt1", "rt2")}.}

\item{tol}{A non-negative number or \code{Inf}. Used to define "exceptional"
points when \code{type = "interval"}. \code{cases[-1]} is highlighted
according to \code{control$points_short} if \code{diff(date) < (1-tol)*m}
and according to \code{control$points_long} if \code{diff(date) > (1+tol)*m},
where \code{m = median(diff(date))}. In both cases, the value of
\code{diff(date)} is printed above the point. Assign 0 to highlight
all deviations from \code{m}. Assign \code{Inf} to disable highlighting.}

\item{legend}{A logical scalar. If \code{TRUE}, then a legend is displayed in the
right margin. Unused by \code{type != "rt2"}.}

\item{level}{A number in the interval (0,1). The confidence level represented
by confidence intervals on doubling times and confidence bands on
predicted values. Unused by \code{type = "rt2"}.}

\item{bands}{A logical scalar. If \code{TRUE}, then confidence bands on predicted
curves are displayed. Longer run times can be anticipated in this
case. Unused by \code{type = "rt2"}.}

\item{control}{A list of lists defining the appearance of various plot elements,
or otherwise \code{NULL} (see Details).}

\item{per_plot}{A positive integer. One plot will display this many time series
when \code{type = "rt2"}.}

\item{bw_panels}{A non-negative number. The spacing between panels as a number of
margin lines used when \code{type = "rt2"}.}

\item{...}{Optional arguments specifying additional graphical parameters
to be recycled for all plots. Currently, only \code{xlim}, \code{ylim},
\code{ylab}, and \code{main} are used; see \code{\link[graphics:plot.default]{graphics::plot()}}. \code{xlim}
can be numeric, Date, or character coercible to Date with
\code{as.Date(xlim)}. If \code{main} is a character string, then flags
of the form \code{"\%f"}, where \code{f} is the name of any factor present
in \code{group_by}, are replaced in a given plot by the factor level
relevant to that plot. Use \code{"\\n"} within \code{main} to separate
title from subtitle.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Methods for plotting \code{"egf"} objects.
}
\details{
\code{control} must be \code{NULL} or a named list containing some subset
of the elements ("control parameters") below:

\describe{
\item{\code{box}}{
A named list of arguments to \code{\link[graphics:box]{graphics::box()}}, affecting
the appearance of the box drawn around the plot region.
}
\item{\code{xax}, \code{yax}}{
Named lists of arguments to \code{\link[graphics:axis]{graphics::axis()}}, affecting the
appearance of the bottom and left axes. Option \code{mgp2} defines
the second component of the usual \code{mgp} argument. The minor
(day or month) and major (month or year) bottom axes can be
controlled separately by listing vectors of length 2.
For example, setting \code{xax = list(col.axis = c("black", "red"))}
makes the minor axis black and major axis red.
}
\item{\code{main}, \code{ylab}}{
Named lists of arguments to \code{\link[graphics:title]{graphics::title()}}, affecting
the appearance of axis titles.
}
\item{\code{rect}}{
A named list of arguments to \code{\link[graphics:rect]{graphics::rect()}}, affecting
the appearance of fitting windows.
}
\item{\code{polygon}}{
A named list of arguments to \code{\link[graphics:polygon]{graphics::polygon()}}, affecting
the appearance of confidence bands on predicted curves.
}
\item{\code{lines}}{
A named list of arguments to \code{\link[graphics:lines]{graphics::lines()}}, affecting
the appearance of predicted curves.
}
\item{\code{points}}{
A named list of arguments to \code{\link[graphics:points]{graphics::points()}}, affecting
the appearance of observed data.
}
\item{\code{points_short}, \code{points_long}}{
Alternatives to \code{points} used to highlight certain points
when \code{type = "interval"}. See argument \code{tol}.
}
\item{\code{text_short_long}}{
A named list of arguments to \code{\link[graphics:text]{graphics::text()}}, affecting
the appearance of text above highlighted points when
\code{type = "interval"}. See argument \code{tol}.
}
\item{\code{text_tdoubling}}{
A named list of arguments to \code{\link[graphics:text]{graphics::text()}}, affecting
the appearance of printed initial doubling times when
\code{x$curve \%in\% c("exponential", "logistic", "richards")}.
The estimate, confidence interval, and caption can be
controlled separately by listing vectors of length 3.
}
\item{\code{abline}}{
A named list of arguments to \code{\link[graphics:abline]{graphics::abline()}}, affecting
the appearance of the line drawn at \code{y = 0} when \code{type = "rt1"}.
}
\item{\code{segments}}{
A named list of arguments to \code{\link[graphics:segments]{graphics::segments()}}, affecting
the appearance of the line segments drawn at \code{y = r} when
\code{type = "rt1"} and
\code{x$curve \%in\% c("exponential", "logistic", "richards")}.
}
\item{\code{colorRamp}}{
A named list of arguments to \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}},
affecting the color palette used when \code{type = "rt2"}.
(Currently, this is the only control parameter for
\code{type = "rt2"}.)
}
}

A catalogue of modifiable options for each control parameter can
be obtained with \code{get_control_default("plot.egf", type = type)}.
Unspecified options take their values in this list. Unsupported
options are discarded.

To suppress a plot element, set the corresponding control parameter
to \code{NULL}, as in \code{control = list(box = NULL)}.
}
