% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\alias{plot.egf_init}
\alias{get_style_default}
\title{Plot models of epidemic growth}
\usage{
\method{plot}{egf_init}(
  x,
  inc = "interval",
  xty = "date",
  log = TRUE,
  add = FALSE,
  annotate = FALSE,
  tol = 0,
  style = get_style_default(x),
  ...
)

\method{plot}{egf}(
  x,
  inc = "interval",
  xty = "date",
  log = TRUE,
  add = FALSE,
  annotate = FALSE,
  tol = 0,
  style = get_style_default(x),
  ...
)

get_style_default(x)
}
\arguments{
\item{x}{An "egf_init" or "egf" object.}

\item{inc}{One of \code{"cumulative"} and \code{"interval"},
indicating a type of incidence to plot.}

\item{xty}{One of \code{"Date"} and \code{"numeric"},
indicating how time is displayed on the bottom axis.}

\item{log}{A logical scalar. If \code{TRUE}, then incidence
is displayed on a logarithmic scale on the left axis.}

\item{add}{A logical scalar. If \code{TRUE}, then plot elements are
added to the current graphics device, which is assumed to
have been started by a previous invocation of the plot method.}

\item{annotate}{A logical scalar. If \code{TRUE}, then a legend and
a list of parameter estimates are displayed in the right margin.}

\item{tol}{A non-negative number, used only if \code{inc = "interval"}.
\code{cases[i]} is highlighted according to \code{style$points_short} if
\code{diff(time)[i] < (1-tol)*m} and according to \code{style$points_long}
if \code{diff(time)[i] > (1+tol)*m}, where \code{m = median(diff(time))}.
In both cases, the value of \code{diff(time)[i]} is printed above
the point. Assign 0 to ensure that all deviations from \code{m}
are highlighted. Assign \code{Inf} to disable highlighting.}

\item{style}{A list of lists defining the appearance of various
plot elements (see Details 2).}

\item{...}{Optional arguments specifying additional
graphical parameters. Currently, only \code{xlim}, \code{ylim},
\code{xlab}, \code{ylab}, and \code{main} are implemented.
See \code{\link[graphics:plot.default]{graphics::plot()}} and \code{\link[graphics:par]{graphics::par()}} for a catalogue
of graphical parameters.}
}
\value{
\code{plot.egf_init()} and \code{plot.egf()} return \code{NULL} (invisibly).

\code{get_style_default()} returns a list of lists specifying the
default appearance of all (modifiable) plot elements.
}
\description{
Methods for plotting objects of class "egf_init" or "egf".
}
\details{
\subsection{1. Plot elements}{

\emph{Below, \code{date}, \code{time}, \code{cases}, \code{first}, \code{last}, \code{theta_init},
and \code{theta_fit} refer to the so-named elements of \code{x} or \code{x$init}.}

If \code{xty = "date"}, then the bottom axis displays the dates specified
by \code{date}. If \code{xty = "numeric"}, then the bottom axis displays the
number of days since \code{date[1]}. Regardless of \code{xty},
numeric coordinates are used, hence the left and right boundaries of
the plot region are defined by \code{range(time)}.

The left axis measures interval or cumulative incidence (depending
on \code{inc}). When incidence is displayed on a logarithmic scale, zeros
are plotted as a positive number less than 1. They are therefore
distinguished from nonzero counts, which are always at least 1.

Observed data, specified by \code{date} and either \code{cases} or
\code{cumsum(cases)} (depending on \code{inc}), are plotted as points.
\code{cases[i]} gives the number of cases observed between \code{date[i]}
and \code{date[i+1]}, while \code{cumsum(cases)[i]} gives the number
observed between \code{date[1]} and \code{date[i+1]}. Both are plotted
at \code{time[i+1]}.

The fitting window, specified by indices \code{first} and \code{last} of
\code{cases}, is displayed as a shaded rectangle behind the other plot
elements. The left and right boundaries occur at \code{time[first]}
and \code{time[last+1]}. (\code{cases[first]} is a count from \code{time[first]}
to \code{time[first+1]}, and \code{cases[last]} is a count from \code{time[last]}
to \code{time[last+1]}. Hence the fitting window starts at \code{time[first]}
and ends at \code{time[last+1]}.)

The incidence curve predicted by parameter estimates \code{theta_init}
or \code{theta_fit} (depending on the plot method) is displayed as a line
supported on grid points \code{wgrid = seq(time[first], time[last+1], by)},
where \code{by = 1} for cumulative incidence and \code{by = median(diff(time))}
for interval incidence (ensuring that the interval incidence curve
has the correct scale; see below). The predicted curve is obtained
with \code{predict(x, time = wgrid)}.

Careful interpretation of interval incidence is required if the
plotted time series is not equally spaced, because \code{cases} roughly
scales with \code{diff(time)}. That is, certain observations may vary
from the predicted curve not due to chance, but because they
represent a count over fewer or more days than the typical
observation interval, namely \code{median(diff(time))}. Observations for
which \code{diff(time)} differs from \code{median(diff(time))} are highlighted
according to argument \code{tol} and labeled with the value of \code{diff(time)}.

If \code{annotate = TRUE} and \code{add = FALSE}, then a legend and the
parameter estimates \code{theta_init} or \code{theta_fit} (depending on
the plot method) are displayed in the right margin.

The plot method for class "egf" displays, in addition,
the doubling time associated with the "egf" object and
the associated 95\% confidence interval.
}

\subsection{2. Customization}{

The appearance of most of the plot elements can be controlled
using argument \code{style}. \code{style} must be a list containing some
subset of the elements below.

\describe{
\item{\code{date}}{
A named list of arguments to \code{\link[=daxis]{daxis()}}
(a subset of \code{tcl}, \code{mgp2}, \code{col.axis}, and \code{cex.axis}),
affecting the appearance of the bottom axis if \code{xty = "date"}.
}
\item{\code{points_main}}{
A named list of arguments to \code{\link[graphics:points]{graphics::points()}},
affecting the appearance of the observed data. Currently,
only \code{pch}, \code{col}, \code{bg} and \code{cex} are implemented.
}
\item{\code{points_short}, \code{points_long}}{
Alternatives to \code{points_main} used to highlight certain
points when \code{inc = "interval"} (see argument \code{tol}).
}
\item{\code{lines}}{
A named list of arguments to \code{\link[graphics:lines]{graphics::lines()}},
affecting the appearance of the predicted incidence curve.
Currently, only \code{lty}, \code{lwd}, and \code{col} are implemented.
}
\item{\code{window}, \code{confband}}{
Named lists of arguments to \code{\link[graphics:polygon]{graphics::polygon()}},
affecting the appearance of the fitting window and confidence
bands. Currently, only \code{col} and \code{border} are implemented.
}
\item{\code{text_hl}, \code{text_dbl}}{
Named lists of arguments to \code{\link[graphics:text]{graphics::text()}},
affecting the appearance of text above highlighted points
(see argument \code{tol}) and text giving doubling times.
Currently, only \code{pos}, \code{offset}, \code{col}, \code{cex}, and \code{font}
are implemented. \code{text_dbl} can further specify coordinates
\code{x} and \code{y}.
}
}

List elements not specified by \code{style} are taken from
\code{get_style_default(x)}.
}
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \code{\link[=egf]{egf()}}
}
\keyword{internal}
