% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\alias{plot.egf_init}
\alias{get_style_default}
\title{Plot models of epidemic growth}
\usage{
\method{plot}{egf_init}(
  x,
  inc = "interval",
  xty = "date",
  log = TRUE,
  add = FALSE,
  annotate = FALSE,
  tol = 0,
  style = get_style_default(),
  ...
)

\method{plot}{egf}(
  x,
  inc = "interval",
  xty = "date",
  log = TRUE,
  add = FALSE,
  annotate = FALSE,
  tol = 0,
  style = get_style_default(),
  ...
)

get_style_default()
}
\arguments{
\item{x}{An "egf_init" or "egf" object.}

\item{inc}{One of \code{"cumulative"} and \code{"interval"},
indicating a type of incidence to plot.}

\item{xty}{One of \code{"date"} and \code{"numeric"},
indicating how time is displayed on the bottom axis
(calendar or number of days).}

\item{log}{A logical scalar. If \code{TRUE}, then incidence is displayed
on a logarithmic scale on the left axis.}

\item{add}{A logical scalar. If \code{TRUE}, then plot elements are added
to the current graphics device, which is assumed to have
been started by a previous invocation of the plot method.}

\item{annotate}{A logical scalar. If \code{TRUE}, then a legend and a list of
parameter estimates are displayed in the right margin.}

\item{tol}{A non-negative number, used only if \code{inc = "interval"}.
\code{cases[-1]} is highlighted according to \code{style$points_short}
if \code{diff(date) < (1-tol)*m} and according to \code{style$points_long}
if \code{diff(date) > (1+tol)*m}, where \code{m = median(diff(date))}.
In both cases, the value of \code{diff(date)} is printed above
the point. Assign 0 to ensure that all deviations from \code{m}
are highlighted. Assign \code{Inf} to disable highlighting.}

\item{style}{A list of lists defining the appearance of various plot elements.}

\item{...}{Optional arguments specifying additional graphical parameters.
Currently, only \code{xlim}, \code{ylim}, \code{xlab}, \code{ylab}, and \code{main}
are implemented. \code{xlim} can be numeric, Date, or character
coercible to Date with \code{as.Date(xlim)}. See \code{\link[graphics:plot.default]{graphics::plot()}}
and \code{\link[graphics:par]{graphics::par()}} for a catalogue of graphical parameters.}
}
\value{
\code{plot.egf_init()} and \code{plot.egf()} return \code{NULL} (invisibly).

\code{get_style_default()} returns a list of lists specifying the
default appearance of all modifiable plot elements.
}
\description{
Methods for plotting objects of class "egf_init" or "egf".
}
\details{
\subsection{1. Plot elements}{

\emph{Below, \code{date}, \code{cases}, \code{first}, \code{last}, \code{theta_init}, and
\code{theta_fit} refer to the so-named elements of \code{x} or \code{x$data}.}

If \code{xty = "date"}, then the bottom axis is a calendar with ticks at
equally spaced Dates. If \code{xty = "numeric"}, then the bottom axis
measures the number of days since \code{date[1]}. Regardless of \code{xty},
Dates \code{d} are assigned numeric user coordinates \code{d - date[1]}.

The left axis measures interval or cumulative incidence (depending
on \code{inc}). When incidence is displayed on a logarithmic scale, zeros
are plotted as a positive number less than 1. They are therefore
distinguished from nonzero counts, which are always at least 1.

Observed data, specified by \code{date} and either \code{cases[-1]} or
\code{cumsum(cases[-1])} (depending on \code{inc}), are plotted as points.
\code{cases[-1][i]} gives the number of cases observed between
\code{date[i]} and \code{date[i+1]}, while \code{cumsum(cases[-1])[i]} gives
the number observed between \code{date[1]} and \code{date[i+1]}. Both are
plotted at coordinate \code{date[i+1]-date[1]}.

The fitting window is displayed as a shaded rectangle behind the
other plot elements. The left and right boundaries are \code{date[first]}
and \code{date[last]}.

The incidence curve predicted by parameter estimates \code{theta_init}
or \code{theta_fit} (depending on \code{class(x)}) is displayed as a line
supported on grid points \code{wgrid = seq(date[first], date[last], by)},
where \code{by = 1} for cumulative incidence and \code{by = median(diff(date))}
for interval incidence (ensuring that the interval incidence curve
has the correct scale; see below). The predicted curve is obtained
with \code{predict(x, time = as.integer(wgrid - date[first]))}.

Careful interpretation of interval incidence is required if the
plotted time series is not equally spaced, because \code{cases}
roughly scales with \code{diff(date)}. That is, certain observations
may vary from the predicted curve not due to chance, but because
they give counts over fewer or more days than the typical
observation interval, namely \code{median(diff(date))}. Observations
for which \code{diff(date)} differs from \code{median(diff(date))} are
highlighted according to argument \code{tol} and labeled with the
value of \code{diff(date)}.

If \code{annotate = TRUE} and \code{add = FALSE}, then a legend and the
parameter estimates \code{theta_init} or \code{theta_fit} (depending on
\code{class(x)}) are displayed in the right margin.

The plot method for class "egf" displays, in addition,
the doubling time associated with the "egf" object and
the associated 95\% confidence interval, obtained with
\code{confint(x, parm = "doubling_time", level = 0.95, method = "linear")}.
}

\subsection{2. Customization}{

The appearance of most of the plot elements can be controlled
using argument \code{style}. \code{style} must be a list containing some
subset of the elements below.

\describe{
\item{\code{box}}{
A named list of arguments to \code{\link[graphics:box]{graphics::box()}}, affecting the
appearance of the box drawn around the plot region. Currently,
only \code{bty}, \code{lty}, \code{col}, and \code{lwd} are implemented.
}
\item{\code{xax}, \code{yax}}{
Named lists of arguments to \code{\link[graphics:axis]{graphics::axis()}}, affecting
the appearance of the bottom and left axes. Currently, only
\code{tcl}, \code{mgp2}, \code{col.axis}, \code{cex.axis}, and \code{font.axis} are
implemented, where \code{mgp2} is the second component of the
usual \code{mgp} argument. If \code{xty = "date"}, then the appearance
of the minor (day or month) and major (month or year) bottom
axes can be controlled separately by assigning vectors of
length 2 to the elements of \code{xax}. For example, setting
\code{xax = list(col.axis = c("black", "red"))} will make the
minor axis black and major axis red.
}
\item{\code{xlab}, \code{ylab}, \code{main}}{
Named lists of arguments to \code{\link[graphics:title]{graphics::title()}}, affecting
the appearance of axis titles. Currently, only \code{line}, \code{adj},
\code{col.lab}, \code{cex.lab}, \code{font.lab},
\code{col.main}, \code{cex.main}, and \code{font.main} are implemented.
}
\item{\code{points_main}}{
A named list of arguments to \code{\link[graphics:points]{graphics::points()}},
affecting the appearance of the observed data. Currently,
only \code{pch}, \code{col}, \code{bg} and \code{cex} are implemented.
}
\item{\code{points_short}, \code{points_long}}{
Alternatives to \code{points_main} used to highlight certain
points when \code{inc = "interval"} (see argument \code{tol}).
}
\item{\code{lines}}{
A named list of arguments to \code{\link[graphics:lines]{graphics::lines()}},
affecting the appearance of the predicted incidence curve.
Currently, only \code{lty}, \code{lwd}, and \code{col} are implemented.
}
\item{\code{window}, \code{confband}}{
Named lists of arguments to \code{\link[graphics:polygon]{graphics::polygon()}},
affecting the appearance of the fitting window and confidence
bands. Currently, only \code{col} and \code{border} are implemented.
}
\item{\code{text_hl}, \code{text_dbl}}{
Named lists of arguments to \code{\link[graphics:text]{graphics::text()}},
affecting the appearance of text above highlighted points
(see argument \code{tol}) and text giving doubling times.
Currently, only \code{pos}, \code{offset}, \code{col}, \code{cex}, and \code{font}
are implemented. \code{text_dbl} can further specify alignment
\code{adj} and coordinates \code{x} and \code{y}. \code{x} can be numeric,
Date, or character coercible to Date with \code{as.Date(x)}.
\code{y} must be numeric.
}
}

List elements not specified by \code{style} are taken from
\code{get_style_default()}.

Assigning \code{NULL} to any of the above possible \code{style} elements
(instead of a named list) will suppress the corresponding plot
element. Setting \code{points_main = NULL} will not suppress points
styled according to \code{points_short} and \code{points_long}. Hence,
to suppress all points, it may be necessary
to also set \code{points_short = NULL} and \code{points_long = NULL}.
}
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \code{\link[=egf]{egf()}}
}
\keyword{internal}
