% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\title{Plot models of epidemic growth}
\usage{
\method{plot}{egf}(
  x,
  type = c("interval", "cumulative", "rt1", "rt2"),
  subset = NULL,
  order = NULL,
  cache = NULL,
  plot = TRUE,
  time_as = c("Date", "numeric"),
  log = TRUE,
  show_fits = TRUE,
  show_bands = FALSE,
  show_tdoubling = FALSE,
  show_legend = FALSE,
  level = 0.95,
  per_plot = 6L,
  control = NULL,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  ylab_outer = NULL,
  plab = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{type}{A character string indicating a type of plot. The options are:
interval incidence (\code{"interval"}),
cumulative incidence (\code{"cumulative"}),
per capita growth rate as curve (\code{"rt1"}), and
per capita growth rate as heat map (\code{"rt2"}).
\code{type != "rt2"} displays one time series per plot.
\code{type = "rt2"} displays \code{per_plot} time series per plot.}

\item{subset}{An expression to be evaluated in the combined model frame.
Must evaluate to a logical vector or list of logical vectors
indexing rows of the model frame, and thus fitting windows.
Only indexed fitting windows are plotted. The default (\code{NULL})
is to plot all fitting windows.}

\item{order}{An expression to be evaluated in the combined model frame,
typically a call to \code{\link[=order]{order()}}, determining the order in which
time series are plotted. The default (\code{NULL}) is equivalent
to \code{seq_len(nrow(frame))}.}

\item{cache}{An \code{"egf_plot_cache"} object returned by a previous evaluation
of \code{plot.egf(x)}. Predicted values and standard errors stored
in \code{cache} will be reused (rather than recomputed) if relevant
to the current function call (i.e., to \code{type} and \code{subset}).}

\item{plot}{A logical scalar. If \code{FALSE}, then nothing is plotted.
Useful when only the return value is desired.}

\item{time_as}{A character string indicating how time is displayed on the
bottom axis. The options are: as a calendar (\code{"Date"}) and
as a number of days since the earliest time point (\code{"numeric"}).}

\item{log}{A logical scalar. If \code{TRUE}, then the dependent variable is
plotted on a logarithmic scale. Unsupported for \code{type = "rt1"}.}

\item{show_fits}{A logical scalar. If \code{TRUE}, then predicted curves are drawn.
Unused by \code{type = "rt2"}.}

\item{show_bands}{A logical scalar. If \code{TRUE}, then confidence bands on
predicted curves are drawn. Unused by \code{type = "rt2"}.}

\item{show_tdoubling}{A logical scalar. If \code{TRUE}, then initial doubling time
estimates and corresponding confidence intervals are printed
in the top margin. Unused by \code{type = "rt2"}. Unsupported
for \code{x$curve} not in \code{c("exponential", "logistic", "richards")}.}

\item{show_legend}{A logical scalar. If \code{TRUE}, then a legend is displayed in
the right margin. Unsupported for \code{type != "interval"}.}

\item{level}{A number in the interval (0,1). The confidence level desired
when \code{show_tdoubling = TRUE} or \code{show_bands = TRUE}.}

\item{per_plot}{A positive integer. The number of time series displayed
in one plot. Unsupported for \code{type != "rt2"}.}

\item{control}{A named list controlling the appearance of various plot elements
(see Details). The default (\code{NULL}) is to use
\code{get_control_default("plot.egf", type = type, time_as = time_as)}.}

\item{xlim, ylim}{Optional numeric vectors specifying axis limits, which are
recycled for all plots. If \code{time_as = "Date"}, then \code{xlim}
can instead be a Date vector or a character vector coercible
to Date via \code{as.Date(xlim)}. \code{ylim} is unused by \code{type = "rt2"}.}

\item{main, sub, xlab, ylab, ylab_outer, plab}{Optional character strings or expressions used to generate plot
(\code{main}, \code{sub}), axis (\code{xlab}, \code{ylab}, \code{ylab_outer}), and panel
(\code{plab}) titles. \code{sub} is unsupported by \code{type = "rt2"}.
\code{ylab_outer} is used only by \code{type = "rt[12]"}. \code{plab} is used
only by \code{type = "rt2"}. When \code{type != "rt2"}, \code{main} and \code{sub}
are evaluated in the combined model frame to generate unique
(sub)titles for each plot. When \code{type = "rt2"}, \code{plab} is
evaluated similarly to generate unique titles for each panel.
Note that \code{\link[grDevices:plotmath]{plotmath}} expressions are
not supported for \code{main}, \code{sub}, and \code{plab} in these cases.}

\item{...}{Unused optional arguments.}
}
\value{
An \code{"egf_plot_cache"} object. If \code{cache} was used, then this object
is the result of augmenting \code{cache} with any new computations.

Note that this object is returned \emph{in spite of} any errors thrown
during plotting, to avoid waste of computation time.
}
\description{
Methods for plotting \code{"egf"} objects.
}
\details{
The combined model frame is \code{do.call(cbind, unname(x$frame_par))}.
If a variable occurs more than once there, because it appears
in multiple model frames, then only the earliest instance is
retained. Except in unusual cases, all instances of a variable
are identical, and no information is lost.

\code{plot.egf()} will \emph{not} detect mismatch between \code{x} and \code{cache}.
In other words, constructions like \code{plot(x2, cache = plot(x1))}
should \emph{not} be expected to produce correct results.

\code{control} is a named list defining some subset of the available
control parameters and their respective components (see section
"Control parameters"). A list of modifiable options for each
control parameter and their default values can be obtained with
\code{get_control_default("plot.egf", type = type, time_as = time_as)}.
Unspecified options take their values in this list. Unsupported
options are silently discarded. To suppress a plot element, set
the corresponding control parameter to \code{NULL},
as in \code{control = list(box = NULL)}.
}
\section{Control parameters}{
For all \code{type}:
\describe{
\item{\code{axis}}{
A named list of the form \code{list(x, y)} affecting the appearance of
the bottom and left axes, respectively.
If \code{time_as = "Date"}, in which case there are minor (day or month)
and major (month or year) bottom axes, then \code{x} should be a named
list of the form \code{list(minor, major)}, and \code{minor}, \code{major}, and
\code{y} should all be named lists of arguments to \code{\link[graphics:axis]{graphics::axis()}}.
If \code{time_as = "numeric"}, in which case there is one bottom axis,
then \code{x} should itself be a named list of arguments to
\code{\link[graphics:axis]{graphics::axis()}}.
}
\item{\code{title}}{
A named list of the form \code{list(main, sub, xlab, ylab, plab)},
affecting the appearance of plot and axis titles. \code{main}, \code{sub},
\code{xlab}, \code{ylab}, and \code{plab} should all be lists of arguments to
\code{\link[graphics:title]{graphics::title()}}. Note that subtitles are handled specially:
they are placed under main titles in the top margin, rather than
in the bottom margin (the usual behavior of \code{\link[graphics:title]{graphics::title()}}).
}
\item{\code{rect}}{
A named list of arguments to \code{\link[graphics:rect]{graphics::rect()}},
affecting the appearance of fitting windows.
}
}
For \code{type != "rt2"}:
\describe{
\item{\code{box}}{
A named list of arguments to \code{\link[graphics:box]{graphics::box()}},
affecting the appearance of the box drawn around the plot region.
}
\item{\code{polygon}}{
A named list of arguments to \code{\link[graphics:polygon]{graphics::polygon()}},
affecting the appearance of confidence bands on predicted curves.
}
\item{\code{lines}}{
A named list of arguments to \code{\link[graphics:lines]{graphics::lines()}},
affecting the appearance of predicted curves.
}
\item{\code{points}}{
A named list of arguments to \code{\link[graphics:points]{graphics::points()}},
affecting the appearance of observed data.
}
\item{\code{tdoubling}}{
A named list of the form \code{list(caption, estimate, ci)},
affecting the appearance of initial doubling times printed
when \code{x$curve \%in\% c("exponential", "logistic", "richards")}.
\code{caption}, \code{estimate}, and \code{ci} should each be named lists
of arguments to \code{\link[graphics:text]{graphics::text()}}.
}
}
For \code{type = "interval"}:
\describe{
\item{\code{special}}{
A named list of the form
\code{list(tol, points = list(short, long), text)},
affecting the appearance of "exceptional" points.
\code{tol} should be a non-negative number or \code{Inf}. \code{short} and \code{long}
should be named lists of arguments to \code{\link[graphics:points]{graphics::points()}}, and
\code{text} should be a named list of arguments to \code{\link[graphics:text]{graphics::text()}}.
Within a time series \code{data.frame(time, x)}, \code{x[-1]} is highlighted
according to \code{short} if \code{diff(time) < (1-tol)*m} and
according to \code{long}  if \code{diff(time) > (1+tol)*m},
where \code{m = median(diff(time))}.
In both cases, the value of \code{diff(time)} is printed above the point
according to \code{text}. Use \code{tol = 0} to highlight all deviations from
\code{m}. Use \code{tol = Inf} to disable highlighting.
}
}
For \code{type = "rt1"}:
\describe{
\item{\code{abline}}{
A named list of arguments to \code{\link[graphics:abline]{graphics::abline()}},
affecting the appearance of the line drawn at \code{y = 0}.
}
\item{\code{segments}}{
A named list of arguments to \code{\link[graphics:segments]{graphics::segments()}},
affecting the appearance of line segments drawn at \code{y = r}
when \code{x$curve \%in\% c("exponential", "logistic", "richards")}.
}
}
For \code{type = "rt2"}:
\describe{
\item{\code{colorRamp}}{
A named list of arguments to \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}}
defining the heat map's color palette.
}
\item{\code{ips}}{
A non-negative number defining the space between panels
as a number of margin lines.
}
}
}

