% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\title{Plot nonlinear mixed effects models of epidemic growth}
\usage{
\method{plot}{egf}(
  x,
  type = c("interval", "cumulative", "rt", "rt_heat"),
  time_as = c("Date", "numeric"),
  log = TRUE,
  dt = 1,
  per_plot = min(6L, nlevels(x$frame$ts)),
  subset = NULL,
  order = NULL,
  cache = NULL,
  plot = TRUE,
  show_predict = TRUE,
  show_tdoubling = TRUE,
  level = 0.95,
  control = egf_plot_control(),
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  ylab_outer = NULL,
  plab = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{"\link{egf}"} object.}

\item{type}{A \link{character} string indicating a type of plot. The options are:
interval incidence (\code{"interval"}),
cumulative incidence (\code{"cumulative"}),
per capita growth rate (\code{"rt"}), and
per capita growth rate \emph{as heat map} (\code{"rt_heat"}).
\code{"rt_heat"} displays \code{per_plot} time series in each plot.
The rest display one time series in each plot.}

\item{time_as}{A \link{character} string indicating how numeric times are displayed
on the bottom axis. The options are:
as is (\code{"numeric"})
and with a calendar (\code{"Date"}).
In the latter case, numeric times are interpreted as numbers of days
since \code{1970-01-01 00:00:00}.}

\item{log}{A \link{logical} flag. If \code{TRUE}, then the dependent variable
is plotted on a logarithmic scale.}

\item{dt}{A positive number indicating an observation interval.
Predicted curves are evaluated on grids with this spacing.
When \code{type = "interval"}, counts observed over shorter
or longer intervals \code{dt0} are scaled by a factor of
\code{dt / dt0} so that their scale matches that of the curves.
Scaled counts can be highlighted via \code{control}.
If \code{x} specifies a model with day of week effects
(\code{x$model$day_of_week > 0}), then setting \code{dt}
has no effect as it is set to 1 internally.}

\item{per_plot}{A positive integer giving the number of time series displayed in
each plot.
[\code{type = "rt_heat"} only.]}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{egf_combine_frames}}). It must evaluate
to a valid index vector for the rows of the data frame
(see \code{\link{[.data.frame}}), and thus fitting windows.
Only indexed fitting windows are highlighted in plots.
Only time series with indexed fitting windows are plotted.
The default (\code{\link{NULL}}) is to plot all time series
and display all fitting windows in each time series.}

\item{order}{An expression to be evaluated in the combined model frame
(see \code{\link{egf_combine_frames}}), typically a
call to \code{\link{order}}, determining the order in which
time series are plotted. It must evaluate to a permutation of
\code{\link{seq_len}(\link{nrow}(combined))}.
The default (\code{\link{NULL}}) is equivalent to
\code{\link{seq_len}(\link{nrow}(combined))}.}

\item{cache}{An \code{"egf_plot_cache"} object returned by a previous evaluation
of \code{plot(x)}. Fitted and predicted values and standard errors
stored in \code{cache} are reused if possible to avoid recomputation.}

\item{plot}{A \link{logical} flag. If \code{FALSE}, then nothing is plotted.
Useful when only the returned \code{"egf_plot_cache"} object is desired.}

\item{show_predict}{An integer flag: 2 is to draw predicted curves with confidence bands,
1 is draw predicted curves only, 0 is to draw neither.
\link[=logical]{Logical} values are coerced to integer.
[\code{type != "rt_heat"} only.]}

\item{show_tdoubling}{An integer flag: 2 is to print initial doubling time estimates in
the top margin with confidence intervals, 1 is to print estimates
only, 0 is to print neither. \link[=logical]{Logical} values are
coerced to integer. Supported only if \code{x$model$curve} is
\code{"exponential"}, \code{"logistic"}, or \code{"richards"}.
[\code{type != "rt_heat"} only.]}

\item{level}{A number in the interval (0,1). This is the confidence level used
when \code{show_predict = 2} or \code{show_tdoubling = 2}.
[\code{type != "rt_heat"} only.]}

\item{control}{An \code{"\link{egf_plot_control}"} object controlling the appearance
of most plot elements.}

\item{xlim, ylim}{\link[=numeric]{Numeric} vectors of length 2 specifying axis limits,
which are recycled for all plots. If \code{time_as = "Date"}, then
\code{xlim} can instead be a \link{Date} vector or a \link{character}
vector coercible to Date via \code{\link{as.Date}(xlim)}. \code{ylim}
is unused by \code{type = "rt_heat"}.}

\item{main, sub, xlab, ylab, ylab_outer, plab}{\link[=character]{Character} strings or expressions used
to generate plot (\code{main}, \code{sub}), axis (\code{xlab},
\code{ylab}, \code{ylab_outer}), and panel (\code{plab}) titles.
\code{main}, \code{xlab}, and \code{ylab} are supported for all
values of \code{type}.
\code{sub} is unused by \code{type = "rt_heat"}.
\code{plab} is used by \code{type = "rt_heat"} only.
\code{ylab_outer} is used by \code{type = "rt(_heat)?"} only.
When \code{type != "rt_heat"}, \code{main} and \code{sub} are evaluated
in the combined model frame (see \code{\link{egf_combine_frames}})
in order to generate unique (sub)titles for each plot.
When \code{type = "rt_heat"}, \code{plab} is evaluated similarly
in order to generate unique titles for each panel.
\code{\link{plotmath}} expressions are not supported
for \code{main}, \code{sub}, and \code{plab} in these cases.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data.frame} inheriting from \link{class}
\code{"egf_plot_cache"}. If argument \code{cache} was supplied
in the function call, then this data frame is the result of
augmenting \code{cache} with new computations.
}
\description{
Methods for plotting \code{"\link{egf}"} objects.
}
\details{
Computation of fitted and predicted values and standard errors
is performed before any plots are created. To avoid waste of
computation time, cached computations are returned \emph{even if}
an error is thrown during plotting. Hence the cache will be
available temporarily via \code{\link{.Last.value}}. To ensure
that the cache is available permanently, assign the result of
the call to \code{\link{plot}} to a name:
\code{cache <- plot(x, \dots)}.

Caching functionality must be used with care, as mismatch between
\code{x} and \code{cache} will not be detected. Constructions such
as \code{plot(y, cache = plot(x, \dots), \dots)}, where \code{x}
and \code{y} are different objects, should not be expected to produce
correct results.

See topic \code{\link{egf_eval}} for details on nonstandard evaluation
of \code{subset}, \code{order}, \code{main}, \code{sub}, and \code{plab}.
}
