% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.egf}
\alias{plot.egf}
\title{Plot models of epidemic growth}
\usage{
\method{plot}{egf}(
  x,
  type = c("interval", "cumulative", "rt1", "rt2"),
  subset = NULL,
  log = TRUE,
  tol = 0,
  legend = FALSE,
  level = 0.95,
  bands = FALSE,
  control = NULL,
  per_plot = 6L,
  ...
)
}
\arguments{
\item{x}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{type}{A character string indicating a type of plot. Options are
\code{"interval"} (interval incidence),
\code{"cumulative"} (cumulative incidence),
\code{"rt1"}, and \code{"rt2"} (instantaneous exponential growth rate).}

\item{subset}{A named list of atomic vectors with elements specifying levels of
factors in \code{x$frame}. Only the subset of fitting windows belonging
to these levels is plotted. Use \code{NULL} (the default) to plot all
fitting windows or if \code{x$frame} has no factors.}

\item{log}{A logical scalar. If \code{TRUE}, then the dependent variable is plotted
on a logarithmic scale. Unused by \code{type \%in\% c("rt1", "rt2")}.}

\item{tol}{A non-negative number or \code{Inf}. Used to define "exceptional"
points when \code{type = "interval"}. \code{cases[-1]} is highlighted
according to \code{control$points_short} if \code{diff(date) < (1-tol)*m}
and according to \code{control$points_long} if \code{diff(date) > (1+tol)*m},
where \code{m = median(diff(date))}. In both cases, the value of
\code{diff(date)} is printed above the point. Assign 0 to highlight
all deviations from \code{m}. Assign \code{Inf} to disable highlighting.}

\item{legend}{A logical scalar. If \code{TRUE}, then a legend is displayed in the
right margin. Unused by \code{type != "rt2"}.}

\item{level}{A number in the interval (0,1). The confidence level represented
by confidence intervals on doubling times and confidence bands on
predicted values. Unused by \code{type = "rt2"}.}

\item{bands}{A logical scalar. If \code{TRUE}, then confidence bands on predicted
values are displayed. Longer run times can be anticipated in this
case. Unused by \code{type = "rt2"}.}

\item{control}{A list of lists defining the appearance of various plot elements,
or otherwise \code{NULL} (see Details). Unused by \code{type = "rt2"}.}

\item{per_plot}{A positive integer. One plot will display this many time series
when \code{type = "rt2"}.}

\item{...}{Optional arguments specifying additional graphical parameters
to be recycled for all plots. Currently, only \code{xlim}, \code{ylim},
\code{ylab}, and \code{main} are used; see \code{\link[graphics:plot.default]{graphics::plot()}}. \code{xlim}
can be numeric, Date, or character coercible to Date with
\code{as.Date(xlim)}. If \code{main} is a character string, then flags
of the form \code{"\%f"}, where \code{f} is the name of any factor present
in \code{group_by}, are replaced in a given plot by the factor level
relevant to that plot. Use \code{"\\n"} within \code{main} to separate
title from subtitle.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Methods for plotting \code{"egf"} objects.
}
\details{
\code{control} must be \code{NULL} or a named list containing some subset
of the elements below:

\describe{
\item{\code{box}}{
A named list of arguments to \code{\link[graphics:box]{graphics::box()}}, affecting
the appearance of the box drawn around the plot region.
Only \code{bty}, \code{lty}, \code{lwd}, and \code{col} are used.
}
\item{\code{xax}, \code{yax}}{
Named lists of arguments to \code{\link[graphics:axis]{graphics::axis()}}, affecting the
appearance of the bottom and left axes. Only \code{tcl}, \code{mgp2},
\code{col.axis}, \code{cex.axis}, and \code{font.axis} are used, where \code{mgp2}
is the second component of the usual \code{mgp} argument. The
appearance of the minor (day or month) and major (month or year)
bottom axes can be controlled separately by listing vectors of
length 2. For example, setting
\code{xax = list(col.axis = c("black", "red"))}
will make the minor axis black and major axis red.
}
\item{\code{ylab}, \code{main}}{
Named lists of arguments to \code{\link[graphics:title]{graphics::title()}}, affecting
the appearance of axis titles. Only \code{adj}, \code{col.lab}, \code{cex.lab},
\code{font.lab}, \code{col.main}, \code{cex.main}, and \code{font.main} are used.
}
\item{\code{points}}{
A named list of arguments to \code{\link[graphics:points]{graphics::points()}}, affecting
the appearance of observed data. Only \code{pch}, \code{col}, \code{bg} and
\code{cex} are used.
}
\item{\code{points_short}, \code{points_long}}{
Alternatives to \code{points_main} used to highlight certain points
when \code{type = "interval"}. See argument \code{tol}.
}
\item{\code{lines}}{
A named list of arguments to \code{\link[graphics:lines]{graphics::lines()}}, affecting
the appearance of predicted curves. Only \code{lty}, \code{lwd}, and
\code{col} are used.
}
\item{\code{zero}}{
A named list of arguments to \code{\link[graphics:lines]{graphics::lines()}}, affecting
the appearance of the line drawn at \code{y = 0} when \code{type = "rt1"}.
}
\item{\code{windows}, \code{bands}}{
Named lists of arguments to \code{\link[graphics:polygon]{graphics::polygon()}}, affecting
the appearance of fitting windows and confidence bands on
predicted incidence curves. Only \code{col} and \code{border} are used.
}
\item{\code{text_td}, \code{text_hl}}{
Named lists of arguments to \code{\link[graphics:text]{graphics::text()}}, affecting
the appearance of text giving initial doubling times and
text above highlighted points. Only \code{col} and \code{cex} are
used by \code{text_td}. \code{text_hl} also uses \code{font}, \code{pos} and
\code{offset}.
}
}

If \code{control = NULL}, then it defaults to
\code{get_control_default("plot.egf", type = type)}.

If \code{control} is a list and one of its elements is a partially
specified list
(e.g., \code{box = list(bty)} rather than \code{box = c(bty, lty, lwd, col)}),
then that element is filled out with values taken from
\code{get_control_default("plot.egf", type = type)}.

If \code{control} is a list and one if its elements is \code{NULL}
(e.g., \code{box = NULL}), then the corresponding plot element
is suppressed.
}
