% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{plot.egf_confint}
\alias{plot.egf_confint}
\title{Plot confidence intervals}
\usage{
\method{plot}{egf_confint}(
  x,
  type = c("bars", "boxes"),
  subset = NULL,
  order = NULL,
  label = NULL,
  per_plot = switch(type, bars = 12L, 4L),
  ...
)
}
\arguments{
\item{x}{An \code{"egf_confint"} object returned by \code{\link[=confint.egf]{confint.egf()}}.}

\item{type}{A character string determining how confidence intervals
are displayed (see Details).}

\item{subset}{An expression to be evaluated in \code{x}. Must evaluate to a
logical vector or list of logical vectors indexing rows
of \code{x}. Only indexed confidence intervals are plotted.
The default (\code{NULL}) is to plot all confidence intervals.}

\item{order}{An expression to be evaluated in \code{x}, typically a call
to \code{\link[=order]{order()}}, determining the order in which confidence
intervals and time series (depending on \code{type}) are
plotted. Must evaluate to a permutation of \code{seq_len(nrow(x))}.
The default (\code{NULL}) is equivalent to \code{seq_len(nrow(x))}.}

\item{label}{An expression to be evaluated in \code{x}, typically a factor,
interaction, or call to \code{\link[=sprintf]{sprintf()}}. Used to create
appropriate labels for confidence intervals and time series
(depending on \code{type}). The default (\code{NULL}) is to take
labels from \code{x$window} and \code{x$ts}, respectively.}

\item{per_plot}{A positive integer. One plot will display at most this many
confidence intervals or time series (depending on \code{type}).}

\item{...}{Unused optional arguments.}
}
\value{
\code{NULL} (invisibly).
}
\description{
A method for graphically comparing confidence intervals
on fitted values of nonlinear model parameters across
fitting windows.
}
\details{
\code{type = "bars"} creates a one-dimensional plot with
confidence intervals drawn as stacked horizontal line segments.

\code{type = "boxes"} creates a two-dimensional plot for each
time series (level of \code{x$ts}), each containing shaded boxes.
Projection of boxes onto the horizontal and vertical axes
yields fitting windows and corresponding confidence intervals.

If an endpoint of a confidence interval is \code{NA}, then dashed
lines are drawn from the point estimate to the boundary of the
plotting region to indicate missingness.
}
