% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{plot.egf_profile}
\alias{plot.egf_profile}
\title{Plot likelihood profiles}
\usage{
\method{plot}{egf_profile}(x, subset = NULL, sqrt = FALSE, level = NULL, ...)
}
\arguments{
\item{x}{An \code{"egf_profile"} object returned by \code{\link[=profile.egf]{profile.egf()}}.}

\item{subset}{An expression to be evaluated in \code{x}. Must evaluate to a
logical vector or list of logical vectors indexing rows of
\code{x}. Only indexed profiles are plotted. The default (\code{NULL})
is to plot all profiles.}

\item{sqrt}{A logical scalar. If \code{TRUE}, then square root-transformed
deviance is plotted.}

\item{level}{A numeric vector with elements in (0,1), or otherwise \code{NULL}.
If non-\code{NULL} and \code{sqrt = FALSE}, then line segments are
drawn showing the intersection of the profile with lines
\code{deviance = qchisq(level, df = 1)}.}

\item{...}{Optional graphical parameters passed to \code{\link[graphics:plot.default]{graphics::plot()}},
such as \code{type = "o"}. Note that \code{axes = FALSE} and \code{ann = FALSE}
are hard-coded, so axes and axis titles cannot be modified.}
}
\value{
\code{NULL} (invisibly).
}
\description{
A method for inspecting computed likelihood profiles.
}
