% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.smooth_cases}
\alias{plot.smooth_cases}
\title{Plot smoothed incidence data}
\usage{
\method{plot}{smooth_cases}(x, ...)
}
\arguments{
\item{x}{A "smooth_cases" object.}

\item{...}{Unused optional arguments.}
}
\value{
\code{NULL} (invisibly).
}
\description{
A method for plotting objects of class "smooth_cases".
}
\details{
\subsection{Plot elements}{

There is one panel for each element of \code{x$spar}. The bottom axis
measures time as a number of days since \code{x$date[1]}. The left axis
measures interval incidence, log(1+x)-transformed if \code{x$log = TRUE}.
Plotted as points are the interval incidence data specified by
\code{x$date} and \code{x$cases}. Plotted as a solid line is a cubic spline
fit to the (possibly transformed) data using \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.
The spline in panel \code{i} is specified by \code{x$ss[[i]]} and uses
smoothing parameter \code{x$spar[i]}. Dotted lines are drawn at peaks
(red) and troughs (blue) in the spline. These lines are labeled in
the top margin with the index of \code{cases} corresponding to the time
of the peak or trough in the spline.
}
}
