% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.egf}
\alias{predict.egf}
\alias{predict.egf_init}
\title{Evaluate models of epidemic growth}
\usage{
\method{predict}{egf_init}(object, time = with(object, time[first:last]), ...)

\method{predict}{egf}(object, time = with(object$init, time[first:last]), ...)
}
\arguments{
\item{object}{An "egf_init" or "egf" object.}

\item{time}{A numeric vector listing increasing time points.
Times must be expressed as numbers of days since
\code{with(object, date[first])} for "egf_init" objects
and since
\code{with(object$init, date[first])} for "egf" objects.}

\item{...}{Unused optional arguments.}
}
\value{
A list with numeric elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{refdate}}{
Matches \code{with(object, date[first])}
or \code{with(object$init, date[first])}.
}
\item{\code{cum_inc}}{
Expected cumulative incidence at time points \code{time}.
Equal to \code{object$eval_cum_inc(time)}.
}
\item{\code{int_inc}}{
Expected interval incidence given interval endpoints \code{time}.
Equal to \code{c(NA, diff(cum_inc))}.
}
}
}
\description{
Methods for evaluating incidence models specified
by objects of class "egf_init" or "egf".
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \code{\link[=egf]{egf()}}
}
