% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.egf}
\alias{predict.egf}
\title{Compute predicted values}
\usage{
\method{predict}{egf}(
  object,
  what = c("interval", "cumulative", "rt"),
  time,
  window,
  append = NULL,
  log = TRUE,
  se = FALSE,
  .append = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{what}{A \link{character} vector listing one or more variables for which
predicted values are sought.}

\item{time}{A \link{numeric} vector supplying time points at which predicted
values are sought. \link{Date} and \link{POSIXt} vectors are
tolerated and coerced to numeric with \code{\link{julian}(time)}.
When \link{time} is missing, the original time points stored in
\code{object} are used.}

\item{window}{A \link{factor} of length \code{\link{length}(time)} grouping
the elements of \code{time} by fitting window. Levels not
belonging to \code{\link{levels}(object$frame$window)} are ignored.
\code{window} is ignored altogether if \code{time} is missing.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link{egf_make_combined}}) to be included with the
result. The default (\code{\link{NULL}}) is to append nothing.}

\item{log}{A \link{logical} flag. If \code{FALSE},
then inverse log-transformed predicted values are returned.}

\item{se}{A \link{logical} flag. If \code{log = TRUE} and \code{se = TRUE},
then approximate (delta method) standard errors on predicted values
are reported. Standard errors are required for subsequent use of
\code{\link{confint.egf_predict}}.}

\item{.append}{A \link{character} vector listing variable names to be used
(if non-\code{\link{NULL}}) in place of the result of evaluating
\code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_predict"}, with variables:
\item{var}{
  Predicted variable, from \code{what}.
}
\item{ts}{
  Time series, from \code{\link{levels}(object$frame_windows$ts)}.
}
\item{window}{
  Fitting window, from \code{\link{levels}(object$frame_windows$window)}.
}
\item{time}{
  Time, after possible coercion to \link{numeric}.
}
\item{estimate}{
  Predicted value of \code{var} at \code{time} in \code{window},
  conditional on the mixed effects data \code{object$frame_parameters}
  and the fitted model \code{object$best}.
}
\item{se}{
  (If \code{log = TRUE} and \code{se = TRUE}.)
  Approximate (delta method) standard error on \code{estimate}.
}
}
\description{
Computes predicted values of interval incidence, cumulative incidence,
and the per capita growth rate, conditional on observed data and a fitted
nonlinear mixed effects model of epidemic growth.
}
\details{
In the result, \code{estimate[i]} can be interpreted as follows
assuming \code{log = FALSE}:
\describe{
\item{\code{interval}}{
  The expected number of cases observed from \code{time[i-1]}
  to \code{time[i]} in \code{window[i]} (interval incidence).
}
\item{\code{cumulative}}{
  The expected number of cases observed up to \code{time[i]}
  in \code{window[i]} (cumulative incidence).
}
\item{\code{rt}}{
  The predicted per capita growth rate at \code{time[i]}.
  This is computed exactly from the differential equation model
  associated with \code{object$model$curve}.
}
}
See topic \code{\link{nse}} for details on nonstandard evaluation
of \code{append}.
}
