% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.egf}
\alias{predict.egf}
\alias{predict.egf_init}
\title{Evaluate models of epidemic growth}
\usage{
\method{predict}{egf_init}(object, time = object$time, ...)

\method{predict}{egf}(object, time = object$init$time, ...)
}
\arguments{
\item{object}{An "egf_init" or "egf" object.}

\item{time}{A numeric vector listing increasing time points,
expressed as numbers of days since \code{object$date[1]}
for "egf_init" objects and since \code{object$init$date[1]}
for "egf" objects. Missing values are not tolerated.}

\item{...}{Unused optional arguments.}
}
\value{
A list with numeric elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{refdate}}{Matches \code{object$date[1]} or \code{object$init$date[1]}.}
\item{\code{cum_inc}}{Expected cumulative incidence at time points
\code{time}. Equal to \code{object$eval_cum_inc(time)}.
}
\item{\code{int_inc}}{Expected interval incidence given interval
endpoints \code{time}. Equal to \code{diff(object$eval_cum_inc(time))}
if \code{length(time) >= 2} and omitted otherwise.
}
}
}
\description{
Methods for evaluating incidence models specified
by objects of class "egf_init" or "egf".
}
\seealso{
\code{\link[=egf_init]{egf_init()}}, \code{\link[=egf]{egf()}}
}
