% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.egf}
\alias{predict.egf}
\title{Compute predicted incidence time series}
\usage{
\method{predict}{egf}(object, time = NULL, se = FALSE, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{time}{A numeric vector listing increasing time points in days since
the start of the fitting window specified by \code{list(...)}.}

\item{se}{A logical scalar. If \code{TRUE}, then standard errors on predicted
values are returned.}

\item{...}{Atomic vectors together specifying a unique fitting window.
See Details.}
}
\value{
A list inheriting from class \code{"egf_predict"}, containing two data
frames \code{log_cum_inc} and \code{log_int_inc}, each with variables \code{time},
\code{estimate}, and \code{se} (if \code{se = TRUE}).

\code{log_cum_inc$estimate[i]} stores log cumulative incidence at
\code{time[i]}. \code{log_int_inc$estimate[i]} stores log incidence during
the interval from \code{time[i-1]} to \code{time[i]}.
}
\description{
Computes predicted values of cumulative and interval incidence
given user-specified time points.
}
\details{
Elements of \code{list(...)} must be named and have the form
\code{factor_name = c(level_names)}, where \code{factor_name} must
be the name of a factor in \code{object$frame} and \code{level_names}
must be a subset of \code{levels(object$frame[[factor_name]])}.
}
