% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.egf}
\alias{profile.egf}
\title{Compute likelihood profiles}
\usage{
\method{profile}{egf}(
  fitted,
  which = NULL,
  A = NULL,
  top = egf_get_names_top(fitted, link = TRUE),
  subset = NULL,
  append = NULL,
  max_level = 0.95,
  grid_len = 12,
  trace = TRUE,
  parallel = egf_parallel(),
  .subset = NULL,
  .append = NULL,
  ...
)
}
\arguments{
\item{fitted}{An \code{"\link{egf}"} object.}

\item{which}{An integer vector indexing coefficients in \code{fitted$best}
to be profiled. Must be a subset of \code{fitted$nonrandom}.}

\item{A}{A \link{numeric} \link{matrix} with
\code{\link{length}(fitted$nonrandom)} columns.
Each row specifies a linear combination of the elements
of \code{fitted$best[fitted$nonrandom]} to be profiled.
Ignored if \code{which} is non-\code{\link{NULL}}.}

\item{top}{A subset of \code{\link{egf_get_names_top}(fitted, link = TRUE)}
naming top level nonlinear model parameters whose population
fitted values (see \code{\link{fitted.egf}}) should be profiled.
Ignored if \code{which} or \code{A} is non-\code{\link{NULL}}.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{egf_combine_frames}}). It must evaluate to
to a valid index vector for the rows of the data frame
(see \code{\link{[.data.frame}}), and thus fitting windows.
Only population fitted values for indexed windows are profiled.
The default (\code{\link{NULL}}) is to consider all windows.
Ignored if \code{which} or \code{A} is non-\code{\link{NULL}}.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link{egf_combine_frames}}) to be included with the
result. The default (\code{\link{NULL}}) is to append nothing.
Ignored if \code{which} or \code{A} is non-\code{\link{NULL}}.}

\item{max_level}{A number in the interval (0,1) indicating a confidence level.
Profiles will be computed up to a deviance
of \code{\link{qchisq}(max_level, df = 1)}.}

\item{grid_len}{A positive integer. Step sizes chosen adaptively by
\code{\link[TMB]{tmbprofile}} will generate approximately
this many points on each side of a profile's minimum point.}

\item{trace}{A \link{logical} flag.
If \code{TRUE}, then basic tracing messages indicating progress
are printed.
Depending on \code{fitted$control$trace}, these may be mixed with
optimizer output.}

\item{parallel}{An \code{"\link{egf_parallel}"} object defining options for \R level
parallelization.}

\item{.subset, .append}{Index vectors to be used (if non-\code{\link{NULL}}) in place of
the result of evaluating \code{subset} and \code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_profile"}, with variables:
\item{top}{
  (\code{top}-based calls only.)
  Top level nonlinear model parameter,
  from \code{\link{egf_get_names_top}(fitted, link = link)}.
}
\item{ts}{
  (\code{top}-based calls only.)
  Time series, from \code{\link{levels}(fitted$frame_windows$ts)}.
}
\item{window}{
  (\code{top}-based calls only.)
  Fitting window, from \code{\link{levels}(fitted$frame_windows$window)}.
}
\item{linear_combination}{
  Row index of linear combination,
  from \code{\link{seq_len}(\link{nrow}(A))}.
}
\item{value}{
  Value of linear combination being profiled.
}
\item{deviance}{
  Deviance of the restricted model that assumes \code{value}
  for the linear combination being profiled.
}
\code{A}, \code{x = fitted$best[fitted$nonrandom]}, and \code{max_level}
are retained as \link{attributes}.
}
\description{
Computes univariate likelihood profiles of fixed effects
coefficients, random effects covariance parameters, and
linear combinations thereof.
}
\details{
Computation of likelihood profiles is expensive as it requires estimation
of many restricted models. It is parallelized at the C++ level when there
is OpenMP support and \code{fitted$control$omp_num_threads} is set to
an integer greater than 1. If there is no OpenMP support, then computation
may still be parallelized at the \R level with appropriate setting of
\code{parallel}.

\code{which} is mapped to an \code{A} matrix composed of unit row vectors,
with 1 at array index \code{[i, which[i]]} for all \code{i}, such that
\code{A \link[=matmult]{\%*\%} fitted$best[fitted$nonrandom]} is precisely
\code{fitted$best[which]}.

\code{top} and \code{subset} are mapped to a block \code{A} matrix.
Block \code{k} (from the top) is composed of rows (those indexed
by \code{subset}) of the fixed effects design matrix for parameter
\code{top[k]} .
\code{A \link[=matmult]{\%*\%} fitted$best[fitted$nonrandom]}
is a vector listing the population fitted values for each parameter
named in \code{top}, for each fitting window indexed by \code{subset}.

See topic \code{\link{egf_eval}} for details on nonstandard evaluation
of \code{subset} and \code{append}.
}
\examples{
example("egf", "epigrowthfit")
zz <- profile(object, subset = (country == "A" & wave == 1))
str(zz)

}
\seealso{
\code{\link{confint.egf_profile}}, \code{\link{plot.egf_profile}}
}
