% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.egf}
\alias{profile.egf}
\title{Compute likelihood profiles}
\usage{
\method{profile}{egf}(
  fitted,
  which = NULL,
  A = NULL,
  par = get_names_top(fitted, link = TRUE),
  subset = NULL,
  append = NULL,
  max_level = 0.99,
  grid_len = 12,
  parallel = egf_parallel(),
  trace = TRUE,
  .subset = NULL,
  .append = NULL,
  ...
)
}
\arguments{
\item{fitted}{An \code{"\link{egf}"} object.}

\item{which}{An integer vector indexing coefficients in \code{fitted$best}
to be profiled. Must be a subset of \code{fitted$nonrandom}.}

\item{A}{A \link{numeric} \link{matrix} with
\code{\link{length}(fitted$nonrandom)} columns.
Each row specifies a linear combination of the elements
of \code{fitted$best[fitted$nonrandom]} to be profiled.
Ignored if \code{which} is non-\code{\link{NULL}}.}

\item{par}{A subset of \code{\link{get_names_top}(fitted, link = TRUE)}
naming top level nonlinear model parameters whose population
fitted values (see \code{\link{fitted.egf}}) should be profiled.
Ignored if \code{which} or \code{A} is non-\code{\link{NULL}}.}

\item{subset}{An expression to be evaluated in the combined model frame
(see \code{\link{make_combined}}). Must evaluate to
a \link{logical} vector indexing rows of the data frame,
and thus fitting windows. Only population fitted values for
indexed windows are profiled. The default (\code{\link{NULL}})
is to consider all windows. Ignored if \code{which} or \code{A}
is non-\code{\link{NULL}}.}

\item{append}{An expression indicating variables in the combined model frame
(see \code{\link{make_combined}}) to be included with the result.
The default (\code{\link{NULL}}) is to append nothing.
Ignored if \code{which} or \code{A} is non-\code{\link{NULL}}.}

\item{max_level}{A number in the interval (0,1) indicating a confidence level.
Profiles will be computed up to a deviance
of \code{\link{qchisq}(max_level, df = 1)}.}

\item{grid_len}{A positive integer. Step sizes chosen adaptively by
\code{\link[TMB]{tmbprofile}} will generate approximately
this many points on each side of a profile's minimum point.}

\item{parallel}{An \code{"\link{egf_parallel}"} object defining parallelization options.}

\item{trace}{A \link{logical} flag. If \code{TRUE}, then basic tracing messages
indicating progress are printed. Depending on \code{fitted$control$trace},
these may be mixed with optimization output.}

\item{.subset}{A \link{logical} vector to be used (if non-\code{\link{NULL}})
in place of the result of evaluating \code{subset}.}

\item{.append}{A \link{character} vector listing variable names to be used
(if non-\code{\link{NULL}}) in place of the result of evaluating
\code{append}.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_profile"}, with variables:
\item{par}{
  (\code{par}-based calls only.)
  Top level nonlinear model parameter,
  from \code{\link{get_names_top}(fitted, link = link)}.
}
\item{ts}{
  (\code{par}-based calls only.)
  Time series, from \code{\link{levels}(fitted$frame_windows$ts)}.
}
\item{window}{
  (\code{par}-based calls only.)
  Fitting window, from \code{\link{levels}(fitted$frame_windows$window)}.
}
\item{linear_combination}{
  Row index of linear combination,
  from \code{\link{seq_len}(\link{nrow}(A))}.
}
\item{value}{
  Value of linear combination being profiled.
}
\item{deviance}{
  Deviance of the restricted model that assumes \code{value}
  for the linear combination being profiled.
}
\code{A}, \code{x = fitted$best[fitted$nonrandom]}, and \code{max_level}
are retained as \link{attributes}.
}
\description{
Computes the univariate likelihood profile of fixed effects
coefficients, random effects covariance parameters, and
linear combinations thereof.
}
\details{
\code{which} is mapped to an \code{A} matrix composed of unit row vectors,
with 1 at array index \code{[i, which[i]]} for all \code{i}, such that
\code{A \link[=matmult]{\%*\%} fitted$best[fitted$nonrandom]} is precisely
\code{fitted$best[which]}.

\code{par} and \code{subset} are mapped to a block \code{A} matrix.
Block \code{k} (from the top) is composed of rows of the fixed effects
design matrix for parameter \code{par[k]} (those indexed by \code{subset}).
\code{A \link[=matmult]{\%*\%} fitted$par[fitted$nonrandom]}
is a vector listing the population fitted values for each parameter
named in \code{par}, for each fitting window indexed by \code{subset}.

See topic \code{\link{nse}} for details on nonstandard evaluation
of \code{subset} and \code{append}.
}
\seealso{
\code{\link{confint.egf_profile}}, \code{plot.egf_profile}
}
