% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.egf}
\alias{profile.egf}
\title{Compute likelihood profiles}
\usage{
\method{profile}{egf}(
  fitted,
  index = NULL,
  lin_comb = NULL,
  parm = get_par_names(fitted),
  decontrast = FALSE,
  max_level = 0.99,
  grid_len = 12,
  trace = TRUE,
  parallel = c("serial", "multicore", "snow"),
  cores = getOption("egf.cores", 2L),
  outfile = NULL,
  cl = NULL,
  ...
)
}
\arguments{
\item{fitted}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{index}{An integer vector indexing elements of \code{fitted$par} to be profiled.
Must be a subset of \code{fitted$nonrandom}.}

\item{lin_comb}{A numeric matrix with \code{length(fitted$nonrandom)} columns,
whose rows specify linear combinations of the elements of
\code{fitted$par[fitted$nonrandom]}. Ignored if \code{index} is non-\code{NULL}.}

\item{parm}{A character vector giving a subset of the response variables
listed in \code{get_par_names(fitted, link = TRUE)}. The elements
of \code{fitted$par[nonrandom]} belonging to the mixed effects model
for these response variables, or linear combinations thereof
(see \code{decontrast}), will be profiled. Ignored if \code{index} or
\code{lin_comb} are non-\code{NULL}.}

\item{decontrast}{A logical scalar. If \code{TRUE}, then those linear combinations of
the fixed effects coefficients (zero-sum contrasts) that recover
within-group means are profiled instead of the fixed effects
coefficients themselves. Ignored if \code{index} or \code{lin_comb} are
non-\code{NULL}.}

\item{max_level}{A number in the interval (0,1). Profiles will be computed up to
a deviance of \code{qchisq(max_level, df = 1)}.}

\item{grid_len}{A positive integer. Step sizes chosen adaptively by
\code{\link[TMB:tmbprofile]{TMB::tmbprofile()}} will generate approximately this
many points on each side of the minimum point.}

\item{trace}{A logical scalar. If \code{TRUE}, then basic tracing information is
printed.}

\item{parallel}{A character string indicating a library for parallel computation.
\code{"serial"} indicates no parallelization. \code{"multicore"} forks and
is intended for use from a terminal rather than, e.g., RStudio.
On Windows, it is equivalent to \code{"serial"}. \code{"snow"} creates
socket clusters and parallelizes on both Unix-alikes and Windows.}

\item{cores}{A positive integer. The number of worker processes spawned when
\code{parallel != "serial"}. See also \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{outfile}{A character string containing a file path or otherwise \code{NULL}.
If a file path, then console output is diverted there. If \code{NULL},
then there is no diversion. When \code{parallel = "snow"}, diversion
may be necessary to view output.}

\item{cl}{A optional socket cluster created by
\code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}}, to be used
when \code{parallel = "snow"}. If non-\code{NULL},
then \code{cores} and \code{outfile} are ignored.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame with variables:
\item{\code{name}}{
A factor. If \code{index} or \code{parm} are used, then the name of the
parameter or linear combination being profiled, taken or formed
from \code{names(fitted$par)}. If \code{lin_comb} is used, then the factor
levels are \code{seq_len(nrow(lin_comb))} corresponding to the rows
of \code{lin_comb}.
}
\item{\code{value}}{
A numeric vector. The value of the parameter
or linear combination being profiled.
}
\item{\code{deviance}}{
A numeric vector. The deviance of the restricted
model that assumes \code{value} for the parameter or
linear combination being profiled.
}
}
\description{
Computes the (univariate) likelihood profile of fixed effect
coefficients, log standard deviations of random effects, and
linear combinations thereof.
}
\seealso{
\code{\link[=confint.egf_profile]{confint.egf_profile()}}, \code{\link[=plot.egf_profile]{plot.egf_profile()}}
}
