% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.egf}
\alias{ranef.egf}
\alias{ranef}
\title{Extract random effect conditional modes}
\usage{
\method{ranef}{egf}(object, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_ranef"}, with one row per coefficient and variables:
\item{cov}{
  Name of a covariance matrix.
  This is the interaction of \code{term} and \code{group}.
}
\item{vec}{
  Name of a random vector.
  This is the interaction of \code{term}, \code{group}, and \code{level}.
}
\item{bottom}{
  Name of a bottom level mixed effects model parameter
  (a random effect coefficient).
}
\item{top}{
  Name of the top level nonlinear model parameter whose fitted value
  is a function of \code{bottom}.
}
\item{term, group}{
  Random effects term from mixed effects model \link{formula}
  for parameter \code{top}. \code{term} and \code{group} give
  the left and right hand sides of the \code{`|`} operator.
}
\item{level}{
  \link[=levels]{Level} of \link{factor} or interaction indicated
  by \code{group}.
}
\item{colname}{
  Column name in the random effects design matrix
  \code{object$tmb_out$env$Z}.
}
\item{mode}{
  Conditional mode (unit variance scale).
}
\item{sd}{
  Estimated standard deviation.
}
Attribute \code{Sigma} lists estimated unstructured random effect
covariance matrices, corresponding to the \link{levels} of \code{cov}.
}
\description{
Retrieve the modes of the coefficients (unit variance scale)
of the random effects component of a mixed effects model,
conditional on data and parameter estimates.
}
\examples{
example("egf", package = "epigrowthfit", local = TRUE, echo = FALSE)
object <- readRDS(system.file("exdata", "egf.rds", package = "epigrowthfit", mustWork = TRUE))

zz <- ranef(object)
str(zz)

}
\seealso{
Other coefficient extractors: 
\code{\link{fitted.egf}()},
\code{\link{fixef.egf}()}
}
\concept{coefficient extractors}
