% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.egf}
\alias{ranef.egf}
\alias{ranef}
\title{Extract random effects conditional modes}
\usage{
\method{ranef}{egf}(object, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{...}{Unused optional arguments.}
}
\value{
A data frame inheriting from class \code{"egf_ranef"},
with one row per coefficient and variables:
\item{\code{name}}{
Coefficient name in the full parameter vector \code{object$best}.
}
\item{\code{par}}{
Nonlinear model parameter, from
\code{get_par_names(object, link = TRUE)}.
}
\item{\code{term}, \code{group}}{
Term from random effects component of mixed effects formula
for nonlinear model parameter \code{par}. \code{term} and \code{group}
give the left and right hand sides of the \code{|} operator,
respectively.
}
\item{\code{level}}{
Level of the interaction specified by \code{group}.
}
\item{\code{colname}}{
Corresponding column name in the random effects design matrix
\code{object$tmb_args$data$Z}.
}
\item{\code{mode}}{
Conditional mode (unit variance scale).
}
\item{\code{sd}}{
Estimated standard deviation.
}

Attribute \code{cov} lists estimated random effect covariance matrices.
There is one matrix per random effects term and, in a given matrix
one row per nonlinear model parameter whose mixed effects formula
includes the term.
}
\description{
Retrieve the modes of the coefficients (unit variance scale)
of the random effects component of a mixed effects model,
conditional on data and parameter estimates.
}
