% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.egf}
\alias{ranef.egf}
\alias{ranef}
\title{Extract random effects conditional modes}
\usage{
\method{ranef}{egf}(object, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_ranef"}, with one row per coefficient and variables:
\item{name}{
  Coefficient name in the full parameter vector \code{object$best}.
}
\item{par}{
  Nonlinear or dispersion model parameter,
  from \code{\link{get_par_names}(object, link = TRUE)}.
}
\item{term, group}{
  Random effects term from mixed effects model formula for parameter
  \code{par}. \code{term} and \code{group} give the left and right
  hand sides of the \code{`|`} operator, respectively.
}
\item{level}{
  Level of interaction \code{group}.
}
\item{colname}{
  Corresponding column name in the random effects design matrix
  \code{object$tmb_args$data$Z}.
}
\item{mode}{
  Conditional mode (unit variance scale).
}
\item{sd}{
  Estimated standard deviation.
}
Attribute \code{cov} lists estimated unstructured random effects
covariance matrices. There is one matrix per random effects term.
The number of rows of a matrix is equal to the number of nonlinear
and dispersion model parameters whose formulae
contain the corresponding random effects term.
}
\description{
Retrieve the modes of the coefficients (unit variance scale)
of the random effects component of a mixed effects model,
conditional on data and parameter estimates.
}
