% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsplit}
\alias{rsplit}
\title{Recursively split data frames}
\usage{
rsplit(x, by = integer(0L), drop = FALSE)
}
\arguments{
\item{x}{A \link[=data.frame]{data frame}.}

\item{by}{A subset of \code{\link{seq_along}(x)} or \code{\link{names}(x)}
indicating (in order) variables on which to split \code{x}.}

\item{drop}{A \link{logical} scalar passed to \code{\link{split}}.}
}
\value{
A recursive \link{list} of \link[=data.frame]{data frame}s
with \code{length(by)} "layers", unless \code{length(by) = 0},
in which case \code{x} is returned as is.
}
\description{
\link[=split]{Split}s a supplied \link[=data.frame]{data frame}
on one or more variables in turn.
}
\examples{
f <- function() sample(letters[1:5], 20L, replace = TRUE)
d <- as.data.frame(replicate(3L, f()))
# rsplit(d, 1:3)

}
\keyword{internal}
