% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.egf}
\alias{simulate.egf}
\title{Simulate models of epidemic growth}
\usage{
\method{simulate}{egf}(object, nsim = 1, seed = NULL, time = object$init$time, ...)
}
\arguments{
\item{object}{An "egf" object.}

\item{nsim}{A positive integer specifying a number of simulations.}

\item{seed}{An integer specifying a seed for RNG, otherwise \code{NULL}.}

\item{time}{A numeric vector listing increasing time points,
expressed as numbers of days since \code{object$date[1]}
for "egf_init" objects and since \code{object$init$date[1]}
for "egf" objects. Missing values are not tolerated and
\code{length(time) >= 2} is required.}

\item{...}{Unused optional arguments.}
}
\value{
An "egf_sim" object. A list with elements:

\describe{
\item{\code{time}}{Matches argument.}
\item{\code{refdate}}{Matches \code{object$init$date[1]}.}
\item{\code{int_inc}}{A matrix with \code{length(time)-1} rows and \code{nsim}
columns, such that \code{int_inc[i, j]} is the number of cases
observed between \code{time[i]} and \code{time[i+1]} in simulation \code{j}
of \code{nsim}. Row vector \code{int_inc[i, ]} is sampled from a
Poisson or negative binomial distribution (depending on
\code{object$init$distr}) with mean \code{predict(object, time)$int_inc[i]}.
The negative binomial dispersion parameter is taken from
\code{object$theta_fit[["nbdisp"]]}.
}
\item{\code{cum_inc}}{A matrix with \code{length(time)} rows and \code{nsim}
columns, such that \code{cum_inc[i, j]} is the number of cases
observed up to \code{time[i]} in simulation \code{j}. Column vector
\code{cum_inc[, j]} is computed as \code{c0 + cumsum(c(0, int_inc[, j]))},
where \code{c0 = predict(object, time)$cum_inc[1]} is the expected
value of cumulative incidence at \code{time[1]} conditional on
parameter vector \code{object$theta_fit}.
}
\item{\code{object}}{Matches argument.}
}
}
\description{
A method for simulating incidence models specified
by objects of class "egf".
}
\seealso{
\code{\link[=egf]{egf()}}, \code{\link[=plot.egf_sim]{plot.egf_sim()}}
}
