% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.egf}
\alias{simulate.egf}
\title{Simulate incidence from a fitted model}
\usage{
\method{simulate}{egf}(object, nsim = 1L, seed = NULL, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object specifying a fitted nonlinear
mixed effects model.}

\item{nsim}{A positive integer indicating a number of replications.}

\item{seed}{An integer used to set the \link{RNG} state before simulation.
The default (\code{\link{NULL}}) is to use the state at the
time of the function call. The RNG state is preserved as an
\link[=attributes]{attribute} of the result (see Value).}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_simulate"}, with variables \code{ts}, \code{window},
and \code{time}, and \code{nsim} further variables with names
of the form \code{x[0-9]+}.

\code{ts}, \code{window}, and \code{time} are copied from the rows
of \code{object$frame} corresponding observations in fitting windows.
\link[=attributes]{Attribute} \code{subset} is an \link{integer} vector
indexing these rows.
}
\description{
Simulates new incidence data conditional on a fitted nonlinear
mixed effects model. Only observations during fitting windows
are simulated.
}
