% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.egf}
\alias{simulate.egf}
\title{Simulate incidence data from a fitted model}
\usage{
\method{simulate}{egf}(object, nsim = 1L, seed = NULL, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object specifying a fitted nonlinear
mixed effects model.}

\item{nsim}{A positive integer indicating a number of replications.}

\item{seed}{An integer used to set the \link{RNG} state before simulation.
The default (\code{\link{NULL}}) is to use the state at the
time of the function call. The RNG state
(either a \link{list} of arguments to \code{\link{set.seed}}
or a value of \code{\link{.Random.seed}}) is preserved as an
\link[=attributes]{attribute} of the result.}

\item{...}{Unused optional arguments.}
}
\value{
A \link[=data.frame]{data frame} inheriting from \link{class}
\code{"egf_simulate"}, with variables \code{ts}, \code{window},
and \code{time}, and \code{nsim} further variables with names
of the form \code{x[0-9]+}.
It corresponds rowwise to \code{object$frame[!is.na(object$frame$window), ]}.
}
\description{
Simulates new incidence data conditional on a fitted nonlinear
mixed effects model of epidemic growth. Only observations within
fitting windows are simulated.
}
