% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.egf_model}
\alias{simulate.egf_model}
\title{Simulate incidence time series}
\usage{
\method{simulate}{egf_model}(
  object,
  nsim = 1L,
  seed = NULL,
  mu,
  Sigma = NULL,
  tol = 1e-06,
  tmax = 100,
  cstart = 0,
  ...
)
}
\arguments{
\item{object}{An \code{"\link{egf_model}"} specifying a top level nonlinear model
to be simulated.}

\item{nsim}{A positive integer indicating a number of simulated time series.}

\item{seed}{An integer used to set the \link{RNG} state before simulation.
The default (\code{NULL}) is to use the state at the time of
the function call. The RNG state (either a list of arguments to
\code{\link{set.seed}} or a value of \code{\link{.Random.seed}})
is preserved as an attribute of the result.}

\item{mu}{A numeric vector listing means across time series
of top level nonlinear model parameters (link scale).
It is assumed that elements are ordered as in
\code{\link{egf_get_names_top}(object, link = TRUE)}.}

\item{Sigma}{A symmetric positive definite numeric matrix to be used
as the covariance matrix corresponding to \code{mu}.
The default (\code{NULL}) is equivalent (conceptually)
to a zero matrix and is handled specially.}

\item{tol}{A non-negative number indicating a tolerance for lack of positive
definiteness of \code{Sigma}. All eigenvalues of \code{Sigma}
must exceed \code{-tol * rho}, where \code{rho} is the spectral
radius of \code{Sigma}.
(However, regardless of \code{tol}, \code{\link{diag}(Sigma)} must
be positive, as standard deviations are processed on the log scale.)}

\item{tmax}{A positive number. Simulated time series run from 0 days to 1 day
after the inflection time of the expected cumulative incidence
function indicated by \code{object$curve}. However, if there is no
inflection time, as in the exponential and subexponential models
of cumulative incidence, then simulated time series run from 0 days
to \code{tmax} days.}

\item{cstart}{A number indicating a threshold value of cumulative incidence.
Left endpoints of suggested fitting windows are those times
when cumulative incidence first exceeds this threshold.}

\item{...}{Unused optional arguments.}
}
\value{
A list inheriting from class \code{"egf_model_simulate"}, with elements:
\item{model, mu, Sigma, origin}{
  Copies of the so-named arguments.
  (\code{model} is a copy of argument \code{object}.)
}
\item{Y}{
  A numeric matrix with \code{nsim} rows and \code{length(mu)} columns
  listing the top level nonlinear model parameter values underlying each
  time series.
  If \code{Sigma} is \code{NULL}, then the row vectors of \code{Y}
  are all \code{mu}. Otherwise, \code{Y} is (conceptually) the result
  of \code{MASS::mvrnorm(nsim, mu, Sigma, tol)}.
}
\item{formula_ts}{
  A formula, \code{cbind(time, x) ~ ts}, expressing how simulated
  time series are stored in \code{data_ts}.
}
\item{formula_windows}{
  A formula, \code{cbind(start, end) ~ ts}, expressing how
  fitting window endpoints are stored in \code{data_windows}.
}
\item{formula_parameters}{
  A formula specifying the generative model.
  If \code{Sigma = NULL}, then the formula is
  \code{~1} if \code{nsim = 1} and \code{~ts} if \code{nsim > 1}.
  Otherwise, it is \code{~(1 | ts)}.
}
\item{data_ts}{
  A data frame with variables \code{ts}, \code{time}, and \code{x}
  storing \code{nsim} simulated time series in long format.
}
\item{data_windows}{
  A data frame with \code{nsim} rows and variables \code{ts},
  \code{start}, and \code{end} suggesting a fitting window for each
  simulated time series. Start times are determined according to
  \code{cstart}. End times are precisely the last time point in the
  corresponding time series.
}
\item{actual}{
  A numeric vector giving the condensed bottom level parameter vector
  of the generative model, corresponding to
  \code{\link[=coef.egf]{coef}(<"egf" object>, full = FALSE)}.
}
\item{random}{
  A logical vector indexing the elements of \code{actual} that are
  random effects.
}
\item{call}{
  The \link{call} to \code{simulate.egf_model}, allowing for updates
  to the \code{"egf_model_simulate"} object via \code{\link{update}}.
}
Attribute \code{RNGstate} preserves the RNG state prior to simulation,
making the result reproducible.
}
\description{
Simulates incidence time series with a unit observation interval
according to a specified nonlinear model. Top level nonlinear model
parameters vary between time series according to a fixed intercept
model \code{~ts} or random intercept model \code{~(1 | ts)}.
}
\examples{
r <- 0.04
c0 <- 400

mu <- log(c(r, c0))
Sigma <- diag(rep_len(0.2^2, length(mu)))

root <- system.file(package = "epigrowthfit", mustWork = TRUE)
path_to_cache <- file.path(root, "exdata", "simulate-egf_model.rds")
if (file.exists(path_to_cache)) {
  zz <- readRDS(path_to_cache)
} else {
  zz <- simulate(
    object = egf_model(curve = "exponential", family = "pois"),
    nsim = 20L,
    seed = 202737L,
    mu = mu,
    Sigma = Sigma,
    cstart = 10
  )
  dir.create(dirname(path_to_cache), showWarnings = FALSE)
  saveRDS(zz, file = path_to_cache)
}

}
\seealso{
\code{\link{egf.egf_model_simulate}} for estimating the generative model
from simulated time series,\cr
\code{\link{simulate.egf}} for simulating time series from \emph{fitted}
models
}
