% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cases.R
\name{smooth_cases}
\alias{smooth_cases}
\alias{smooth_cases.default}
\alias{smooth_cases.egf_init}
\alias{smooth_cases.egf}
\title{Fit cubic splines to interval incidence}
\usage{
smooth_cases(object = NULL, date, cases, log = TRUE, spar = 0.5, ...)

\method{smooth_cases}{default}(object, date, cases, log, spar, ...)

\method{smooth_cases}{egf_init}(object, date, cases, log, spar, ...)

\method{smooth_cases}{egf}(object, date, cases, log, spar, ...)
}
\arguments{
\item{object}{An "egf_init" or "egf" object, in which case \code{date} and \code{cases}
are ignored, or \code{NULL}, in which case \code{date} and \code{cases} are
required.}

\item{date}{A Date vector of length 5 or greater listing increasing time
points. Should start at or before the date of the first observed
case in an epidemic wave.}

\item{cases}{A numeric vector of length \code{length(date)}. For \code{i > 1}, \code{cases[i]}
must specify the number of cases observed between \code{date[i-1]} and
\code{date[i]}. \code{cases[1]} is ignored. Missing values are tolerated
unless \code{sum(!is.na(cases[-1])) < 4}.}

\item{log}{A logical scalar. If \code{TRUE}, then \code{log10(1+cases)} is smoothed
instead of \code{cases}.}

\item{spar}{A number in the interval (0,1]. Degree of smoothing increases
with \code{spar} (smoothing parameter). See \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}

\item{...}{Optional arguments to \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}
}
\value{
The default method returns a "smooth_cases" object, which is
a list containing copies of arguments \code{date}, \code{cases}, \code{log},
and \code{spar}, with these additional elements:

\describe{
\item{\code{ss}}{
The list output of \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.
}
\item{\code{peaks}, \code{troughs}}{
Integer vectors indexing dates of peaks and troughs,
respectively, in the cubic spline specified by \code{ss}.
Dates are obtained as \code{date[peaks]} and \code{date[troughs]}.
}
\item{\code{call}}{
The call to \code{smooth_cases()}, allowing the output
to be updated using \code{\link[stats:update]{stats::update()}}.
}
}

The method for class "egf_init" applies the default method
to \code{date = object$date} and \code{cases = object$cases}.

The method for class "egf" applies the default method
to \code{date = object$init$date} and \code{cases = object$init$cases}.
}
\description{
Fits a cubic spline to an interval incidence time series
using \code{\link[stats:smooth.spline]{stats::smooth.spline()}}, then looks for local extrema
by checking when the first derivative changes sign. Intended
for use in conjunction with \code{\link[=egf_init]{egf_init()}} to facilitate
fitting window selection.
}
\details{
A cubic spline is fit to \code{x = as.numeric(date[-1]-date[1])} and
\code{y = cases[-1]} or \code{y = log(1+cases[-1])}. The points are weighted
by \code{diff(date)} to account for possibly unequal spacing.
}
\examples{
data(canadacovid)
ontario <- subset(canadacovid, province == "ON")
x <- smooth_cases(
  date = ontario$date,
  cases = ontario$new_confirmed,
  log = TRUE,
  spar = 0.7
)
plot(x)

}
\seealso{
\code{\link[=plot.smooth_cases]{plot.smooth_cases()}}
}
