% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formula.R
\name{split_effects}
\alias{split_effects}
\title{Split fixed and random effects terms}
\usage{
split_effects(x)
}
\arguments{
\item{x}{A \link{formula}.}
}
\value{
A \link{list} with elements:
\item{fixed}{
  A fixed effects model \link{formula}.
}
\item{random}{
  A \link{list} of random effects terms,
  i.e., \link{call}s to the \code{`|`} operator.
}
}
\description{
Retrieves from a mixed effects model \link{formula} the corresponding
fixed effects model formula and a \link{list} of all random effects terms.
}
\examples{
# split_effects(y ~ 0 + x + (1 | f) + (a | g))

}
\keyword{internal}
