% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not}
\alias{stop_if_not}
\alias{warn_if_not}
\title{Utilities for object validation}
\usage{
stop_if_not(..., m = "", n = 1L)

warn_if_not(..., m, n = 1L)
}
\arguments{
\item{...}{Logical vectors.}

\item{m}{A character string used as an error or warning message.}

\item{n}{A non-negative integer specifying a call to be associated with
the error or warning. \code{n} indicates a number of generations
backwards relative to the function call. The default \code{n = 1L}
corresponds to the call from the parent frame.}

\item{x}{An object to be checked.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Utilities for object validation
}
\details{
\code{stop_if_not()} is a replacement for \code{\link[=stopifnot]{stopifnot()}} allowing
(requiring) the user to specify an error message and function call.

\code{warn_if_not()} behaves identically to \code{stop_if_not()} but issues
a warning instead of an error.
}
\examples{
# x <- 1.1
# stop_if_not(
#   is.vector(x, "numeric"),
#   length(x) == 1L,
#   x > 0,
#   x < 1,
#   m = "`x` must be a number in the interval (0,1)."
# )

}
\keyword{internal}
