% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.egf}
\alias{vcov.egf}
\title{Extract covariance matrix of fixed effects coefficients}
\usage{
\method{vcov}{egf}(object, full = FALSE, cor = FALSE, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{full}{A logical scalar. If \code{TRUE}, then covariances involving log
standard deviations of random effects are retained.}

\item{cor}{A logical scalar. If \code{TRUE}, then correlations are returned
instead of covariances.}

\item{...}{Unused optional arguments.}
}
\value{
A matrix of covariances or correlations. The number of rows is
equal to the length of the \code{"beta"} component of \code{object$par}
plus, if \code{full = TRUE}, the length the \code{"log_sd_b"} component.
}
\description{
Retrieves the covariance or correlation matrix of the fixed effects
coefficients and, optionally, the log standard deviations of random
effects.
}
\details{
Contrary to the documentation of generic function \code{\link[stats:vcov]{stats::vcov()}},
\code{vcov.egf()} does \emph{not} return the covariance matrix corresponding
to the output of \code{\link[=coef.egf]{coef.egf()}}, which returns fitted values,
not coefficients, of the generalized linear mixed effects models.
}
