% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.egf}
\alias{vcov.egf}
\title{Extract fixed effects covariance matrix}
\usage{
\method{vcov}{egf}(object, full = FALSE, cor = FALSE, ...)
}
\arguments{
\item{object}{An \code{"egf"} object returned by \code{\link[=egf]{egf()}}.}

\item{full}{A logical scalar. If \code{TRUE}, then covariances involving log
standard deviations of random effects coefficients are retained.}

\item{cor}{A logical scalar. If \code{TRUE}, then a correlation matrix is returned.}

\item{...}{Unused optional arguments.}
}
\value{
A covariance or correlation matrix inheriting from class
\code{"egf_vcov"}. The number of rows is equal to the length
of the \code{"beta"} component of \code{object$par} plus the length
the \code{"log_sd_b"} component (if \code{full = TRUE}).
}
\description{
Retrieves the covariance or correlation matrix of fixed effects
coefficients and, optionally, log standard deviations of random
effects coefficients.
}
\details{
Contrary to the documentation of generic function \code{\link[stats:vcov]{stats::vcov()}},
\code{vcov.egf()} returns the covariance matrix corresponding to the
output of \code{\link[=fixef.egf]{fixef.egf()}}, not \code{\link[=coef.egf]{coef.egf()}}. Currently, \code{\link[=coef.egf]{coef.egf()}}
is an alias for \code{\link[=fitted.egf]{fitted.egf()}} and so returns fitted values of
nonlinear model parameters, not mixed effects model coefficients.
}
