% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.egf}
\alias{vcov.egf}
\title{Extract fixed effects covariance matrix}
\usage{
\method{vcov}{egf}(object, full = FALSE, cor = FALSE, ...)
}
\arguments{
\item{object}{An \code{"\link{egf}"} object.}

\item{full}{A \link{logical} flag. If \code{TRUE}, then covariances
of random effects covariance parameters are retained.}

\item{cor}{A \link{logical} flag. If \code{TRUE}, then a correlation
matrix is returned.}

\item{...}{Unused optional arguments.}
}
\value{
A covariance or correlation \link{matrix} inheriting from \link{class}
\code{"egf_vcov"}. The number of rows is equal to the length of segment
\code{beta} of the full parameter vector \code{c(beta, theta, b)} plus
(if \code{full = TRUE}) the length of segment \code{theta}.
}
\description{
Computes the covariance matrix of fixed effects coefficients
and (optionally) random effects covariance parameters.
}
\details{
If \code{object} was constructed by \code{\link{egf}(se = TRUE)},
then the full covariance matrix has already been computed and
is stored in the \link{list} \code{object$sdreport}.
\code{vcov} reuses this matrix to avoid needless recomputation.
}
\section{Note}{

Contrary to the documentation of generic function \code{\link{vcov}},
\code{vcov.egf} returns the covariance matrix corresponding
to the output of \code{\link{fixef.egf}}, not \code{\link{coef.egf}}.
Currently, \code{\link{coef.egf}} is an alias for \code{\link{fitted.egf}}:
it returns fitted values of top level nonlinear model parameters,
not fixed effects coefficients.
}

