% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ymd}
\alias{ymd}
\title{Decompose a vector of dates}
\usage{
ymd(x, which = 1:3, drop = TRUE)
}
\arguments{
\item{x}{A Date vector.}

\item{which}{A subset of \code{1:3} indicating which of year, month, and day
should be returned.}

\item{drop}{A logical scalar. If \code{drop = TRUE}, then an integer vector
is returned instead of a matrix when one of \code{x} and \code{which}
has length less than 2,}
}
\value{
\code{X[, which, drop]}, where \code{X} is an integer matrix with
\code{length(x)} rows and 3 columns listing year, month, and day.
}
\description{
Extracts year, month, and day from a Date vector.
}
\keyword{internal}
