% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datetime.R
\name{ymd}
\alias{ymd}
\title{Decompose a Date vector}
\usage{
ymd(x, which = "ymd", drop = TRUE)
}
\arguments{
\item{x}{A \link{Date} vector.}

\item{which}{A \link{character} string composed of the letters \samp{y m d},
indicating which of (and in what order) year, month, and day
should be returned.}

\item{drop}{A logical scalar. If \code{drop = TRUE}, then the matrix result
is coerced to a vector when one of its dimensions is less than 2.}
}
\value{
\code{X[, j, drop = drop]}, where \code{X} is an \link{integer}
\link{matrix} with \code{length(x)} rows and 3 columns listing
year, month, and day, and \code{j} is an index vector derived
from \code{which}.
}
\description{
Extracts year, month, and day from a \link{Date} vector.
}
\examples{
# x <- .Date(sample.int(1e4L, 10L))
# ymd(x)
# ymd(x, which = "d", drop = TRUE)

}
\keyword{internal}
